Function smoldot::author::aura::next_slot_claim

pub fn next_slot_claim<'a>(
    config: Config<'a, impl Iterator<Item = &'a [u8; 32]>>
) -> Option<SlotClaim>

Calculates the earliest one of the authorities in Config::local_authorities is allowed to produce a block.

Returns None if none of the local authorities are allowed to produce blocks.

The value returned by this function is entirely deterministic based on the Config and never changes until Config::now_from_unix_epoch gets past the value returned in SlotClaim::slot_end_from_unix_epoch.

However, keep in mind that, as the best block changes, the list of authorities (Config::current_authorities) might change, in which case this function should be called again.