Module smoldot::author::build

Structs

ApplyExtrinsic
More transactions can be added.

AuthoringStart
Ready to start producing blocks.

AuthoringStartConfig
Configuration to pass when the actual block authoring is started.

Config
Configuration for a block generation.

NextKey
Fetching the key that follows a given one in the parent storage is required in order to continue.

PrefixKeys
Fetching the list of keys with a given prefix from the parent storage is required in order to continue.

Seal
Block has been produced and must now be sealed.

StorageGet
Loading a storage value from the parent storage is required in order to continue.

WaitSlot
Block production is idle, waiting for a slot.

Enums

Builder
Current state of the block building process.

BuilderAuthoring
Current state of the block building process.

ConfigConsensus
Extension to Config.

Error
Error that can happen during the block production.