Module smoldot::chain::async_tree

Data structure containing trees of nodes.

The AsyncTree data structure can be used in situations where there exists a finalized block, plus a tree of non-finalized blocks that all descend from that finalized block.

The AsyncTree only contains the non-finalized blocks. The finalized block, virtual root of the tree, is only implicitly there.

When a block is inserted in the AsyncTree, it is marked as “pending”.

Re-exports

pub use fork_tree::NodeIndex;

Structs

AsyncOpId

Identifier for an asynchronous operation in the AsyncTree.

AsyncOpParams

Information about an operation that must be started.

AsyncTree

OutputUpdateBlock

See OutputUpdate.

Enums

NextNecessaryAsyncOp

OutputUpdate

See AsyncTree::try_advance_output.