Module smoldot::chain::blocks_tree

Finalized block header, plus tree of authenticated non-finalized block headers.

This module provides the NonFinalizedTree type. This type is a data structure containing a valid tree of block headers, plus the state necessary to verify new blocks with the intent to add them to that tree. Each block header additionally holds a user-chosen opaque data.

The state in the NonFinalizedTree consists of:

One “latest finalized” block and various information about its ancestors, akin to a chain_information::ChainInformation.

Zero or more blocks that descend from that latest finalized block.

The latest finalized block is a block that is guaranted to never be reverted. While it can always be set to the genesis block of the chain, it is preferable, in order to reduce memory utilization, to maintain it to a block that is as high as possible in the chain.

Note: While mechanisms such as GrandPa provide a network-wide way to designate a block as final, the concept of GrandPa-provided finality doesn’t necessarily have to match the concept of finality in the NonFinalizedTree. For example, an API user might decide to optimistically assume that the block whose number is highest_block - 5 is automatically finalized, and fall back to rebuilding a new NonFinalizedTree if that assumption turns out to not be true. The finalized block in the NonFinalizedTree only represents a block that the NonFinalizedTree itself cannot remove, not a block that cannot be removed in the absolute.

A block can be added to the chain by calling NonFinalizedTree::verify_header or NonFinalizedTree::verify_body. As explained in details in the verify module, verifying the header only verifies the authenticity of a block and not its correctness. Verifying both the header and body provides the strongest guarantee, but requires knowledge of the storage of the block that is parent of the block to verify.

Note: There typically exists two kinds of clients: full and light. Full clients store the state of the storage, while light clients don’t. For this reason, light clients can only verify the header of new blocks. Both full and light clients should wait for a block to be finalized if they want to be certain that it will forever remain part of the chain.

Additionally, a NonFinalizedTree::verify_justification method is provided in order to verify the correctness of a justification.

Structs

BlockAccess
Access to a block’s information and hierarchy.

BodyInsert
Holds the NonFinalizedTree and allows insert a successfully-verified block into it.

BodyVerifyRuntimeRequired
Verification is pending. In order to continue, a host::HostVmPrototype of the runtime of the parent block must be provided.

Config
Configuration for the NonFinalizedTree.

FinalityApply
Returned by NonFinalizedTree::verify_justification and NonFinalizedTree::verify_grandpa_commit_message on success.

HeaderInsert
Mutably borrows the NonFinalizedTree and allows insert a successfully-verified block into it.

NonFinalizedTree
Holds state about the current state of the chain for the purpose of verifying headers.

RuntimeCompilation
A new runtime must be compiled.

SetFinalizedBlockIter
Iterator producing the newly-finalized blocks removed from the state when the finalized block is updated.

StorageGet
Loading a storage value is required in order to continue.

StorageNextKey
Fetching the key that follows a given one is required in order to continue.

StoragePrefixKeys
Fetching the list of keys with a given prefix is required in order to continue.

Enums

BodyVerifyError
Error while verifying a block body.

BodyVerifyStep1
Block verification, either just finished or still in progress.

BodyVerifyStep2
Block verification, either just finished or still in progress.

CommitVerifyError
Error that can happen when verifying a Grandpa commit.

FinalityVerifyError
Error that can happen when verifying a proof of finality.

HeaderVerifyError
Error that can happen when verifying a block header.

HeaderVerifySuccess

JustificationVerifyError
Error that can happen when verifying a justification.

SetFinalizedError
Error that can happen when setting the finalized block.