Struct smoldot::chain::blocks_tree::NonFinalizedTree

pub struct NonFinalizedTree<T> { /* private fields */ }

Holds state about the current state of the chain for the purpose of verifying headers.

Implementations

impl<T> NonFinalizedTree<T>

pub fn finality_checkpoints(&self) -> impl Iterator<Item = (u64, &[u8; 32])>

Returns a list of blocks (by their height and hash) that need to be finalized before any of their descendants can be finalized.

In other words, blocks in the NonFinalizedTree can be immediately finalized by call to NonFinalizedTree::verify_justification or NonFinalizedTree::verify_grandpa_commit_message, unless they descend from any of the blocks returned by this function, in which case that block must be finalized beforehand.

pub fn verify_justification(
&mut self,
consensus_engine_id: [u8; 4],
scale_encoded_justification: &[u8],
randomness_seed: [u8; 32]
) -> Result<FinalityApply<'_, T>, JustificationVerifyError>

Verifies the given justification.

The verification is performed in the context of the chain. In particular, the verification will fail if the target block isn’t already in the chain.

If the verification succeeds, a FinalityApply object will be returned which can be used to apply the finalization.

A randomness seed must be provided and will be used during the verification. Note that the verification is nonetheless deterministic.

pub fn verify_grandpa_commit_message(
&mut self,
scale_encoded_commit: &[u8],
randomness_seed: [u8; 32]
) -> Result<FinalityApply<'_, T>, CommitVerifyError>

Verifies the given Grandpa commit message.

The verification is performed in the context of the chain. In particular, the verification will fail if the target block isn’t already in the chain or if one of the voted blocks is unknown locally.

If the verification succeeds, a FinalityApply object will be returned which can be used to apply the finalization.

A randomness seed must be provided and will be used during the verification. Note that the verification is nonetheless deterministic.

pub fn set_finalized_block(
&mut self,
block_hash: &[u8; 32]
) -> Result<SetFinalizedBlockIter<'_, T>, SetFinalizedError>

Sets the latest known finalized block. Trying to verify a block that isn’t a descendant of that block will fail.

The block must have been passed to NonFinalizedTree::verify_header.

Returns an iterator containing the now-finalized blocks in decreasing block numbers. In other words, the first element of the iterator is always the block whose hash is the block_hash passed as parameter.

Note: This function returns blocks in decreasing block number, because any other ordering would incur a performance cost. While returning blocks in increasing block number would often be more convenient, the overhead of doing so is moved to the user.

The pruning is completely performed, even if the iterator is dropped eagerly.

If necessary, the current best block will be updated to be a descendant of the newly-finalized block.

impl<T> NonFinalizedTree<T>

pub fn verify_header(
&mut self,
scale_encoded_header: Vec<u8>,
now_from_unix_epoch: Duration
) -> Result<HeaderVerifySuccess<'_, T>, HeaderVerifyError>

Verifies the given block.

The verification is performed in the context of the chain. In particular, the verification will fail if the parent block isn’t already in the chain.

If the verification succeeds, an HeaderInsert object might be returned which can be used to then insert the block in the chain.

Must be passed the current UNIX time in order to verify that the block doesn’t pretend to come from the future.

pub fn verify_body(
self,
scale_encoded_header: Vec<u8>,
now_from_unix_epoch: Duration
) -> BodyVerifyStep1<T>

Verifies the given block.

The verification is performed in the context of the chain. In particular, the verification will fail if the parent block isn’t already in the chain.

This method takes ownership of both the block’s information and the NonFinalizedTree. It turns an object that must be driver by the user, until either the verification is finished or the process aborted, at which point the NonFinalizedTree can be retrieved back. The state of the NonFinalizedTree isn’t modified until BodyInsert::insert is called after the end of the verification.

Must be passed the current UNIX time in order to verify that the block doesn’t pretend to come from the future.

impl<T> NonFinalizedTree<T>

pub fn new(config: Config) -> Self

Initializes a new queue.

Panic

Panics if the chain information is incorrect.

pub fn clear(&mut self)

Removes all non-finalized blocks from the tree.

pub fn is_empty(&self) -> bool

Returns true if there isn’t any non-finalized block in the chain.

pub fn len(&self) -> usize

Returns the number of non-finalized blocks in the chain.

pub fn iter_unordered(&self) -> impl Iterator<Item = HeaderRef<'_>> + '_

Returns the header of all known non-finalized blocks in the chain without any specific order.

pub fn iter_ancestry_order(&self) -> impl Iterator<Item = HeaderRef<'_>> + '_

Returns the header of all known non-finalized blocks in the chain.

The returned items are guaranteed to be in an order in which the parents are found before their children.

pub fn reserve(&mut self, additional: usize)

Reserves additional capacity for at least additional new blocks without allocating.

pub fn shrink_to_fit(&mut self)

Shrink the capacity of the chain as much as possible.

pub fn block_number_bytes(&self) -> usize

Returns the value that was initially passed in Config::block_number_bytes.

pub fn as_chain_information(&self) -> ValidChainInformationRef<'_>

Builds a chain_information::ChainInformationRef struct that might later be used to build a new NonFinalizedTree.

pub fn finalized_block_header(&self) -> HeaderRef<'_>

Returns the header of the latest finalized block.

pub fn finalized_block_hash(&self) -> [u8; 32]

Returns the hash of the latest finalized block.

pub fn best_block_header(&self) -> HeaderRef<'_>

Returns the header of the best block.

pub fn best_block_hash(&self) -> [u8; 32]

Returns the hash of the best block.

pub fn best_block_consensus(&self) -> ChainInformationConsensusRef<'_>

Returns consensus information about the current best block of the chain.

pub fn contains_non_finalized_block(&self, hash: &[u8; 32]) -> bool

Returns true if the block with the given hash is in the NonFinalizedTree.

pub fn non_finalized_block_user_data(&self, hash: &[u8; 32]) -> Option<&T>

Gives access to the user data of a block stored by the NonFinalizedTree, identified by its hash.

Returns None if the block can’t be found.

pub fn non_finalized_block_by_hash(
&mut self,
hash: &[u8; 32]
) -> Option<BlockAccess<'_, T>>

Gives access to a block stored by the NonFinalizedTree, identified by its hash.