Module smoldot::chain::chain_information

Data structures containing the finalized state of the chain, except for its storage.

The types provided in this module contain the state of the chain, other than its storage, that has been finalized.

Note: These data structures only provide a way to communicate that finalized state, but the existence of a ChainInformation alone does in no way mean that its content is accurate. As an example, one use case of ChainInformation is to be written to disk then later reloaded. It is possible for the user to modify the data on disk, in which case the loaded ChainInformation might be erroneous.

These data structures contain all the information that is necessary to verify the authenticity (but not the correctness) of blocks that descend from the finalized block contained in the structure.

They do not, however, contain the storage of the finalized block, which is necessary to verify the correctness of new blocks. It possible possible, though, for instance to download the storage of the finalized block from another node. This downloaded storage can be verified to make sure that it matches the content of the ChainInformation.

They also do not contain the past history of the chain. It is, however, similarly possible to for instance download the history from other nodes.

Modules

aura_config

babe_config

babe_fetch_epoch

Structs

BabeEpochInformation
Information about a Babe epoch.

BabeEpochInformationRef
Information about a Babe epoch.

ChainInformation
Information about the latest finalized block and state found in its ancestors.

ChainInformationRef
Equivalent to a ChainInformation but referencing an existing structure. Cheap to copy.

ValidChainInformation
Information about the latest finalized block and state found in its ancestors.

ValidChainInformationRef
Information about the latest finalized block and state found in its ancestors.

Enums

ChainInformationConsensus
Extra items that depend on the consensus engine.

ChainInformationConsensusRef
Extra items that depend on the consensus engine.

ChainInformationFinality
Extra items that depend on the finality engine.

ChainInformationFinalityRef
Extra items that depend on the finality engine.

FromGenesisStorageError
Error when building the chain information from the genesis storage.

ValidityError
Error when turning a ChainInformation into a ValidChainInformation.