Module smoldot::chain::chain_information::babe_fetch_epoch

Structs

Config
Configuration for babe_fetch_epoch.

NextKey
Fetching the key that follows a given one is required in order to continue.

StorageGet
Loading a storage value is required in order to continue.

StorageRoot
Fetching the storage trie root is required in order to continue.

Enums

BabeEpochToFetch
The Babe epoch to fetch.

Error
Problem encountered during a call to babe_fetch_epoch.

Query
Current state of the operation.

Functions

babe_fetch_epoch
Fetches a Babe epoch using BabeApi_current_epoch or BabeApi_next_epoch.