Struct smoldot::chain_spec::ChainSpec

pub struct ChainSpec { /* private fields */ }

A configuration of a chain. Can be used to build a genesis block.

Implementations

impl ChainSpec

pub fn from_json_bytes(json: impl AsRef<[u8]>) -> Result<Self, ParseError>

Parse JSON content into a ChainSpec.

pub fn as_chain_information(
&self
) -> Result<(ValidChainInformation, HostVmPrototype), FromGenesisStorageError>

Builds the ChainInformation corresponding to the genesis block contained in this chain spec.

In addition to the information, also returns the virtual machine of the runtime of the genesis block.

pub fn name(&self) -> &str

Returns the name of the chain. Meant to be displayed to the user.

pub fn id(&self) -> &str

Returns the identifier of the chain. Similar to the name, but a bit more “system-looking”. For example, if the name is “Flaming Fir 7”, then the id could be flamingfir7. To be used for example in file system paths.

pub fn chain_type(&self) -> &str

Returns a string indicating the type of chain.

This value doesn’t have any meaning in the absolute and is only meant to be shown to the user.

pub fn block_number_bytes(&self) -> u8

Returns the number of bytes that the “block number” field of various data structures uses.

pub fn has_live_network(&self) -> bool

Returns true if the chain is of a type for which a live network is expected.

pub fn bad_blocks_hashes(&self) -> impl Iterator<Item = &[u8; 32]> + '_

Returns a list of hashes of block headers that should always be considered as invalid.

pub fn boot_nodes(&self) -> impl ExactSizeIterator<Item = Bootnode<'_>> + '_

Returns the list of bootnode addresses found in the chain spec.

Bootnode addresses that have failed to be parsed are returned as well in the form of a Bootnode::UnrecognizedFormat.

pub fn telemetry_endpoints(
&self
) -> impl Iterator<Item = impl AsRef<str> + '_> + '_

Returns the list of libp2p multiaddresses of the default telemetry servers of the chain.

pub fn protocol_id(&self) -> &str

Returns the network protocol id that uniquely identifies a chain. Used to prevent nodes from different blockchain networks from accidentally connecting to each other.

It is possible for the JSON chain specs to not specify any protocol id, in which case a default value is returned.

pub fn fork_id(&self) -> Option<&str>

Returns the “fork id” of the chain. This is arbitrary string that can be used in order to segregate nodes in case when multiple chains have the same genesis hash. Nodes should only synchronize with nodes that have the same “fork id”.

pub fn relay_chain(&self) -> Option<(&str, u32)>

pub fn genesis_storage(&self) -> GenesisStorage<'_>

Gives access to what is known about the storage of the genesis block of the chain.

pub fn properties(&self) -> &str

Returns a list of arbitrary properties contained in the chain specs, such as the name of the token or the number of decimals.

The value of these properties is never interpreted by the local node, but can be served to a UI.

The returned value is a JSON-formatted map, for example {"foo":"bar"}.

pub fn light_sync_state(&self) -> Option<LightSyncState>