Module smoldot::database::full_sled

Filesystem-backed database containing all the information about a chain.

This module handles the persistent storage of the chain on disk.

Usage

Use the open() function to create a new database or open an existing one. open() returns a DatabaseOpen enum. This enum will contain either a SledFullDatabase object, representing an access to the database, or a DatabaseEmpty if the database didn't exist or is empty. If that is the case, use DatabaseEmpty::initialize in order to populate it and obtain a SledFullDatabase.

Use SledFullDatabase::insert to insert a new block in the database. The block is assumed to have been successfully verified prior to insertion. An error is returned if this block is already in the database or isn't a descendant or ancestor of the latest finalized block.

Use SledFullDatabase::set_finalized to mark a block already in the database as finalized. Any block that isn't an ancestor or descendant will be removed. Reverting finalization is not possible.

Due to the database's schema, it is not possible to efficiently retrieve the storage items of blocks that are ancestors of the finalized block. When a block is finalized, the storage of its ancestors is lost, and the only way to reconstruct it is to execute all blocks starting from the genesis to the desired one.

About errors handling

Most of the functions and methods in this module return a Result containing notably an AccessError. This kind of errors can happen if the operating system returns an error when accessing the file system, or if the database has been corrupted, for example by the user manually modifying it.

There isn't much that can be done to properly handle an AccessError. The only reasonable solutions are either to stop the program, or to delete the entire database and recreate it.

Schema

Each section below corresponds to a "tree" in the sled database.

meta

Contains all the meta-information about the content.

Keys in that tree are:

best: Hash of the best block.

finalized: Height of the finalized block, as a 64bits big endian number.

grandpa_authorities_set_id: A 64bits big endian number representing the id of the authorities set that must finalize the block right after the finalized block. The value is 0 at the genesis block, and increased by 1 at every authorities change. Missing if and only if the chain doesn't use Grandpa.

grandpa_triggered_authorities: List of public keys and weights of the GrandPa authorities that must finalize the children of the finalized block. Consists in 40bytes values concatenated together, each value being a 32bytes ed25519 public key and a 8bytes little endian weight. Missing if and only if the chain doesn't use Grandpa.

grandpa_scheduled_target: A 64bits big endian number representing the block where the authorities found in grandpa_scheduled_authorities will be triggered. Blocks whose height is strictly higher than this value must be finalized using the new set of authorities. This authority change must have been scheduled in or before the finalized block. Missing if no change is scheduled or if the chain doesn't use Grandpa.

grandpa_scheduled_authorities: List of public keys and weights of the GrandPa authorities that will be triggered at the block found in grandpa_scheduled_target. Consists in 40bytes values concatenated together, each value being a 32bytes ed25519 public key and a 8bytes little endian weight. Missing if no change is scheduled or if the chain doesn't use Grandpa.

aura_slot_duration: A 64bits big endian number indicating the duration of an Aura slot. Missing if and only if the chain doesn't use Aura.

aura_finalized_authorities: List of public keys of the Aura authorities that must author the children of the finalized block. Consists in 32bytes values concatenated together. Missing if and only if the chain doesn't use Aura.

babe_slots_per_epoch: A 64bits big endian number indicating the number of slots per Babe epoch. Missing if and only if the chain doesn't use Babe.

babe_finalized_epoch: SCALE encoding of a structure that contains the information about the Babe epoch used for the finalized block. Missing if and only if the finalized block is block #0 or the chain doesn't use Babe.

babe_finalized_next_epoch: SCALE encoding of a structure that contains the information about the Babe epoch that follows the one described by babe_finalized_epoch. If the finalized block is block #0, then this contains information about epoch #0. Missing if and only if the chain doesn't use Babe.

block_hashes_by_number

For each possible block number, stores a list of block hashes having that number.

If the key is inferior or equal to the value in finalized, guaranteed to only contain on block.

Keys in that tree are 64-bits-big-endian block numbers, and values are a concatenation of 32-bytes block hashes (without any encoding). If the value is for example 96 bytes long, that means there are 3 blocks in the database with that block number.

Never contains any empty value.

block_headers

Contains an entry for every known block that is an ancestor or descendant of the finalized block. When the finalized block is updated, entries that aren't ancestors or descendants of the new finalized block are automatically purged.

Keys are block hashes, and values are SCALE-encoded block headers.

block_bodies

Entries are the same as for block_headers_tree.

Keys are block hashes, and values are SCALE-encoded Vecs containing the extrinsics. Each extrinsic is itself a SCALE-encoded Vec<u8>.

block_justifications

Entries are a subset of the ones of block_headers_tree. Not all blocks have a justification. Only finalized blocks have a justification.

Keys are block hashes, and values are SCALE-encoded Vecs containing the justification.

storage_top_trie

Contains the key-value storage at the finalized block.

Keys are storage keys, and values are storage values.

non_finalized_changes_keys

For each hash of non-finalized block, contains the list of keys in the storage that this block modifies.

Keys are a 32 bytes block hash. Values are a list of SCALE-encoded Vec<u8> concatenated together. In other words, each value is a length (SCALE-compact-encoded), a key of that length, a length, a key of that length, and so on.

non_finalized_changes

For each element in non_finalized_changes_keys_tree, contains the new value for this storage modification. If an entry is found in non_finalized_changes_keys_tree and not in non_finalized_changes_tree, that means that the storage entry must be removed.

Keys are a 32 bytes block hash followed with a storage key.

Structs

Config
Configuration for the database.

DatabaseEmpty
An open database. Holds file descriptors.

SledError
Low-level database error, such as an error while accessing the file system.

SledFullDatabase
An open database. Holds file descriptors.

VarLenBytes
Bytes in the database.

Enums

AccessError
Error while accessing some information.

ConfigTy
Type of database.

CorruptedError
Error in the content of the database.

DatabaseOpen
Either existing database or database prototype.

FinalizedAccessError
Error while accessing the storage of the finalized block.

InsertError
Error while calling SledFullDatabase::insert.

SetFinalizedError
Error while calling SledFullDatabase::set_finalized.

Functions

open
Opens the database using the given Config.