Module smoldot::executor

WebAssembly runtime code execution.

WebAssembly (often abbreviated Wasm) plays a big role in Substrate/Polkadot. The storage of each block in the chain has a special key named :code which contains the WebAssembly code of what we call the runtime.

The runtime is a program (in WebAssembly) that decides, amongst other things, whether transactions are valid and how to apply them on the storage, and whether blocks themselves are valid.

This module contains everything necessary to execute runtime code. The highest-level sub-module is runtime_host.

Modules

host
Wasm virtual machine specific to the Substrate/Polkadot Runtime Environment.

read_only_runtime_host
Wasm virtual machine, with automatic storage overlay and logs management.

runtime_host
Wasm virtual machine, with automatic storage overlay and logs management.

vm
General-purpose WebAssembly virtual machine.

Structs

CoreVersion
Buffer storing the SCALE-encoded core version.

CoreVersionApi
One API that the runtime supports.

CoreVersionApisRefIter
Iterator to a list of APIs. See CoreVersionRef::apis.

CoreVersionRef
Runtime specification, once decoded.

Enums

CoreVersionError
Error while executing core_version.

InvalidHeapPagesError
Error potentially returned by storage_heap_pages_to_value.

Constants

DEFAULT_HEAP_PAGES
Default number of heap pages if the storage doesn’t specify otherwise.

Functions

core_version
Runs the Core_version function using the given virtual machine prototype, and returns the output.

storage_heap_pages_to_value
Converts a value of the key :heappages found in the storage to an actual number of heap pages.