Enum smoldot::executor::host::Error

pub enum Error {

    Trap(Trap),
    BadReturnValue {
        actual: Option<ValueType>,
    },
    ReturnedPtrOutOfRange {
        pointer: u32,
        size: u32,
        memory_size: u32,
    },
    ReturnValueTypeMismatch,
    ParamsCountMismatch {
        function: &'static str,
        expected: usize,
        actual: usize,
    },
    ParamDecodeError(Error),
    WrongParamTy {
        function: &'static str,
        param_num: usize,
        expected: ValueType,
        actual: ValueType,
    },
    ParamOutOfRange {
        function: &'static str,
        param_num: usize,
        pointer: u32,
        length: u32,
    },
    Utf8Error {
        function: &'static str,
        param_num: usize,
        error: Utf8Error,
    },
    NestedTransaction,
    NoActiveTransaction,
    FinishedWithPendingTransaction,
    OutOfMemory {
        function: &'static str,
        requested_size: u32,
    },
    FreeError {
        pointer: u32,
    },

}

Reason why the Wasm blob isn’t conforming to the runtime environment.

Variants

Trap(Trap)

Error in the Wasm code execution.

BadReturnValue

A non-i64 value has been returned by the Wasm entry point.

Fields of BadReturnValue

actual: Option<ValueType>

Type that has actually gotten returned. None for “void”.

ReturnedPtrOutOfRange

The pointer and size returned by the Wasm entry point function are invalid.

Fields of ReturnedPtrOutOfRange

pointer: u32

Pointer that got returned.

size: u32

Size that got returned.

memory_size: u32

Size of the virtual memory.

ReturnValueTypeMismatch

An host_fn wants to returns a certain value, but the Wasm code expects a different one.

ParamsCountMismatch

Mismatch between the number of parameters expected and the actual number.

Fields of ParamsCountMismatch

function: &'static str

Name of the function being called whose number of parameters mismatches.

expected: usize

Expected number of parameters.

actual: usize

Number of parameters that have been passed.

ParamDecodeError(Error)

Failed to decode a SCALE-encoded parameter.

WrongParamTy

The type of one of the parameters is wrong.

Fields of WrongParamTy

function: &'static str

Name of the function being called where a type mismatch happens.

param_num: usize

Index of the invalid parameter. The first parameter has index 0.

expected: ValueType

Type of the value that was expected.

actual: ValueType

Type of the value that got passed.

ParamOutOfRange

One parameter is expected to point to a buffer, but the pointer is out of range of the memory of the Wasm VM.

Fields of ParamOutOfRange

function: &'static str

Name of the function being called where a type mismatch happens.

param_num: usize

Index of the invalid parameter. The first parameter has index 0.

pointer: u32

Pointer passed as parameter.

length: u32

Expected length of the buffer.

Depending on the function, this can either be an implicit length or a length passed as parameter.

Utf8Error

One parameter is expected to point to a UTF-8 string, but the buffer isn’t valid UTF-8.

Fields of Utf8Error

function: &'static str

Name of the function being called where a type mismatch happens.

param_num: usize

Index of the invalid parameter. The first parameter has index 0.

error: Utf8Error

Decoding error that happened.

NestedTransaction

Called ext_storage_start_transaction_version_1 with a transaction was already in progress.

NoActiveTransaction

Called ext_storage_rollback_transaction_version_1 or ext_storage_commit_transaction_version_1 but no transaction was in progress.

FinishedWithPendingTransaction

Execution has finished while a transaction started with ext_storage_start_transaction_version_1 was still in progress.

OutOfMemory

Error when allocating memory for a return type.

Fields of OutOfMemory

function: &'static str

Name of the function being called.

requested_size: u32

Size of the requested allocation.

FreeError

Called ext_allocator_free_version_1 with an invalid pointer.

Fields of FreeError

pointer: u32

Pointer that was expected to be free’d.