Enum smoldot::executor::host::HostVm

pub enum HostVm {

    ReadyToRun(ReadyToRun),
    Finished(Finished),
    Error {
        prototype: HostVmPrototype,
        error: Error,
    },
    ExternalStorageGet(ExternalStorageGet),
    ExternalStorageSet(ExternalStorageSet),
    ExternalStorageAppend(ExternalStorageAppend),
    ExternalStorageClearPrefix(ExternalStorageClearPrefix),
    ExternalStorageRoot(ExternalStorageRoot),
    ExternalStorageNextKey(ExternalStorageNextKey),
    ExternalOffchainIndexSet(ExternalOffchainIndexSet),
    ExternalOffchainStorageGet(ExternalOffchainStorageGet),
    ExternalOffchainStorageSet(ExternalOffchainStorageSet),
    OffchainTimestamp(OffchainTimestamp),
    OffchainRandomSeed(OffchainRandomSeed),
    OffchainSubmitTransaction(OffchainSubmitTransaction),
    SignatureVerification(SignatureVerification),
    CallRuntimeVersion(CallRuntimeVersion),
    StartStorageTransaction(StartStorageTransaction),
    EndStorageTransaction {
        resume: EndStorageTransaction,
        rollback: bool,
    },
    GetMaxLogLevel(GetMaxLogLevel),
    LogEmit(LogEmit),

}

Running virtual machine.

Variants

ReadyToRun(ReadyToRun)

Wasm virtual machine is ready to be run. Call ReadyToRun::run to make progress.

Finished(Finished)

Function execution has succeeded. Contains the return value of the call.

The trie root hash of all the child tries must be recalculated and written to the main trie similar to when a ExternalStorageRoot with a child_trie of None is generated. See the documentation of ExternalStorageRoot.

Error

The Wasm blob did something that doesn’t conform to the runtime environment.

Fields

prototype: HostVmPrototype

Virtual machine ready to be used again.

error: Error

Error that happened.

ExternalStorageGet(ExternalStorageGet)

Must load an storage value.

ExternalStorageSet(ExternalStorageSet)

Must set an storage value.

ExternalStorageAppend(ExternalStorageAppend)

See documentation of ExternalStorageAppend.

ExternalStorageClearPrefix(ExternalStorageClearPrefix)

Must remove all the storage values starting with a certain prefix.

ExternalStorageRoot(ExternalStorageRoot)

Must provide the trie root hash of the storage and write the trie root hash of child tries to the main trie.

ExternalStorageNextKey(ExternalStorageNextKey)

Need to provide the storage key that follows a specific one.

ExternalOffchainIndexSet(ExternalOffchainIndexSet)

Must set off-chain index value.

ExternalOffchainStorageGet(ExternalOffchainStorageGet)

Must load an offchain storage value.

ExternalOffchainStorageSet(ExternalOffchainStorageSet)

Must set value of an off-chain storage entry.

OffchainTimestamp(OffchainTimestamp)

Need to provide the current timestamp.

OffchainRandomSeed(OffchainRandomSeed)

Must return random seed.

OffchainSubmitTransaction(OffchainSubmitTransaction)

Submit a transaction from offchain worker.

SignatureVerification(SignatureVerification)

Need to verify whether a signature is valid.

CallRuntimeVersion(CallRuntimeVersion)

Need to call Core_version on the given Wasm code and return the raw output (i.e. still SCALE-encoded), or an error if the call has failed.

StartStorageTransaction(StartStorageTransaction)

Declares the start of a storage transaction. See HostVm::EndStorageTransaction.

EndStorageTransaction

Ends a storage transaction. All changes made to the storage (e.g. through a HostVm::ExternalStorageSet) since the previous HostVm::StartStorageTransaction must be rolled back if rollback is true.

Guaranteed by the code in this module to never happen if no transaction is in progress. If the runtime attempts to end a non-existing transaction, an HostVm::Error is generated instead.

Fields

resume: EndStorageTransaction

Object used to resume execution.

rollback: bool

If true, changes must be rolled back.

GetMaxLogLevel(GetMaxLogLevel)

Need to provide the maximum log level.

LogEmit(LogEmit)

Runtime has emitted a log entry.

Implementations

impl HostVm

pub fn into_prototype(self) -> HostVmPrototype

Cancels execution of the virtual machine and returns back the prototype.