smoldot::executor::host::runtime_version

Function find_encoded_embedded_runtime_version_apis

pub fn find_encoded_embedded_runtime_version_apis(
    binary_wasm_module: &[u8],
) -> Result<EmbeddedRuntimeVersionApis<'_>, FindEncodedEmbeddedRuntimeVersionApisError>

Tries to find the custom sections containing the runtime version and APIs.

This function does not attempt to decode the content of the custom sections.