Module smoldot::executor::read_only_runtime_host

Wasm virtual machine, with automatic storage overlay and logs management.

Structs

Config
Configuration for run.

Error
Error that can happen during the execution.

NextKey
Fetching the key that follows a given one is required in order to continue.

StorageGet
Loading a storage value is required in order to continue.

StorageRoot
Fetching the storage trie root is required in order to continue.

Success
Execution is successful.

SuccessVirtualMachine
Function execution has succeeded. Contains the return value of the call.

Enums

ErrorDetail
See Error::detail.

RuntimeHostVm
Current state of the execution.

Functions

run
Start running the WebAssembly virtual machine.