smoldot::executor::runtime_call

Struct StorageChanges

pub struct StorageChanges { /* private fields */ }

See Success::storage_changes.

Implementations

impl StorageChanges

pub fn empty() -> StorageChanges

Returns an empty StorageChanges, as if the execution didn’t modify anything.

pub fn main_trie_diff_get(&self, key: &[u8]) -> Option<Option<&[u8]>>

Returns the change, if any, at the given key of the main trie.

Returns None if the runtime hasn’t modified this entry in the main trie, and Some(None) if the runtime has removed the storage item of this entry.

pub fn main_trie_storage_changes_iter_unordered( &self, ) -> impl Iterator<Item = (&[u8], Option<&[u8]>)> + Clone

Returns an iterator to all the entries of the main trie that are modified by this runtime call.

Each value is either Some if the runtime overwrites this value, or None if it erases the underlying value.

Note: This function is equivalent to StorageChanges::storage_changes_iter_unordered, except that it only returns changes made to the main trie.

pub fn tries_with_storage_changes_unordered( &self, ) -> impl Iterator<Item = &[u8]>

Returns the list of all child tries whose content has been modified in some way.

pub fn child_trie_storage_changes_iter_unordered( &self, child_trie: &[u8], ) -> impl Iterator<Item = (&[u8], Option<&[u8]>)> + Clone

Returns an iterator to all the entries of the given child trie that are modified by this runtime call.

Each value is either Some if the runtime overwrites this value, or None if it erases the underlying value.

Note: This function is equivalent to StorageChanges::storage_changes_iter_unordered, except that it only returns changes made to the given child trie.

pub fn storage_changes_iter_unordered( &self, ) -> impl Iterator<Item = (Option<&[u8]>, &[u8], Option<&[u8]>)> + Clone

Returns an iterator to all the entries that are modified by this runtime call.

Each value is either Some if the runtime overwrites this value, or None if it erases the underlying value.

pub fn trie_changes_iter_ordered( &self, ) -> Option<impl Iterator<Item = (Option<&[u8]>, &[Nibble], TrieChange<'_>)>>

Returns an iterator over the list of all changes performed to the tries (main trie and child tries included).

Returns Some if and only if Config::calculate_trie_changes was true or if the StorageChanges was created using StorageChanges::empty.

pub fn into_main_trie_diff(self) -> TrieDiff

Returns a diff of the main trie.