Module smoldot::executor::runtime_host

Wasm virtual machine, with automatic storage overlay and logs management.

The code in this module builds upon the functionalities of the host module and implements some of the host function calls. In other words, it is an easier-to-use version of the host module.

Most of the documentation of the host module also applies here.

In addition to the functionalities provided by the host module, the runtime_host module:

Keeps track of the changes to the storage and off-chain storage made by the execution, and provides them at the end. Any storage access takes into account the intermediary list of changes.

Keeps track of the logs generated by the call and concatenates them into a String.

Automatically handles some externalities, such as calculating the Merkle root or storage transactions.

These additional features considerably reduces the number of externals concepts to plug to the virtual machine.

Structs

Config

Configuration for run.

Error

Error that can happen during the execution.

NextKey

Fetching the key that follows a given one is required in order to continue.

PrefixKeys

Fetching the list of keys with a given prefix is required in order to continue.

SignatureVerification

Verifying whether a signature is correct is required in order to continue.

StorageGet

Loading a storage value is required in order to continue.

Success

Execution is successful.

SuccessVirtualMachine

Function execution has succeeded. Contains the return value of the call.

Enums

ErrorDetail

See Error::detail.

RuntimeHostVm

Current state of the execution.

Functions

run

Start running the WebAssembly virtual machine.