Struct smoldot::executor::storage_diff::StorageDiff

pub struct StorageDiff { /* private fields */ }

Implementations

impl StorageDiff

pub fn empty() -> Self

Builds a new empty diff.

pub fn clear(&mut self)

Removes all the entries within this diff.

pub fn diff_insert(
&mut self,
key: impl Into<Vec<u8>>,
value: impl Into<Vec<u8>>
) -> Option<Option<Vec<u8>>>

Inserts the given key-value combination in the diff.

Returns the value associated to this key that was previously in the diff, if any.

pub fn diff_insert_erase(
&mut self,
key: impl Into<Vec<u8>>
) -> Option<Option<Vec<u8>>>

Inserts in the diff an entry at the given key that delete the value that is located in the base storage.

Returns the value associated to this key that was previously in the diff, if any.

pub fn diff_remove(&mut self, key: impl AsRef<[u8]>) -> Option<Option<Vec<u8>>>

Removes from the diff the entry corresponding to the given key.

Returns the value associated to this key that was previously in the diff, if any.

pub fn diff_get(&self, key: &[u8]) -> Option<Option<&[u8]>>

Returns the diff entry at the given key.

Returns None if the diff doesn’t have any entry for this key, and Some(None) if the diff has an entry that deletes the storage item.

pub fn diff_iter_unordered(
&self
) -> impl ExactSizeIterator<Item = (&[u8], Option<&[u8]>)> + Clone

Returns an iterator to all the entries in the diff.

Each value is either Some if the diff overwrites this diff, or None if it erases the underlying value.

pub fn diff_into_iter_unordered(
self
) -> impl ExactSizeIterator<Item = (Vec<u8>, Option<Vec<u8>>)>

Returns an iterator to all the entries in the diff.

Each value is either Some if the diff overwrites this diff, or None if it erases the underlying value.

pub fn storage_get<'a, 'b>(
&'a self,
key: &'b [u8],
or_parent: impl FnOnce() -> Option<&'a [u8]>
) -> Option<&'a [u8]>

Returns the storage value at the given key. None if this key doesn’t have any value.

pub fn storage_next_key<'a, 'b>(
&'a self,
key: &'b [u8],
in_parent_next_key: Option<&'a [u8]>
) -> StorageNextKey<'a>

Returns the storage key that immediately follows the provided key. Must be passed the storage key that immediately follows the provided key according to the base storage this diff is based upon.

If StorageNextKey::Found is returned, it contains the desired key. If StorageNextKey::NextOf is returned, then this function should be called again but by passing the key found in the StorageNextKey::NextOf (and of course the corresponding in_parent_next_key).

Panic

Panics if in_parent_next_key is provided and is inferior or equal to key.

pub fn storage_prefix_keys_ordered<'a>(
&'a self,
prefix: &'a [u8],
in_parent_ordered: impl Iterator<Item = impl AsRef<[u8]> + 'a> + 'a
) -> impl Iterator<Item = impl AsRef<[u8]> + 'a> + 'a

Takes as parameter a list (in_parent_ordered) of all the keys that are present in the storage this diff is based upon and that start with the given prefix.

Returns another iterator that provides the list of all keys that start with the given prefix after this diff has been applied.

The list must be lexicographically ordered. The returned list is ordered lexicographically as well.

pub fn merge(&mut self, other: &StorageDiff)

Applies the given diff on top of the current one.