Enum smoldot::executor::vm::NewErr

pub enum NewErr {
    UnresolvedFunctionImport {
        function: String,
        module_name: String,
    },
    StartFunctionNotSupported,
    MemoryIsntMemory,
    MemoryNotNamedMemory,
    NoMemory,
    TwoMemories,
    IndirectTableIsntTable,
    CouldntAllocateMemory,
    ModuleError(ModuleError),
}

Error that can happen when initializing a VirtualMachinePrototype.

Variants

UnresolvedFunctionImport

Fields

function: String

Name of the function that was unresolved.

module_name: String

Name of module associated with the unresolved function.

Failed to resolve a function imported by the module.

StartFunctionNotSupported

Smoldot doesnโt support wasm runtime that have a start function. It is unclear whether this is allowed in the Substrate/Polkadot specification.

MemoryIsntMemory

If a โmemoryโ symbol is provided, it must be a memory.

MemoryNotNamedMemory

Wasm module imports a memory that isnโt named โmemoryโ.

NoMemory

Wasm module doesnโt contain any memory.

TwoMemories

Wasm module both imports and exports a memory.

IndirectTableIsntTable

If a __indirect_function_table symbol is provided, it must be a table.

CouldntAllocateMemory

Failed to allocate memory for the virtual machine.

ModuleError(ModuleError)

Error while parsing or compiling the WebAssembly code.