Struct smoldot::header::HeaderRef

pub struct HeaderRef<'a> {
    pub parent_hash: &'a [u8; 32],
    pub number: u64,
    pub state_root: &'a [u8; 32],
    pub extrinsics_root: &'a [u8; 32],
    pub digest: DigestRef<'a>,
}

Header of a block, after decoding.

Note that the information in there are not guaranteed to be exact. The exactness of the information depends on the context.

Fields
parent_hash: &'a [u8; 32]

Hash of the parent block stored in the header.

number: u64

Block number stored in the header.

state_root: &'a [u8; 32]

The state trie Merkle root

extrinsics_root: &'a [u8; 32]

The Merkle root of the extrinsics.

You can use the extrinsics_root function to compute this value.

digest: DigestRef<'a>

List of auxiliary data appended to the block header.

Implementations

impl<'a> HeaderRef<'a>

pub fn scale_encoding( &self, block_number_bytes: usize ) -> impl Iterator<Item = impl AsRef<[u8]> + Clone + 'a> + Clone + 'a

Returns an iterator to list of buffers which, when concatenated, produces the SCALE encoding of the header.

pub fn scale_encoding_with_extra_digest_item( &self, block_number_bytes: usize, extra_item: DigestItemRef<'a> ) -> impl Iterator<Item = impl AsRef<[u8]> + Clone + 'a> + Clone + 'a

Same as HeaderRef::scale_encoding, but with an extra digest item added to the end of the list.

Note: Keep in mind that this can produce a potentially invalid header, for example if the digest contains duplicate items that shouldn’t be duplicate.

pub fn scale_encoding_vec(&self, block_number_bytes: usize) -> Vec<u8>

Equivalent to HeaderRef::scale_encoding but returns the data in a Vec.

pub fn hash(&self, block_number_bytes: usize) -> [u8; 32]

Builds the hash of the header.