Struct smoldot::identity::keystore::Keystore

pub struct Keystore { /* private fields */ }

Collection of key pairs.

This module doesn’t give you access to the content of private keys, only to signing capabilities.

Implementations

impl Keystore

pub async fn new(
keys_directory: Option<PathBuf>,
randomness_seed: [u8; 32]
) -> Result<Self, Error>

Initializes a new keystore.

Must be passed bytes of entropy that are used to avoid hash collision attacks and to generate private keys.

An error is returned if the keys_directory couldn’t be opened because, for example, of some missing permission or because it isn’t a directory. If the keys_directory doesn’t exist, it will be created using fs::create_dir_all.

pub fn insert_sr25519_memory(
&mut self,
namespaces: impl Iterator<Item = KeyNamespace>,
private_key: &[u8; 64]
) -> [u8; 32]

Inserts an Sr25519 private key in the keystore.

Returns the corresponding public key.

This is meant to be called with publicly-known private keys. Use Keystore::generate_sr25519 if the private key is meant to actually be private.

The key is not saved on disk.

Panic

Panics if the key isn’t a valid Sr25519 private key. This function is meant to be used with hard coded values which are known to be correct. Please do not call it with any sort of user input.

pub async fn generate_ed25519(
&self,
namespace: KeyNamespace,
save: bool
) -> Result<[u8; 32], Error>

Generates a new Ed25519 key and inserts it in the keystore.

If save is true, the generated key is saved in the file system. This function returns an error only if save is true and the key couldn’t be written to the file system. The value of save is silently ignored if no path was provided to Keystore::new.

Returns the corresponding public key.

pub async fn keys(&self) -> impl Iterator<Item = (KeyNamespace, [u8; 32])>

Returns the list of all keys known to this keystore.

Note: Keep in mind that this function is racy, as keys can be added and removed in parallel of this function being called.

pub async fn generate_sr25519(
&self,
namespace: KeyNamespace,
save: bool
) -> Result<[u8; 32], Error>

Generates a new Sr25519 key and inserts it in the keystore.

If save is true, the generated key is saved in the file system. This function returns an error only if save is true and the key couldn’t be written to the file system. The value of save is silently ignored if no path was provided to Keystore::new.

Returns the corresponding public key.

pub async fn sign(
&self,
key_namespace: KeyNamespace,
public_key: &[u8; 32],
payload: &[u8]
) -> Result<[u8; 64], SignError>

Signs the given payload using the private key associated to the public key passed as parameter.

An error is returned if the key-namespace combination is not in the keystore, or if the key couldn’t be loaded from disk. In the case when a key couldn’t be loaded from disk, it is automatically removed from the keystore.

pub fn sign_sr25519_vrf<'a>(
&'a self,
key_namespace: KeyNamespace,
public_key: &'a [u8; 32],
label: &'static [u8],
transcript_items: impl Iterator<Item = (&'static [u8], Either<&'a [u8], u64>)> + 'a
) -> impl Future<Output = Result<VrfSignature, SignVrfError>> + 'a

Note that the labels must be 'static due to requirements from the underlying library.