smoldot::identity::seed_phrase
Function bip39_to_seed

pub fn bip39_to_seed(
    phrase: &str,
    password: &str,
) -> Result<Box<[u8; 32]>, Bip39ToSeedError>

Turns a BIP39 seed phrase into a 32 bytes cryptographic seed.

Note: The key is returned within a Box in order to guarantee that no trace of the secret key is accidentally left in memory due to automatic copies of stack data.