Enum smoldot::json_rpc::methods::MethodCall

pub enum MethodCall<'a> {

    account_nextIndex {},
    author_hasKey {},
    author_hasSessionKeys {},
    author_insertKey {},
    author_pendingExtrinsics {},
    author_removeExtrinsic {},
    author_rotateKeys {},
    author_submitAndWatchExtrinsic {
        transaction: HexString,
    },
    author_submitExtrinsic {
        transaction: HexString,
    },
    author_unwatchExtrinsic {
        subscription: &'a str,
    },
    babe_epochAuthorship {},
    chain_getBlock {
        hash: Option<HashHexString>,
    },
    chain_getBlockHash {
        height: Option<u64>,
    },
    chain_getFinalizedHead {},
    chain_getHeader {
        hash: Option<HashHexString>,
    },
    chain_subscribeAllHeads {},
    chain_subscribeFinalizedHeads {},
    chain_subscribeNewHeads {},
    chain_unsubscribeAllHeads {
        subscription: String,
    },
    chain_unsubscribeFinalizedHeads {
        subscription: String,
    },
    chain_unsubscribeNewHeads {
        subscription: String,
    },
    childstate_getKeys {},
    childstate_getStorage {},
    childstate_getStorageHash {},
    childstate_getStorageSize {},
    grandpa_roundState {},
    offchain_localStorageGet {},
    offchain_localStorageSet {},
    payment_queryInfo {
        extrinsic: HexString,
        hash: Option<HashHexString>,
    },
    rpc_methods {},
    state_call {},
    state_getKeys {},
    state_getKeysPaged {
        prefix: Option<HexString>,
        count: u32,
        start_key: Option<HexString>,
        hash: Option<HashHexString>,
    },
    state_getMetadata {},
    state_getPairs {},
    state_getReadProof {},
    state_getRuntimeVersion {
        at: Option<HashHexString>,
    },
    state_getStorage {
        key: HexString,
        hash: Option<HashHexString>,
    },
    state_getStorageHash {},
    state_getStorageSize {},
    state_queryStorage {},
    state_queryStorageAt {
        keys: Vec<HexString>,
        at: Option<HashHexString>,
    },
    state_subscribeRuntimeVersion {},
    state_subscribeStorage {
        list: Vec<HexString>,
    },
    state_unsubscribeRuntimeVersion {},
    state_unsubscribeStorage {
        subscription: &'a str,
    },
    system_accountNextIndex {
        account: AccountId,
    },
    system_addReservedPeer {},
    system_chain {},
    system_chainType {},
    system_dryRun {},
    system_health {},
    system_localListenAddresses {},
    system_localPeerId {},
    system_name {},
    system_networkState {},
    system_nodeRoles {},
    system_peers {},
    system_properties {},
    system_removeReservedPeer {},
    system_version {},

}

Variants

account_nextIndex

Fields of account_nextIndex

author_hasKey

Fields of author_hasKey

author_hasSessionKeys

Fields of author_hasSessionKeys

author_insertKey

Fields of author_insertKey

author_pendingExtrinsics

Fields of author_pendingExtrinsics

author_removeExtrinsic

Fields of author_removeExtrinsic

author_rotateKeys

Fields of author_rotateKeys

author_submitAndWatchExtrinsic

Fields of author_submitAndWatchExtrinsic

transaction: HexString

author_submitExtrinsic

Fields of author_submitExtrinsic

transaction: HexString

author_unwatchExtrinsic

Fields of author_unwatchExtrinsic

subscription: &'a str

babe_epochAuthorship

Fields of babe_epochAuthorship

chain_getBlock

Fields of chain_getBlock

hash: Option<HashHexString>

chain_getBlockHash

Fields of chain_getBlockHash

height: Option<u64>

chain_getFinalizedHead

Fields of chain_getFinalizedHead

chain_getHeader

Fields of chain_getHeader

hash: Option<HashHexString>

chain_subscribeAllHeads

Fields of chain_subscribeAllHeads

chain_subscribeFinalizedHeads

Fields of chain_subscribeFinalizedHeads

chain_subscribeNewHeads

Fields of chain_subscribeNewHeads

chain_unsubscribeAllHeads

Fields of chain_unsubscribeAllHeads

subscription: String

chain_unsubscribeFinalizedHeads

Fields of chain_unsubscribeFinalizedHeads

subscription: String

chain_unsubscribeNewHeads

Fields of chain_unsubscribeNewHeads

subscription: String

childstate_getKeys

Fields of childstate_getKeys

childstate_getStorage

Fields of childstate_getStorage

childstate_getStorageHash

Fields of childstate_getStorageHash

childstate_getStorageSize

Fields of childstate_getStorageSize

grandpa_roundState

Fields of grandpa_roundState

offchain_localStorageGet

Fields of offchain_localStorageGet

offchain_localStorageSet

Fields of offchain_localStorageSet

payment_queryInfo

Fields of payment_queryInfo

extrinsic: HexStringhash: Option<HashHexString>

rpc_methods

Returns a list of all JSON-RPC methods that are available.

Fields of rpc_methods

state_call

Fields of state_call

state_getKeys

Fields of state_getKeys

state_getKeysPaged

Fields of state_getKeysPaged

prefix: Option<HexString>count: u32start_key: Option<HexString>hash: Option<HashHexString>

state_getMetadata

Fields of state_getMetadata

state_getPairs

Fields of state_getPairs

state_getReadProof

Fields of state_getReadProof

state_getRuntimeVersion

Fields of state_getRuntimeVersion

at: Option<HashHexString>

state_getStorage

Fields of state_getStorage

key: HexStringhash: Option<HashHexString>

state_getStorageHash

Fields of state_getStorageHash

state_getStorageSize

Fields of state_getStorageSize

state_queryStorage

Fields of state_queryStorage

state_queryStorageAt

Fields of state_queryStorageAt

keys: Vec<HexString>at: Option<HashHexString>

state_subscribeRuntimeVersion

Fields of state_subscribeRuntimeVersion

state_subscribeStorage

Fields of state_subscribeStorage

list: Vec<HexString>

state_unsubscribeRuntimeVersion

Fields of state_unsubscribeRuntimeVersion

state_unsubscribeStorage

Fields of state_unsubscribeStorage

subscription: &'a str

system_accountNextIndex

Fields of system_accountNextIndex

account: AccountId

system_addReservedPeer

Fields of system_addReservedPeer

system_chain

Fields of system_chain

system_chainType

Fields of system_chainType

system_dryRun

Fields of system_dryRun

system_health

Fields of system_health

system_localListenAddresses

Fields of system_localListenAddresses

system_localPeerId

Fields of system_localPeerId

system_name

Returns, as an opaque string, the name of the client serving these JSON-RPC requests.

Fields of system_name

system_networkState

Fields of system_networkState

system_nodeRoles

Fields of system_nodeRoles

system_peers

Fields of system_peers

system_properties

Fields of system_properties

system_removeReservedPeer

Fields of system_removeReservedPeer

system_version

Returns, as an opaque string, the version of the client serving these JSON-RPC requests.

Fields of system_version

Implementations

impl<'a> MethodCall<'a>

pub fn method_names() -> impl ExactSizeIterator<Item = &'static str>

Returns a list of RPC method names of all the methods in the MethodCall enum.