Function smoldot::json_rpc::parse::build_call

pub fn build_call(call: Call<'_>) -> String

Builds a JSON call.

method must be the name of the method to call. params_json must be the JSON-formatted object or array containing the parameters of the call.

Panic

Panics if the Call::id_json or Call::params_json isn’t valid JSON.