Struct smoldot::json_rpc::requests_subscriptions::RequestsSubscriptions

pub struct RequestsSubscriptions { /* private fields */ }

Implementations

impl RequestsSubscriptions

pub fn new(config: Config) -> Self

Creates a new empty state machine.

pub fn set_max_clients(&self, max_clients: usize)

Changes the maximum number of allowed clients.

Note that the ordering of this operation is Relaxed, meaning that this change will apply some time in the future and not necessarily immediately.

Lowering the maximum number of allowed clients below the current number will not remove existing clients. If this is desired, this must be done manually by calling RequestsSubscriptions::remove_client.

Note: This function can typically be used at runtime to adjust the maximum number of clients based on the resource consumption of the binary.

pub async fn add_client(&self) -> Result<ClientId, AddClientError>

Adds a new client to the state machine. A new ClientId is attributed.

Can return an error if the maximum simultaneous number of clients has been reached.

A single instance of RequestsSubscriptions will never allocate multiple times the same ClientId.

pub fn add_client_mut(&mut self) -> Result<ClientId, AddClientError>

Similar to RequestsSubscriptions::add_client, but non-async and takes self as &mut.

Note: This function is notably useful for adding clients at initialization, when outside of an asynchronous context.

pub async fn remove_client(
&self,
client: &ClientId
) -> Option<(Vec<RequestId>, Vec<SubscriptionId>)>

Removes from the state machine the client with the given id.

Returns None if this ClientId is stale or invalid.

This function invalidates all active requests and subscriptions that relate to this client. The concerned RequestIds and SubscriptionIds are returned by this function.

Note however that functions such as RequestsSubscriptions::respond and RequestsSubscriptions::push_notification intentionally have no effect if you pass an invalid RequestId or SubscriptionId. There is therefore no need to cancel any parallel task that might currently be responding to requests or pushing notification messages.

pub async fn next_response(&self, client: &ClientId) -> String

Waits until a message is available in the queue of messages to send back to the given client, then extracts it.

This asynchronous function can take a long time, as it blocks until a message is available.

If the ClientId no longer exists because of a call to RequestsSubscriptions, then this function never returns. It is the responsibility of the API user to not call this function or to interrupt a pending function call with a stale ClientId.

pub async fn queue_client_request(&self, client: &ClientId, request: String)

Waits until a slot in the queue of requests is available, then queues the given request passed as parameter.

The request can later be retrieved by calling RequestsSubscriptions::next_request.

This asynchronous function can take a long time, as it blocks until a queue entry is available.

Slots in the queue of requests are only reclaimed after RequestsSubscriptions::next_response has returned a response to a previous request.

Has no effect if the ClientId is stale or invalid.

pub fn try_queue_client_request(
&self,
client: &ClientId,
request: String
) -> Result<(), TryQueueClientRequestError>

Similar to RequestsSubscriptions::queue_client_request, but succeeds or fails instantly depending on whether there is enough room in the queue.

Slots in the queue of requests are only reclaimed after RequestsSubscriptions::next_response has returned a response to a previous request.

Returns Ok and silently discards the request if the ClientId is stale or invalid.

pub async fn next_request(&self) -> (String, RequestId)

Waits until a request has been queued using RequestsSubscriptions::queue_client_request and returns it, alongside with an identifier to later pass back when answering the request.

Note that the request’s body, as a String has no guarantee to be valid. The String is simply the value that was passed to RequestsSubscriptions::queue_client_request and isn’t parsed or validated by the state machine in any way.

pub async fn respond(&self, request: &RequestId, response: String)

Sets the response to a request previously returned by RequestsSubscriptions::next_request.

Has no effect if the RequestId is invalid, for example if it has already been responded or if the client has been removed.

pub async fn start_subscription(
&self,
client: &RequestId,
messages_capacity: usize
) -> Result<SubscriptionId, StartSubscriptionError>

Adds a new subscription to the state machine, associated with the client that started the given request.

Returns an error if the client has reached the maximum number of allowed subscriptions per client.

If the given RequestId is stale or invalid, this function always succeeds and returns a stale SubscriptionId.

About the messages capacity

The messages_capacity parameter contains the number of notifications related to this notification that can be queued for send back simultaneously.

This value is one of the parameters that bound the total memory usage of this state machine. In other words, if there was no maximum, a malicious JSON-RPC client could intentionally create a memory leak on the server.

It is therefore important that this value isn’t too large never gets decided by the JSON-RPC client.

For some JSON-RPC functions, the value of this constant can easily be deduced from the logic of the function. For other functions, the value of this constant should be hard coded.

pub async fn stop_subscription(&self, subscription: &SubscriptionId)

Destroys the given subscription.

All messages already queued will still be available through RequestsSubscriptions::next_response, but no new subscription message can be pushed.

This function should be seen as a way to clean up the internal state of the state machine and prevent new notifications from being pushed.

Has no effect if the SubscriptionId is stale or invalid.

pub async fn set_queued_notification(
&self,
subscription: &SubscriptionId,
index: usize,
message: String
)

Overwrites the notification whose index is index in the queue of notifications destined for the user.

The queue of notifications can be sent as equivalent to a Vec<Option<String>> whose length is equal to the messages_capacity that was passed to RequestsSubscriptions::start_subscription, and this function does queue[index] = Some(message);. It discards the message that was previously there, and works no matter the presence or not of other messages in the queue.

Note that notifications are not provided to RequestsSubscriptions::next_response in the order of their index, but in the order in which they entered the queue. If there was no notification overwritten, then this notification is now at the end of the list of notifications to send back. If a notification is overwritten, then only its content is modified but not its position in the queue.

This function isn’t meant to interact well with RequestsSubscriptions::try_push_notification. For each subscription, you are expected to either push notifications one behind the other, or track which notification queue index corresponds to what, but not both at the same time.

Has no effect if the SubscriptionId is stale or invalid.

Panic

Panics if the index is superior or equal to the messages_capacity that was passed to RequestsSubscriptions::start_subscription.

pub async fn push_notification(
&self,
subscription: &SubscriptionId,
message: String
)

Adds the given notification to the queue of notifications to send out for this subscription.

This function will choose an index with no notification, and write the notification to it. If the queue is full, this function will wait for a slot to be available. Slots will only become available if RequestsSubscriptions::next_response is called in parallel.

This function isn’t meant to interact well with RequestsSubscriptions::set_queued_notification. For each subscription, you are expected to either push notifications one behind the other, or track which notification queue index corresponds to what, but not both at the same time.

Has no effect and silently discards the message if the SubscriptionId is stale or invalid.

pub async fn try_push_notification(
&self,
subscription: &SubscriptionId,
message: String
) -> Result<(), ()>

Adds the given notification to the queue of notifications to send out for this subscription.

This function will choose an index with no notification, and write the notification to it. Returns an error if the queue is full.

This function isn’t meant to interact well with RequestsSubscriptions::set_queued_notification. For each subscription, you are expected to either push notifications one behind the other, or track which notification queue index corresponds to what, but not both at the same time.

Note that notifications are not provided to RequestsSubscriptions::next_response in the order of their index, but in the order in which they entered the queue.

Returns Ok and silently discards the message if the SubscriptionId is stale or invalid.