Enum smoldot::json_rpc::websocket_server::Event

pub enum Event<'a, T> {
    ConnectionOpen {
        address: SocketAddr,
    },
    ConnectionError {
        connection_id: ConnectionId,
        user_data: T,
    },
    TextFrame {
        connection_id: ConnectionId,
        user_data: &'a mut T,
        message: String,
    },
}

Event that has happened on a WsServer.

Variants

ConnectionOpen

Fields

address: SocketAddr

Address of the remote, as provided by the operating system.

A new TCP connection has arrived on the listening socket.

The connection must be accepted or rejected using WsServer::accept or WsServer::reject. No other Event::ConnectionOpen event will be generated until the current pending connection has been either accepted or rejected.

ConnectionError

Fields

connection_id: ConnectionId

Identifier of the connection. This identifier might be reused by the WsServer for another connection.

user_data: T

User data associated with the connection.

An error has happened on a connection. The connection is now closed and its ConnectionId is now invalid.

TextFrame

Fields

connection_id: ConnectionId

Identifier of the connection that sent the frame.

user_data: &'a mut T

User data associated with the connection.

message: String

Message sent by the remote. Its content is entirely decided by the client, and nothing must be assumed about the validity of this message.

A text frame has been received on a connection.