smoldot::libp2p

Module cid

CID is a data type for referencing content in IPFS, including Bitswap protocol.

See https://github.com/multiformats/cid/blob/master/README.md.

For simplicity, we support only CIDv1 version used in Substrate Bitswap server implementation (without support for CIDv0).

Binary representation of CIDv1

cidv1 ::= <leb128(cid_version)><leb128(codec)><leb128(mh_type)><leb128(mh_len)><hash_digest>

Subparts of the CIDv1 binary representation we are interested in are CID prefix & multihash correspondingly:

cid_prefix ::= <leb128(cid_version)><leb128(codec)><leb128(mh_type)><leb128(mh_len)> multihash ::= <leb128(mh_type)><leb128(mh_len)><hash_digest>

Here leb128() means LEB128 varint encoded values.

In Bitswap requests, complete cidv1 of requested data block is sent, while the Bitswap responses only include the cid_prefix. This means we should manually compute the hash digest identified by mh_type in order to recover the complete CID and match the received data block to the request (request and response messages are sent asynchronously in Bitswap through independent substreams).

Structs

Cid

CID: IPFS Content Identifier.

CidPrefix

CID prefix: everything in a CID except the hash digest.

Enums

MultihashType

Multihash algorithm types supported.

ParseError

Error when parsing string/binary Cid.