Enum smoldot::libp2p::collection::Event

pub enum Event<TConn> {

    HandshakeFinished {
        id: ConnectionId,
        peer_id: PeerId,
    },
    StartShutdown {
        id: ConnectionId,
        reason: ShutdownCause,
    },
    Shutdown {
        id: ConnectionId,
        was_established: bool,
        user_data: TConn,
    },
    InboundError {
        id: ConnectionId,
        error: InboundError,
    },
    Response {
        substream_id: SubstreamId,
        response: Result<Vec<u8>, RequestError>,
    },
    RequestIn {
        id: ConnectionId,
        substream_id: SubstreamId,
        protocol_index: usize,
        request_payload: Vec<u8>,
    },
    RequestInCancel {
        substream_id: SubstreamId,
    },
    NotificationsOutResult {
        substream_id: SubstreamId,
        result: Result<Vec<u8>, NotificationsOutErr>,
    },
    NotificationsOutCloseDemanded {
        substream_id: SubstreamId,
    },
    NotificationsOutReset {
        substream_id: SubstreamId,
    },
    NotificationsInOpen {
        id: ConnectionId,
        substream_id: SubstreamId,
        notifications_protocol_index: usize,
        remote_handshake: Vec<u8>,
    },
    NotificationsInOpenCancel {
        substream_id: SubstreamId,
    },
    NotificationsIn {
        substream_id: SubstreamId,
        notification: Vec<u8>,
    },
    NotificationsInClose {
        substream_id: SubstreamId,
        outcome: Result<(), NotificationsInClosedErr>,
    },
    PingOutSuccess {
        id: ConnectionId,
    },
    PingOutFailed {
        id: ConnectionId,
    },

}

Event generated by Network::next_event.

Variants

HandshakeFinished

Fields

id: ConnectionId

Identifier of the connection whose handshake is finished.

peer_id: PeerId

Identity of the peer on the other side of the connection.

Handshake of the given connection has completed.

This event can only happen once per connection and only for single-stream connections.

StartShutdown

Fields

id: ConnectionId

Identifier of the connection that is starting its shutdown.

reason: ShutdownCause

Reason why the connection is starting its shutdown. Because this event is not generated when the shutdown is initiated locally, the reason is always cause by the remote.

A transport-level connection (e.g. a TCP socket) is starting its shutdown.

It is no longer possible to start requests, open notification substreams, or open notifications on this connection, and no new incoming requests or notification substreams will be reported as events.

Further events will close all existing substreams (requests and notifications) one by one. Once all substreams have been closed, a Event::Shutdown is reported.

Keep in mind that this event can also happen for connections that haven’t finished their handshake.

This event is not generated when Network::start_shutdown is called.

Shutdown

Fields

id: ConnectionId

Identifier of the connection that has finished its shutdown.

was_established: bool

true if the connection was in its established phase before the shutdown.

user_data: TConn

User data that was stored in the state machine for this connection.

A transport-level connection (e.g. a TCP socket) has been shut down.

This ConnectionId is no longer valid, and using it will result in panics.

InboundError

Fields

id: ConnectionId

Identifier of the connection that has received the substream.

error: InboundError

Error that happened.

Received an incoming substream, but this substream has produced an error.

Note: This event exists only for diagnostic purposes. No action is expected in return.

Response

Fields

substream_id: SubstreamId

Substream that was returned by Network::start_request.

response: Result<Vec<u8>, RequestError>

If the request is successful, contains the response sent back by the remote. Otherwise, contains the reason why the request isn’t successful.

Outcome of a request started using Network::start_request.

All requests always lead to an outcome, even if the connection has been closed while the request was in progress.

RequestIn

Fields

id: ConnectionId

Identifier of the connection that has received the request.

substream_id: SubstreamId

Substream on which the request has been received. Must be passed back when providing the response.

protocol_index: usize

Index of the negotiated protocol within Config::request_response_protocols.

request_payload: Vec<u8>

Payload that has been sent by the remote. Its interpretation is beyond the scope of this module.

Received a request from a request-response protocol.

RequestInCancel

Fields

substream_id: SubstreamId

Request received earlier has been canceled by the remote.

The SubstreamId is now invalid.

NotificationsOutResult

Fields

substream_id: SubstreamId

result: Result<Vec<u8>, NotificationsOutErr>

If Ok, contains the handshake sent back by the remote. Its interpretation is out of scope of this module.

Outcome of trying to open a substream with Network::open_out_notifications.

If Ok, it is now possible to send notifications on this substream. If Err, the substream no longer exists and the SubstreamId becomes invalid.

NotificationsOutCloseDemanded

Fields

substream_id: SubstreamId

Remote has closed an outgoing notifications substream, meaning that it demands the closing of the substream. Use Network::close_out_notifications as soon as possible, which is typically after all outbound notifications that need to be queued have been queued.

This event is only generated for notification substreams that are fully open.

NotificationsOutReset

Fields

substream_id: SubstreamId

A previously open outbound substream has been closed, by the remote or as a consequence of the connection shutting down.

This event is only generated for notification substreams that are fully open.

The substream no longer exists and the SubstreamId becomes invalid.

NotificationsInOpen

Fields

id: ConnectionId

Identifier of the connection that has received the notification substream request.

substream_id: SubstreamId

Newly-generated identifier for the substream on which the request has been received. Must be passed back when accepting or refusing the substream.

notifications_protocol_index: usize

Index of the negotiated protocol within Config::notification_protocols.

remote_handshake: Vec<u8>

Handshake that has been sent by the remote. Its interpretation is beyond the scope of this module.

The remote would like to open a notifications substream.

The substream needs to be accepted or refused using Network::accept_in_notifications or Network::reject_in_notifications.

NotificationsInOpenCancel

Fields

substream_id: SubstreamId

Substream that has been closed. Guaranteed to match a substream that was earlier reported with a Event::NotificationsInOpen.

The remote has canceled the opening an incoming notifications substream.

This can only happen before the notification substream has been accepted or refused.

NotificationsIn

Fields

substream_id: SubstreamId

Substream on which the notification has been received. Guaranteed to be a substream that has been accepted with Network::accept_in_notifications.

notification: Vec<u8>

Notification that the remote has sent. The meaning of this data is out of scope of this module.

Received a notification on a notifications substream of a connection.

NotificationsInClose

Fields

substream_id: SubstreamId

Substream that has been closed. Guaranteed to match a substream that was earlier reported with a Event::NotificationsInOpen.

outcome: Result<(), NotificationsInClosedErr>

Reason why the substream has been closed.

The remote has closed an incoming notifications substream.

This can only happen after the notification substream has been accepted.

PingOutSuccess

Fields

id: ConnectionId

An outgoing ping has succeeded. This event is generated automatically over time for each connection in the collection.

PingOutFailed

Fields

id: ConnectionId

An outgoing ping has failed. This event is generated automatically over time for each connection in the collection.