Struct smoldot::libp2p::collection::MultiStreamConnectionTask

pub struct MultiStreamConnectionTask<TNow, TSubId> { /* private fields */ }

State machine dedicated to a single multi-stream connection.

Implementations

impl<TNow, TSubId> MultiStreamConnectionTask<TNow, TSubId>
where TNow: Clone + Add<Duration, Output = TNow> + Sub<TNow, Output = Duration> + Ord, TSubId: Clone + PartialEq + Eq + Hash,

pub fn pull_message_to_coordinator( self ) -> (Option<Self>, Option<ConnectionToCoordinator>)

Pulls a message to send back to the coordinator.

This function takes ownership of self and optionally yields it back. If the first option contains None, then no more message will be generated and the MultiStreamConnectionTask has vanished. This will happen after the connection has been shut down or reset. It is possible for self to not be yielded back even if substreams are still open, in which case the API user should abruptly reset the connection, for example by sending a TCP RST flag.

If any message is returned, it is the responsibility of the API user to send it to the coordinator. Do not attempt to buffer the message being returned, as it would work against the back-pressure strategy used internally. As soon as a message is returned, it should be delivered. If the coordinator is busy at the moment a message should be delivered, then the entire thread of execution dedicated to this MultiStreamConnectionTask should be paused until the coordinator is ready and the message delivered.

Messages aren’t generated spontaneously. In other words, you don’t need to periodically call this function just in case there’s a new message. Messages are always generated after MultiStreamConnectionTask::substream_read_write, MultiStreamConnectionTask::add_substream, or MultiStreamConnectionTask::reset has been called. Multiple messages can happen in a row.

Because this function frees space in a buffer, processing substreams again after it has returned might read/write more data and generate an event again. In other words, the API user should call MultiStreamConnectionTask::substream_read_write and MultiStreamConnectionTask::pull_message_to_coordinator repeatedly in a loop until no more message is generated.

pub fn inject_coordinator_message( &mut self, now: &TNow, message: CoordinatorToConnection )

Injects a message that has been pulled from the coordinator.

Calling this function might generate data to send to the connection. You should call MultiStreamConnectionTask::desired_outbound_substreams and MultiStreamConnectionTask::substream_read_write after this function has returned.

pub fn desired_outbound_substreams(&self) -> u32

Returns the number of new outbound substreams that the state machine would like to see opened.

This value doesn’t change automatically over time but only after a call to MultiStreamConnectionTask::substream_read_write, MultiStreamConnectionTask::inject_coordinator_message, MultiStreamConnectionTask::add_substream, or MultiStreamConnectionTask::reset_substream.

Note that the user is expected to track the number of substreams that are currently being opened. For example, if this function returns 2 and there are already 2 substreams currently being opened, then there is no need to open any additional one.

pub fn add_substream(&mut self, id: TSubId, outbound: bool)

Notifies the state machine that a new substream has been opened.

outbound indicates whether the substream has been opened by the remote (false) or locally (true).

If outbound is true, then the value returned by MultiStreamConnectionTask::desired_outbound_substreams will decrease by one.

Panic

Panics if there already exists a substream with an identical identifier.

pub fn reset(&mut self)

Sets the state of the connection to “reset”.

This should be called if the remote abruptly closes the connection, such as with a TCP/IP RST flag.

After this function has been called, it is illegal to call MultiStreamConnectionTask::substream_read_write or MultiStreamConnectionTask::reset again.

Calling this function might have generated messages for the coordinator. MultiStreamConnectionTask::pull_message_to_coordinator should be called afterwards in order to process these messages.

Panic

Panics if MultiStreamConnectionTask::reset has been called in the past.

pub fn is_reset_called(&self) -> bool

Returns true if MultiStreamConnectionTask::reset has been called in the past.

pub fn reset_substream(&mut self, substream_id: &TSubId)

Immediately destroys the substream with the given identifier.

The given identifier is now considered invalid by the state machine.

Panic

Panics if there is no substream with that identifier.

pub fn substream_read_write( &mut self, substream_id: &TSubId, read_write: &mut ReadWrite<TNow> ) -> SubstreamFate

Reads/writes data on the substream.

If the method returns SubstreamFate::Reset, then the substream is now considered dead according to the state machine and its identifier is now invalid. If the reading or writing side of the substream was still open, then the user should reset that substream.

In the case of a WebRTC connection, the ReadWrite::incoming_buffer and ReadWrite::write_bytes_queueable must always be Some.

Panic

Panics if there is no substream with that identifier. Panics if this is a WebRTC connection, and the reading or writing side is closed.