Struct smoldot::libp2p::collection::Network

pub struct Network<TConn, TNow> { /* fields omitted */ }

Data structure containing the list of all connections, pending or not, and their latest known state. See also the module-level documentation.

Implementations

impl<TConn, TNow> Network<TConn, TNow> where
TConn: Clone,
TNow: Clone + Add<Duration, Output = TNow> + Sub<TNow, Output = Duration> + Ord,

pub fn new(config: Config) -> Self

Initializes a new network data structure.

pub async fn insert(
&self,
when_connected: TNow,
is_initiator: bool,
user_data: TConn
) -> ConnectionId

Adds a new connection to the collection.

Must be passed the moment (as a TNow) when the connection as been established, in order to determine when the handshake timeout expires.

is_initiator must be true if the connection has been initiated locally, or false if it has been initiated by the remote.

pub async fn len(&self) -> usize

Returns the number of connections in the collection.

Note: Keep in mind that this method is prone to race conditions, as the user can add and remove connections at the same time as this is being called.

pub fn noise_key(&self) -> &NoiseKey

Returns the Noise key originalled passed as Config::noise_key.

pub fn notification_protocols(
&self
) -> impl ExactSizeIterator<Item = &NotificationProtocolConfig>

Returns the list the overlay networks originally passed as Config::notification_protocols.

pub fn request_response_protocols(
&self
) -> impl ExactSizeIterator<Item = &ConfigRequestResponse>

Returns the list the request-response protocols originally passed as Config::request_response_protocols.

pub async fn request(
&self,
now: TNow,
target: ConnectionId,
protocol_index: usize,
request_data: Vec<u8>
) -> Result<Vec<u8>, RequestError>

Sends a request to the given peer, and waits for a response.

This consists in:

Opening a substream on an established connection with the target.

Negotiating the requested protocol (protocol_index) on this substream using the multistream-select protocol.

Sending the request (request_data parameter), prefixed with its length.

Waiting for the response (prefixed with its length), which is then returned.

An error happens if the provided ConnectionId is invalid, if the connection closes while the request is in progress, if the request or response doesn’t respect the protocol limits (see ConfigRequestResponse), or if the remote takes too much time to answer.

As the API of this module is inherently subject to race conditions, it is never possible to guarantee that this function will succeed. RequestError::ConnectionClosed should be handled by retrying the same request again.

Note: This function doesn’t return before the remote has answered. It is strongly recommended to await the returned Future in the background, and not block any important task on this.

Panic

Panics if protocol_index isn’t a valid index in Config::request_response_protocols.

pub async fn open_notifications_substream(
&self,
connection_id: ConnectionId,
overlay_network_index: usize,
now: TNow,
handshake: impl Into<Vec<u8>>
) -> Result<SubstreamId, OpenNotificationsSubstreamError>

Start opening a notifications substream.

pub async fn close_notifications_substream(
&self,
connection_id: ConnectionId,
substream_id: SubstreamId
)

Start closing a previously-open notifications substream, or cancels opening a notifications substream.

Calling this method will emit a Event::NotificationsOutClose if there was indeed a notifications substream being open or already open.

The interaction of this method with Event::NotificationsOutResult and Event::NotificationsOutClose is non-trivial.

It is possible for a Event::NotificationsOutResult to be emitted after this method has been called. In any situation, if a Event::NotificationsOutResult regarding this notifications protocol is emitted in the past, then a corresponding Event::NotificationsOutClose will always be emitted.

Similarly, it is possible for a Event::NotificationsOutClose to be emitted right before this method is called. For this reason, this method has no effect if no notifications substream is open.

pub async fn queue_notification(
&self,
connection: ConnectionId,
substream_id: SubstreamId,
notification: impl Into<Vec<u8>>
) -> Result<(), QueueNotificationError>

Adds a notification to the queue of notifications to send to the given peer.

Each substream maintains a queue of notifications to be sent to the remote. This method attempts to push a notification to this queue.

As the API of this module is inherently subject to race conditions, it is possible for connections and substreams to no longer exist with this peer. This is a normal situation, and the error returned by this function should simply be ignored apart from diagnostic purposes.

An error is also returned if the queue exceeds a certain size in bytes, for two reasons:

Since the content of the queue is transferred at a limited rate, each notification pushed at the end of the queue will take more time than the previous one to reach the destination. Once the queue reaches a certain size, the time it would take for newly-pushed notifications to reach the destination would start being unreasonably large.

If the remote deliberately applies back-pressure on the substream, it is undesirable to increase the memory usage of the local node.

Similarly, the queue being full is a normal situations and notification protocols should be designed in such a way that discarding notifications shouldn’t have a too negative impact.

Regardless of the success of this function, no guarantee exists about the successful delivery of notifications.

pub async fn accept_notifications_in(
&self,
connection: ConnectionId,
substream_id: SubstreamId,
handshake: Vec<u8>
) -> Result<(), ()>

Accepts a request for an inbound notifications substream reported by an Event::NotificationsInOpen.

Returns Ok if the substream has been successfully accepted (and thus opened), or Err if the request was obsolete, for example because the connection has been closed.

pub async fn reject_notifications_in(
&self,
connection: ConnectionId,
substream_id: SubstreamId
) -> Result<(), ()>

Rejects a request for an inbound notifications substream reported by an Event::NotificationsInOpen.

Returns Ok if the substream has been successfully refused, or Err if the request was obsolete, for example because the connection has been closed.

pub async fn respond_in_request(
&self,
id: ConnectionId,
substream_id: SubstreamId,
response: Result<Vec<u8>, ()>
)

Responds to an incoming request. Must be called in response to a Event::RequestIn.

Passing an Err corresponds, on the other side, to a established::RequestError::SubstreamClosed.

Has no effect if the connection has been closed in the meanwhile.

pub async fn next_event(&self) -> Event<TConn>

Returns the next event produced by the service.

This function should be called at a high enough rate that Network::read_write can continue pushing events to the internal buffer of events. Failure to call this function often enough will lead to connections being back-pressured. See also Config::pending_api_events_buffer_size.

It is technically possible to call this function multiple times simultaneously, in which case the events will be distributed amongst the multiple calls in an unspecified way. Keep in mind that some Events have logic attached to the order in which they are produced, and calling this function multiple times is therefore discouraged.

pub async fn read_write(
&self,
connection_id: ConnectionId,
read_write: &mut ReadWrite<'_, TNow>
) -> Result<(), ConnectionError>

Reads data coming from the connection, updates the internal state machine, and writes data destined to the connection through the ReadWrite.

If an error is returned, the connection should be destroyed altogether and the ConnectionId is no longer valid.

Panic

Panics if connection_id isn’t a valid connection.