Struct smoldot::libp2p::collection::SingleStreamConnectionTask

pub struct SingleStreamConnectionTask<TNow> { /* private fields */ }

State machine dedicated to a single single-stream connection.

Implementations

impl<TNow> SingleStreamConnectionTask<TNow>
where TNow: Clone + Add<Duration, Output = TNow> + Sub<TNow, Output = Duration> + Ord,

pub fn pull_message_to_coordinator( self ) -> (Option<Self>, Option<ConnectionToCoordinator>)

Pulls a message to send back to the coordinator.

This function takes ownership of self and optionally yields it back. If the first option contains None, then no more message will be generated and the SingleStreamConnectionTask has vanished. This will happen after the connection has been shut down or reset. It is possible for self to not be yielded back even if the ReadWrite that was last passed to SingleStreamConnectionTask::read_write is still fully open, in which case the API user should abruptly reset the connection, for example by sending a TCP RST flag. This can happen for example if the connection seems unresponsive and that an attempt at closing the connection in a clean way is futile.

If any message is returned, it is the responsibility of the API user to send it to the coordinator. Do not attempt to buffer the message being returned, as it would work against the back-pressure strategy used internally. As soon as a message is returned, it should be delivered. If the coordinator is busy at the moment a message should be delivered, then the entire thread of execution dedicated to this SingleStreamConnectionTask should be paused until the coordinator is ready and the message delivered.

Messages aren’t generated spontaneously. In other words, you don’t need to periodically call this function just in case there’s a new message. Messages are always generated after either SingleStreamConnectionTask::read_write or SingleStreamConnectionTask::reset has been called. Multiple messages can happen in a row.

Because this function frees space in a buffer, calling SingleStreamConnectionTask::read_write again after it has returned might read/write more data and generate an event again. In other words, the API user should call SingleStreamConnectionTask::read_write and SingleStreamConnectionTask::pull_message_to_coordinator repeatedly in a loop until no more message is generated.

pub fn inject_coordinator_message( &mut self, now: &TNow, message: CoordinatorToConnection )

Injects a message that has been pulled from the coordinator.

Calling this function might generate data to send to the connection. You should call SingleStreamConnectionTask::read_write after this function has returned (unless you have called SingleStreamConnectionTask::reset in the past).

pub fn reset(&mut self)

Sets the state of the connection to “reset”.

This should be called if the remote abruptly closes the connection, such as with a TCP/IP RST flag.

After this function has been called, it is illegal to call SingleStreamConnectionTask::read_write or SingleStreamConnectionTask::reset again.

Calling this function might have generated messages for the coordinator. SingleStreamConnectionTask::pull_message_to_coordinator should be called afterwards in order to process these messages.

Panic

Panics if SingleStreamConnectionTask::reset has been called in the past.

pub fn is_reset_called(&self) -> bool

Returns true if SingleStreamConnectionTask::reset has been called in the past.

pub fn read_write(&mut self, read_write: &mut ReadWrite<TNow>)

Reads data coming from the connection, updates the internal state machine, and writes data destined to the connection through the ReadWrite.

Calling this function might have generated messages for the coordinator. SingleStreamConnectionTask::pull_message_to_coordinator should be called afterwards in order to process these messages.

Panic

Panics if SingleStreamConnectionTask::reset has been called in the past.