Enum smoldot::libp2p::connection::established::Event

pub enum Event<TRqUd, TNotifUd> {

    NewOutboundSubstreamsForbidden,
    InboundError(InboundError),
    RequestIn {
        id: SubstreamId,
        protocol_index: usize,
        request: Vec<u8>,
    },
    Response {
        response: Result<Vec<u8>, RequestError>,
        id: SubstreamId,
        user_data: TRqUd,
    },
    NotificationsInOpen {
        id: SubstreamId,
        protocol_index: usize,
        handshake: Vec<u8>,
    },
    NotificationsInOpenCancel {
        id: SubstreamId,
    },
    NotificationIn {
        id: SubstreamId,
        notification: Vec<u8>,
    },
    NotificationsInClose {
        id: SubstreamId,
        outcome: Result<(), NotificationsInClosedErr>,
    },
    NotificationsOutResult {
        id: SubstreamId,
        result: Result<Vec<u8>, (NotificationsOutErr, TNotifUd)>,
    },
    NotificationsOutCloseDemanded {
        id: SubstreamId,
    },
    NotificationsOutReset {
        id: SubstreamId,
        user_data: TNotifUd,
    },
    PingOutSuccess,
    PingOutFailed,

}

Event that happened on the connection. See SingleStream::read_write and MultiStream::pull_event.

Variants

NewOutboundSubstreamsForbidden

The connection is now in a mode where opening new substreams (i.e. starting requests and opening notifications substreams) is forbidden, but the remote is still able to open substreams and messages on existing substreams are still allowed to be sent and received.

InboundError(InboundError)

Received an incoming substream, but this substream has produced an error.

Note: This event exists only for diagnostic purposes. No action is expected in return.

RequestIn

Fields

id: SubstreamId

Identifier of the request. Needs to be provided back when answering the request.

protocol_index: usize

Index of the request-response protocol the request was sent on.

The index refers to the position of the protocol in Config::request_protocols.

request: Vec<u8>

Bytes of the request. Its interpretation is out of scope of this module.

Received a request in the context of a request-response protocol.

Response

Fields

response: Result<Vec<u8>, RequestError>

Bytes of the response. Its interpretation is out of scope of this module.

id: SubstreamId

Identifier of the request. Value that was returned by SingleStream::add_request or MultiStream::add_request.

user_data: TRqUd

Value that was passed to SingleStream::add_request or MultiStream::add_request.

Received a response to a previously emitted request on a request-response protocol.

NotificationsInOpen

Fields

id: SubstreamId

Identifier of the substream. Needs to be provided back when accept or rejecting the substream.

protocol_index: usize

Index of the notifications protocol concerned by the substream.

The index refers to the position of the protocol in Config::notifications_protocols.

handshake: Vec<u8>

Handshake sent by the remote. Its interpretation is out of scope of this module.

Remote has opened an inbound notifications substream.

Either SingleStream::accept_in_notifications_substream or SingleStream::reject_in_notifications_substream, or MultiStream::accept_in_notifications_substream or MultiStream::reject_in_notifications_substream must be called in the near future in order to accept or reject this substream.

NotificationsInOpenCancel

Fields

id: SubstreamId

Identifier of the substream.

Remote has canceled an inbound notifications substream opening.

This can only happen after Event::NotificationsInOpen. SingleStream::accept_in_notifications_substream or SingleStream::reject_in_notifications_substream, or MultiStream::accept_in_notifications_substream or MultiStream::reject_in_notifications_substream should not be called on this substream.

NotificationIn

Fields

id: SubstreamId

Identifier of the substream.

notification: Vec<u8>

Notification sent by the remote.

Remote has sent a notification on an inbound notifications substream. Can only happen after the substream has been accepted.

NotificationsInClose

Fields

id: SubstreamId

Identifier of the substream.

outcome: Result<(), NotificationsInClosedErr>

If Ok, the substream has been closed gracefully. If Err, a problem happened.

Remote has closed an inbound notifications substream.Can only happen after the substream has been accepted.

NotificationsOutResult

Fields

id: SubstreamId

Identifier of the substream. Value that was returned by SingleStream::open_notifications_substream or MultiStream::open_notifications_substream.

result: Result<Vec<u8>, (NotificationsOutErr, TNotifUd)>

If Ok, contains the handshake sent back by the remote. Its interpretation is out of scope of this module.

Outcome of trying to open a substream with SingleStream::open_notifications_substream or MultiStream::open_notifications_substream.

If Ok, it is now possible to send notifications on this substream. If Err, the substream no longer exists.

NotificationsOutCloseDemanded

Fields

id: SubstreamId

Identifier of the substream. Value that was returned by SingleStream::open_notifications_substream or MultiStream::open_notifications_substream.

Remote has closed an outgoing notifications substream, meaning that it demands the closing of the substream.

NotificationsOutReset

Fields

id: SubstreamId

Identifier of the substream. Value that was returned by SingleStream::open_notifications_substream.

user_data: TNotifUd

Value that was passed to SingleStream::open_notifications_substream or MultiStream::open_notifications_substream.

Remote has reset an outgoing notifications substream. The substream is instantly closed.

PingOutSuccess

An outgoing ping has succeeded. This event is generated automatically over time.

PingOutFailed

An outgoing ping has failed. This event is generated automatically over time.