Struct smoldot::libp2p::connection::established::MultiStream

pub struct MultiStream<TNow, TSubId, TRqUd, TNotifUd> { /* private fields */ }

State machine of a fully-established connection where substreams are handled externally.

Implementations

impl<TNow, TSubId, TRqUd, TNotifUd> MultiStream<TNow, TSubId, TRqUd, TNotifUd>where
TNow: Clone + Add<Duration, Output = TNow> + Sub<TNow, Output = Duration> + Ord,
TSubId: Clone + PartialEq + Eq + Hash,

pub fn webrtc(
config: Config<TNow>
) -> MultiStream<TNow, TSubId, TRqUd, TNotifUd>

Creates a new connection from the given configuration.

pub fn pull_event(&mut self) -> Option<Event<TRqUd, TNotifUd>>

Removes an event from the queue of events and returns it.

This method should be called after MultiStream::substream_read_write or MultiStream::reset_substream is called.

pub fn desired_outbound_substreams(&self) -> u32

Returns the number of new outbound substreams that the state machine would like to see opened.

This value doesn’t change automatically over time but only after a call to MultiStream::substream_read_write, MultiStream::add_substream, MultiStream::reset_substream, MultiStream::add_request, or MultiStream::open_notifications_substream.

Note that the user is expected to track the number of substreams that are currently being opened. For example, if this function returns 2 and there are already 2 substreams currently being opened, then there is no need to open any additional one.

pub fn add_substream(&mut self, id: TSubId, outbound: bool)

Notifies the state machine that a new substream has been opened.

outbound indicates whether the substream has been opened by the remote (false) or locally (true).

If outbound is true, then the value returned by MultiStream::desired_outbound_substreams will decrease by one.

Panic

Panics if there already exists a substream with an identical identifier.

pub fn reset_substream(&mut self, substream_id: &TSubId)

Immediately destroys the substream with the given identifier.

The given identifier is now considered invalid by the state machine.

Panic

Panics if there is no substream with that identifier.

pub fn substream_read_write(
&mut self,
substream_id: &TSubId,
read_write: &mut ReadWrite<'_, TNow>
) -> SubstreamFate

Reads/writes data on the substream.

If the method returns SubstreamFate::Reset, then the substream is now considered dead according to the state machine and its identifier is now invalid. If the reading or writing side of the substream was still open, then the user should reset that substream.

This method will refuse to accept data if too many events are already queued. Use MultiStream::pull_event to empty the queue of events between calls to this method.

In the case of a WebRTC connection, the ReadWrite::incoming_buffer and ReadWrite::outgoing_buffer must always be Some.

Panic

Panics if there is no substream with that identifier. Panics if this is a WebRTC connection, and the reading or writing side is closed.

pub fn add_request(
&mut self,
protocol_index: usize,
request: Vec<u8>,
timeout: TNow,
user_data: TRqUd
) -> SubstreamId

Sends a request to the remote.

Must pass the index of the protocol within Config::request_protocols.

This method only inserts the request into the connection object. The request will later be sent out through MultiStream::substream_read_write.

Assuming that the remote is using the same implementation, an Event::RequestIn will be generated on its side.

After the remote has sent back a response, an Event::Response event will be generated locally. The user_data parameter will be passed back.

The timeout is the time between the moment the substream is opened and the moment the response is sent back. If the emitter doesn’t send the request or if the receiver doesn’t answer during this time window, the request is considered failed.

pub fn notifications_substream_user_data_mut(
&mut self,
id: SubstreamId
) -> Option<&mut TNotifUd>

Returns the user data associated to a notifications substream.

Returns None if the substream doesn’t exist or isn’t a notifications substream.

pub fn open_notifications_substream(
&mut self,
now: TNow,
protocol_index: usize,
handshake: Vec<u8>,
user_data: TNotifUd
) -> SubstreamId

Opens a outgoing substream with the given protocol, destined for a stream of notifications.

Must pass the index of the protocol within Config::notifications_protocols.

The remote must first accept (or reject) the substream before notifications can be sent on it.

This method only inserts the opening handshake into the connection object. The handshake will later be sent out through MultiStream::substream_read_write.

Assuming that the remote is using the same implementation, an Event::NotificationsInOpen will be generated on its side.

pub fn accept_in_notifications_substream(
&mut self,
substream_id: SubstreamId,
handshake: Vec<u8>,
user_data: TNotifUd
)

Accepts an inbound notifications protocol. Must be called in response to a Event::NotificationsInOpen.

# Panic

Panics if the substream id is not valid or the substream is of the wrong type.

pub fn reject_in_notifications_substream(&mut self, substream_id: SubstreamId)

Rejects an inbound notifications protocol. Must be called in response to a Event::NotificationsInOpen.

# Panic

Panics if the substream id is not valid or the substream is of the wrong type.

pub fn write_notification_unbounded(
&mut self,
substream_id: SubstreamId,
notification: Vec<u8>
)

Queues a notification to be written out on the given substream.

About back-pressure

This method unconditionally queues up data. You must be aware that the remote, however, can decide to delay indefinitely the sending of that data, which can potentially lead to an unbounded increase in memory.

As such, you are encouraged to call this method only if the amount of queued data (as determined by calling MultiStream::notification_substream_queued_bytes) is below a certain threshold. If above, the notification should be silently discarded.

Panic

Panics if the SubstreamId doesn’t correspond to a notifications substream, or if the notifications substream isn’t in the appropriate state.

pub fn notification_substream_queued_bytes(
&self,
substream_id: SubstreamId
) -> usize

Returns the number of bytes waiting to be sent out on that substream.

See the documentation of MultiStream::write_notification_unbounded for context.

Panic

Panics if the SubstreamId doesn’t correspond to a notifications substream, or if the notifications substream isn’t in the appropriate state.

pub fn close_notifications_substream(&mut self, substream_id: SubstreamId)

Closes a notifications substream opened after a successful Event::NotificationsOutResult or that was accepted using MultiStream::accept_in_notifications_substream.

In the case of an outbound substream, this can be done even when in the negotiation phase, in other words before the remote has accepted/refused the substream.

Panic

Panics if the SubstreamId doesn’t correspond to a notifications substream, or if the notifications substream isn’t in the appropriate state.

pub fn respond_in_request(
&mut self,
substream_id: SubstreamId,
response: Result<Vec<u8>, ()>
) -> Result<(), RespondInRequestError>

Responds to an incoming request. Must be called in response to a Event::RequestIn.

Returns an error if the SubstreamId is invalid.