Struct smoldot::libp2p::connection::established::SingleStream

pub struct SingleStream<TNow, TRqUd, TNotifUd> { /* private fields */ }

State machine of a fully-established connection.

Implementations

impl<TNow, TRqUd, TNotifUd> SingleStream<TNow, TRqUd, TNotifUd>where
TNow: Clone + Add<Duration, Output = TNow> + Sub<TNow, Output = Duration> + Ord,

pub fn read_write(
self,
read_write: &mut ReadWrite<'_, TNow>
) -> Result<(SingleStream<TNow, TRqUd, TNotifUd>, Option<Event<TRqUd, TNotifUd>>), Error>

Reads data coming from the socket, updates the internal state machine, and writes data destined to the socket through the ReadWrite.

In order to avoid unnecessary memory allocations, only one Event is returned at a time. Consequently, this method returns as soon as an event is available, even if the buffers haven’t finished being read. Call this method in a loop until the number of bytes read and written are both 0, and the returned Event is None.

If an error is returned, the socket should be entirely shut down.

pub fn deny_new_incoming_substreams(&mut self)

Close the incoming substreams, automatically denying any new substream request from the remote.

Note that this does not prevent incoming-substreams-related events (such as Event::RequestIn) from being generated, as it is possible that the remote has already opened a substream but has no sent all the necessary handshake messages yet.

Panic

Panic if this function has been called before. It is illegal to call SingleStream::deny_new_incoming_substreams more than one on the same connections.

pub fn add_request(
&mut self,
protocol_index: usize,
request: Vec<u8>,
timeout: TNow,
user_data: TRqUd
) -> SubstreamId

Sends a request to the remote.

Must pass the index of the protocol within Config::request_protocols.

This method only inserts the request into the connection object. Use SingleStream::read_write in order to actually send out the request.

Assuming that the remote is using the same implementation, an Event::RequestIn will be generated on its side.

After the remote has sent back a response, an Event::Response event will be generated locally. The user_data parameter will be passed back.

The timeout is the time between the moment the substream is opened and the moment the response is sent back. If the emitter doesn’t send the request or if the receiver doesn’t answer during this time window, the request is considered failed.

Panic

Panics if a Event::NewOutboundSubstreamsForbidden event has been generated in the past.

pub fn notifications_substream_user_data_mut(
&mut self,
id: SubstreamId
) -> Option<&mut TNotifUd>

Returns the user data associated to a notifications substream.

Returns None if the substream doesn’t exist or isn’t a notifications substream.

pub fn open_notifications_substream(
&mut self,
protocol_index: usize,
handshake: Vec<u8>,
timeout: TNow,
user_data: TNotifUd
) -> SubstreamId

Opens a outgoing substream with the given protocol, destined for a stream of notifications.

Must pass the index of the protocol within Config::notifications_protocols.

The remote must first accept (or reject) the substream before notifications can be sent on it.

This method only inserts the opening handshake into the connection object. Use SingleStream::read_write in order to actually send out the request.

Assuming that the remote is using the same implementation, an Event::NotificationsInOpen will be generated on its side.

Panic

Panics if a Event::NewOutboundSubstreamsForbidden event has been generated in the past.

pub fn accept_in_notifications_substream(
&mut self,
substream_id: SubstreamId,
handshake: Vec<u8>,
user_data: TNotifUd
)

Accepts an inbound notifications protocol. Must be called in response to a Event::NotificationsInOpen.

# Panic

Panics if the substream id is not valid or the substream is of the wrong type.

pub fn reject_in_notifications_substream(&mut self, substream_id: SubstreamId)

Rejects an inbound notifications protocol. Must be called in response to a Event::NotificationsInOpen.

# Panic

Panics if the substream id is not valid or the substream is of the wrong type.

pub fn write_notification_unbounded(
&mut self,
substream_id: SubstreamId,
notification: Vec<u8>
)

Queues a notification to be written out on the given substream.

About back-pressure

This method unconditionally queues up data. You must be aware that the remote, however, can decide to delay indefinitely the sending of that data, which can potentially lead to an unbounded increase in memory.

As such, you are encouraged to call this method only if the amount of queued data (as determined by calling SingleStream::notification_substream_queued_bytes) is below a certain threshold. If above, the notification should be silently discarded.

Panic

Panics if the SubstreamId doesn’t correspond to a notifications substream, or if the notifications substream isn’t in the appropriate state.

pub fn notification_substream_queued_bytes(
&self,
substream_id: SubstreamId
) -> usize

Returns the number of bytes waiting to be sent out on that substream.

See the documentation of SingleStream::write_notification_unbounded for context.

Panic

Panics if the SubstreamId doesn’t correspond to a notifications substream, or if the notifications substream isn’t in the appropriate state.

pub fn close_notifications_substream(&mut self, substream_id: SubstreamId)

Closes a notifications substream opened after a successful Event::NotificationsOutResult or that was accepted using SingleStream::accept_in_notifications_substream.

In the case of an outbound substream, this can be done even when in the negotiation phase, in other words before the remote has accepted/refused the substream.

Panic

Panics if the SubstreamId doesn’t correspond to a notifications substream, or if the notifications substream isn’t in the appropriate state.

pub fn respond_in_request(
&mut self,
substream_id: SubstreamId,
response: Result<Vec<u8>, ()>
) -> Result<(), RespondInRequestError>

Responds to an incoming request. Must be called in response to a Event::RequestIn.

Passing an Err corresponds, on the other side, to a substream::RequestError::SubstreamClosed.

Returns an error if the SubstreamId is invalid.