smoldot::libp2p::connection::established
Module substream

Individual substream within an established connection.

This module contains the Substream struct, a state machine containing the state of a single substream. When the remote sends data on that substream, or when the remote is ready to accept more data on that substream, the state machine can be updated by calling Substream::read_write. This optionally produces an event that indicates what happened on the substream as a result of the call.

Structs

Substream

State machine containing the state of a single substream of an established connection.

Enums

Event

Event that happened on the connection. See Substream::read_write.

InboundError

Error that can happen while processing an inbound substream.

InboundTy

Type of inbound protocol.

NotificationsInClosedErr

Reason why an inbound notifications substream has been closed.

NotificationsOutErr

Error that can happen when trying to open an outbound notifications substream.

RequestError

Error that can happen during a request in a request-response scheme.

RespondInRequestError

Error potentially returned by Substream::respond_in_request.