Enum smoldot::libp2p::connection::established::substream::Event

pub enum Event<TRqUd, TNotifUd> {
    InboundError(InboundError),
    InboundNegotiated(String),
    RequestIn {
        protocol_index: usize,
        request: Vec<u8>,
    },
    Response {
        response: Result<Vec<u8>, RequestError>,
        user_data: TRqUd,
    },
    NotificationsInOpen {
        protocol_index: usize,
        handshake: Vec<u8>,
    },
    NotificationsInOpenCancel {
        protocol_index: usize,
    },
    NotificationIn {
        notification: Vec<u8>,
    },
    NotificationsInClose {
        outcome: Result<(), NotificationsInClosedErr>,
        protocol_index: usize,
    },
    NotificationsOutResult {
        result: Result<Vec<u8>, (NotificationsOutErr, TNotifUd)>,
    },
    NotificationsOutCloseDemanded,
    NotificationsOutReset {
        user_data: TNotifUd,
    },
}

Event that happened on the connection. See Substream::read_write.

Variants

InboundError(InboundError)

Error while receiving an inbound substream.

InboundNegotiated(String)

An inbound substream has successfully negotiated a protocol. Call Substream::set_inbound_ty in order to resume.

RequestIn

Received a request in the context of a request-response protocol.

Fields of RequestIn

protocol_index: usize

Index of the request-response protocol the request was sent on.

request: Vec<u8>

Bytes of the request. Its interpretation is out of scope of this module.

Response

Received a response to a previously emitted request on a request-response protocol.

Fields of Response

response: Result<Vec<u8>, RequestError>

Bytes of the response. Its interpretation is out of scope of this module.

user_data: TRqUd

Value that was passed to Substream::request_out.

NotificationsInOpen

Remote has opened an inbound notifications substream.

Either Substream::accept_in_notifications_substream or Substream::reject_in_notifications_substream must be called in the near future in order to accept or reject this substream.

Fields of NotificationsInOpen

protocol_index: usize

Index of the notifications protocol concerned by the substream.

handshake: Vec<u8>

Handshake sent by the remote. Its interpretation is out of scope of this module.

NotificationsInOpenCancel

Remote has canceled an inbound notifications substream opening.

This can only happen after Event::NotificationsInOpen. Substream::accept_in_notifications_substream or Substream::reject_in_notifications_substream should not be called on this substream.

Fields of NotificationsInOpenCancel

protocol_index: usize

Index of the notifications protocol concerned by the substream.

NotificationIn

Remote has sent a notification on an inbound notifications substream. Can only happen after the substream has been accepted.

Fields of NotificationIn

notification: Vec<u8>

Notification sent by the remote.

NotificationsInClose

Remote has closed an inbound notifications substream opening. No more notifications will be received.

This can only happen after the substream has been accepted.

Fields of NotificationsInClose

outcome: Result<(), NotificationsInClosedErr>

If Ok, the substream has been closed gracefully. If Err, a problem happened.

protocol_index: usize

Index of the notifications protocol concerned by the substream.

NotificationsOutResult

Remote has accepted or refused a substream opened with Substream::notifications_out.

If Ok, it is now possible to send notifications on this substream.

Fields of NotificationsOutResult

result: Result<Vec<u8>, (NotificationsOutErr, TNotifUd)>

If Ok, contains the handshake sent back by the remote. Its interpretation is out of scope of this module.

NotificationsOutCloseDemanded

Remote has closed an outgoing notifications substream, meaning that it demands the closing of the substream.

NotificationsOutReset

Remote has reset an outgoing notifications substream. The substream is instantly closed.

Fields of NotificationsOutReset

user_data: TNotifUd

Value that was passed to Substream::notifications_out.