Module smoldot::libp2p::connection::single_stream_handshake

State machine handling the handshake with a TCP or WebSocket libp2p connection.

A connection handshake consists of three steps:

A multistream-select negotiation to negotiate the encryption protocol. Only the noise protocol is supported at the moment.

A noise protocol handshake, where public keys are exchanged and symmetric encryption is initialized.

A multistream-select negotiation to negotiate the Yamux protocol. Only the Yamux protocol is supported at the moment. This negotiation is performed on top of the noise cipher.

This entire handshake requires in total either three or five TCP packets (not including the TCP handshake), depending on the strategy used for the multistream-select protocol.

Structs

HealthyHandshake

Connection handshake in progress.

Enums

Handshake

Current state of a connection handshake.

HandshakeError

Error during a connection handshake. The connection should be shut down.