Struct smoldot::libp2p::connection::yamux::SubstreamMut

pub struct SubstreamMut<'a, T> { /* fields omitted */ }

Reference to a substream within the Yamux.

Implementations

impl<'a, T> SubstreamMut<'a, T>

pub fn id(&self) -> SubstreamId

Identifier of the substream.

pub fn user_data(&mut self) -> &mut T

Returns the user data associated to this substream.

pub fn into_user_data(self) -> &'a mut T

Returns the user data associated to this substream.

pub fn write(&mut self, data: Vec<u8>)

Appends data to the buffer of data to send out on this substream.

Panic

Panics if SubstreamMut::close has already been called on this substream.

pub fn reserve_window(&mut self, bytes: u64)

Allow the remote to send up to bytes bytes at once in the next packet.

This method sets the number of allowed bytes to at least this value. In other words, if this method was to be twice with the same parameter, the second call would have no effect.

Context

In order to properly handle back-pressure, the yamux protocol only allows the remote to send a certain number of bytes before the local node grants the authorization to send more data. This method grants the authorization to the remote to send up to bytes bytes.

Call this when you expect a large payload with the maximum size this payload is allowed to be.

pub fn queued_bytes(&self) -> usize

Returns the number of bytes queued for writing on this substream.

pub fn is_remote_closed(&self) -> bool

Returns true if the remote has closed their writing side of this substream.

pub fn is_closed(&self) -> bool

Returns true if SubstreamMut::close has been called on this substream.

pub fn close(self) -> Option<T>

Marks the substream as closed. It is no longer possible to write data on it.

If the remote writing side is still open, this method returns None and the remote can continue to send data.

If the remote writing side is already closed, this method returns Some with the user data, and the substream is now destroyed.

Panic

Panics if SubstreamMut::close has already been called on this substream.

pub fn reset(self) -> T

Abruptly shuts down the substream. Its identifier is now invalid. Sends a frame with the RST flag to the remote.

Use this method when a protocol error happens on a substream.