Struct smoldot::libp2p::connection::yamux::Yamux

pub struct Yamux<T> { /* private fields */ }

Implementations

impl<T> Yamux<T>

pub fn new(config: Config) -> Yamux<T>

Initializes a new Yamux state machine.

pub fn is_empty(&self) -> bool

Returns true if there is no substream in the state machine.

Note: After a substream has been closed or reset, it must be removed using Yamux::remove_dead_substream before this function can return true.

pub fn len(&self) -> usize

Returns the number of substreams in the Yamux state machine. Includes substreams that are dead but haven’t been removed yet.

pub fn num_inbound(&self) -> usize

Returns the number of inbound substreams in the Yamux state machine. Includes substreams that are dead but haven’t been removed yet.

pub fn open_substream(&mut self, user_data: T) -> SubstreamMut<'_, T>

Opens a new substream.

This method only modifies the state of self and reserves an identifier. No message needs to be sent to the remote before data is actually being sent on the substream.

Note: Importantly, the remote will not be notified of the substream being open before the local side sends data on this substream. As such, protocols where the remote is expected to send data in response to a substream being open, without the local side first sending some data on that substream, will not work. In practice, while this is technically out of concern of the Yamux protocol, all substreams in the context of libp2p start with a multistream-select negotiation, and this scenario can therefore never happen.

Panic

Panics if all possible substream IDs are already taken. This happen if there exists more than approximately 2^31 substreams, which is very unlikely to happen unless there exists a bug in the code.

Panics if a IncomingDataDetail::GoAway event has been generated. This can also be checked by calling Yamux::received_goaway.

pub fn received_goaway(&self) -> Option<GoAwayErrorCode>

Returns Some if a IncomingDataDetail::GoAway event has been generated in the past, in which case the code is returned.

If Some is returned, it is forbidden to open new outbound substreams.

pub fn user_datas(&self) -> impl ExactSizeIterator<Item = (SubstreamId, &T)>

Returns an iterator to the list of all substream user datas.

pub fn user_datas_mut(
&mut self
) -> impl ExactSizeIterator<Item = (SubstreamId, &mut T)>

Returns an iterator to the list of all substream user datas.

pub fn substream_by_id(&self, id: SubstreamId) -> Option<SubstreamRef<'_, T>>

Returns a reference to a substream by its ID. Returns None if no substream with this ID is open.

pub fn substream_by_id_mut(
&mut self,
id: SubstreamId
) -> Option<SubstreamMut<'_, T>>

Returns a reference to a substream by its ID. Returns None if no substream with this ID is open.

pub fn queue_ping(&mut self)

Queues sending out a ping to the remote.

pub fn goaway_queued_or_sent(&self) -> bool

Returns true if Yamux::send_goaway has been called in the past.

In other words, returns true if a GoAway frame has been either queued for sending (and is available through Yamux::extract_out) or has already been sent out.

pub fn goaway_sent(&self) -> bool

Returns true if Yamux::send_goaway has been called in the past and that this GoAway frame has been extracted through Yamux::extract_out.

pub fn send_goaway(&mut self, code: GoAwayErrorCode)

Queues a GoAway frame, requesting the remote to no longer open any substream.

If the state of Yamux is currently waiting for a confirmation to accept/reject a substream, then this function automatically implies calling Yamux::reject_pending_substream.

All follow-up requests for new substreams from the remote are automatically rejected. IncomingDataDetail::IncomingSubstream events can no longer happen.

Panic

Panics if this function has already been called in the past. This can be verified using Yamux::goaway_queued_or_sent. It is illegal to call this function more than once on the same instance of Yamux.

pub fn dead_substreams(
&self
) -> impl Iterator<Item = (SubstreamId, DeadSubstreamTy, &T)> + '_

Returns the list of all substreams that have been closed or reset.

This function does not remove dead substreams from the state machine. In other words, if this function is called multiple times in a row, it will always return the same substreams. Use Yamux::remove_dead_substream to remove substreams.

pub fn remove_dead_substream(&mut self, id: SubstreamId) -> T

Removes a dead substream from the state machine.

Panic

Panics if the substream with that id doesn’t exist or isn’t dead.

pub fn incoming_data(self, data: &[u8]) -> Result<IncomingDataOutcome<T>, Error>

Process some incoming data.

This function takes ownership of self and yields it back if everything goes well. If, on the other hand, a malformed packet is received, an error is yielded and self is destroyed.

This function might not process all the data available for one of the following reasons:

Not all outgoing data has been extracted. In order to process incoming messages, the Yamux might have to queue data to be written out. For example, incoming pings must be replied to. In order to avoid queue an infinite amount of data, processing incoming messages might be blocked if there is data to be sent out.

It is currently waiting for either Yamux::accept_pending_substream or Yamux::reject_pending_substream to be called.

If the remote opens a substream whose ID is equal to a previous substream that is now dead. Use Yamux::dead_substreams and Yamux::remove_dead_substream to remove dead substreams before continuing.

If the return value contains IncomingDataDetail::IncomingSubstream, then either Yamux::accept_pending_substream or Yamux::reject_pending_substream must be called in order to accept or reject the pending substream. API users are encouraged to enforce a limit to the total number of substreams in order to clamp the memory usage of this state machine.

pub fn extract_out(&mut self, size_bytes: usize) -> ExtractOut<'_, T>

Returns an object that provides an iterator to a list of buffers whose content must be sent out on the socket.

The buffers produced by the iterator will never yield more than size_bytes bytes of data. The user is expected to pass an exact amount of bytes that the next layer is ready to accept.

After the ExtractOut has been destroyed, the Yamux state machine will automatically consider that these size_bytes have been sent out, even if the iterator has been destroyed before finishing. It is a logic error to mem::forget the ExtractOut.

Note: Most other objects in the networking code have a “read_write” method that writes the outgoing data to a buffer. This is an idiomatic way to do things in situations where the data is generated on the fly. In the context of Yamux, however, this would be rather sub-optimal considering that buffers to send out are already stored in their final form in the state machine.

pub fn accept_pending_substream(&mut self, user_data: T) -> SubstreamMut<'_, T>

Accepts an incoming substream.

Either Yamux::accept_pending_substream or Yamux::reject_pending_substream must be called after IncomingDataDetail::IncomingSubstream is returned.

Note that there is no expiration window after IncomingDataDetail::IncomingSubstream is returned until the substream is no longer valid. However, reading will be blocked until the substream is either accepted or rejected. This function should thus be called as soon as possible.

Panic

Panics if no incoming substream is currently pending.

pub fn reject_pending_substream(&mut self)

Rejects an incoming substream.

Either Yamux::accept_pending_substream or Yamux::reject_pending_substream must be called after IncomingDataDetail::IncomingSubstream is returned.

Note that there is no expiration window after IncomingDataDetail::IncomingSubstream is returned until the substream is no longer valid. However, reading will be blocked until the substream is either accepted or rejected. This function should thus be called as soon as possible.

Panic

Panics if no incoming substream is currently pending.