Module smoldot::libp2p::discovery

Mechanisms related to discovering nodes that are part of a certain overlay network.

Internet is a network of computers. The list of machines that are running a Substrate/Polkadot client is a subset of the Internet. This module provides the mechanisms that help finding out which machines (more precisely, which IP address and port) run a Substrate/Polkadot client.

Details

Substrate-compatible chains use two discovery mechanisms:

Kademlia. All nodes that belong to a certain chain are encouraged to participate in the Kademlia DHT of that chain, making it possible to ask a node for the nodes it has learned about in the past.

mDNS. By broadcasting UDP packets over the local network, one can find other nodes running the same chain. // TODO: not implemented yet

The main discovery mechanism is the DHT. In order to bootstrap this mechanism, a list of nodes known to always be part of the chain is hardcoded in the chain specification. These nodes are called boostrap nodes.

Modules

kademlia