smoldot::libp2p::multiaddr

Enum Protocol

pub enum Protocol<T = Vec<u8>> {

    Dns(DomainName<T>),
    Dns4(DomainName<T>),
    Dns6(DomainName<T>),
    DnsAddr(DomainName<T>),
    Ip4([u8; 4]),
    Ip6([u8; 16]),
    P2p(Multihash<T>),
    Quic,
    Tcp(u16),
    Tls,
    Udp(u16),
    Ws,
    Wss,
    Memory(u64),
    WebRtcDirect,
    Certhash(Multihash<T>),

}

Variants

Dns(DomainName<T>)

Dns4(DomainName<T>)

Dns6(DomainName<T>)

DnsAddr(DomainName<T>)

Ip4([u8; 4])

Ip6([u8; 16])

P2p(Multihash<T>)

Quic

Tcp(u16)

Tls

Udp(u16)

Ws

Wss

Memory(u64)

WebRtcDirect

Certhash(Multihash<T>)

Contains the multihash of the TLS certificate.

Implementations

impl<'a> Protocol<Cow<'a, [u8]>>

pub fn from_str_parts( iter: impl Iterator<Item = &'a str>, ) -> Result<Self, ParseError>

Attempts to extract a protocol from an iterator of /-separated components.

impl<T: AsRef<[u8]>> Protocol<T>

pub fn as_bytes(&self) -> impl Iterator<Item = impl AsRef<[u8]>>

Returns an iterator to a list of buffers that, when concatenated together, form the binary representation of this protocol.