Enum smoldot::libp2p::multiaddr::ProtocolRef

pub enum ProtocolRef<'a> {

    Dns(DomainNameRef<'a>),
    Dns4(DomainNameRef<'a>),
    Dns6(DomainNameRef<'a>),
    DnsAddr(DomainNameRef<'a>),
    Ip4([u8; 4]),
    Ip6([u8; 16]),
    P2p(Cow<'a, [u8]>),
    Quic,
    Tcp(u16),
    Tls,
    Udp(u16),
    Ws,
    Wss,
    Memory(u64),
    WebRTC,
    Certhash(Cow<'a, [u8]>),

}

Variants

Dns(DomainNameRef<'a>)

Dns4(DomainNameRef<'a>)

Dns6(DomainNameRef<'a>)

DnsAddr(DomainNameRef<'a>)

Ip4([u8; 4])

Ip6([u8; 16])

P2p(Cow<'a, [u8]>)

Quic

Tcp(u16)

Tls

Udp(u16)

Ws

Wss

Memory(u64)

WebRTC

Certhash(Cow<'a, [u8]>)

Contains the multihash of the TLS certificate.

Implementations

impl<'a> ProtocolRef<'a>

pub fn from_str_parts(
iter: impl Iterator<Item = &'a str>
) -> Result<Self, ParseError>

Attempts to extract a protocol from an iterator of /-separated components.

pub fn as_bytes(&self) -> impl Iterator<Item = impl AsRef<[u8]>>

Returns an iterator to a list of buffers that, when concatenated together, form the binary representation of this protocol.