Trait smoldot::libp2p::multiaddr::multihash::MultihashDigest

pub trait MultihashDigest<T> where
    T: TryFrom<u64>, {
    pub fn code(&self) -> T;

    pub fn digest(&self, data: &[u8]) -> MultihashGeneric<T>;

    pub fn input(&mut self, data: &[u8]);

    pub fn result(self) -> MultihashGeneric<T>;

    pub fn result_reset(&mut self) -> MultihashGeneric<T>;

    pub fn reset(&mut self);
}

The MultihashDigest trait specifies an interface common for all multihash functions.

Required methods

pub fn code(&self) -> T

The Mutlihash byte value.

pub fn digest(&self, data: &[u8]) -> MultihashGeneric<T>

Hash some input and return the digest.

Panics

Panics if the digest length is bigger than 2^32. This only happens for identity hasing.

pub fn input(&mut self, data: &[u8])

Digest data, updating the internal state.

This method can be called repeatedly for use with streaming messages.

Panics

Panics if the digest length is bigger than 2^32. This only happens for identity hashing.

pub fn result(self) -> MultihashGeneric<T>

Retrieve the computed MultihashGeneric, consuming the hasher.

pub fn result_reset(&mut self) -> MultihashGeneric<T>

Retrieve result and reset hasher instance.

This method sometimes can be more efficient compared to hasher re-creation.

pub fn reset(&mut self)

Reset hasher instance to its initial state.

Loading content...