Trait smoldot::libp2p::multiaddr::multihash::Multihasher

pub trait Multihasher<T> where
    T: TryFrom<u64> + Copy, {
    pub const CODE: T;

    pub fn digest(data: &[u8]) -> MultihashGeneric<T>;
}

The Multihasher trait specifies an interface common for all multihash functions that does not require allocating a Box<dyn MultihashDigest<T>>.

Associated Constants

pub const CODE: T

The multihash code.

Loading content...
Required methods

pub fn digest(data: &[u8]) -> MultihashGeneric<T>

Hash some input and return the digest.

Loading content...