Module smoldot::libp2p::peers

Network of peers.

The Peers state machine builds on top of the libp2p module and provides an abstraction over the network based on network identities (i.e. PeerIds). One can set the list of peers to be connected to and through which notification protocols, and the Peers struct will try to open or re-open connections with these peers. Once connected, one can use the Peers to send request or notifications with these peers.

Detailed usage

The Peers struct contains six different collections:

A list, decided by the API user, of peers that are marked as “desired”.

A list, decided by the API user, of (peer_id, notification_protocol) tuples that are marked as “desired”.

A list of connections identified by ConnectionIds.

A list of requests for inbound substreams, identified by a DesiredInNotificationId. When a peer desired to open a notification substream with the local node, a DesiredInNotificationId is generated by the Peers. The API user must answer by either accepting or refusing the request. The requests can automatically become obsolete if the remote decides to withdraw their request or the connection closes. A request becoming obsolete does not invalidate its DesiredInNotificationId.

A list of requests for outbound substreams emitted by the local node, identified by a DesiredOutNotificationId. Must be responded using Peers::open_out_notification.

A list of requests that have been received, identified by a RequestId. The API user must answer by calling Peers::respond. Requests can automatically become obsolete if the remote decides to withdraw their request or the connection closes. A request becoming obsolete does not invalidate its RequestId.

Re-exports

pub use collection::ConfigRequestResponse;

pub use collection::ConfigRequestResponseIn;

pub use collection::ConnectionError;

pub use collection::ConnectionReadyFuture;

pub use collection::InboundError;

pub use collection::NotificationProtocolConfig;

pub use collection::NotificationsInClosedErr;

pub use collection::NotificationsOutErr;

pub use collection::ReadWrite;

pub use collection::ConnectionId;

Structs

Config

Configuration for a Peers.

DesiredInNotificationId

See Event::DesiredInNotification.

DesiredOutNotificationId

See Event::DesiredOutNotification.

Peers

RequestId

See Event::RequestIn.

Enums

DesiredState

See Peers::set_peer_notifications_out_desired.

Event

Event happening over the network. See Peers::next_event.

InNotificationAcceptError

Error potentially returned by Peers::in_notification_accept.

QueueNotificationError

Error potentially returned by Peers::queue_notification.

RequestError

Error potentially returned by Peers::request.