Enum smoldot::libp2p::peers::Event

pub enum Event<TConn> {
    Connected {
        peer_id: PeerId,
        num_peer_connections: NonZeroU32,
    },
    Disconnected {
        peer_id: PeerId,
        peer_is_desired: bool,
        num_peer_connections: u32,
        user_data: TConn,
    },
    InboundError {
        peer_id: PeerId,
        connection_id: ConnectionId,
        error: InboundError,
    },
    RequestIn {
        request_id: RequestId,
        peer_id: PeerId,
        connection_id: ConnectionId,
        connection_user_data: TConn,
        protocol_index: usize,
        request_payload: Vec<u8>,
    },
    RequestInCancel {
        id: RequestId,
    },
    DesiredInNotification {
        id: DesiredInNotificationId,
        peer_id: PeerId,
        notifications_protocol_index: usize,
        handshake: Vec<u8>,
    },
    DesiredInNotificationCancel {
        id: DesiredInNotificationId,
    },
    DesiredOutNotification {
        id: DesiredOutNotificationId,
        peer_id: PeerId,
        notifications_protocol_index: usize,
    },
    NotificationsOutResult {
        peer_id: PeerId,
        notifications_protocol_index: usize,
        result: Result<Vec<u8>, NotificationsOutErr>,
    },
    NotificationsOutClose {
        peer_id: PeerId,
        notifications_protocol_index: usize,
    },
    NotificationsIn {
        peer_id: PeerId,
        notifications_protocol_index: usize,
        notification: Vec<u8>,
    },
    NotificationsInClose {
        peer_id: PeerId,
        notifications_protocol_index: usize,
        outcome: Result<(), NotificationsInClosedErr>,
    },
}

Event happening over the network. See Peers::next_event.

Variants

Connected

Established a new connection to the given peer.

Fields of Connected

peer_id: PeerId

Identity of the peer on the other side of the connection.

num_peer_connections: NonZeroU32

Number of other established connections with the same peer, including the one that has just been established.

Disconnected

A connection has stopped.

Fields of Disconnected

peer_id: PeerId

Identity of the peer on the other side of the connection.

peer_is_desired: bool

true if the peer is marked as desired. If this is true and num_peer_connections is 0, then calling Peers::unfulfilled_desired_peers will now return this peer.

Note: Keep in mind that everything is subject to race conditions. For example, a parallel thread might at the same time remove the “desired” marker of this peer, or an incoming connection might at the same time finish negotiating a new connection with this peer. When sharing the Peers between multiple threads, there is no guarantee that Peers::unfulfilled_desired_peers will in fact return this peer.

num_peer_connections: u32

Number of other established connections with the same peer remaining after the disconnection.

user_data: TConn

User data that was associated to this connection.

InboundError

Received an incoming substream, but this substream has produced an error.

Note: This event exists only for diagnostic purposes. No action is expected in return.

Fields of InboundError

peer_id: PeerId

Peer which opened the substream.

connection_id: ConnectionId

Identifier of the connection on which the problem happened.

error: InboundError

Error that happened.

RequestIn

Received a request from a request-response protocol.

Fields of RequestIn

request_id: RequestId

Identifier for this request. Must be passed back when calling Peers::respond.

peer_id: PeerId

Peer which sent the request.

connection_id: ConnectionId

Identifier of the connection that has sent the request.

connection_user_data: TConn

Copy of the user data associated to this connection.

protocol_index: usize

Request-response protocol the request is about.

request_payload: Vec<u8>

Payload of the request, opaque to this state machine.

Note: Keep in mind that this data is untrusted.

RequestInCancel

A previously-emitted Event::RequestIn is now obsolete. This event is for informative purpose and does not invalidate the RequestId.

Fields of RequestInCancel

id: RequestId

Identifier for this request.

DesiredInNotification

A peer would like to open a notifications substream with the local node, in order to send notifications.

Only one inbound notifications substream can exist per peer and per protocol. Any additional one will be automatically refused.

Fields of DesiredInNotification

id: DesiredInNotificationId

Identifier for this request. Must be passed back when calling Peers::in_notification_accept or Peers::in_notification_refuse.

peer_id: PeerId

Peer which tries to open an inbound substream.

notifications_protocol_index: usize

Notifications protocol the substream is about.

handshake: Vec<u8>

Handshake of the request sent by the peer. Opaque to this state machine.

Note: Keep in mind that this data is untrusted.

DesiredInNotificationCancel

A previously-emitted DesiredInNotificationId is now obsolete. This event is for informative purpose and does not invalidate the DesiredInNotificationId. Use Peers::in_notification_refuse if you no longer care about this request.

Fields of DesiredInNotificationCancel

id: DesiredInNotificationId

Identifier for this request.

DesiredOutNotification

Local node would like to open a notifications substream with the given peer. This can only happen if the combination of peer and notification protocol was marked as desired.

Fields of DesiredOutNotification

id: DesiredOutNotificationId

Identifier for this request. Must be passed back when calling Peers::open_out_notification.

peer_id: PeerId

Peer which tries to open an outbound substream.

notifications_protocol_index: usize

Notifications protocol the substream is about.

NotificationsOutResult

A handshaking outbound substream has been accepted by the remote.

Will happen for combinations of PeerId and notification protocols that have been marked as desired. Can also happen for other combinations, if there were marked as desired in the past but no longer are.

If Ok, it is now possible to send notifications on this substream.

Fields of NotificationsOutResult

peer_id: PeerId

Peer the substream is open with.

notifications_protocol_index: usize

Notifications protocol the substream is about.

result: Result<Vec<u8>, NotificationsOutErr>

If Ok, contains the handshake sent back by the remote. Its interpretation is out of scope of this module. If Err, the state machine will not automatically try to re-open a substream again. Use Peers::set_peer_notifications_out_desired with DesiredState::DesiredReset in order to try again.

NotificationsOutClose

A previously open outbound substream has been closed by the remote. Can only happen after a corresponding successful Event::NotificationsOutResult event has been emitted in the past.

This combination of PeerId and notification protocol will now be returned when calling Peers::refused_notifications_out.

Fields of NotificationsOutClose

peer_id: PeerId

Peer the subtream is no longer open with.

notifications_protocol_index: usize

Notifications protocol the substream is about.

NotificationsIn

Received a notification on a notifications substream of a connection.

Fields of NotificationsIn

peer_id: PeerId

Peer that sent the notification.

notifications_protocol_index: usize

Notifications protocol the substream is about.

notification: Vec<u8>

Payload of the notification. Opaque to this state machine.

Note: Keep in mind that this data is untrusted.

NotificationsInClose

Remote has closed a previously-open inbound notifications substream.

Fields of NotificationsInClose

peer_id: PeerId

Peer the substream is no longer with.

notifications_protocol_index: usize

Notifications protocol the substream is about.

outcome: Result<(), NotificationsInClosedErr>

If Ok, the substream has been closed gracefully. If Err, a problem happened.