Struct smoldot::libp2p::peers::Config

pub struct Config {
    pub randomness_seed: [u8; 32],
    pub connections_capacity: usize,
    pub peers_capacity: usize,
    pub max_inbound_substreams: usize,
    pub notification_protocols: Vec<NotificationProtocolConfig>,
    pub request_response_protocols: Vec<ConfigRequestResponse>,
    pub ping_protocol: String,
    pub handshake_timeout: Duration,
    pub noise_key: NoiseKey,
}

Configuration for a Peers.

Fields
randomness_seed: [u8; 32]

Seed for the randomness within the networking state machine.

connections_capacity: usize

Capacity to initially reserve to the list of connections.

peers_capacity: usize

Capacity to initially reserve to the list of peers.

max_inbound_substreams: usize

Maximum number of substreams that each remote can have simultaneously opened on each connection.

If there exists multiple connections with the same remote, the limit is enforced for each connection separately.

Note: This limit is necessary in order to avoid DoS attacks where a remote opens too many substreams.

notification_protocols: Vec<NotificationProtocolConfig>request_response_protocols: Vec<ConfigRequestResponse>ping_protocol: String

Name of the ping protocol on the network.

handshake_timeout: Duration

Amount of time after which a connection handshake is considered to have taken too long and must be aborted.

noise_key: NoiseKey

Key used for the encryption layer. This is a Noise static key, according to the Noise specification. Signed using the actual libp2p key.