Struct smoldot::libp2p::peers::Peers

pub struct Peers<TConn, TNow> { /* fields omitted */ }

Implementations

impl<TConn, TNow> Peers<TConn, TNow> where
TConn: Clone,
TNow: Clone + Add<Duration, Output = TNow> + Sub<TNow, Output = Duration> + Ord,

pub fn new(config: Config) -> Self

Creates a new Peers.

pub fn notification_protocols(
&self
) -> impl ExactSizeIterator<Item = &NotificationProtocolConfig>

Returns the list the overlay networks originally passed as Config::notification_protocols.

pub fn request_response_protocols(
&self
) -> impl ExactSizeIterator<Item = &ConfigRequestResponse>

Returns the list the request-response protocols originally passed as Config::request_response_protocols.

pub fn noise_key(&self) -> &NoiseKey

Returns the Noise key originalled passed as Config::noise_key.

pub async fn next_event(&self) -> Event<TConn>

Returns the next event produced by the service.

This function should be called at a high enough rate that Peers::read_write can continue pushing events to the internal buffer of events. Failure to call this function often enough will lead to connections being back-pressured. See also Config::pending_api_events_buffer_size.

It is technically possible to call this function multiple times simultaneously, in which case the events will be distributed amongst the multiple calls in an unspecified way. Keep in mind that some Events have logic attached to the order in which they are produced, and calling this function multiple times is therefore discouraged.

pub async fn add_incoming_connection(
&self,
when_connected: TNow,
user_data: TConn
) -> ConnectionId

Inserts an incoming connection in the state machine.

This connection hasn’t finished handshaking and the PeerId of the remote isn’t known yet.

Must be passed the moment (as a TNow) when the connection as been established, in order to determine when the handshake timeout expires.

After this function has returned, you must process the connection with Peers::read_write.

pub async fn add_outgoing_connection(
&self,
when_connected: TNow,
expected_peer_id: &PeerId,
user_data: TConn
) -> ConnectionId

Inserts an outgoing connection in the state machine.

This connection hasn’t finished handshaking, and the PeerId of the remote isn’t known yet, but it is expected to be unfulfilled_desired_peers. After this function has been called, the provided expected_peer_id will no longer be part of the return value of Peers::unfulfilled_desired_peers.

Must be passed the moment (as a TNow) when the connection as been established, in order to determine when the handshake timeout expires.

After this function has returned, you must process the connection with Peers::read_write.

pub async fn unfulfilled_desired_peers(&self) -> impl Iterator<Item = PeerId>

Returns the list of PeerIds that have been marked as desired, but that don’t have any associated connection. An associated connection is either a fully established connection with that peer, or an outgoing connection that is still handshaking but expects to reach that peer.

pub async fn set_peer_desired(&self, peer_id: &PeerId, desired: bool)

Sets the “desired” flag of the given PeerId.

When a peer is marked as “desired” and there isn’t any pending or established connection towards it, it is returned when calling Peers::unfulfilled_desired_peers.

pub async fn set_peer_notifications_out_desired(
&self,
peer_id: &PeerId,
notification_protocol: usize,
new_desired_state: DesiredState
)

Sets the given combinations of notification protocol and PeerId as “desired”.

When a peer is marked as “desired” and there isn’t any pending or established connection towards it, it is returned when calling Peers::unfulfilled_desired_peers.

When a combination of network protocol and PeerId is marked as “desired”, the state machine will try to maintain open an outbound substream. If the remote refuses the substream, it will be returned when calling Peers::refused_notifications_out.

pub async fn refused_notifications_out(
&self
) -> impl Iterator<Item = (PeerId, usize)>

Returns the combinations of notification and PeerId that are marked as “desired”, but where the remote has refused the request for a notifications substream.

Use Peers::set_peer_notifications_out_desired with DesiredState::DesiredReset in order to try again.

pub async fn in_notification_accept(
&self,
id: DesiredInNotificationId,
handshake_back: Vec<u8>
) -> Result<(), InNotificationAcceptError>

Responds to an Event::DesiredInNotification by accepting the request for an inbound substream.

If Ok is returned, the substream is now considered open. If Err is returned, then no substream has been opened.

Panic

Panics if the DesiredInNotificationId is invalid. Note that these ids remain valid forever until Peers::in_notification_accept or Peers::in_notification_refuse is called.

pub async fn in_notification_refuse(&self, id: DesiredInNotificationId)

Responds to an Event::DesiredInNotification by refusing the request for an inbound substream.

Panic

Panics if the DesiredInNotificationId is invalid. Note that these ids remain valid forever until Peers::in_notification_accept or Peers::in_notification_refuse is called.

pub async fn open_out_notification(
&self,
id: DesiredOutNotificationId,
now: TNow,
handshake: Vec<u8>
)

Responds to an Event::DesiredOutNotification by indicating the handshake to send to the remote.

Panic

Panics if the DesiredOutNotificationId is invalid. Note that these ids remain valid forever until Peers::open_out_notification is called.

pub async fn queue_notification(
&self,
target: &PeerId,
notifications_protocol_index: usize,
notification: impl Into<Vec<u8>>
) -> Result<(), QueueNotificationError>

pub async fn broadcast_notification(
&self,
notifications_protocol_index: usize,
notification: impl Into<Vec<u8>>
)

Equivalent to calling Peers::queue_notification for all peers an outbound notifications substream is open with.

Individual errors that would have occured when calling Peers::queue_notification are silently discarded.

pub async fn request(
&self,
now: TNow,
target: &PeerId,
protocol_index: usize,
request_data: Vec<u8>
) -> Result<Vec<u8>, RequestError>

Sends a request to the given peer, and waits for a response.

This consists in:

Opening a substream on an established connection with the target.

Negotiating the requested protocol (protocol_index) on this substream using the multistream-select protocol.

Sending the request (request_data parameter), prefixed with its length.

Waiting for the response (prefixed with its length), which is then returned.

An error happens if there is no suitable connection for that request, if the connection closes while the request is in progress, if the request or response doesn’t respect the protocol limits (see ConfigRequestResponse), or if the remote takes too much time to answer.

As the API of this module is inherently subject to race conditions, it is never possible to guarantee that this function will succeed. RequestError::ConnectionClosed should be handled by retrying the same request again.

Note: This function doesn’t return before the remote has answered. It is strongly recommended to await the returned Future in the background, and not block any important task on this.

Panic

Panics if protocol_index isn’t a valid index in Config::request_response_protocols.

pub async fn respond(&self, id: RequestId, response: Result<Vec<u8>, ()>)

Responds to a previously-emitted Event::RequestIn.

Panic

Panics if the RequestId is invalid. Note that these ids remain valid forever until Peers::respond is called.

pub async fn read_write(
&self,
connection_id: ConnectionId,
read_write: &mut ReadWrite<'_, TNow>
) -> Result<(), ConnectionError>

Reads data coming from the connection, updates the internal state machine, and writes data destined to the connection through the ReadWrite.

If an error is returned, the connection should be destroyed altogether and the ConnectionId is no longer valid.

Panic

Panics if connection_id isn’t a valid connection.

pub async fn peers_list(&self) -> impl Iterator<Item = PeerId>

Returns an iterator to the list of PeerIds that we have an established connection with.

pub async fn num_outgoing_substreams(
&self,
notifications_protocol_index: usize
) -> usize

Returns the number of connections we have a substream with.