Module smoldot::libp2p::peerset

Data structure storing a networking state. Helper for managing connectivity to overlay networks.

The Peerset is a data structure that holds a list of node identities (PeerIds) and a list of overlay networks. Each PeerId is associated with:

A list of active inbound connections.

A list of Multiaddrs onto which the node is believed to be reachable.

For each multiaddr, optionally an active connection or pending dialing attempt.

A list of overlay networks the node is believed to belong to.

For each overlay network the node belongs to, one optional inbound and one optional outbound substream.

Each substream can be either “pending” or “established”.

Note: The Peerset does not do anything by itself, such as opening new connections. it is purely a data structure that helps organize and maintain information about the network.

Usage

The Peerset must be initialized with a list of overlay networks the node is interested in.

It is assumed that some discovery mechanism, not covered by this module, is in place in order to discover the identities and addresses of nodes that belong to these various overlay networks.

Structs

Config
Configuration for a Peerset.

ConnectionId
Identifier for a connection in a Peerset.

ConnectionMut
Access to a connection in the Peerset.

NodeMutKnown
Access to a node is already known to the data structure.

NodeMutUnknown
Access to a node that isn’t known to the data structure.

OverlayNetworkId
Identifier for an overlay network.

Peerset
See the module-level documentation.

PendingMut
Access to a connection in the Peerset.

Enums

NodeMut
Access to a node in the Peerset.

PendingOrConnectionMut
Access to a connection in the Peerset.

SubstreamDirection
Direction of a substream.