smoldot::libp2p

Module websocket

Implementation of a WebSocket client that wraps around an abstract representation of a TCP socket through the AsyncRead and AsyncWrite traits.

Structs

Config

Configuration for websocket_client_handshake.

Connection

Negotiated WebSocket connection.

Functions

websocket_client_handshake

Negotiates the WebSocket protocol (including the HTTP-like request) on the given socket, and returns an object that translates reads and writes into WebSocket binary frames.