Module smoldot::metadata::events

Events retrieval and decoding.

Overview

While this behaviour is not part of the specification and is thus not a strict requirement, Substrate-compatible blockchains built using the Substrate framework provide a storage item (in other words, an entry in the storage, with a specific key) containing a list of so-called events. This storage item is updated at each block with the events that have happened in the latest block.

Note: Events include, for example, a transfer between two accounts, a change in the validators nominated by a nominator, the beginning of a referundum, etc.

This module provides the tooling necessary to help retrieve said events.

In order to determine which storage key holds the list of events, one needs to provide the metadata. See the metadata module for information about what the metadata is and how to obtain it.

Usage

In order to know the events that have happened during a specific block:

Obtain the metadata of the runtime used by the desired block. This is out of scope of this module. See the metadata module for more information.

Call events_storage_key in order to obtain a key where to find the list of events in the storage. If events_storage_key returns an error, the runtime most likely doesn’t support events.

Obtain the storage value corresponding to the key obtained at the previous step. This is out of scope of this module. If there is no storage value at this key, this most likely indicates a bug somewhere, either in substrate-lite or in the runtime.

Decode the events. This isn’t implemented in this module yet. See the next section.

Flaw in the design

In the SCALE codec, each field is put one behind the other in an unstructured way. The start of a new field is known by adding the length of all the preceding fields, and the length of a field depends on the type of data it contains. The type of data is not explicitly laid out and is only known through context.

The list of events (encoded using SCALE) contains, amongst other things, the values of the parameters of said event. In order to be able decode this list of events, one has to know the types of these parameter values. This information is found in the metadata, in the form of a string representing the type as written out in the original Rust source code. This type could be anything, from a primitive type (e.g. u32) to a type alias (type Foo = ...;), or a locally-defined struct.

In order to properly decode events, one has to parse these strings representing Rust types and hard-code a list of types known to be used in the runtime code. This ranges from u32 to for example EthereumAddress.

Runtime upgrades can introduce new types and (albeit unlikely) modify the definition of existing types. As such, a function that decodes events can stop working after any runtime upgrade. For this reason, smoldot doesn’t provide any such function.

In the future, it is planned to add, in Substrate, runtime type information to the metadata, thus allowing a proper decoding function to be implemented. When that is done, smoldot can be updated to support decoding events.

Enums

EventsStorageKeyError
Error potentially returned by events_storage_key.

Functions

events_storage_key
Returns the key in the storage at which events can be found.