Enum smoldot::network::basic_peering_strategy::InsertAddressResult

pub enum InsertAddressResult {
    Inserted {
        address_removed: Option<Vec<u8>>,
    },
    AlreadyKnown,
    UnknownPeer,
}

See BasicPeeringStrategy::insert_address.

Variants

Inserted

Fields

address_removed: Option<Vec<u8>>

If the maximum number of addresses is reached, an old address might have been removed. If so, this contains the address.

Address has been successfully inserted.

AlreadyKnown

Address was already known.

UnknownPeer

The peer isn’t associated to any chain, and as such the address was not inserted.