smoldot::network::basic_peering_strategy

Struct BasicPeeringStrategy

pub struct BasicPeeringStrategy<TChainId, TInstant> { /* private fields */ }

Implementations

impl<TChainId, TInstant> BasicPeeringStrategy<TChainId, TInstant>
where TChainId: PartialOrd + Ord + Eq + Hash + Clone, TInstant: PartialOrd + Ord + Eq + Clone,

pub fn new(config: Config) -> Self

Creates a new empty BasicPeeringStrategy.

Must be passed a seed for randomness used in BasicPeeringStrategy::pick_assignable_peer.

pub fn remove_chain_peers(&mut self, chain: &TChainId)

Removes all the chain assignments for the given chain.

If a peer isn’t assigned to any chain anymore and doesn’t have any connected address, all of its addresses are also removed from the collection.

pub fn insert_chain_peer( &mut self, chain: TChainId, peer_id: PeerId, max_peers_per_chain: usize, ) -> InsertChainPeerResult

Inserts a chain-peer combination to the collection, indicating that the given peer belongs to the given chain.

Has no effect if the peer is already assigned to the given chain, in which case InsertChainPeerResult::Duplicate is returned.

A maximum number of peers per chain must be provided. If the peer is inserted and the limit is exceeded, a peer (other than the one that has just been inserted) that belongs to the given chain is randomly chosen and removed. Peers that have slots assigned to them are never removed.

pub fn unassign_slot_and_remove_chain_peer( &mut self, chain: &TChainId, peer_id: &PeerId, ) -> UnassignSlotAndRemoveChainPeer<TInstant>

Removes a peer-chain associated previously inserted with BasicPeeringStrategy::insert_chain_peer.

Has no effect if the peer-chain association didn’t exist.

If the peer isn’t assigned to any chain anymore and doesn’t have any connected address, all of its addresses are also removed from the collection.

pub fn chain_peers_unordered( &self, chain: &TChainId, ) -> impl Iterator<Item = &PeerId>

Returns the list of all peers that are known to belong to the given chain, in other words peers added through BasicPeeringStrategy::insert_chain_peer.

The order of the yielded elements is unspecified.

pub fn insert_address( &mut self, peer_id: &PeerId, address: Vec<u8>, max_addresses: usize, ) -> InsertAddressResult

Inserts a new address for the given peer.

If the address wasn’t known yet, its number of connections is set to zero.

If the peer doesn’t belong to any chain (see BasicPeeringStrategy::insert_chain_peer), then this function has no effect, unless the peer has at least one connected address. This is to avoid accidentally collecting addresses for peers that will never be removed and create a memory leak. For this reason, you most likely want to call BasicPeeringStrategy::insert_chain_peer before calling this function.

A maximum number of addresses that are maintained for this peer must be passed as parameter. If this number is exceeded, an address with zero connections (other than the one passed as parameter) is randomly removed.

pub fn increase_address_connections( &mut self, peer_id: &PeerId, address: Vec<u8>, max_addresses: usize, ) -> InsertAddressConnectionsResult

Increases the number of connections of the given address. If the address isn’t known, it is inserted.

Contrary to BasicPeeringStrategy::insert_address, the address is inserted anyway if the PeerId isn’t known.

Note: Use this function if you establish a connection and accidentally reach a certain PeerId.

Panic

Panics if the number of connections is equal to u32::MAX.

pub fn peer_addresses(&self, peer_id: &PeerId) -> impl Iterator<Item = &[u8]>

Returns the list of all addresses that have been inserted for the given peer.

pub fn pick_assignable_peer( &mut self, chain: &TChainId, now: &TInstant, ) -> AssignablePeer<'_, TInstant>

Chooses a PeerId that is known to belong to the given chain, that is not banned, and that doesn’t have a slot assigned to it.

A TInstant must be provided in order to determine whether past bans have expired.

If multiple peers can be assigned a slot, the one returned is chosen randomly. Calling this function multiple times might return different peers. For this reason, this function requires &mut self.

Note that this function might return a peer for which no address is present. While this is often not desirable, it is preferable to keep the API simple and straight-forward rather than try to be smart about function behaviours.

pub fn assign_slot(&mut self, chain: &TChainId, peer_id: &PeerId)

Assigns a slot to the given peer on the given chain.

Acts as an implicit call to BasicPeeringStrategy::insert_chain_peer.

A slot is assigned even if the peer is banned. API users that call this function are expected to be aware of that.

pub fn unassign_slot_and_ban( &mut self, chain: &TChainId, peer_id: &PeerId, when_unban: TInstant, ) -> UnassignSlotAndBan<TInstant>

Unassign the slot that has been assigned to the given peer and bans the peer, preventing it from being assigned a slot on this chain for a certain amount of time.

Has no effect if the peer isn’t assigned to the given chain.

If the peer was already banned, the new ban expiration is max(existing_ban, when_unban).

Returns what this function did.

pub fn unassign_slots_and_ban( &mut self, peer_id: &PeerId, when_unban: TInstant, ) -> UnassignSlotsAndBanIter<'_, TChainId, TInstant>

Unassigns all the slots that have been assigned to the given peer and bans the peer, preventing it from being assigned a slot for all of the chains it had a slot on for a certain amount of time.

Has no effect on chains the peer isn’t assigned to.

If the peer was already banned, the new ban expiration is max(existing_ban, when_unban).

Returns an iterator to the list of chains where the peer is now banned, and the details of what has happened.

Note: This function is a shortcut for calling BasicPeeringStrategy::unassign_slot_and_ban for all existing chains.

pub fn pick_address_and_add_connection( &mut self, peer_id: &PeerId, ) -> Option<&[u8]>

Picks an address from the list with zero connections, and sets the number of connections to one. Returns None if no such address is available.

pub fn decrease_address_connections( &mut self, peer_id: &PeerId, address: &[u8], ) -> Result<(), DecreaseAddressConnectionsError>

Removes one connection from the given address.

Returns an error if the address isn’t known to the data structure, or if there was no connection.

pub fn decrease_address_connections_and_remove_if_zero( &mut self, peer_id: &PeerId, address: &[u8], ) -> Result<(), DecreaseAddressConnectionsError>

Removes one connection from the given address. If this decreases the number of connections from one to zero, the address is removed entirely.

Returns an error if the address isn’t known to the data structure, or if there was no connection.