smoldot::network

Module bitswap_peering_strategy

Simple slot assignment algorithm for Bitswap protocol.

The strategy tries to open Bitswap substreams with all connected peers. In case a substream can’t be opened (for example, the remote doesn’t support Bitswap protocol) or the peer misbehaves, the peer is banned for a certain period of time, preventing it from being returned as a candidate for a Bitswap connection.

The strategy user must update its state when:

A connection is established with a peer — by calling BitswapPeeringStrategy::increase_peer_connections

A connection is terminated with a peer — by calling BitswapPeeringStrategy::decrease_peer_connections

BitswapPeeringStrategy contains all currently connected peers, identified by PeerId.

Each peer can be in one of these three states:

Normal.

Banned until a certain instant represented by TInstant.

Has a slot.

“Normal” and “banned” peers represent the potential peers to connect to over Bitswap protocol, while “slot” represent pending or established Bitswap protocol connection.

Use BitswapPeeringStrategy::pick_assignable_peer in order to get a randomly-chosen candidate for slot assignment from the peers that don’t currently have a slot assigned and are not banned. Use BitswapPeeringStrategy::assign_slot to assign a slot.

If a Bitswap connection fails to be established with a certain peer, or if the peer misbehaves, use BitswapPeeringStrategy::unassign_slot_and_ban to ban the peer, preventing it from obtaining a slot for a provided amount of time.

Re-exports

pub use crate::libp2p::PeerId;

Structs

BitswapPeeringStrategy

Config

Configuration passed to BitswapPeeringStrategy::new.

Enums

AssignSlotError

See BitswapPeeringStrategy::assign_slot.

AssignablePeer

See BitswapPeeringStrategy::pick_assignable_peer.

DecreasePeerConnectionsError

See BitswapPeeringStrategy::decrease_peer_connections.

UnassignSlotAndBan

See BitswapPeeringStrategy::unassign_slot_and_ban.