smoldot::network::codec

Function build_bitswap_cancel_message

pub fn build_bitswap_cancel_message(
    cids: impl Iterator<Item = impl AsRef<[u8]>>,
) -> Vec<u8> 

Builds a Bitswap message that withdraws (cancels) previously-issued wantlist entries.

Each entry in the resulting message has the cancel flag set, instructing peers to drop any pending want-list state they were holding for the listed CIDs. want_type is set to Block (the wire default) since cancel applies to all variants of a want-list entry for a given CID.

Arguments

cids - Iterator of CIDs to cancel