Module smoldot::network::kademlia

Structs

Kademlia

Data structure containing the k-buckets and the state of the current Kademlia queries.

ProtobufDecodeError

Error while decoding the protobuf encoding.

Enums

DecodeFindNodeResponseError

Error potentially returned by decode_find_node_response.

Functions

build_find_node_request

Builds a wire message to send on the Kademlia request-response protocol to ask the target to return the nodes closest to the parameter.

decode_find_node_response

Decodes a response to a request built using build_find_node_request.