Module smoldot::network::protocol

Substrate/Polkadot-specific protocols on top of libp2p.

Libp2p provides a framework on top of which higher-level, domain-specific protocols can be used. This module contains the domain-specific protocols specific to Substrate and Polkadot.

This module only provides the tools to encode/decode messages.

Re-exports

pub use crate::finality::grandpa::commit::decode::CommitMessageRef;

pub use crate::finality::grandpa::commit::decode::UnsignedPrecommitRef;

Structs

BlockAnnounceRef
Decoded block announcement notification.

BlockAnnouncesHandshakeDecodeError
Error potentially returned by decode_block_announces_handshake.

BlockAnnouncesHandshakeRef
Decoded handshake sent or received when opening a block announces notifications substream.

BlockData
Block sent in a block response.

BlocksRequestConfig
Description of a block request that can be sent to a peer.

BlocksRequestFields
Which fields should be present in the response.

CallProofRequestConfig
Description of a call proof request that can be sent to a peer.

CatchUpRef

CatchUpRequest

DecodeBlockAnnounceError
Error potentially returned by decode_block_announces_handshake.

DecodeGrandpaNotificationError
Error potentially returned by decode_grandpa_notification.

DecodeGrandpaWarpSyncResponseError
Error potentially returned by decode_grandpa_warp_sync_response.

GrandpaWarpSyncResponse
Response to a GrandPa warp sync request.

GrandpaWarpSyncResponseFragment
Response to a GrandPa warp sync request.

IdentifyResponse
Description of a response to an identify request.

NeighborPacket

PrevoteRef

PrimaryProposeRef

ProtobufDecodeError
Error while decoding the protobuf encoding.

StorageProofRequestConfig
Description of a storage proof request that can be sent to a peer.

UnsignedPrevoteRef

VoteMessageRef

Enums

BlocksRequestConfigStart
Which block the remote must return first.

BlocksRequestDirection
Whether the first block should be the one with the highest number, of the one with the lowest number.

DecodeBlockResponseError
Error potentially returned by decode_block_response.

DecodeCallProofResponseError
Error potentially returned by decode_call_proof_response.

DecodeStorageProofResponseError
Error potentially returned by decode_storage_proof_response.

GrandpaNotificationRef

MessageRef

Role
Role a node reports playing on the network.

Functions

build_block_request
Builds the bytes corresponding to a block request.

build_call_proof_request
Builds the bytes corresponding to a call proof request.

build_identify_response
Builds the bytes corresponding to a block request.

build_storage_proof_request
Builds the bytes corresponding to a storage proof request.

decode_block_announce
Decodes a block announcement.

decode_block_announces_handshake
Decodes a SCALE-encoded block announces handshake.

decode_block_response
Decodes a response to a block request.

decode_call_proof_response
Decodes a response to a call proof request.

decode_grandpa_notification
Attempt to decode the given SCALE-encoded Grandpa notification.

decode_grandpa_warp_sync_response
Decodes a SCALE-encoded GrandPa warp sync response.

decode_storage_proof_response
Decodes a response to a storage proof request.

encode_block_announce
Turns a block announcement into its SCALE-encoding ready to be sent over the wire.

encode_block_announces_handshake
Turns a block announces handshake into its SCALE-encoding ready to be sent over the wire.