Enum smoldot::network::service::Event

pub enum Event {

    Connected(PeerId),
    Disconnected {
        peer_id: PeerId,
        chain_indices: Vec<usize>,
    },
    ChainConnected {
        chain_index: usize,
        peer_id: PeerId,
        slot_ty: SlotTy,
        role: Role,
        best_number: u64,
        best_hash: [u8; 32],
    },
    ChainDisconnected {
        peer_id: PeerId,
        chain_index: usize,
        unassigned_slot_ty: SlotTy,
    },
    ChainConnectAttemptFailed {
        chain_index: usize,
        peer_id: PeerId,
        error: NotificationsOutErr,
        unassigned_slot_ty: SlotTy,
    },
    RequestResult {
        request_id: OutRequestId,
        response: RequestResult,
    },
    InboundSlotAssigned {
        chain_index: usize,
        peer_id: PeerId,
    },
    BlockAnnounce {
        peer_id: PeerId,
        chain_index: usize,
        announce: EncodedBlockAnnounce,
    },
    GrandpaCommitMessage {
        peer_id: PeerId,
        chain_index: usize,
        message: EncodedGrandpaCommitMessage,
    },
    ProtocolError {
        peer_id: PeerId,
        error: ProtocolError,
    },
    IdentifyRequestIn {
        peer_id: PeerId,
        request_id: InRequestId,
    },
    BlocksRequestIn {
        peer_id: PeerId,
        chain_index: usize,
        config: BlocksRequestConfig,
        request_id: InRequestId,
    },
    RequestInCancel {
        request_id: InRequestId,
    },
    KademliaDiscoveryResult {
        operation_id: KademliaOperationId,
        result: Result<Vec<(PeerId, Vec<Multiaddr>)>, DiscoveryError>,
    },

}

Event generated by ChainNetwork::next_event.

Variants

Connected(PeerId)

Established a transport-level connection (e.g. a TCP socket) with the given peer.

Disconnected

Fields

peer_id: PeerId

chain_indices: Vec<usize>

A transport-level connection (e.g. a TCP socket) has been closed.

This event is called unconditionally when a connection with the given peer has been closed. If chain_indices isn’t empty, this event is also equivalent to one or more Event::ChainDisconnected events.

ChainConnected

Fields

chain_index: usize

peer_id: PeerId

slot_ty: SlotTy

Type of the slot that the peer has.

role: Role

Role the node reports playing on the network.

best_number: u64

Height of the best block according to this node.

best_hash: [u8; 32]

Hash of the best block according to this node.

ChainDisconnected

Fields

peer_id: PeerId

chain_index: usize

unassigned_slot_ty: SlotTy

Type of the slot that the peer had and no longer has.

ChainConnectAttemptFailed

Fields

chain_index: usize

peer_id: PeerId

error: NotificationsOutErr

Problem that happened.

unassigned_slot_ty: SlotTy

Type of the slot that the peer had and no longer has.

An attempt has been made to open the given chain, but a problem happened.

RequestResult

Fields

request_id: OutRequestId

response: RequestResult

InboundSlotAssigned

Fields

chain_index: usize

peer_id: PeerId

The given peer has opened a block announces substream with the local node, and an inbound slot has been assigned locally to this peer.

A Event::ChainConnected or Event::ChainConnectAttemptFailed will later be generated for this peer.

BlockAnnounce

Fields

peer_id: PeerId

Identity of the sender of the block announce.

chain_index: usize

Index of the chain the block relates to.

announce: EncodedBlockAnnounce

Received a new block announce from a peer.

Can only happen after a Event::ChainConnected with the given PeerId and chain index combination has happened.

GrandpaCommitMessage

Fields

peer_id: PeerId

Identity of the sender of the message.

chain_index: usize

Index of the chain the commit message relates to.

message: EncodedGrandpaCommitMessage

Received a GrandPa commit message from the network.

ProtocolError

Fields

peer_id: PeerId

Peer that has caused the protocol error.

error: ProtocolError

Error that happened.

Error in the protocol in a connection, such as failure to decode a message. This event doesn’t have any consequence on the health of the connection, and is purely for diagnostic purposes.

IdentifyRequestIn

Fields

peer_id: PeerId

Remote that has sent the request.

request_id: InRequestId

Identifier of the request. Necessary to send back the answer.

A remote has sent a request for identification information.

You are strongly encouraged to call ChainNetwork::respond_identify.

BlocksRequestIn

Fields

peer_id: PeerId

Remote that has sent the request.

chain_index: usize

Index of the chain concerned by the request.

config: BlocksRequestConfig

Information about the request.

request_id: InRequestId

Identifier of the request. Necessary to send back the answer.

A remote has sent a request for blocks.

Can only happen for chains where ChainConfig::allow_inbound_block_requests is true.

You are strongly encouraged to call ChainNetwork::respond_blocks.

RequestInCancel

Fields

request_id: InRequestId

KademliaDiscoveryResult

Fields

operation_id: KademliaOperationId

result: Result<Vec<(PeerId, Vec<Multiaddr>)>, DiscoveryError>