Struct smoldot::network::service::ChainNetwork

pub struct ChainNetwork<TNow> { /* fields omitted */ }

Data structure containing the list of all connections, pending or not, and their latest known state. See also the module-level documentation.

Implementations

impl<TNow> ChainNetwork<TNow> where
TNow: Clone + Add<Duration, Output = TNow> + Sub<TNow, Output = Duration> + Ord,

pub fn new(config: Config) -> Self

Initializes a new ChainNetwork.

pub async fn num_established_connections(&self) -> usize

Returns the number of established TCP connections, both incoming and outgoing.

pub async fn num_peers(&self, chain_index: usize) -> usize

Returns the number of peers we have a substream with.

pub fn num_chains(&self) -> usize

Returns the number of chains. Always equal to the length of Config::chains.

pub async fn add_incoming_connection(
&self,
when_connected: TNow,
remote_addr: Multiaddr
) -> ConnectionId

Adds an incoming connection to the state machine.

This connection hasn’t finished handshaking and the PeerId of the remote isn’t known yet.

Must be passed the moment (as a TNow) when the connection as been established, in order to determine when the handshake timeout expires.

After this function has returned, you must process the connection with ChainNetwork::read_write.

The remote_addr is the address used to reach back the remote. In the case of TCP, it contains the TCP dialing port of the remote. The remote can ask, through the identify libp2p protocol, its own address, in which case we send it.

pub async fn set_local_best_block(
&self,
chain_index: usize,
best_hash: [u8; 32],
best_number: u64
)

Modifies the best block of the local node. See ChainConfig::best_hash and ChainConfig::best_number.

Panic

Panics if chain_index is out of range.

pub async fn set_local_grandpa_state(
&self,
chain_index: usize,
grandpa_state: GrandpaState
)

Update the state of the local node with regards to GrandPa rounds.

Calling this method does two things:

Send on all the active GrandPa substreams a “neighbor packet” indicating the state of the local node.

Update the neighbor packet that is automatically sent to peers when a GrandPa substream gets opened.

In other words, calling this function atomically informs all the present and future peers of the state of the local node regarding the GrandPa protocol.

Note: The information passed as parameter isn’t validated in any way by this method.

Panic

Panics if chain_index is out of range, or if the chain has GrandPa disabled.

pub async fn blocks_request(
&self,
now: TNow,
target: &PeerId,
chain_index: usize,
config: BlocksRequestConfig
) -> Result<Vec<BlockData>, BlocksRequestError>

Sends a blocks request to the given peer.

pub async fn grandpa_warp_sync_request(
&self,
now: TNow,
target: &PeerId,
chain_index: usize,
begin_hash: [u8; 32]
) -> Result<GrandpaWarpSyncResponse, GrandpaWarpSyncRequestError>

pub async fn storage_proof_request(
&self,
now: TNow,
target: &PeerId,
chain_index: usize,
config: StorageProofRequestConfig<impl Iterator<Item = impl AsRef<[u8]>>>
) -> Result<Vec<Vec<u8>>, StorageProofRequestError>

Sends a storage request to the given peer.

pub async fn call_proof_request(
&self,
now: TNow,
target: &PeerId,
chain_index: usize,
config: CallProofRequestConfig<'_, impl Iterator<Item = impl AsRef<[u8]>>>
) -> Result<Vec<Vec<u8>>, CallProofRequestError>

Sends a call proof request to the given peer.

This request is similar to ChainNetwork::storage_proof_request. Instead of requesting specific keys, we request the list of all the keys that are accessed for a specific runtime call.

There exists no guarantee that the proof is complete (i.e. that it contains all the necessary entries), as it is impossible to know this from just the proof itself. As such, this method is just an optimization. When performing the actual call, regular storage proof requests should be performed if the key is not present in the call proof response.

pub async fn announce_transaction(
&self,
target: &PeerId,
chain_index: usize,
extrinsic: &[u8]
) -> Result<(), QueueNotificationError>

Must be passed the double-SCALE-encoded transaction.

pub async fn pending_outcome_ok(&self, id: PendingId) -> ConnectionId

After calling ChainNetwork::next_start_connect, notifies the ChainNetwork of the success of the dialing attempt.

See also ChainNetwork::pending_outcome_err.

After this function has returned, you must process the connection with ChainNetwork::read_write.

Panic

Panics if the PendingId is invalid.

pub async fn pending_outcome_err(&self, id: PendingId)

After calling ChainNetwork::next_start_connect, notifies the ChainNetwork of the failure of the dialing attempt.

See also ChainNetwork::pending_outcome_ok.

Panic

Panics if the PendingId is invalid.

pub async fn next_event(&self, now: TNow) -> Event<'_, TNow>

Returns the next event produced by the service.

This function should be called at a high enough rate that ChainNetwork::read_write can continue pushing events to the internal buffer of events. Failure to call this function often enough will lead to connections being back-pressured. See also Config::pending_api_events_buffer_size.

It is technically possible to call this function multiple times simultaneously, in which case the events will be distributed amongst the multiple calls in an unspecified way. Keep in mind that some Events have logic attached to the order in which they are produced, and calling this function multiple times is therefore discouraged.

pub async fn kademlia_discovery_round(
&self,
now: TNow,
chain_index: usize
) -> Result<DiscoveryInsert<'_, TNow>, DiscoveryError>

Performs a round of Kademlia discovery.

This future yields once a list of nodes on the network has been discovered, or a problem happened.

pub async fn kademlia_find_node(
&self,
target: &PeerId,
now: TNow,
chain_index: usize,
close_to_key: &[u8]
) -> Result<Vec<(PeerId, Vec<Multiaddr>)>, KademliaFindNodeError>

Sends a Kademlia “find node” request to a single peer, and waits for it to answer.

Returns an error if there is no active connection with that peer.

pub async fn next_start_connect<'a>(&self, now: TNow) -> StartConnect<TNow>

Allocates a PendingId and returns a StartConnect indicating a multiaddress that the API user must try to dial.

Later, the API user must use ChainNetwork::pending_outcome_ok or ChainNetwork::pending_outcome_err to report how the connection attempt went.

The returned StartConnect contains the StartConnect::timeout field. It is the responsibility of the API user to ensure that ChainNetwork::pending_outcome_err is called if this timeout is reached.

If no outgoing connection is desired, the method waits until there is one.

pub async fn read_write(
&self,
connection_id: ConnectionId,
read_write: &mut ReadWrite<'_, TNow>
) -> Result<(), ConnectionError>

Reads data coming from the connection, updates the internal state machine, and writes data destined to the connection through the ReadWrite.

If an error is returned, the connection should be destroyed altogether and the ConnectionId is no longer valid.

Panic

Panics if connection_id isn’t a valid connection.

pub async fn peers_list(&self) -> impl Iterator<Item = PeerId>

Returns an iterator to the list of PeerIds that we have an established connection with.