Module smoldot::sync::all

All syncing strategies (optimistic, warp sync, all forks) grouped together.

This state machine combines GrandPa warp syncing, optimistic syncing, and all forks syncing into one state machine.

Overview

This state machine acts as a container of sources, blocks (verified or not), and requests. In order to initialize it, you need to pass, amongst other things, a chain_information::ChainInformation struct indicating the known state of the finality of the chain.

A request represents a query for information from a source. Once the request has finished, call one of the methods of the AllSync in order to notify the state machine of the oucome.

Structs

AllSync

BlockRequestSuccessBlock

Config
Configuration for the AllSync.

ConfigFull
See Config::full.

HeaderBodyVerify

HeaderVerify

RequestId
Identifier for a request in the AllSync.

SourceId
Identifier for a source in the AllSync.

StorageGet
Loading a storage value is required in order to continue.

StorageNextKey
Fetching the key that follows a given one is required in order to continue.

StoragePrefixKeys
Fetching the list of keys with a given prefix is required in order to continue.

WarpSyncFragmentVerify

Enums

Action
Start or cancel a request.

BlockAnnounceOutcome
Outcome of calling AllSync::block_announce.

BlockVerification
State of the processing of blocks.

BlocksRequestFirstBlock

HeaderVerifyError
Error that can happen when verifying a block header.

HeaderVerifyOutcome
Outcome of calling HeaderVerify::perform.

ProcessOne
Outcome of calling AllSync::process_one.

RequestDetail
See Action::Start::detail.

ResponseOutcome
Outcome of injecting a response in the AllSync.