Enum smoldot::sync::all::DesiredRequest

pub enum DesiredRequest {
    BlocksRequest {
        first_block_height: u64,
        first_block_hash: [u8; 32],
        num_blocks: NonZeroU64,
        request_headers: bool,
        request_bodies: bool,
        request_justification: bool,
    },
    WarpSync {
        sync_start_block_hash: [u8; 32],
    },
    StorageGetMerkleProof {
        block_hash: [u8; 32],
        state_trie_root: [u8; 32],
        keys: Vec<Vec<u8>>,
    },
    RuntimeCallMerkleProof {
        block_hash: [u8; 32],
        function_name: Cow<'static, str>,
        parameter_vectored: Cow<'static, [u8]>,
    },
}

See AllSync::desired_requests.

Variants

BlocksRequest

Fields

first_block_height: u64

Height of the first block to request.

first_block_hash: [u8; 32]

Hash of the first block to request.

num_blocks: NonZeroU64

Number of blocks the request should return.

Note that this is only an indication, and the source is free to give fewer blocks than requested.

This might be equal to u64::MAX in case no upper bound is required. The API user is responsible for clamping this value to a reasonable limit.

request_headers: bool

True if headers should be included in the response.

request_bodies: bool

True if bodies should be included in the response.

request_justification: bool

True if the justification should be included in the response, if any.

Requesting blocks from the source is requested.

The blocks should be provided in decreasing number, with first_block_hash as the highest number.

WarpSync

Fields

sync_start_block_hash: [u8; 32]

Hash of the known finalized block. Starting point of the request.

Sending a Grandpa warp sync request is requested.

StorageGetMerkleProof

Fields

block_hash: [u8; 32]

Hash of the block whose storage is requested.

state_trie_root: [u8; 32]

Merkle value of the root of the storage trie of the block.

keys: Vec<Vec<u8>>

Keys whose values is requested.

Sending a storage query is requested.

RuntimeCallMerkleProof

Fields

block_hash: [u8; 32]

Hash of the block whose call is made against.

function_name: Cow<'static, str>

Name of the function to be called.

parameter_vectored: Cow<'static, [u8]>

Concatenated SCALE-encoded parameters to provide to the call.

Sending a call proof query is requested.