Enum smoldot::sync::all::RequestDetail

pub enum RequestDetail {
    BlocksRequest {
        first_block_height: u64,
        first_block_hash: Option<[u8; 32]>,
        ascending: bool,
        num_blocks: NonZeroU64,
        request_headers: bool,
        request_bodies: bool,
        request_justification: bool,
    },
    GrandpaWarpSync {
        sync_start_block_hash: [u8; 32],
    },
    StorageGet {
        block_hash: [u8; 32],
        keys: Vec<Vec<u8>>,
    },
    RuntimeCallMerkleProof {
        block_hash: [u8; 32],
        function_name: Cow<'static, str>,
        parameter_vectored: Cow<'static, [u8]>,
    },
}

See AllSync::desired_requests.

Variants

BlocksRequest

Fields

first_block_height: u64

Height of the first block to request.

first_block_hash: Option<[u8; 32]>

Hash of the first block to request. None if not known.

ascending: bool

True if the first_block_hash is the response should contain blocks in an increasing number, starting from first_block_hash with the lowest number. If false, the blocks should be in decreasing number, with first_block_hash as the highest number.

num_blocks: NonZeroU64

Number of blocks the request should return.

Note that this is only an indication, and the source is free to give fewer blocks than requested.

This might be equal to u64::max_value() in case no upper bound is required. The API user is responsible for clamping this value to a reasonable limit.

request_headers: bool

True if headers should be included in the response.

request_bodies: bool

True if bodies should be included in the response.

request_justification: bool

True if the justification should be included in the response, if any.

Requesting blocks from the source is requested.

GrandpaWarpSync

Fields

sync_start_block_hash: [u8; 32]

Hash of the known finalized block. Starting point of the request.

Sending a Grandpa warp sync request is requested.

StorageGet

Fields

block_hash: [u8; 32]

Hash of the block whose storage is requested.

keys: Vec<Vec<u8>>

Keys whose values is requested.

Sending a storage query is requested.

RuntimeCallMerkleProof

Fields

block_hash: [u8; 32]

Hash of the block whose call is made against.

function_name: Cow<'static, str>

Name of the function to be called.

parameter_vectored: Cow<'static, [u8]>

Concatenated SCALE-encoded parameters to provide to the call.

Sending a call proof query is requested.

Implementations

impl RequestDetail

pub fn num_blocks_clamp(&mut self, max: NonZeroU64)

Caps the number of blocks to request to max.

pub fn with_num_blocks_clamp(self, max: NonZeroU64) -> Self

Caps the number of blocks to request to max.