Struct smoldot::sync::all::Config

pub struct Config {
    pub chain_information: ValidChainInformation,
    pub block_number_bytes: usize,
    pub allow_unknown_consensus_engines: bool,
    pub sources_capacity: usize,
    pub blocks_capacity: usize,
    pub max_disjoint_headers: usize,
    pub max_requests_per_block: NonZeroU32,
    pub download_ahead_blocks: NonZeroU32,
    pub full: Option<ConfigFull>,
}

Configuration for the AllSync.

Fields
chain_information: ValidChainInformation

Information about the latest finalized block and its ancestors.

block_number_bytes: usize

Number of bytes used when encoding/decoding the block number. Influences how various data structures should be parsed.

allow_unknown_consensus_engines: bool

If false, blocks containing digest items with an unknown consensus engine will fail to verify.

Passing true can lead to blocks being considered as valid when they shouldn’t. However, even if true is passed, a recognized consensus engine must always be present. Consequently, both true and false guarantee that the number of authorable blocks over the network is bounded.

sources_capacity: usize

Pre-allocated capacity for the number of block sources.

blocks_capacity: usize

Pre-allocated capacity for the number of blocks between the finalized block and the head of the chain.

Should be set to the maximum number of block between two consecutive justifications.

max_disjoint_headers: usize

Maximum number of blocks of unknown ancestry to keep in memory.

See all_forks::Config::max_disjoint_headers for more information.

max_requests_per_block: NonZeroU32

Maximum number of simultaneous pending requests made towards the same block.

See all_forks::Config::max_requests_per_block for more information.

download_ahead_blocks: NonZeroU32

Number of blocks to download ahead of the best verified block.

Whenever the latest best block is updated, the state machine will start block requests for the block best_block_height + download_ahead_blocks and all its ancestors. Considering that requesting blocks has some latency, downloading blocks ahead of time ensures that verification isn’t blocked waiting for a request to be finished.

The ideal value here depends on the speed of blocks verification speed and latency of block requests.

full: Option<ConfigFull>

If Some, the block bodies and storage are also synchronized. Contains the extra configuration.