Module smoldot::sync::all_forks

All-forks header and body syncing.

Overview

This state machine holds:

A list of sources of blocks, maintained by the API user.

For each source, a list of blocks hashes known by the source.

The latest known finalized block.

A tree of valid non-finalized blocks that all descend from the latest known finalized block.

(if full mode) A list of block headers whose body is currently being downloaded.

A list of block header waiting to be verified and whose ancestry with the latest finalized block is currently unknown.

The state machine has the objective to synchronize the tree of non-finalized blocks with its equivalent on the sources added by the API user.

Because it is not possible to predict which block in this tree is going to be finalized in the future, the entire tree needs to be synchronized.

Example: If the latest finalized block is block number 4, and the tree contains blocks 5, 6, and 7, and a source announces a block 5 that is different from the locally-known block 5, a block request will be emitted for this block 5, even if it is certain that this “other” block 5 will not become the local best block. This is necessary in case it is this other block 5 that will end up being finalized.

Full vs non-full

The Config::full option configures whether the state machine only holds headers of the non-finalized blocks (full equal to false), or the headers, and bodies, and storage (full equal to true).

In full mode, .

Bounded and unbounded containers

It is important to limit the memory usage of this state machine no matter how the potentially-malicious sources behave.

The state in this state machine can be put into three categories:

Each source of blocks has a certain fixed-size state associated to it (containing for instance its best block number and height). Each source also has up to one in-flight request, which might incur more memory usage. Managing this additional request is out of scope of this module. The user of this module is expected to limit the number of simultaneous sources.

A set of verified blocks that descend from the latest finalized block. This set is unbounded. The consensus and finalization algorithms of the chain are supposed to limit the number of possible blocks in this set.

A set of blocks that can’t be verified yet. Receiving a block announce inserts an element in this set. In order to handle situations where a malicious source announces lots of invalid blocks, this set must be bounded. Once it has reached a certain size, the blocks with the highest block number are discarded if their parent is also in this set or being downloaded from a source.

Consequently, and assuming that the number of simultaneous sources is bounded, and that the consensus and finalization algorithms of the chain are properly configured, malicious sources can’t indefinitely grow the state in this state machine. Malicious sources, however, can potentially increase the number of block requests required to download a long fork. This is, at most, an annoyance, and not a vulnerability.

Structs

AllForksSync

Config
Configuration for the AllForksSync.

HeaderVerify
Header verification to be performed.

RequestId
Identifier for a request in the [PendingBlocks].

RequestParams
Information about a blocks request to be performed on a source.

RequestSuccessBlock
Struct to pass back when a block request has succeeded.

SourceId
Identifier for a source in the [AllForksSources].

Enums

AncestrySearchResponseOutcome
Outcome of calling AllForksSync::finish_ancestry_search.

BlockAnnounceOutcome
Outcome of calling AllForksSync::block_announce.

BlockBodyVerify
State of the processing of blocks.

HeaderVerifyError
Error that can happen when verifying a block header.

HeaderVerifyOutcome
Outcome of calling HeaderVerify::perform.

JustificationVerification
Information about the verification of a justification that was stored for this block.

ProcessOne
State of the processing of blocks.