Struct smoldot::sync::all_forks::sources::AllForksSources

pub struct AllForksSources<TSrc> { /* fields omitted */ }

Collection of sources and which blocks they know about.

Implementations

impl<TSrc> AllForksSources<TSrc>

pub fn new(sources_capacity: usize, finalized_block_height: u64) -> Self

Creates a new container. Must be passed the height of the known finalized block.

pub fn keys(&self) -> impl ExactSizeIterator<Item = SourceId> + '_

Returns the list of all SourceIds.

pub fn is_empty(&self) -> bool

Returns true if the data structure is empty.

pub fn len(&self) -> usize

Returns the number of sources in the data structure.

pub fn num_blocks(&self) -> usize

Returns the number of unique blocks in the data structure.

pub fn finalized_block_height(&self) -> u64

Returns the finalized block height this state machine knows about.

pub fn add_source(
&mut self,
best_block_number: u64,
best_block_hash: [u8; 32],
user_data: TSrc
) -> SourceId

Add a new source to the container.

The user_data parameter is opaque and decided entirely by the user. It can later be retrieved using AllForksSources::user_data.

Returns the newly-created source entry.

pub fn remove(&mut self, source_id: SourceId) -> TSrc

Removes the source from the AllForksSources.

Returns the user data that was originally passed to AllForksSources::add_source.

Panic

Panics if the SourceId is out of range.

pub fn set_finalized_block_height(&mut self, height: u64)

Updates the height of the finalized block.

This removes from the collection, and will ignore in the future, all blocks whose height is inferior or equal to this value.

Panic

Panics if the new height is inferior to the previous value.

pub fn add_known_block(
&mut self,
source_id: SourceId,
height: u64,
hash: [u8; 32]
)

Registers a new block that the source is aware of.

Has no effect if height is inferior or equal to the finalized block height.

Panic

Panics if the SourceId is out of range.

pub fn remove_known_block(&mut self, height: u64, hash: &[u8; 32])

Removes a block from the list of blocks the sources are aware of.

Note: Alongside with AllForksSources::set_finalized_block_height, this method can be used to prevent the data structure from growing indefinitely.

pub fn source_remove_known_block(
&mut self,
source_id: SourceId,
height: u64,
hash: &[u8; 32]
)

Removes a block from the list of blocks the source is aware of.

Has no effect if the source didn’t know this block.

Note: Can be used when a request is sent to a node, and the node answers that it doesn’t know about the requested block contrary to previously believed.

Panic

Panics if the SourceId is out of range.

pub fn set_best_block(
&mut self,
source_id: SourceId,
height: u64,
hash: [u8; 32]
)

Sets the best block of this source.

Panic

Panics if the SourceId is out of range.

pub fn best_block(&self, source_id: SourceId) -> (u64, &[u8; 32])

Returns the current best block of the given source.

This corresponds either the latest call to AllForksSources::set_best_block, or to the parameter passed to AllForksSources::add_source.

Panic

Panics if the SourceId is invalid.

pub fn knows_non_finalized_block<'a>(
&'a self,
height: u64,
hash: &[u8; 32]
) -> impl Iterator<Item = SourceId> + 'a

Returns the list of sources for which AllForksSources::knows_non_finalized_block would return true.

Panic

Panics if height is inferior or equal to the finalized block height. Finalized blocks are intentionally not tracked by this data structure, and panicking when asking for a potentially-finalized block prevents potentially confusing or erroneous situations.

pub fn source_knows_non_finalized_block(
&self,
source_id: SourceId,
height: u64,
hash: &[u8; 32]
) -> bool

Returns true if AllForksSources::add_known_block or AllForksSources::set_best_block has earlier been called on this source with this height and hash, or if the source was originally created (using AllForksSources::add_source) with this height and hash.

Panic

Panics if the SourceId is out of range.

Panics if height is inferior or equal to the finalized block height. Finalized blocks are intentionally not tracked by this data structure, and panicking when asking for a potentially-finalized block prevents potentially confusing or erroneous situations.

pub fn contains(&self, source_id: SourceId) -> bool

Returns true if the SourceId is present in the collection.

pub fn user_data(&self, source_id: SourceId) -> &TSrc

Returns the user data associated to the source. This is the value originally passed through AllForksSources::add_source.

Panic

Panics if the SourceId is out of range.

pub fn user_data_mut(&mut self, source_id: SourceId) -> &mut TSrc

Returns the user data associated to the source. This is the value originally passed through AllForksSources::add_source.

Panic

Panics if the SourceId is out of range.