Module smoldot::sync::download_tree

Data structure. Inputs blocks, tries to fetch their runtime, and outputs blocks whose runtime is known.

Usage

This data structure holds a tree of blocks whose runtime is either known, or not. This tree can be updated by using DownloadTree::input_insert_block and DownloadTree::input_finalize.

The data structure also holds a list of on-going runtime parameter downloads. Use DownloadTree::next_necessary_download to insert an ongoing download in the data structure. It is the responsibility of the API user to actually perform the download and use DownloadTree::runtime_download_finished_existing, DownloadTree::runtime_download_finished_new, or DownloadTree::runtime_download_failure when this download is finished.

Whenever it is updated, the DownloadTree can also update the block that it considers as the “output best block” and the block that it considers as the “output finalized block”. These blocks are guaranteed to have their runtime known. At initialization, the data structure is initially in a “not ready” state, in which case it doesn’t have any output best or finalized block. Use DownloadTree::has_output to determine whether the data structure is ready.

Finally, the data structure also holds a list of runtimes injected through DownloadTree::runtime_download_finished_new. You are strongly encouraged to periodically call DownloadTree::drain_unused_runtimes in order to free up resources.

About the runtime environment upgrade log item

The code in this state machine parses block headers, and checks whether the “runtime enviroment changed” log item is present in its digests. This log item has only been added to Substrate on September 15th 2021. At the time of the writing of this comment, this change is planned to be released as part of Polkadot v0.9.11. Consquently, blocks older than this version never have this log item.

This means that, for older blocks, this state machine isn’t able to detect changes in the runtime. Please be aware of this and do not use this state machine on old blocks.

Structs

DownloadId

Identifier for a download in the DownloadTree.

DownloadParams

Information about a download that must be performed.

DownloadTree

OutputUpdateBlock

See OutputUpdate.

RuntimeId

Identifier for a runtime in the DownloadTree.

Enums

NextNecessaryDownload

OutputUpdate

See DownloadTree::try_advance_output.

OutputUpdateBlockBest

See OutputUpdateBlock.

RuntimeError

Error when analyzing the runtime.