Struct smoldot::sync::download_tree::DownloadTree

pub struct DownloadTree<TNow, TRt> { /* fields omitted */ }

Implementations

impl<TNow, TRt> DownloadTree<TNow, TRt> where
TNow: Clone + Add<Duration, Output = TNow> + Ord,

pub fn from_finalized_block(
finalized_block_scale_encoded_header: Vec<u8>
) -> Self

Returns a new DownloadTree containing one “input” finalized block.

This DownloadTree is in a non-ready state.

pub fn from_finalized_block_and_runtime<'a>(
finalized_block_scale_encoded_header: Vec<u8>,
finalized_runtime: TRt
) -> Self

Returns a new DownloadTree containing one “input” finalized block. The runtime of this finalized block will be the one passed as parameter, meaning that this block also becomes the “output” finalized block and the “output” best block.

This DownloadTree is immediately in a ready state. No OutputUpdate::FirstFinalized will be generated.

pub fn has_output(&self) -> bool

Returns true if the state machine is in a ready state, meaning that it has an “output” finalized block and an “output” best block. In other words, returns true if calling DownloadTree::try_advance_output has returned OutputUpdate::FirstFinalized in the past.

Several methods panic if the state machine isn’t in a ready state.

pub fn finalized_block_hash(&self) -> &[u8; 32]

Returns the hash of the current “output” finalized block.

Panic

Panics if DownloadTree::has_output isn’t true.

pub fn best_block_hash(&self) -> &[u8; 32]

Returns the hash of the current “output” best block.

Panic

Panics if DownloadTree::has_output isn’t true.

pub fn finalized_block_header(&self) -> &[u8]

Returns the SCALE-encoded header of the current “output” finalized block.

Panic

Panics if DownloadTree::has_output isn’t true.

pub fn best_block_header(&self) -> &[u8]

Returns the SCALE-encoded header of the current “output” best block.

Panic

Panics if DownloadTree::has_output isn’t true.

pub fn finalized_block_runtime_id(&self) -> RuntimeId

Returns the runtime of the current “output” finalized block. Returns an error if the runtime had failed to compile.

Panic

Panics if DownloadTree::has_output isn’t true.

pub fn best_block_runtime_id(&self) -> RuntimeId

Returns the runtime of the current “output” best block. Returns an error if the runtime had failed to compile.

Panic

Panics if DownloadTree::has_output isn’t true.

pub fn finalized_block_runtime(&self) -> &TRt

Returns the runtime of the current “output” finalized block. Returns an error if the runtime had failed to compile.

Panic

Panics if DownloadTree::has_output isn’t true.

pub fn best_block_runtime(&self) -> &TRt

Returns the runtime of the current “output” best block. Returns an error if the runtime had failed to compile.

Panic

Panics if DownloadTree::has_output isn’t true.

pub fn finalized_block_runtime_mut(&mut self) -> &mut TRt

Returns the runtime of the current “output” finalized block. Returns an error if the runtime had failed to compile.

Panic

Panics if DownloadTree::has_output isn’t true.

pub fn best_block_runtime_mut(&mut self) -> &mut TRt

Returns the runtime of the current “output” best block. Returns an error if the runtime had failed to compile.

Panic

Panics if DownloadTree::has_output isn’t true.

pub fn block_header(&self, hash: &[u8; 32]) -> Option<&[u8]>

Returns the SCALE-encoded header of the block with the given hash. Returns None if no block with that hash is part of the output.

The returned header is guaranteed to be valid, as otherwise the block wouldn’t be part of the output.

Panic

Panics if DownloadTree::has_output isn’t true.

pub fn block_runtime(&self, hash: &[u8; 32]) -> Option<&TRt>

Returns the runtime of the block with the given hash. Returns None if no block with that hash is part of the output.

Panic

Panics if DownloadTree::has_output isn’t true.

pub fn block_runtime_mut(&mut self, hash: &[u8; 32]) -> Option<&mut TRt>

Returns the runtime of the block with the given hash. Returns None if no block with that hash is part of the output.

Panic

Panics if DownloadTree::has_output isn’t true.

pub fn non_finalized_blocks_headers_ancestry_order(
&self
) -> impl Iterator<Item = (&[u8], bool)> + '_

Returns the list of all SCALE-encoded headers of non-finalized blocks, plus a boolean indicating whether this is the best block.

The returned items are guaranteed to be in an order in which the parents are found before their children.

pub fn runtimes_iter(&self) -> impl Iterator<Item = (RuntimeId, &TRt)> + '_

Iterates over all the runtimes stored in this data structure.

pub fn runtime_user_data(&self, id: RuntimeId) -> &TRt

Returns the user data associated to the runtime with the given identifier.

Panic

Panics if the RuntimeId is invalid.

pub fn runtime_user_data_mut(&mut self, id: RuntimeId) -> &mut TRt

Returns the user data associated to the runtime with the given identifier.

Panic

Panics if the RuntimeId is invalid.

pub fn drain_unused_runtimes(
&mut self
) -> impl Iterator<Item = (RuntimeId, TRt)>

Removes from the data structure the runtimes that aren’t used by any block.

pub fn runtime_download_finished_existing(
&mut self,
download_id: DownloadId,
runtime_id: RuntimeId
)

Injects into the state of the data structure a completed runtime download, and that the downloaded runtime is the same as one that already exists in the data structure.

Panic

Panics if the DownloadId is invalid. Panics if the RuntimeId is invalid.

pub fn runtime_download_finished_new(
&mut self,
download_id: DownloadId,
user_data: TRt
) -> RuntimeId

Injects into the state of the data structure a completed runtime download, and that the downloaded runtime is new.

Panic

Panics if the DownloadId is invalid.

pub fn runtime_download_failure(&mut self, download_id: DownloadId, now: &TNow)

Injects into the state of the state machine a failed runtime download.

This same download will not be repeated for the next few seconds. Thanks to this, it is possible to immediately call this function in response to a new necessary download.

Panic

Panics if the DownloadId is invalid.

pub fn next_necessary_download(
&mut self,
now: &TNow
) -> NextNecessaryDownload<TNow>

Examines the state of self and, if a block’s runtime should be downloaded, changes the state of the block to “downloading” and returns the parameters of the download.

pub fn input_insert_block(
&mut self,
scale_encoded_header: Vec<u8>,
parent_hash: &[u8; 32],
is_new_best: bool
)

Updates the state machine with a new block.

Panic

Panics if parent_hash wasn’t inserted before.

pub fn input_finalize(
&mut self,
hash_to_finalize: [u8; 32],
new_best_block_hash: [u8; 32]
)

Updates the state machine to take into account that the input of blocks has finalized the given block.

new_best_block_hash is the hash of the best block after the finalization.

Note: Finalizing a block might have to modify the current best block if the block being finalized isn’t an ancestor of the current best block.

Panic

Panics if hash_to_finalize or new_best_block_hash weren’t inserted before. Panics if trying to finalize the parent of a block that is already finalized. Panics if new_best_block_hash is not a descendant of hash_to_finalize.

pub fn try_advance_output(&mut self) -> Option<OutputUpdate<'_>>

Tries to update the output blocks to follow the input.

Should be called after inserting a new block, finalizing a block, or when a download is finished.

Returns None if the state machine doesn’t have any output yet.

If Some is returned, this method should be called repeatedly until OutputUpdate::None is returned. Each call can perform an additional update.