Module smoldot::sync::grandpa_warp_sync

Grandpa warp syncing.

Overview

The Grandpa warp syncing algorithm consists in the following steps:

Downloading a warp sync proof from a source. This proof contains a list of fragments. Each fragment represents a change in the list of Grandpa authorities, and a list of signatures of the previous authorities that certify that this change is correct.

Verifying the fragments. Each fragment that is successfully verified progresses towards towards the head of the chain. Even if one fragment is invalid, all the previously-verified fragments can still be kept, and the warp syncing can resume from there.

Downloading from a source the runtime code of the final block of the proof.

Performing some runtime calls in order to obtain the current consensus-related parameters of the chain. This might require obtaining some storage items, in which case they must also be downloaded from a source.

At the end of the syncing, a ValidChainInformation corresponding to the head of the chain is yielded.

Usage

Use the grandpa_warp_sync function to start a Grandpa warp syncing state machine.

At any given moment, this state machine holds a list of sources that it might use to download the warp sync proof or the runtime code. Sources must be added and removed by the API user by calling one of the various add_source and remove_source functions.

Sources are identified through a SourceId. Each source has an opaque so-called “user data” of type TSrc associated to it. The content of this “user data” is at the discretion of the API user.

The InProgressGrandpaWarpSync enum must be examined in order to determine how to make the warp syncing process.

At the end of the process, a Success is returned and can be used to kick-off another syncing phase.

Re-exports

pub use warp_sync::Error as FragmentError;

Structs

Config
The configuration for grandpa_warp_sync.

NextKey
Fetching the key that follows a given one is required in order to continue.

SourceId
Identifier for a source in the GrandpaWarpSync.

StorageGet
Loading a storage value is required in order to continue.

Success
The result of a successful warp sync.

Verifier
Verifying the warp sync response is required to continue.

VirtualMachineParamsGet
Fetching the parameters for the virtual machine is required to continue.

WaitingForSources
Adding more sources of GrandPa warp sync data to is required to continue.

WarpSyncRequest
Requesting GrandPa warp sync data from a source is required to continue.

Enums

Error
Problem encountered during a call to grandpa_warp_sync.

GrandpaWarpSync
The GrandPa warp sync state machine.

InProgressGrandpaWarpSync

Functions

grandpa_warp_sync
Starts syncing via GrandPa warp sync.