Enum smoldot::sync::grandpa_warp_sync::InProgressGrandpaWarpSync

pub enum InProgressGrandpaWarpSync<TSrc> {
    StorageGet(StorageGet<TSrc>),
    NextKey(NextKey<TSrc>),
    Verifier(Verifier<TSrc>),
    WarpSyncRequest(WarpSyncRequest<TSrc>),
    VirtualMachineParamsGet(VirtualMachineParamsGet<TSrc>),
    WaitingForSources(WaitingForSources<TSrc>),
}

Variants

StorageGet(StorageGet<TSrc>)

Loading a storage value is required in order to continue.

NextKey(NextKey<TSrc>)

Fetching the key that follows a given one is required in order to continue.

Verifier(Verifier<TSrc>)

Verifying the warp sync response is required to continue.

WarpSyncRequest(WarpSyncRequest<TSrc>)

Requesting GrandPa warp sync data from a source is required to continue.

VirtualMachineParamsGet(VirtualMachineParamsGet<TSrc>)

Fetching the parameters for the virtual machine is required to continue.

WaitingForSources(WaitingForSources<TSrc>)

Adding more sources of GrandPa warp sync data to is required to continue.

Implementations

impl<TSrc> InProgressGrandpaWarpSync<TSrc>

pub fn as_chain_information(&self) -> ValidChainInformationRef<'_>

Returns the chain information that is considered verified.

pub fn sources(&self) -> impl Iterator<Item = SourceId> + '_

Returns a list of all known sources stored in the state machine.

pub fn source_user_data(&self, source_id: SourceId) -> &TSrc

Returns the user data (TSrc) corresponding to the given source.

Panic

Panics if the SourceId is invalid.

pub fn source_user_data_mut(&mut self, source_id: SourceId) -> &mut TSrc

Returns the user data (TSrc) corresponding to the given source.

Panic

Panics if the SourceId is invalid.

pub fn remove_source(
self,
to_remove: SourceId
) -> (TSrc, InProgressGrandpaWarpSync<TSrc>)

Remove a source from the list of sources.

Panic

Panics if the source wasn’t added to the list earlier.