Enum smoldot::sync::optimistic::BlockVerification

pub enum BlockVerification<TRq, TSrc, TBl> {
    Reset {
        sync: OptimisticSync<TRq, TSrc, TBl>,
        previous_best_height: u64,
        reason: ResetCause,
    },
    NewBest {
        sync: OptimisticSync<TRq, TSrc, TBl>,
        new_best_number: u64,
        new_best_hash: [u8; 32],
    },
    Finalized {
        sync: OptimisticSync<TRq, TSrc, TBl>,
        finalized_blocks: Vec<Block<TBl>>,
    },
    FinalizedStorageGet(StorageGet<TRq, TSrc, TBl>),
    FinalizedStoragePrefixKeys(StoragePrefixKeys<TRq, TSrc, TBl>),
    FinalizedStorageNextKey(StorageNextKey<TRq, TSrc, TBl>),
}

State of the processing of blocks.

Variants

Reset

An issue happened when verifying the block or its justification, resulting in resetting the chain to the latest finalized block.

Note: The latest finalized block might be a block imported during the same operation.

Fields of Reset

sync: OptimisticSync<TRq, TSrc, TBl>

The state machine. The OptimisticSync::process_one method takes ownership of the OptimisticSync. This field yields it back.

previous_best_height: u64

Height of the best block before the reset.

reason: ResetCause

Problem that happened and caused the reset.

NewBest

Processing of the block is over.

There might be more blocks remaining. Call OptimisticSync::process_one again.

Fields of NewBest

sync: OptimisticSync<TRq, TSrc, TBl>

The state machine. The OptimisticSync::process_one method takes ownership of the OptimisticSync. This field yields it back.

new_best_number: u64new_best_hash: [u8; 32]

Finalized

Processing of the block is over. The block has been finalized.

There might be more blocks remaining. Call OptimisticSync::process_one again.

Fields of Finalized

sync: OptimisticSync<TRq, TSrc, TBl>

The state machine. The OptimisticSync::process_one method takes ownership of the OptimisticSync. This field yields it back.

finalized_blocks: Vec<Block<TBl>>

Blocks that have been finalized. Includes the block that has just been verified.

FinalizedStorageGet(StorageGet<TRq, TSrc, TBl>)

Loading a storage value of the finalized block is required in order to continue.

FinalizedStoragePrefixKeys(StoragePrefixKeys<TRq, TSrc, TBl>)

Fetching the list of keys of the finalized block with a given prefix is required in order to continue.

FinalizedStorageNextKey(StorageNextKey<TRq, TSrc, TBl>)

Fetching the key of the finalized block storage that follows a given one is required in order to continue.