Module smoldot::sync::para

Parachains and parathreads syncing.

A parachain is a blockchain whose best and finalized blocks are determined by looking at the storage of a different chain called the relay chain.

In order to obtain the current best block of a parachain, you must obtain the parachains persisted validation data from the relay chain. This is done by calling the ParachainHost_persisted_validation_data runtime function. The runtime function returns an opaque set of bytes called the “head data” whose meaning depends on the parachain. Most of the time, it is a block hash.

In order to obtain the current finalized block of a parachain, do the same but on the current finalized block of the relay chain.

See the persisted_validation_data_parameters to obtain the input to pass to the runtime function. The first parameter is a para_id found in the chain specification of the parachain of parathread.

Structs

Error
Error that can happen during the decoding.

PersistedValidationDataRef
Decoded persisted validation data.

Enums

OccupiedCoreAssumption
An assumption being made about the state of an occupied core.

Constants

PERSISTED_VALIDATION_FUNCTION_NAME
Name of the runtime function to call in order to obtain the parachain heads.

Functions

decode_persisted_validation_data_return_value
Attempt to decode the return value of the ParachainHost_persisted_validation_data runtime call.

persisted_validation_data_parameters
Produces the input to pass to the ParachainHost_persisted_validation_data runtime call.