Module smoldot::sync::warp_sync

Warp syncing.

Warp syncing makes it possible to very quickly reach the currently finalized block of a chain. No attempt is made at verifying blocks. The algorithm assumes that the currently finalized block is valid, as the chain is likely bricked if this is not the case.

Long range attack vulnerability

Warp syncing is particularly vulnerable to what is called long range attacks.

The authorities allowed to finalize blocks can generate multiple proofs of finality for multiple different blocks of the same height. In other words, they can finalize more than one chain at a time.

Finalizing multiple chains at the same time is called an equivocation. Client implementations detect equivocations and report them to the runtime. While this is out of scope of the client, the runtime then typically provides a mechanism that punishes the validators that have equivocated. However, this defense mechanism is flawed in case when a long time passes between the multiple finality proofs generated by the same validators. Clients cannot hold an unlimited amount of information in memory, so they might not detect the equivocation, and even if it is detected, the punishment might not be enforceable because validators have moved all their funds.

In other words, it is possible for two thirds of the validators that were active at a certain past block N to collude and decide to finalize a different block N, even when N has been finalized for the first time several weeks or months in the past. When a client then warp syncs, it can be tricked to consider this alternative block N as the finalized one.

There is no fool-proof defense against this attack. However, consider the extremely high investment and high risk for the malicious validators, and the difficulty of pulling off this attack, it is extremely unlikely to happen in reality. The aforementioned punishment system is the only defense against this attack, and in order to be effective, the starting point of the warp syncing shouldn’t be too far in the past. How far exactly depends on the logic of the runtime of the chain.

Overview

The warp syncing algorithm works only if the chain uses Grandpa for its finality. It consists in the following steps:

Downloading a warp sync proof from a source. This proof contains a list of fragments. Each fragment represents a change in the list of Grandpa authorities, and a list of signatures of the previous authorities that certify that this change is correct.

Verifying the fragments. Each fragment that is successfully verified progresses towards the head of the chain. Even if one fragment is invalid, all the previously-verified fragments can still be kept, and the warp syncing can resume from there.

Downloading from a source the runtime code of the final block of the proof.

Performing some runtime calls in order to obtain the current consensus-related parameters of the chain. This might require obtaining some storage items, in which case they must also be downloaded from a source.

At the end of the syncing, a ValidChainInformation corresponding to the head of the chain is yielded.

Usage

Use the start_warp_sync() function to start a Grandpa warp syncing state machine.

At any given moment, this state machine holds a list of sources that it might use to download the warp sync proof or the runtime code. Sources must be added and removed by the API user by calling one of the various add_source and remove_source functions.

Sources are identified through a SourceId. Each source has an opaque so-called “user data” of type TSrc associated to it. The content of this “user data” is at the discretion of the API user.

Similarly, at any given moment, this state machine holds a list of requests that concern these sources. Use WarpSync::desired_requests to determine which requests will be useful to the progress of the warp syncing, then use WarpSync::add_request to update the state machine with a newly-started request.

Use WarpSync::process_one in order to run verifications of the payloads that have previously been downloaded.

Re-exports

pub use trie::Nibble;

Structs

BuildChainInformation

Ready to verify the parameters of the chain against the finalized block.

BuildRuntime

Ready to build the runtime of the finalized chain.

Config

The configuration for start_warp_sync().

ConfigCodeTrieNodeHint

See Config::code_trie_node_hint.

RequestId

Identifier for a request in the warp sync state machine.

RuntimeInformation

SourceId

Identifier for a source in the WarpSync.

SourceMisbehavior

Problem encountered during a call to BuildRuntime::build or BuildChainInformation::build that can be attributed to the source sending invalid data.

VerifyWarpSyncFragment

Ready to verify a warp sync fragment.

WarpSync

Warp syncing process state machine.

WarpSyncFragment

Fragment to be verified.

Enums

BuildChainInformationError

Problem encountered during a call to BuildChainInformation::build.

BuildRuntimeError

Problem encountered during a call to BuildRuntime::build.

DesiredRequest

Information about a request that the warp sync state machine would like to start.

ProcessOne

Return value of WarpSync::process_one.

RequestDetail

Information about a request to add to the state machine.

SourceMisbehaviorTy

See SourceMisbehavior::error.

Status

See WarpSync::status.

VerifyFragmentError

Error potentially returned by VerifyWarpSyncFragment::verify.

WarpSyncInitError

Error potentially returned by start_warp_sync().

Functions

start_warp_sync

Initializes the warp sync state machine.