Enum smoldot::sync::warp_sync::DesiredRequest

pub enum DesiredRequest {
    WarpSyncRequest {
        block_hash: [u8; 32],
    },
    StorageGetMerkleProof {
        block_hash: [u8; 32],
        state_trie_root: [u8; 32],
        keys: Vec<Vec<u8>>,
    },
    RuntimeCallMerkleProof {
        block_hash: [u8; 32],
        function_name: Cow<'static, str>,
        parameter_vectored: Cow<'static, [u8]>,
    },
}

Information about a request that the warp sync state machine would like to start.

See InProgressWarpSync::desired_requests.

Variants

WarpSyncRequest

Fields

block_hash: [u8; 32]

Starting point of the warp syncing. The first fragment of the response should be the of the epoch that the starting point is in.

A warp sync request should be start.

StorageGetMerkleProof

Fields

block_hash: [u8; 32]

Hash of the block to request the parameters against.

state_trie_root: [u8; 32]

State trie root hash found in the header of the block.

keys: Vec<Vec<u8>>

Keys whose values are requested.

A storage request of the runtime code and heap pages should be started.

RuntimeCallMerkleProof

Fields

block_hash: [u8; 32]

Hash of the header of the block the call should be made against.

function_name: Cow<'static, str>

Name of the function of the call proof.

parameter_vectored: Cow<'static, [u8]>

Parameters of the call.

A call proof should be started.