Struct smoldot::sync::warp_sync::Config

pub struct Config {
    pub start_chain_information: ValidChainInformation,
    pub block_number_bytes: usize,
    pub sources_capacity: usize,
    pub requests_capacity: usize,
    pub code_trie_node_hint: Option<ConfigCodeTrieNodeHint>,
    pub num_download_ahead_fragments: usize,
    pub warp_sync_minimum_gap: usize,
}

The configuration for start_warp_sync().

Fields
start_chain_information: ValidChainInformation

The chain information of the starting point of the warp syncing.

block_number_bytes: usize

Number of bytes used when encoding/decoding the block number. Influences how various data structures should be parsed.

sources_capacity: usize

The initial capacity of the list of sources.

requests_capacity: usize

The initial capacity of the list of requests.

code_trie_node_hint: Option<ConfigCodeTrieNodeHint>

Known valid Merkle value and storage value combination for the :code key.

If provided, the warp syncing algorithm will first fetch the Merkle value of :code, and if it matches the Merkle value provided in the hint, use the storage value in the hint instead of downloading it. If the hint doesn’t match, an extra round-trip will be needed, but if the hint matches it saves a big download.

num_download_ahead_fragments: usize

Number of warp sync fragments after which the state machine will pause downloading new ones until the ones that have been downloaded are verified.

A too low value will cause stalls, while a high value will use more memory and runs the risk of wasting more bandwidth in case the downloaded fragments need to be thrown away.

warp_sync_minimum_gap: usize

If the height of the current local finalized block is N, the warp sync state machine will not attempt to warp sync to blocks whose height inferior or equal to N + k where k is the value in this field.

Because warp syncing is a relatively expensive process, it is not worth performing it between two blocks that are too close to each other.

The ideal value of this field depends on the block production rate and the time it takes to answer requests.