Struct smoldot::sync::warp_sync::InProgressWarpSync

pub struct InProgressWarpSync<TSrc, TRq> { /* private fields */ }

Warp syncing process now obtaining the chain information.

Implementations

impl<TSrc, TRq> InProgressWarpSync<TSrc, TRq>

pub fn block_number_bytes(&self) -> usize

Returns the value that was initially passed in Config::block_number_bytes.

pub fn as_chain_information(&self) -> ValidChainInformationRef<'_>

Returns the chain information that is considered verified.

pub fn status(&self) -> Status<'_, TSrc>

Returns the current status of the warp syncing.

pub fn sources(&self) -> impl Iterator<Item = SourceId> + '_

Returns a list of all known sources stored in the state machine.

pub fn add_source(&mut self, user_data: TSrc) -> SourceId

Add a source to the list of sources.

pub fn remove_source(
&mut self,
to_remove: SourceId
) -> (TSrc, impl Iterator<Item = (RequestId, TRq)> + '_)

Removes a source from the list of sources. In addition to the user data associated to this source, also returns a list of requests that were in progress concerning this source. These requests are now considered obsolete.

Panic

Panics if the SourceId is invalid.

pub fn desired_requests(
&self
) -> impl Iterator<Item = (SourceId, &TSrc, DesiredRequest)> + '_

Returns a list of requests that should be started in order to drive the warp syncing process to completion.

Once a request that matches a desired request is added through InProgressWarpSync::add_request, it is no longer returned by this function.

pub fn add_request(
&mut self,
source_id: SourceId,
user_data: TRq,
detail: RequestDetail
) -> RequestId

Inserts a new request in the data structure.

Note: The request doesn’t necessarily have to match a request returned by InProgressWarpSync::desired_requests.

Panic

Panics if the SourceId is out of range.

pub fn fail_request(&mut self, id: RequestId) -> TRq

Removes the given request from the state machine. Returns the user data that was associated to it.

Panic

Panics if the RequestId is invalid.

pub fn storage_get_success(&mut self, id: RequestId, merkle_proof: Vec<u8>) -> TRq

Injects a successful Merkle proof and removes the given request from the state machine. Returns the user data that was associated to it.

Panic

Panics if the RequestId is invalid. Panics if the RequestId doesn’t correspond to a storage get request.

pub fn runtime_call_merkle_proof_success(
&mut self,
request_id: RequestId,
response: Vec<u8>
) -> TRq

Injects a successful response and removes the given request from the state machine. Returns the user data that was associated to it.

Panic

Panics if the RequestId is invalid. Panics if the RequestId doesn’t correspond to a runtime Merkle call proof request.

pub fn warp_sync_request_success(
&mut self,
request_id: RequestId,
fragments: Vec<WarpSyncFragment>,
final_set_of_fragments: bool
) -> TRq

Injects a successful response and removes the given request from the state machine. Returns the user data that was associated to it.

Panic

Panics if the RequestId is invalid. Panics if the RequestId doesn’t correspond to a warp sync request.

pub fn process_one(self) -> ProcessOne<TSrc, TRq>

Start processing one CPU operation.

This function takes ownership of self and yields it back after the operation is finished.