Module smoldot::transactions::validate

Runtime call to obtain the transactions validity status.

Structs

Config
Configuration for a transaction validation process.

DecodeError
Error that can happen during the decoding.

NextKey
Fetching the key that follows a given one is required in order to continue.

PrefixKeys
Fetching the list of keys with a given prefix from the parent storage is required in order to continue.

StorageGet
Loading a storage value is required in order to continue.

ValidTransaction
Information concerning a valid transaction.

Enums

Error
Problem encountered during a call to validate_transaction.

InvalidTransaction
An invalid transaction validity.

Query
Current state of the operation.

TransactionSource
Source of the transaction.

TransactionValidityError
Errors that can occur while checking the validity of a transaction.

UnknownTransaction
An unknown transaction validity.

Constants

VALIDATION_FUNCTION_NAME
Name of the runtime function to call in order to validate a transaction.

Functions

decode_validate_transaction_return_value
Attempt to decode the return value of the TaggedTransactionQueue_validate_transaction runtime call.

validate_transaction
Validates a transaction by calling TaggedTransactionQueue_validate_transaction.

validate_transaction_runtime_parameters
Produces the input to pass to the TaggedTransactionQueue_validate_transaction runtime call.