smoldot::trie
Module branch_search

Allows searching for the closest branch node in a trie when only the storage trie nodes are known.

This module can be used in situations where only the list of trie nodes with a storage value is known, such as in the form of a BTreeMap, and allows finding the list of branch nodes.

Re-exports

pub use crate::trie::Nibble;

Structs

BranchTrieNodeKeyIter

Implementation of Iterator<Item = Nibble>. See BranchSearch::Found::branch_trie_node_key.

Config

Configuration for start_branch_search.

NextKey

In order to continue, the API user must indicate the trie node with a storage value that immediately follows a given one.

Enums

BranchSearch

Progress in the search algorithm.

Functions

start_branch_search

Start the search algorithm.