Module smoldot::trie::calculate_root

Freestanding function that calculates the root of a radix-16 Merkle-Patricia trie.

See the parent module documentation for an explanation of what the trie is.

Usage

Calling the root_merkle_value function creates a RootMerkleValueCalculation object which you have to drive to completion.

Example:

use std::collections::BTreeMap;
use smoldot::trie::calculate_root;

// In this example, the storage consists in a binary tree map.
let mut storage = BTreeMap::<Vec<u8>, Vec<u8>>::new();
storage.insert(b"foo".to_vec(), b"bar".to_vec());

let trie_root = {
    let mut calculation = calculate_root::root_merkle_value(None);
    loop {
        match calculation {
            calculate_root::RootMerkleValueCalculation::Finished { hash, .. } => break hash,
            calculate_root::RootMerkleValueCalculation::AllKeys(keys) => {
                calculation = keys.inject(storage.keys().map(|k| k.iter().cloned()));
            }
            calculate_root::RootMerkleValueCalculation::StorageValue(value_request) => {
                let key = value_request.key().collect::<Vec<u8>>();
                calculation = value_request.inject(storage.get(&key));
            }
        }
    }
};

assert_eq!(
    trie_root,
    [204, 86, 28, 213, 155, 206, 247, 145, 28, 169, 212, 146, 182, 159, 224, 82,
     116, 162, 143, 156, 19, 43, 183, 8, 41, 178, 204, 69, 41, 37, 224, 91]
);

You have the possibility to pass a CalculationCache to the calculation. This cache will be filled with intermediary calculations and can later be passed again to calculate the root in a more efficient way.

When using a cache, be careful to properly invalidate cache entries whenever you perform modifications on the trie associated to it.

Structs

AllKeys
Request to return the list of all the keys in the storage. Call AllKeys::inject to indicate this list.

CalculationCache
Cache containing intermediate calculation steps.

StorageValue
Request the value of the node with a specific key. Call StorageValue::inject to indicate the value.

Enums

RootMerkleValueCalculation
Current state of the RootMerkleValueCalculation and how to continue.

Functions

root_merkle_value
Start calculating the Merkle value of the root node.