smoldot::trie

Function nibbles_to_bytes_truncate

pub fn nibbles_to_bytes_truncate<I: Iterator<Item = Nibble>>(
    nibbles: I,
) -> impl Iterator<Item = u8>

Turns an iterator of nibbles into an iterator of bytes.

If the number of nibbles is uneven, the last nibble is truncated.

Examples

use smoldot::trie::{Nibble, nibbles_to_bytes_truncate};

let input = [Nibble::try_from(0x5).unwrap(), Nibble::try_from(0xa).unwrap()];
assert_eq!(nibbles_to_bytes_truncate(input.into_iter()).collect::<Vec<_>>(), &[0x5a]);

use smoldot::trie::{Nibble, nibbles_to_bytes_truncate};

let input = [Nibble::try_from(0x5).unwrap(), Nibble::try_from(0xa).unwrap(), Nibble::try_from(0x9).unwrap()];
assert_eq!(nibbles_to_bytes_truncate(input.into_iter()).collect::<Vec<_>>(), &[0x5a]);