smoldot::trie::proof_decode
Function decode_and_verify_proof

pub fn decode_and_verify_proof<T>(
    config: Config<T>,
) -> Result<DecodedTrieProof<T>, Error>
where
    T: AsRef<[u8]>,

Verifies whether a proof is correct and returns an object that allows examining its content.

The proof is then stored within the DecodedTrieProof.

Due to the generic nature of this function, the proof can be either a Vec<u8> or a &[u8].

Returns an error if the proof is invalid, or if the proof contains entries that are disconnected from the root node of the trie.