Struct smoldot::trie::proof_encode::ProofBuilder

pub struct ProofBuilder { /* private fields */ }

Prototype for a Merkle proof whose building is in progress.

Implementations

impl ProofBuilder

pub fn new() -> Self

Initializes a new empty proof builder.

This is equivalent to calling ProofBuilder::with_nodes_capacity with a value of 0.

pub fn with_nodes_capacity(capacity: usize) -> Self

Initializes a new empty proof builder.

Memory is allocated to store capacity nodes, in other words the number of nodes added using ProofBuilder::set_node_value.

pub fn set_node_value(
&mut self,
key: &[Nibble],
node_value: &[u8],
unhashed_storage_value: Option<&[u8]>
)

Inserts the node value of a given trie node into the builder.

Overwrites any previously-set value for this key.

The node_value is decoded in order for the proof builder to determine the hierarchy of the trie and know which node values are missing. If the node_value is invalid, this function panics. In order to avoid difficult-to-debug corner cases, this function also panics if it is no possible for node_value to be found at the given key due to a mismatch.

If the node_value contains a storage value as a hash, then a unhashed_storage_value can optionally be provided in order to provide the unhashed version of this value.

The validity of the node_value in the context of the other node values that have been stored in this proof builder isn’t verified. For example, a node value can indicate no children while the node value of a child has been added, or a node value can indicate a child with a specific hash, while the child in question has a different hash. This will lead to an invalid proof being generated.

Panic

Panics if node_value is not a valid node value. Panics if the partial key found in node_value doesn’t match the last elements of key. Panics in case node_value indicates no storage value, but unhashed_storage_value is Some.

pub fn missing_node_values(&self) -> impl Iterator<Item = &[Nibble]>

Returns a list of keys for which the node value must be known in order to be able to build the proof.

For each entry returned by this iterator, ProofBuilder::set_node_value must be called.

This function has a complexity of O(1) and thus can be called repeatedly.

pub fn trie_root_hash(&self) -> Option<[u8; 32]>

Returns the hash of the trie root node.

This function returns None if the proof is empty or if the trie root node is missing from the proof, in which case ProofBuilder::missing_node_values will return it.

In other words, if the proof is not empty and if ProofBuilder::missing_node_values returns an empty iterator, then you are guaranteed that this function returns Some.

This function has a complexity of O(1).

pub fn make_coherent(&mut self)

Modifies the node values that have been inserted in the proof builder in order to make the proof coherent, if necessary.

If all the node values that have been added through ProofBuilder::set_node_value come from a valid trie, then the proof is already coherent and there is no need to call this function. This function is necessary only in situations where the node values have been manually modified.

Calling this function when the node values aren’t coherent will modify the hash of the trie root that is found in the proof. Most of the time, the verification of a trie proof checks whether the hash of the trie root in the proof matches an expected value. When that is the case, then calling this function would produce a proof that is no longer accepted.

This function only updates the hashes of storage values and children. Storage values themselves are always left untouched. This function also updates the presence of children. If a node has been added using ProofBuilder::set_node_value but its parent indicates a lack of child, then calling this function will modify the parent in order to indicate the presence of a child. However, this function never removes children. If a node indicates that it has a child, but the child has not been added using ProofBuilder::set_node_value, it simply indicates that the child in question is (intentionally) not part of the proof, and the proof is still valid.

This function works even if ProofBuilder::missing_node_values returns a non-empty iterator, but will not be able to update everything. You are encouraged to call this function only after having added all missing node values;

pub fn build(self) -> impl Iterator<Item = impl AsRef<[u8]> + Clone>

Builds the Merkle proof.

This function returns an iterator of buffers. The actual Merkle proof consists in the concatenation of all the buffers.

This function will succeed if no entry at all has been inserted in the ProofBuilder.

This function will succeed even if ProofBuilder::missing_node_values returns a non-zero number of elements. However, the proof produced will then be invalid.

pub fn build_to_vec(self) -> Vec<u8>

Similar to ProofBuilder::build, but returns a Vec.

This is a convenience wrapper around ProofBuilder::build.