Function smoldot::trie::proof_node_codec::encode

pub fn encode(
    decoded: Decoded<'_, impl ExactSizeIterator<Item = Nibble> + Clone>
) -> impl Iterator<Item = impl AsRef<[u8]> + '_ + Clone> + Clone + '_

Encodes the components of a node value into the node value itself.

This function returns an iterator of buffers. The actual node value is the concatenation of these buffers put together.

Note: The returned iterator might contain a reference to the storage value and children values in the Decoded. By returning an iterator of buffers, we avoid copying these storage value and children values.

This encoding is independent of the trie version.