Module smoldot::trie::proof_verify

Verification of a trie proof.

A trie proof is a proof that a certain key in the trie has a certain storage value (or lacks a storage value). The proof can be verified by knowing only the Merkle value of the root node.

Details

Note: For reminder, the Merkle value of a node is the hash of its node value, or the node value directly if its length is smaller than 32 bytes.

A trie proof consists in a list of node values of nodes in the trie. For the proof to be valid, the hash of one of these node values must match the expected trie root node value. Since a node value contains the Merkle values of the children of the node, it is possible to iterate down the hierarchy of nodes until the one closest to the desired key is found.

Multiple proofs merged into one

Considering that a trie proof consists in a list of node values, it is possible to reduce the space occupied by multiple trie proofs built from the same trie by merging them into a single list and removing duplicate elements.

In order to support this use case, the verify_proof function intentionally doesn’t return an error if some elements in the proof are unused, as it might be that these elements are part of a different proof that has been merged with the one that is relevant.

Note: The main use case for merging multiple proofs into one is when a machine that has access to the storage of a block sends to a machine that doesn’t all the proofs corresponding to the storage entries necessary for a certain runtime call.

Structs

TrieNodeInfo
Information about a node of the trie.

TrieNodeInfoConfig
Configuration to pass to trie_node_info.

VerifyProofConfig
Configuration to pass to verify_proof.

Enums

Children
See TrieNodeInfo::children.

Error
Possible error returned by verify_proof

Functions

trie_node_info
Find information about the node whose key is requested by TrieNodeInfoConfig::requested_key.

verify_proof
Find the storage value of the requested key (as designated by VerifyProofConfig::requested_key).