Function smoldot::trie::proof_verify::verify_proof

pub fn verify_proof<'a, 'b>(
    config: VerifyProofConfig<'a, impl Iterator<Item = &'b [u8]> + Clone>
) -> Result<Option<&'b [u8]>, Error>

Find the storage value of the requested key (as designated by VerifyProofConfig::requested_key).

Returns an error if the proof couldn’t be verified. If the proof could be verified and the key has an associated storage value, Ok(Some(_)) is returned, containining that storage value. If the proof could be verified but the key does not have an associated storage value, Ok(None) is returned.

Note: This does not fully verify the correctness of the node values provided by proof. Only the minimum amount of information required is fetched from proof, and an error is returned if a problem happens during this process.