smoldot::trie
Module trie_node

Structs

Decoded

Decoded node value. Returned by decode or passed as parameter to encode.

DecodedPartialKey

Iterator to the nibbles of the partial key. See Decoded::partial_key.

MerkleValueOutput

Output of the calculation.

Enums

EncodeError

Error potentially returned by encode.

Error

Possible error returned by decode.

StorageValue

See Decoded::storage_value.

Functions

calculate_merkle_value

Calculates the Merkle value of the given node.

decode

Decodes a node value found in a proof into its components.

encode

Encodes the components of a node value into the node value itself.

encode_to_vec

Encodes the components of a node value into the node value itself.