Function smoldot::trie::trie_node::calculate_merkle_value

pub fn calculate_merkle_value(
    decoded: Decoded<'_, impl ExactSizeIterator<Item = Nibble> + Clone, impl AsRef<[u8]> + Clone>,
    hash_function: HashFunction,
    is_root_node: bool
) -> Result<MerkleValueOutput, EncodeError>

Calculates the Merkle value of the given node.

is_root_node must be true if the encoded node is the root node of the trie.

This is similar to encode, except that the encoding is then optionally hashed.

Hashing is performed if the encoded value is 32 bytes or more, or if is_root_node is true. This is the reason why is_root_node must be provided.