Struct smoldot::trie::trie_structure::StorageNodeAccess

pub struct StorageNodeAccess<'a, TUd> { /* fields omitted */ }

Access to a node within the TrieStructure that is known to have a storage value associated to it.

Implementations

impl<'a, TUd> StorageNodeAccess<'a, TUd>

pub fn node_index(&self) -> NodeIndex

Returns an opaque NodeIndex representing the node in the trie.

It can later be used to retrieve this same node using TrieStructure::node_by_index.

pub fn into_parent(self) -> Option<NodeAccess<'a, TUd>>

Returns the parent of this node, or None if this is the root node.

pub fn parent(&mut self) -> Option<NodeAccess<'_, TUd>>

Returns the parent of this node, or None if this is the root node.

pub fn into_first_child(self) -> Result<NodeAccess<'a, TUd>, Self>

Returns the first child of this node.

Returns back self if this node doesn’t have any children.

pub fn into_next_sibling(self) -> Result<NodeAccess<'a, TUd>, Self>

Returns the next sibling of this node.

Returns back self if this node is the last child of its parent.

pub fn child(&mut self, index: Nibble) -> Option<NodeAccess<'_, TUd>>

Returns the child of this node at the given index.

pub fn child_user_data(&self, index: Nibble) -> Option<&TUd>

Returns the user data of the child at the given index.

Note: This method exists because it accepts &self rather than &mut self. A cleaner alternative would be to split the NodeAccess struct into NodeAccessRef and NodeAccessMut, but that’s a lot of efforts compare to this single method.

pub fn into_child(self, index: Nibble) -> Result<NodeAccess<'a, TUd>, Self>

Returns the child of this node given the given index.

Returns back self if there is no such child at this index.

pub fn is_root_node(&self) -> bool

Returns true if this node is the root node of the trie.

pub fn full_key(&self) -> impl Iterator<Item = Nibble> + '_

Returns the full key of the node.

pub fn partial_key(&self) -> impl ExactSizeIterator<Item = Nibble> + '_

Returns the partial key of the node.

pub fn user_data(&mut self) -> &mut TUd

Returns the user data associated to this node.

pub fn remove(self) -> Remove<'a, TUd>

Removes the storage value and returns what this changes in the trie structure.