smoldot::verify::babe

Enum VerifyError

pub enum VerifyError {

    MissingSeal,
    MissingPreRuntimeDigest,
    ParentIsntBabeConsensus,
    SlotNumberNotIncreasing,
    UnexpectedEpochChangeLog,
    MissingEpochChangeLog,
    InvalidBabeParametersChange(BabeValidityError),
    InvalidAuthorityIndex,
    BadPublicKey,
    BadSignature,
    BadVrfProof,
    BadSecondarySlotAuthor,
    OverPrimaryClaimThreshold,
    ForbiddenSlotType,
    NextEpochStartSlotNumberOverflow,
    EpochIndexOverflow,
    InvalidChainConfiguration(InvalidChainConfiguration),

}

Failure to verify a block.

Variants

MissingSeal

The seal (containing the signature of the authority) is missing from the header.

MissingPreRuntimeDigest

No pre-runtime digest in the block header.

ParentIsntBabeConsensus

Parent block doesnโt contain any Babe information.

SlotNumberNotIncreasing

Slot number must be strictly increasing between a parent and its child.

UnexpectedEpochChangeLog

Block contains an epoch change digest log, but no epoch change is to be performed.

MissingEpochChangeLog

Block is the first block after a new epoch, but it is missing an epoch change digest log.

InvalidBabeParametersChange(BabeValidityError)

The header contains an epoch change that would put the Babe configuration in an non-sensical state.

InvalidAuthorityIndex

Authority index stored within block is out of range.

BadPublicKey

Public key used to for the signature is invalid.

BadSignature

Block header signature is invalid.

BadVrfProof

VRF proof in the block header is invalid.

BadSecondarySlotAuthor

Block is a secondary slot claim and its author is not the expected author.

OverPrimaryClaimThreshold

VRF output is over threshold required to claim the primary slot.

ForbiddenSlotType

Type of slot claim forbidden by current configuration.

NextEpochStartSlotNumberOverflow

Overflow when calculating the starting slot of the next epoch.

EpochIndexOverflow

Overflow when calculating the index of the next epoch.

InvalidChainConfiguration(InvalidChainConfiguration)

The configuration of the chain is invalid. It canโt be determined whether the block is valid or not.