Module smoldot::verify::header_body

Re-exports

pub use runtime_host::Nibble;

pub use runtime_host::TrieEntryVersion;

Structs

Config

Configuration for a block verification.

RuntimeCompilation

A new runtime must be compiled.

StorageClosestDescendantMerkleValue

Obtaining the Merkle value of the closest descendant of a trie node is required in order to continue.

StorageGet

Loading a storage value is required in order to continue.

StorageNextKey

Fetching the key that follows a given one is required in order to continue.

Success

Block successfully verified.

Enums

ConfigConsensus

Extra items of Config that are dependant on the consensus engine of the chain.

Error

Error that can happen during the verification.

SuccessConsensus

Extra items in Success relevant to the consensus engine.

Verify

Current state of the verification.

Functions

verify

Verifies whether a block is valid.