smoldot::author

Module build

Re-exports

pub use runtime::Nibble;

pub use runtime::StorageChanges;

pub use runtime::TrieEntryVersion;

Structs

ApplyExtrinsic

More transactions can be added.

AuthoringStart

Ready to start producing blocks.

AuthoringStartConfig

Configuration to pass when the actual block authoring is started.

ClosestDescendantMerkleValue

Obtaining the Merkle value of the closest descendant of a trie node is required in order to continue.

Config

Configuration for a block generation.

NextKey

Fetching the key that follows a given one in the parent storage is required in order to continue.

OffchainStorageSet

Setting the value of an offchain storage value is required.

Seal

Block has been produced and must now be sealed.

StorageGet

Loading a storage value from the parent storage is required in order to continue.

WaitSlot

Block production is idle, waiting for a slot.

Enums

Builder

Current state of the block building process.

BuilderAuthoring

Current state of the block building process.

ConfigConsensus

Extension to Config.

Error

Error that can happen during the block production.