smoldot::chain

Module async_tree

Performing asynchronous operations on blocks.

This module contains the AsyncTree data structure.

When a block is inserted in the data structure, it is added to the so-called “input tree” and its status is marked as pending. The data structure then starts, for each block marked as pending, an asynchronous operation (what this operation consists of is decided by the API user). Once an asynchronous operation is successful, the status of the block is switched to “finished”. The data structure then puts the blocks in the so-called “output tree”.

The output tree is consistent, meaning that if the asynchronous operation of a child finishes before the one of its parent, the child will be added to the output tree only after its parent has finished its operation. Similarly, if a block is finalized in the input tree, it only gets finalized in the output tree after all of its ancestors have all finished their asynchronous operations.

An example use case is: you insert block headers, then for each block you download its body, and thus obtain as output a tree of block headers and bodies.

Details

The AsyncTree data structure contains two trees of blocks: one input tree and one output tree. The output tree is a subset of the input tree.

Each of the two trees (input and output) has the following properties:

A finalized block.

A tree of non-finalized blocks that all descend from the finalized block.

A best block that can be either the finalized block or one of the non-finalized blocks.

Furthermore, each block has the following properties:

An opaque user data.

A status: pending, in progress, or finished. Once finished, an “asynchronous user data” is also attached to the block. All the blocks of the output tree are always in the “finished” state.

At initialization, both the input and output trees are initialized to the same finalized block (that is also the best block), and don’t have any non-finalized block.

Example

use smoldot::chain::async_tree;
use std::time::{Instant, Duration};

let mut tree = async_tree::AsyncTree::new(async_tree::Config {
    finalized_async_user_data: "hello",
    retry_after_failed: Duration::from_secs(5),
    blocks_capacity: 32,
});

// Insert a new best block, child of the finalized block.
// When doing so, we insert a "user data", a value opaque to the tree and that can be
// retreived later. Here we pass "my block".
let _my_block_index = tree.input_insert_block("my block", None, false, true);

// When calling `next_necessary_async_op`, the tree now generates a new asynchronous
// operation id.
let async_op_id = match tree.next_necessary_async_op(&Instant::now()) {
    async_tree::NextNecessaryAsyncOp::Ready(params) => {
        assert_eq!(params.block_index, _my_block_index);
        assert_eq!(tree[params.block_index], "my block");
        params.id
    }
    async_tree::NextNecessaryAsyncOp::NotReady { when: _ } => {
        // In this example, this variant can't be returned. In practice, however, you need
        // to call `next_necessary_async_op` again after `when`.
        panic!();
    }
};

// The user is now responsible for performing this asynchronous operation.
// When it is finished, call `async_op_finished`.
// Just like when inserting a new block, we insert another "user data" in all the blocks that
// have this asynchronous operation associated to them.
tree.async_op_finished(async_op_id, "world");

// You can now advance the best and finalized block of the tree. Calling this function tries
// to update the tree to match the best and finalized block of the input, except that only
// blocks whose asynchronous operation is finished are considered.
match tree.try_advance_output() {
    Some(async_tree::OutputUpdate::Block(block)) => {
        assert_eq!(block.index, _my_block_index);
        assert!(block.is_new_best);
    }
    _ => unreachable!() // Unreachable in this example.
}

Re-exports

pub use fork_tree::NodeIndex;

Structs

AsyncOpId

Identifier for an asynchronous operation in the AsyncTree.

AsyncOpParams

Information about an operation that must be started.

AsyncTree

See the module-level documentation.

Config

Configuration for AsyncTree::new.

InputIterItem

See AsyncTree::input_output_iter_unordered and AsyncTree::input_output_iter_ancestry_order.

OutputUpdateBlock

See OutputUpdate.

Enums

NextNecessaryAsyncOp

OutputUpdate

See AsyncTree::try_advance_output.