Struct smoldot::chain::async_tree::AsyncTree

pub struct AsyncTree<TNow, TBl, TAsync> { /* private fields */ }

Implementations

impl<TNow, TBl, TAsync> AsyncTree<TNow, TBl, TAsync>where
TNow: Clone + Add<Duration, Output = TNow> + Ord,
TAsync: Clone,

pub fn new(config: Config<TAsync>) -> Self

Returns a new empty AsyncTree.

pub fn num_input_non_finalized_blocks(&self) -> usize

Returns the number of non-finalized blocks.

This is equal to the number of items yielded by AsyncTree::input_iter_unordered.

pub fn map_async_op_user_data<TAsync2>(
self,
map: impl FnMut(TAsync) -> TAsync2
) -> AsyncTree<TNow, TBl, TAsync2>

Replaces all asynchronous operation user data with new values.

The returned tree keeps the same NodeIndexes as self.

pub fn best_block_index(&self) -> Option<(NodeIndex, &TAsync)>

Returns the NodeIndex of the current “output” best block.

Returns None if there is no best block. In terms of logic, this means that the best block is the finalized block, which is out of scope of this data structure.

pub fn block_user_data(&self, node_index: NodeIndex) -> &TBl

Returns the user data associated to the given block.

Panic

Panics if the NodeIndex is invalid.

pub fn block_user_data_mut(&mut self, node_index: NodeIndex) -> &mut TBl

Returns the user data associated to the given block.

Panic

Panics if the NodeIndex is invalid.

pub fn finalized_async_user_data(&self) -> &TAsync

Returns the outcome of the asynchronous operation for the finalized block.

This is the value that was passed at initialization, or is updated after AsyncTree::try_advance_output has returned OutputUpdate::Finalized.

pub fn block_async_user_data(&self, node_index: NodeIndex) -> Option<&TAsync>

Returns the asynchronous operation user data associated to the given block.

Returns None if the asynchronous operation isn’t complete for this block yet.

Panic

Panics if the NodeIndex is invalid.

pub fn block_async_user_data_mut(
&mut self,
node_index: NodeIndex
) -> Option<&mut TAsync>

Returns the asynchronous operation user data associated to the given block.

Returns None if the asynchronous operation isn’t complete for this block yet.

Panic

Panics if the NodeIndex is invalid.

pub fn parent(&self, node: NodeIndex) -> Option<NodeIndex>

Returns the parent of the given node. Returns None if the node doesn’t have any parent, meaning that its parent is the finalized node.

Panic

Panics if the NodeIndex is invalid.

pub fn ancestors(&self, node: NodeIndex) -> impl Iterator<Item = NodeIndex> + '_

Returns the ancestors of the given node. The iterator stops when it reaches the finalized block. The iterator is empty if the parent of the node is the finalized block.

Panic

Panics if the NodeIndex is invalid.

pub fn children(
&self,
node: Option<NodeIndex>
) -> impl Iterator<Item = NodeIndex> + '_

Returns the list of children that have the given node as parent.

Panic

Panics if the NodeIndex is invalid.

pub fn input_iter_ancestry_order(
&self
) -> impl Iterator<Item = InputIterItem<'_, TBl, TAsync>> + '_

Returns the list of all non-finalized blocks that have been inserted.

Similar to AsyncTree::input_iter_unordered, except that the returned items are guaranteed to be in an order in which the parents are found before their children.

pub fn input_iter_unordered(
&self
) -> impl Iterator<Item = InputIterItem<'_, TBl, TAsync>> + '_

Returns the list of all non-finalized blocks that have been inserted, in no particular order.

pub fn async_op_blocks(
&self,
async_op_id: AsyncOpId
) -> impl Iterator<Item = &TBl>

Returns the blocks targeted by this asynchronous operation.

pub fn async_op_finished(
&mut self,
async_op_id: AsyncOpId,
user_data: TAsync
) -> Vec<NodeIndex>where
TAsync: Clone,

Injects into the state of the data structure a completed operation.

This “destroys” the AsyncOpId.

Returns the list of blocks whose state was affected by this asynchronous operation. This can be zero blocks, or be more than one block if blocks were inserted with same_async_op_as_parent as true.

Panic

Panics if the AsyncOpId is invalid.

pub fn async_op_failure(&mut self, async_op_id: AsyncOpId, now: &TNow)

Injects into the state of the state machine a failed operation.

This same operation will not be repeated for the next few seconds. Thanks to this, it is possible to immediately call this function in response to a new necessary operation without worrying about loops.

This “destroys” the AsyncOpId.

Panic

Panics if the AsyncOpId is invalid.

pub fn next_necessary_async_op(
&mut self,
now: &TNow
) -> NextNecessaryAsyncOp<'_, TNow, TBl>

Examines the state of self and, if a block’s asynchronous operation should be started, changes the state of the block to “in progress” and returns the parameters of the operation.

The order in which operations are started is:

The input finalized block.

The input best block.

Any other block.

pub fn input_insert_block(
&mut self,
block: TBl,
parent_index: Option<NodeIndex>,
same_async_op_as_parent: bool,
is_new_best: bool
) -> NodeIndex

Inserts a new block in the state machine.

If same_async_op_as_parent is true, then the asynchronous operation user data is shared with the parent of the block. This “sharing” is done by emitting only one asynchronous operation for both blocks, and/or by cloning the TAsyncOp.

Panic

Panics if parent_index is an invalid node.

pub fn input_set_best_block(&mut self, new_best_block: NodeIndex)

Updates the state machine to take into account that the best block of the input has been modified.

Panic

Panics if new_best_block isn’t a valid node. Panics if new_best_block isn’t equal or a descendant of the input finalized block.

pub fn input_finalize(
&mut self,
node_to_finalize: NodeIndex,
new_best_block: NodeIndex
)

Updates the state machine to take into account that the input of blocks has finalized the given block.

new_best_block is the best block after the finalization.

Note: Finalizing a block might have to modify the current best block if the block being finalized isn’t an ancestor of the current best block.

Panic

Panics if node_to_finalize or new_best_block aren’t valid nodes. Panics if new_best_block is not a descendant of node_to_finalize.

pub fn try_advance_output(&mut self) -> Option<OutputUpdate<'_, TBl, TAsync>>

Tries to update the output blocks to follow the input.

Should be called after inserting a new block, finalizing a block, or when an operation is finished.

Returns None if the state machine doesn’t have any update. This method should be called repeatedly until it returns None. Each call can perform an additional update.