smoldot::chain

Module blocks_tree

Finalized block header, plus tree of authenticated non-finalized block headers.

This module provides the NonFinalizedTree type. This type is a data structure containing a valid tree of block headers, plus the state necessary to verify new blocks with the intent to add them to that tree. Each block header additionally holds a user-chosen opaque data.

The state in the NonFinalizedTree consists of:

One “latest finalized” block and various information about its ancestors, akin to a chain_information::ChainInformation.

Zero or more blocks that descend from that latest finalized block.

The latest finalized block is a block that is guaranteed to never be reverted. While it can always be set to the genesis block of the chain, it is preferable, in order to reduce memory utilization, to maintain it to a block that is as high as possible in the chain.

Note: While mechanisms such as GrandPa provide a network-wide way to designate a block as final, the concept of GrandPa-provided finality doesn’t necessarily have to match the concept of finality in the NonFinalizedTree. For example, an API user might decide to optimistically assume that the block whose number is highest_block - 5 is automatically finalized, and fall back to rebuilding a new NonFinalizedTree if that assumption turns out to not be true. The finalized block in the NonFinalizedTree only represents a block that the NonFinalizedTree itself cannot remove, not a block that cannot be removed in the absolute.

A block can be added to the chain by calling NonFinalizedTree::verify_header then NonFinalizedTree::insert_verified_header. As explained in details in the verify module, verifying the header only verifies the authenticity of a block and not its correctness. Additionally verifying the body of the block provides the strongest guarantee, but is the responsibility of the API user and is out of the scope of this module.

Note: There typically exists two kinds of clients: full and light. Full clients store the state of the storage, while light clients don’t. For this reason, light clients can only verify the header of new blocks. Both full and light clients should wait for a block to be finalized if they want to be certain that it will forever remain part of the chain.

Additionally, a NonFinalizedTree::verify_justification method is provided in order to verify the correctness of a justification.

Structs

Config

Configuration for the NonFinalizedTree.

FinalityApply

Returned by NonFinalizedTree::verify_justification and NonFinalizedTree::verify_grandpa_commit_message on success.

NonFinalizedTree

Holds state about the current state of the chain for the purpose of verifying headers.

RemovedBlock

Block removed from the NonFinalizedTree by a SetFinalizedBlockIter.

SetFinalizedBlockIter

Iterator producing the newly-finalized blocks removed from the state when the finalized block is updated.

VerifiedHeader

Successfully-verified block header that can be inserted into the chain.

Enums

CommitVerifyError

Error that can happen when verifying a Grandpa commit.

FinalityVerifyError

Error that can happen when verifying a proof of finality.

HeaderVerifyError

Error that can happen when verifying a block header.

HeaderVerifySuccess

See NonFinalizedTree::verify_header.

JustificationVerifyError

Error that can happen when verifying a justification.

RemovedBlockType

Reason why a block was removed from the NonFinalizedTree by a SetFinalizedBlockIter.

SetFinalizedError

Error that can happen when setting the finalized block.