smoldot::chain::blocks_tree

Struct NonFinalizedTree

pub struct NonFinalizedTree<T> { /* private fields */ }

Holds state about the current state of the chain for the purpose of verifying headers.

Implementations

impl<T> NonFinalizedTree<T>

pub fn finality_checkpoints(&self) -> impl Iterator<Item = (u64, &[u8; 32])>

Returns a list of blocks (by their height and hash) that need to be finalized before any of their descendants can be finalized.

In other words, blocks in the NonFinalizedTree can be immediately finalized by call to NonFinalizedTree::verify_justification or NonFinalizedTree::verify_grandpa_commit_message, unless they descend from any of the blocks returned by this function, in which case that block must be finalized beforehand.

pub fn verify_justification( &mut self, consensus_engine_id: [u8; 4], scale_encoded_justification: &[u8], randomness_seed: [u8; 32], ) -> Result<FinalityApply<'_, T>, JustificationVerifyError>

Verifies the given justification.

The verification is performed in the context of the chain. In particular, the verification will fail if the target block isn’t already in the chain.

If the verification succeeds, a FinalityApply object will be returned which can be used to apply the finalization.

A randomness seed must be provided and will be used during the verification. Note that the verification is nonetheless deterministic.

pub fn verify_grandpa_commit_message( &mut self, scale_encoded_commit: &[u8], randomness_seed: [u8; 32], ) -> Result<FinalityApply<'_, T>, CommitVerifyError>

Verifies the given Grandpa commit message.

The verification is performed in the context of the chain. In particular, the verification will fail if the target block isn’t already in the chain or if one of the voted blocks is unknown locally.

If the verification succeeds, a FinalityApply object will be returned which can be used to apply the finalization.

A randomness seed must be provided and will be used during the verification. Note that the verification is nonetheless deterministic.

pub fn set_finalized_block( &mut self, block_hash: &[u8; 32], ) -> Result<SetFinalizedBlockIter<'_, T>, SetFinalizedError>

Sets the latest known finalized block. Trying to verify a block that isn’t a descendant of that block will fail.

The block must have been passed to NonFinalizedTree::verify_header.

Returns an iterator containing the now-finalized blocks and the pruned blocks in “reverse hierarchical order”. Each block is yielded by the iterator before its parent.

Note: This function returns blocks in this order, because any other ordering would incur a performance cost. While returning blocks in hierarchical order would often be more convenient, the overhead of doing so is moved to the user.

The pruning is completely performed, even if the iterator is dropped eagerly.

If necessary, the current best block will be updated to be a descendant of the newly-finalized block.

impl<T> NonFinalizedTree<T>

pub fn verify_header( &self, scale_encoded_header: Vec<u8>, now_from_unix_epoch: Duration, ) -> Result<HeaderVerifySuccess, HeaderVerifyError>

Verifies the given block header.

The verification is performed in the context of the chain. In particular, the verification will fail if the parent block isn’t already in the chain.

If the verification succeeds, an VerifiedHeader object might be returned which can be used to then insert the block in the chain using NonFinalizedTree::insert_verified_header.

Must be passed the current UNIX time in order to verify that the block doesn’t pretend to come from the future.

pub fn insert_verified_header( &mut self, verified_header: VerifiedHeader, user_data: T, )

Insert a header that has already been verified to be valid.

Panic

Panics if the parent of the block isn’t in the tree. The presence of the parent is verified when the block is verified, so this can only happen if you remove the parent after having verified the block but before calling this function.

impl<T> NonFinalizedTree<T>

pub fn new(config: Config) -> Self

Initializes a new queue.

Panic

Panics if the chain information is incorrect.

pub fn clear(&mut self)

Removes all non-finalized blocks from the tree.

pub fn is_empty(&self) -> bool

Returns true if there isn’t any non-finalized block in the chain.

pub fn len(&self) -> usize

Returns the number of non-finalized blocks in the chain.

pub fn iter_unordered(&self) -> impl Iterator<Item = HeaderRef<'_>>

Returns the header of all known non-finalized blocks in the chain without any specific order.

pub fn iter_ancestry_order(&self) -> impl Iterator<Item = HeaderRef<'_>>

Returns the header of all known non-finalized blocks in the chain.

The returned items are guaranteed to be in an order in which the parents are found before their children.

pub fn reserve(&mut self, additional: usize)

Reserves additional capacity for at least additional new blocks without allocating.

pub fn shrink_to_fit(&mut self)

Shrink the capacity of the chain as much as possible.

pub fn block_number_bytes(&self) -> usize

Returns the value that was initially passed in Config::block_number_bytes.

pub fn as_chain_information(&self) -> ValidChainInformationRef<'_>

Builds a chain_information::ChainInformationRef struct that might later be used to build a new NonFinalizedTree.

pub fn finalized_block_header(&self) -> &[u8]

Returns the header of the latest finalized block.

pub fn finalized_block_hash(&self) -> &[u8; 32]

Returns the hash of the latest finalized block.

pub fn finalized_block_height(&self) -> u64

Returns the height of the latest finalized block.

pub fn best_block_header(&self) -> &[u8]

Returns the header of the best block.

pub fn best_block_hash(&self) -> &[u8; 32]

Returns the hash of the best block.

pub fn best_block_height(&self) -> u64

Returns the height of the best block.

pub fn best_block_consensus(&self) -> ChainInformationConsensusRef<'_>

Returns consensus information about the current best block of the chain.

pub fn contains_non_finalized_block(&self, hash: &[u8; 32]) -> bool

Returns true if the block with the given hash is in the NonFinalizedTree.

pub fn non_finalized_block_user_data(&self, hash: &[u8; 32]) -> Option<&T>

Gives access to the user data of a block stored by the NonFinalizedTree, identified by its hash.

Returns None if the block can’t be found.

pub fn non_finalized_block_user_data_mut( &mut self, hash: &[u8; 32], ) -> Option<&mut T>

Gives access to the user data of a block stored by the NonFinalizedTree, identified by its hash.

Returns None if the block can’t be found.

pub fn non_finalized_block_header(&self, hash: &[u8; 32]) -> Option<&[u8]>

Returns the SCALE-encoded header of a block stored by the NonFinalizedTree, identified by its hash.

Returns None if the block can’t be found.