Module smoldot::chain::chain_information::build

Build the chain information of a chain given its runtime.

This module contains the ChainInformationBuild struct, a state machine that drives the process of building the chain information of a certain finalized point of a chain.

Re-exports

pub use runtime_call::Nibble;

pub use runtime_call::TrieEntryVersion;

Structs

ClosestDescendantMerkleValue

Obtaining the Merkle value of the closest descendant of a trie node is required in order to continue.

Config

Configuration to provide to ChainInformationBuild::new.

NextKey

Fetching the key that follows a given one is required in order to continue.

StorageGet

Loading a storage value is required in order to continue.

Enums

ChainInformationBuild

Current state of the operation.

ConfigFinalizedBlockHeader

See Config::finalized_block_header.

Error

Problem encountered during the chain building process.

InProgress

Chain information building is still in progress.

RuntimeCall

Function call to perform or being performed.