Module smoldot::chain_spec

Substrate chain configuration.

A chain spec (short for chain specification) is the description of everything that is required for the client to successfully interact with a certain blockchain. For example, the Polkadot chain spec contains all the constants that are needed in order to successfully interact with Polkadot.

Chain specs contain, notably:

The state of the genesis block. In other words, the initial content of the database. This includes the Wasm runtime code of the genesis block.

The list of bootstrap nodes. These are the IP addresses of the machines we need to connect to.

The default telemetry endpoints, to which we should send telemetry information to.

The name of the network protocol, in order to avoid accidentally connecting to a different network.

Multiple other miscellaneous information.

Structs

ChainSpec

A configuration of a chain. Can be used to build a genesis block.

GenesisStorageItems

See GenesisStorage.

LightSyncState

ParseError

Error that can happen when parsing a chain spec JSON.

Enums

Bootnode

See ChainSpec::boot_nodes.

FromGenesisStorageError

Error when building the chain information from the genesis storage.

GenesisStorage

See ChainSpec::genesis_storage.