Module smoldot::database::finalized_serialize

Serializing/deserializing a chain_information::ChainInformation.

This module contains the encode_chain_storage and decode_chain functions that can turn a chain_information::ChainInformation into a string and back, with optionally the state of the finalized block. With no finalized block storage, the string is expected to be in the order of magnitude of a few dozens kilobytes.

The string format designed to be stable even if the structure of chain_information::ChainInformation is later modified.

This feature is expected to be used for example by light clients in order to easily (but inefficiently) store the state of the finalized chain somewhere and later reload it.

Structs

CorruptedError

Opaque error indicating a corruption in the data stored in the local storage.

Functions

decode_chain

Deserializes the information about the chain.

encode_chain

Serializes the given chain information as a JSON string.

encode_chain_storage

Serializes the given chain information and finalized block storage as a string.