smoldot::database::full_sqlite

Struct SqliteFullDatabase

pub struct SqliteFullDatabase { /* private fields */ }

An open database. Holds file descriptors.

Implementations

impl SqliteFullDatabase

pub fn best_block_hash(&self) -> Result<[u8; 32], CorruptedError>

Returns the hash of the block in the database whose storage is currently accessible.

pub fn finalized_block_hash(&self) -> Result<[u8; 32], CorruptedError>

Returns the hash of the finalized block in the database.

pub fn block_scale_encoded_header( &self, block_hash: &[u8; 32], ) -> Result<Option<Vec<u8>>, CorruptedError>

Returns the SCALE-encoded header of the given block, or None if the block is unknown.

Note: If this method is called twice times in a row with the same block hash, it is possible for the first time to return Some and the second time to return None, in case the block has since been removed from the database.

pub fn block_parent( &self, block_hash: &[u8; 32], ) -> Result<Option<[u8; 32]>, CorruptedError>

Returns the hash of the parent of the given block, or None if the block is unknown.

Note: If this method is called twice times in a row with the same block hash, it is possible for the first time to return Some and the second time to return None, in case the block has since been removed from the database.

pub fn block_extrinsics( &self, block_hash: &[u8; 32], ) -> Result<Option<impl ExactSizeIterator<Item = Vec<u8>>>, CorruptedError>

Returns the list of extrinsics of the given block, or None if the block is unknown.

Note: The list of extrinsics of a block is also known as its body.

Note: If this method is called twice times in a row with the same block hash, it is possible for the first time to return Some and the second time to return None, in case the block has since been removed from the database.

pub fn block_hash_by_number( &self, block_number: u64, ) -> Result<impl ExactSizeIterator<Item = [u8; 32]>, CorruptedError>

Returns the hashes of the blocks given a block number.

pub fn best_block_hash_by_number( &self, block_number: u64, ) -> Result<Option<[u8; 32]>, CorruptedError>

Returns the hash of the block of the best chain given a block number.

pub fn to_chain_information( &self, finalized_block_hash: &[u8; 32], ) -> Result<ValidChainInformation, StorageAccessError>

Returns a chain_information::ChainInformation struct containing the information about the current finalized state of the chain.

This method is relatively expensive and should preferably not be called repeatedly.

In order to avoid race conditions, the known finalized block hash must be passed as parameter. If the finalized block in the database doesn’t match the hash passed as parameter, most likely because it has been updated in a parallel thread, a StorageAccessError::IncompleteStorage error is returned.

pub fn insert<'a>( &self, scale_encoded_header: &[u8], is_new_best: bool, body: impl ExactSizeIterator<Item = impl AsRef<[u8]>>, ) -> Result<(), InsertError>

Insert a new block in the database.

Must pass the header and body of the block.

Blocks must be inserted in the correct order. An error is returned if the parent of the newly-inserted block isn’t present in the database.

Note: It is not necessary for the newly-inserted block to be a descendant of the finalized block, unless is_new_best is true.

pub fn insert_trie_nodes<'a>( &self, new_trie_nodes: impl Iterator<Item = InsertTrieNode<'a>>, trie_entries_version: u8, ) -> Result<(), CorruptedError>

pub fn finalized_and_above_missing_trie_nodes_unordered( &self, ) -> Result<Vec<MissingTrieNode>, CorruptedError>

Returns a list of trie nodes that are missing from the database and that belong to the state of a block whose number is superior or equal to the finalized block.

The ordering of the returned trie nodes is unspecified.

Note: This function call is relatively expensive, and the API user is expected to cache the return value.

pub fn set_finalized( &self, new_finalized_block_hash: &[u8; 32], ) -> Result<(), SetFinalizedError>

Changes the finalized block to the given one.

The block must have been previously inserted using SqliteFullDatabase::insert, otherwise an error is returned.

Blocks are expected to be valid in context of the chain. Inserting an invalid block can result in the database being corrupted.

The block must be a descendant of the current finalized block. Reverting finalization is forbidden, as the database intentionally discards some information when finality is applied.

Note: This function doesn’t remove any block from the database but simply moves the finalized block “cursor”.

pub fn purge_finality_orphans(&self) -> Result<(), CorruptedError>

Removes from the database all blocks that aren’t a descendant of the current finalized block.

pub fn block_storage_get( &self, block_hash: &[u8; 32], parent_tries_paths_nibbles: impl Iterator<Item = impl Iterator<Item = u8>>, key_nibbles: impl Iterator<Item = u8>, ) -> Result<Option<(Vec<u8>, u8)>, StorageAccessError>

Returns the value associated with a node of the trie of the given block.

parent_tries_paths_nibbles is a list of keys to follow in order to find the root of the trie into which key_nibbles should be searched.

Beware that both parent_tries_paths_nibbles and key_nibbles must yield nibbles, in other words values strictly inferior to 16.

Returns an error if the block or its storage can’t be found in the database.

Panic

Panics if any of the values yielded by parent_tries_paths_nibbles or key_nibbles is superior or equal to 16.

pub fn block_storage_next_key( &self, block_hash: &[u8; 32], parent_tries_paths_nibbles: impl Iterator<Item = impl Iterator<Item = u8>>, key_nibbles: impl Iterator<Item = u8>, prefix_nibbles: impl Iterator<Item = u8>, branch_nodes: bool, ) -> Result<Option<Vec<u8>>, StorageAccessError>

Returns the key in the storage that immediately follows or is equal to the key passed as parameter in the storage of the block.

key_nibbles must be an iterator to the nibbles of the key.

prefix_nibbles must be an iterator to nibbles. If the result of the function wouldn’t start with this specific list of bytes, None is returned.

parent_tries_paths_nibbles is a list of keys to follow in order to find the root of the trie into which key_nibbles should be searched.

Returns None if parent_tries_paths_nibbles didn’t lead to any trie, or if there is no next key.

The key is returned in the same format as key_nibbles.

If branch_nodes is false, then branch nodes (i.e. nodes with no value associated to them) are ignored during the search.

Note: Contrary to many other similar functions in smoldot, there is no or_equal parameter to this function. Instead, or_equal is implicitly true, and a value of false can be easily emulated by appending a 0 at the end of key_nibbles.

Panics

Panics if any of the values yielded by parent_tries_paths_nibbles, key_nibbles, or prefix_nibbles is superior or equal to 16.

pub fn block_storage_closest_descendant_merkle_value( &self, block_hash: &[u8; 32], parent_tries_paths_nibbles: impl Iterator<Item = impl Iterator<Item = u8>>, key_nibbles: impl Iterator<Item = u8>, ) -> Result<Option<Vec<u8>>, StorageAccessError>

Returns the Merkle value of the trie node in the storage that is the closest descendant of the provided key.

key_nibbles must be an iterator to the nibbles of the key.

parent_tries_paths_nibbles is a list of keys to follow in order to find the root of the trie into which key_nibbles should be searched.

Returns None if parent_tries_paths_nibbles didn’t lead to any trie, or if there is no such descendant.

Panics

Panics if any of the values yielded by parent_tries_paths_nibbles or key_nibbles is superior or equal to 16.

pub fn reset<'a>( &self, finalized_block_header: &[u8], finalized_block_body: impl ExactSizeIterator<Item = &'a [u8]>, finalized_block_justification: Option<Vec<u8>>, ) -> Result<(), CorruptedError>

Inserts a block in the database and sets it as the finalized block.

The parent of the block doesn’t need to be present in the database.

If the block is already in the database, it is replaced by the one provided.