Struct smoldot::database::full_sqlite::SqliteFullDatabase

pub struct SqliteFullDatabase { /* private fields */ }

An open database. Holds file descriptors.

Implementations

impl SqliteFullDatabase

pub fn best_block_hash(&self) -> Result<[u8; 32], AccessError>

Returns the hash of the block in the database whose storage is currently accessible.

pub fn finalized_block_hash(&self) -> Result<[u8; 32], AccessError>

Returns the hash of the finalized block in the database.

pub fn block_scale_encoded_header( &self, block_hash: &[u8; 32] ) -> Result<Option<Vec<u8>>, AccessError>

Returns the SCALE-encoded header of the given block, or None if the block is unknown.

Note: If this method is called twice times in a row with the same block hash, it is possible for the first time to return Some and the second time to return None, in case the block has since been removed from the database.

pub fn block_extrinsics( &self, block_hash: &[u8; 32] ) -> Result<Option<impl ExactSizeIterator<Item = Vec<u8>>>, AccessError>

Returns the list of extrinsics of the given block, or None if the block is unknown.

Note: The list of extrinsics of a block is also known as its body.

Note: If this method is called twice times in a row with the same block hash, it is possible for the first time to return Some and the second time to return None, in case the block has since been removed from the database.

pub fn block_hash_by_number( &self, block_number: u64 ) -> Result<impl ExactSizeIterator<Item = [u8; 32]>, AccessError>

Returns the hashes of the blocks given a block number.

pub fn to_chain_information( &self, finalized_block_hash: &[u8; 32] ) -> Result<ValidChainInformation, StorageAccessError>

Returns a chain_information::ChainInformation struct containing the information about the current finalized state of the chain.

This method is relatively expensive and should preferably not be called repeatedly.

In order to avoid race conditions, the known finalized block hash must be passed as parameter. If the finalized block in the database doesn’t match the hash passed as parameter, most likely because it has been updated in a parallel thread, a StorageAccessError::Pruned error is returned.

pub fn insert( &self, scale_encoded_header: &[u8], is_new_best: bool, body: impl ExactSizeIterator<Item = impl AsRef<[u8]>>, storage_main_trie_changes: impl Iterator<Item = (impl AsRef<[u8]>, Option<impl AsRef<[u8]>>)> + Clone, trie_entries_version: u8 ) -> Result<(), InsertError>

Insert a new block in the database.

Must pass the header and body of the block, and the changes to the storage that this block performs relative to its parent.

Blocks must be inserted in the correct order. An error is returned if the parent of the newly-inserted block isn’t present in the database.

pub fn set_finalized( &self, new_finalized_block_hash: &[u8; 32] ) -> Result<(), SetFinalizedError>

Changes the finalized block to the given one.

The block must have been previously inserted using SqliteFullDatabase::insert, otherwise an error is returned.

Blocks are expected to be valid in context of the chain. Inserting an invalid block can result in the database being corrupted.

The block must be a descendant of the current finalized block. Reverting finalization is forbidden, as the database intentionally discards some information when finality is applied.

pub fn block_storage_main_trie_get( &self, block_hash: &[u8; 32], key: &[u8] ) -> Result<Option<(Vec<u8>, u8)>, StorageAccessError>

Returns the value associated to a key in the storage of either the latest finalized block or a non-finalized block, and the trie entry version.

pub fn block_storage_main_trie_next_key( &self, block_hash: &[u8; 32], key: &[u8], or_equal: bool ) -> Result<Option<Vec<u8>>, StorageAccessError>

Returns the key in the storage that immediately follows the key passed as parameter in the storage of either the latest finalized block or a non-finalized block.

If or_equal, then key itself is returned if it corresponds to a database entry.