Module smoldot::executor::host

Wasm virtual machine specific to the Substrate/Polkadot Runtime Environment.

Contrary to VirtualMachine, this code is not just a generic Wasm virtual machine, but is aware of the Substrate/Polkadot runtime environment. The host functions that the Wasm code calls are automatically resolved and either handled or notified to the user of this module.

Any host function that requires pure CPU computations (for example building or verifying a cryptographic signature) is directly handled by the code in this module. Other host functions (for example accessing the state or printing a message) are instead handled by interrupting the virtual machine and waiting for the user of this module to handle the call.

Note: The ext_offchain_random_seed_version_1 and ext_offchain_timestamp_version_1 functions, which requires the host to respectively produce a random seed and return the current time, must also be handled by the user. While these functions could theoretically be handled directly by this module, it might be useful for testing purposes to have the possibility to return a deterministic value.

Contrary to most programs, runtime code doesn’t have a singe main or start function. Instead, it exposes several entry points. Which one to call indicates which action it has to perform. Not all entry points are necessarily available on all runtimes.

Runtime requirements

See the documentation of the vm module for details about the requirements a runtime must adhere to.

In addition to the requirements described there, the WebAssembly runtime code can also be zstandard-compressed and must also export a global symbol named __heap_base. More details below.

Zstandard compression

The runtime code passed as parameter to HostVmPrototype::new can be compressed using the zstd algorithm.

If the code starts with the magic bytes [82, 188, 83, 118, 70, 219, 142, 5], then it is assumed that the rest of the data is a zstandard-compressed WebAssembly module.

Runtime version

Wasm files can contain so-called custom sections. A runtime can contain two custom sections whose names are "runtime_version" and "runtime_apis", in which case they must contain a so-called runtime version.

The runtime version contains important field that identifies a runtime.

If no "runtime_version" and "runtime_apis" custom sections can be found, the Core_version entry point is used as a fallback in order to obtain the runtime version. This fallback mechanism is maintained for backwards compatibility purposes, but is considered deprecated.

Memory allocations

One of the instructions available in WebAssembly code is the memory.grow instruction, which allows increasing the size of the memory.

WebAssembly code is normally intended to perform its own heap-management logic internally, and use the memory.grow instruction if more memory is needed.

In order to minimize the size of the runtime binary, and in order to accommodate for the API of the host functions that return a buffer of variable length, the Substrate/Polkadot runtimes, however, do not perform their heap management internally. Instead, they use the ext_allocator_malloc_version_1 and ext_allocator_free_version_1 host functions for this purpose. Calling memory.grow is forbidden.

The runtime code must export a global symbol named __heap_base of type i32. Any memory whose offset is below the value of __heap_base can be used at will by the program, while any memory above __heap_base but below __heap_base + heap_pages (where heap_pages is the value passed as parameter to HostVmPrototype::new) is available for use by the implementation of ext_allocator_malloc_version_1.

Entry points

All entry points that can be called from the host (using, for example, HostVmPrototype::run) have the same signature:

(func $runtime_entry(param $data i32) (param $len i32) (result i64))

In order to call into the runtime, one must write a buffer of data containing the input parameters into the Wasm virtual machine’s memory, then pass a pointer and length of this buffer as the parameters of the entry point.

The function returns a 64 bits number. The 32 less significant bits represent a pointer to the Wasm virtual machine’s memory, and the 32 most significant bits a length. This pointer and length designate a buffer containing the actual return value.

Host functions

The list of host functions available to the runtime is long and isn’t documented here. See the official specification for details.

Usage

The first step is to create a HostVmPrototype object from the WebAssembly code. Creating this object performs some initial steps, such as parsing and compiling the WebAssembly code. You are encouraged to maintain a cache of HostVmPrototype objects (one instance per WebAssembly byte code) in order to avoid performing these operations too often.

To start calling the runtime, create a HostVm by calling HostVmPrototype::run.

While the Wasm runtime code has side-effects (such as storing values in the storage), the HostVm itself is a pure state machine with no side effects.

At any given point, you can examine the HostVm in order to know in which state the execution currently is. In case of a HostVm::ReadyToRun (which initially is the case when you create the HostVm), you can execute the Wasm code by calling ReadyToRun::run. No background thread of any kind is used, and calling ReadyToRun::run directly performs the execution of the Wasm code. If you need parallelism, you are encouraged to spawn a background thread yourself and call this function from there. ReadyToRun::run tries to make the execution progress as much as possible, and returns the new state of the virtual machine once that is done.

If the runtime has finished, or has crashed, or wants to perform an operation with side effects, then the HostVm determines what to do next. For example, for HostVm::ExternalStorageGet, you must load a value from the storage and pass it back by calling ExternalStorageGet::resume.

The Wasm execution is fully deterministic, and the outcome of the execution only depends on the inputs. There is, for example, no implicit injection of randomness or of the current time.

Example

use smoldot::executor::host::{Config, HeapPages, HostVm, HostVmPrototype};

// Start executing a function on the runtime.
let mut vm: HostVm = {
    let prototype = HostVmPrototype::new(Config {
        module: &wasm_binary_code,
        heap_pages: HeapPages::from(2048),
        exec_hint: smoldot::executor::vm::ExecHint::Oneshot,
        allow_unresolved_imports: false
    }).unwrap();
    prototype.run_no_param("Core_version").unwrap().into()
};

// We need to answer the calls that the runtime might perform.
loop {
    match vm {
        // Calling `runner.run()` is what actually executes WebAssembly code and updates
        // the state.
        HostVm::ReadyToRun(runner) => vm = runner.run(),

        HostVm::Finished(finished) => {
            // `finished.value()` here is an opaque blob of bytes returned by the runtime.
            // In the case of a call to `"Core_version"`, we know that it must be empty.
            assert!(finished.value().as_ref().is_empty());
            println!("Success!");
            break;
        },

        // Errors can happen if the WebAssembly code panics or does something wrong.
        // In a real-life situation, the host should obviously not panic in these situations.
        HostVm::Error { .. } => {
            panic!("Error while executing code")
        },

        // All the other variants correspond to function calls that the runtime might perform.
        // `ExternalStorageGet` is shown here as an example.
        HostVm::ExternalStorageGet(req) => {
            println!("Runtime requires the storage value at {:?}", req.key().as_ref());
            // Injects the value into the virtual machine and updates the state.
            vm = req.resume(None); // Just a stub
        }
        _ => unimplemented!()
    }
}

Re-exports

pub use runtime_version::CoreVersion;

pub use runtime_version::CoreVersionError;

pub use runtime_version::CoreVersionRef;

pub use vm::HeapPages;

Modules

runtime_version

Wasm runtimes can optionally contain a custom section (as defined in the official WebAssembly core specification).

Structs

CallRuntimeVersion

Must provide the runtime version obtained by calling the Core_version entry point of a Wasm blob.

Config

Configuration for HostVmPrototype::new.

EndStorageTransaction

Declares the end of a transaction.

ExternalOffchainStorageSet

Must set the value of the off-chain storage.

ExternalStorageAppend

Must load a storage value, treat it as if it was a SCALE-encoded container, and put value at the end of the container, increasing the number of elements.

ExternalStorageClearPrefix

Must remove from the storage keys which start with a certain prefix. Use ExternalStorageClearPrefix::max_keys_to_remove to determine the maximum number of keys to remove.

ExternalStorageGet

Must provide the value of a storage entry.

ExternalStorageNextKey

Must provide the storage key that follows, in lexicographic order, a specific one.

ExternalStorageRoot

Must provide the trie root hash of the storage.

ExternalStorageSet

Must set the value of a storage entry.

Finished

Function execution has succeeded. Contains the return value of the call.

GetMaxLogLevel

Queries the maximum log level.

HostVmPrototype

Prototype for an HostVm.

LogEmit

Report about a log entry being emitted.

ReadyToRun

Virtual machine is ready to run.

SignatureVerification

Must verify whether a signature is correct.

StartStorageTransaction

Declares the start of a transaction.

Enums

Error

Reason why the Wasm blob isn’t conforming to the runtime environment.

HostVm

Running virtual machine.

ModuleFormatError

Error possibly returned when decoding a zstd-compressed Wasm blob.

NewErr

Error that can happen when initializing a VM.

StartErr

Error that can happen when starting a VM.