smoldot::executor::host
Module runtime_version

Wasm runtimes can optionally contain a custom section (as defined in the official WebAssembly core specification).

This module is dedicated to finding the custom sections containing the runtime version.

Re-exports

pub use super::TrieEntryVersion;

Structs

CoreVersion

Buffer storing the SCALE-encoded core version.

CoreVersionApi

One API that the runtime supports.

CoreVersionApisFromSliceErr

Error potentially returned by CoreVersionApisRefIter::from_slice_no_length.

CoreVersionApisRefIter

Iterator to a list of APIs. See CoreVersionRef::apis.

CoreVersionRef

Runtime specification, once decoded.

EmbeddedRuntimeVersionApis

Returns by find_encoded_embedded_runtime_version_apis.

Enums

CoreVersionError

Error while executing Core_version.

FindEmbeddedRuntimeVersionError

Error returned by find_embedded_runtime_version.

FindEncodedEmbeddedRuntimeVersionApisError

Error returned by find_encoded_embedded_runtime_version_apis.

Functions

find_embedded_runtime_version

Tries to find the custom section containing the runtime version and checks its validity.

find_encoded_embedded_runtime_version_apis

Tries to find the custom sections containing the runtime version and APIs.

hash_api_name

Hashes the name of an API in order to be able to compare it to CoreVersionApi::name_hash.