smoldot::executor

Module trie_root_calculator

Given as input a partial base trie and a diff, calculates the new root of the trie.

Implementation

The algorithm for calculating the trie root is roughly:

TrieRoot = ClosestDescendantMerkleValue(∅, False) || EmptyTrieRootHash

ClosestDescendantMerkleValue(Key, ForceRecalculate) =
    If !ForceRecalculate AND !DiffContainsEntryWithPrefix(Key)
        MaybeMerkleValue := BaseTrieClosestDescendantMerkleValue(Key)
        If MaybeMerkleValue
            return MaybeMerkleValue

    Btcd := BaseTrieClosestDescendant(Key)
    MaybeNodeKey := LongestCommonDenominator(Btcd, ...DiffInsertsNodesWithPrefix(Key))
    MaybeChildren := Array(ClosestDescendantMerkleValue(Concat(MaybeNodeKey, 0), MaybeNodeKey != Btcd), ..., ClosestDescendantMerkleValue(Concat(MaybeNodeKey, 15), MaybeNodeKey != Btcd))
    MaybeStorageValue := DiffInserts(MaybeNodeKey) || (BaseTrieStorageValue(MaybeNodeKey) - DiffErases(MaybeNodeKey))

    If MaybeStorageValue = ∅ AND NumChildren(MaybeChildren) == 0
        ∅
    ElseIf NumChildren(MaybeChildren) == 1 AND BaseTrieStorageValue(MaybeNodeKey) != ∅ AND MaybeStorageValue = ∅
        # Because the partial key of the child has changed, we have to recalculate it.
        ClosestDescendantMerkleValue(ChildThatExists(MaybeChildren), True)
    ElseIf NumChildren(MaybeChildren) == 1 AND BaseTrieStorageValue(MaybeNodeKey) = ∅ AND MaybeStorageValue = ∅
        ChildThatExists(MaybeChildren)
    Else
        Encode(MaybeStorageValue, MaybeChildren)

The public API functions are thus BaseTrieClosestDescendant, BaseTrieClosestDescendantMerkleValue, and BaseTrieStorageValue.

The algorithm above is recursive. In order to implement it, we instead maintain a stack of nodes that are currently being calculated.

Re-exports

pub use trie::Nibble;

pub use trie::TrieEntryVersion;

Structs

ClosestDescendant

In order to continue the calculation, must find in the base trie the closest descendant (including branch node) to a certain key in the trie. This can be equal to the requested key itself.

ClosestDescendantMerkleValue

In order to continue, must fetch the Merkle value of the closest descendant to the given key in the base trie.

Config

Configuration for trie_root_calculator.

StorageValue

In order to continue, must fetch the storage value of the given key in the base trie.

TrieNodeInsertUpdateEvent

Event indicating that a trie node has been inserted or updated.

TrieNodeRemoveEvent

Event indicating that a trie node has been destroyed.

Enums

InProgress

Trie root calculation in progress.

Functions

trie_root_calculator

Starts a new calculation.