Module smoldot::executor::vm

General-purpose WebAssembly virtual machine.

Contains code related to running a WebAssembly virtual machine. Contrary to (HostVm)super::host::HostVm, this module isn’t aware of any of the host functions available to Substrate runtimes. It only contains the code required to run a virtual machine, with some adjustments explained below.

Usage

Call VirtualMachinePrototype::new in order to parse and/or compile some WebAssembly code. One of the parameters of this function is a function that is passed the name of functions imported by the Wasm code, and must return an opaque usize. This usize doesn’t have any meaning, but will later be passed back to the user through ExecOutcome::Interrupted::id when the corresponding function is called.

The WebAssembly code can export functions in two different ways:

Some functions are exported through an (export) statement. See https://webassembly.github.io/spec/core/bikeshed/#export-section%E2%91%A0.

Some functions are stored in a global table called __indirect_function_table, and are later referred to by their index in this table. This is how the concept of “function pointers” commonly found in low-level programming languages is translated in WebAssembly.

Note: At the time of writing, it isn’t possible to call the second type of functions yet.

Use VirtualMachinePrototype::start in order to start executing a function exported through an (export) statement.

Call VirtualMachine::run on the VirtualMachine returned by start in order to run the WebAssembly code. The run method returns either if the function being called returns, or if the WebAssembly code calls a host function. In the latter case, ExecOutcome::Interrupted is returned and the virtual machine is now paused. Once the logic of the host function has been executed, call run again, passing the return value of that host function.

About __indirect_function_table

At initialization, the virtual machine will look for a table named __indirect_function_table. If present, this table is expected to contain functions. These functions can then be referred to by their index in this table. This is how the concept of “function pointers” commonly found in programming languages is translated in WebAssembly.

Note: When compiling C, C++, Rust, or similar languages to WebAssembly, one must pass the --export-table option to the LLVM linker in order for this symbol to be exported.

About imported vs exported memory

WebAssembly supports, in theory, addressing multiple different memory objects. The WebAssembly module can declare memory in two ways:

Either by exporting a memory object in the (export) section under the name memory.

Or by importing a memory object in its (import) section.

The virtual machine in this module supports both variants. However, no more than one memory object can be exported or imported, and it is illegal to not use any memory.

The first variant used to be the default model when compiling to WebAssembly, but the second variant (importing memory objects) is preferred nowadays.

Structs

HeapPages

Number of heap pages available to the Wasm code.

Module

Compiled Wasm code.

ModuleError

Opaque error indicating an error while parsing or compiling the WebAssembly code.

OutOfBoundsError

Error while reading memory.

Signature

Low-level Wasm function signature.

Trap

Opaque error that happened during execution, such as an unreachable instruction.

UnsupportedTypeError

Error used in the conversions between VM implementation and the public API.

VirtualMachine

VirtualMachinePrototype

Enums

ExecHint

Hint used by the implementation to decide which kind of virtual machine to use.

ExecOutcome

Outcome of the run function.

GlobalValueErr

Error that can happen when calling VirtualMachinePrototype::global_value.

NewErr

Error that can happen when initializing a VirtualMachinePrototype.

RunErr

Error that can happen when resuming the execution of a function.

StartErr

Error that can happen when calling VirtualMachinePrototype::start.

ValueType

Type of a value passed as parameter or returned by a function.

WasmValue

Value that a Wasm function can accept or produce.