smoldot::executor
Module vm

General-purpose WebAssembly virtual machine.

Contains code related to running a WebAssembly virtual machine. Contrary to (HostVm)super::host::HostVm, this module isn’t aware of any of the host functions available to Substrate runtimes. It only contains the code required to run a virtual machine, with some adjustments explained below.

Usage

Call VirtualMachinePrototype::new in order to parse and/or compile some WebAssembly code. One of the parameters of this function is a function that is passed the name of functions imported by the Wasm code, and must return an opaque usize. This usize doesn’t have any meaning, but will later be passed back to the user through ExecOutcome::Interrupted::id when the corresponding function is called.

Use VirtualMachinePrototype::prepare then Prepare::start in order to start executing a function exported through an (export) statement.

Call VirtualMachine::run on the VirtualMachine returned by start in order to run the WebAssembly code. The run method returns either if the function being called returns, or if the WebAssembly code calls a host function. In the latter case, ExecOutcome::Interrupted is returned and the virtual machine is now paused. Once the logic of the host function has been executed, call run again, passing the return value of that host function.

About imported vs exported memory

WebAssembly supports, in theory, addressing multiple different memory objects. The WebAssembly module can declare memory in two ways:

Either by exporting a memory object in the (export) section under the name memory.

Or by importing a memory object in its (import) section.

The virtual machine in this module supports both variants. However, no more than one memory object can be exported or imported, and it is illegal to not use any memory.

The first variant used to be the default model when compiling to WebAssembly, but the second variant (importing memory objects) is preferred nowadays.

Wasm features

The WebAssembly specification is a moving one. The specification as it was when it launched in 2017 is commonly referred to as “the MVP” (minimum viable product). Since then, various extensions have been added to the WebAssembly format.

The code in this module, however, doesn’t allow any of the feature that were added post-MVP. Trying to use WebAssembly code that uses one of these features will result in an error.

Structs

Config

Configuration to pass to VirtualMachinePrototype::new.

HeapPages

Number of heap pages available to the Wasm code.

OutOfBoundsError

Error while reading memory.

Prepare

Signature

Low-level Wasm function signature.

Trap

Opaque error that happened during execution, such as an unreachable instruction.

UnsupportedTypeError

Error used in the conversions between VM implementation and the public API.

VirtualMachine

VirtualMachinePrototype

Virtual machine ready to start executing a function.

Enums

ExecHint

Hint used by the implementation to decide which kind of virtual machine to use.

ExecOutcome

Outcome of the run function.

GlobalValueErr

Error that can happen when calling VirtualMachinePrototype::global_value.

NewErr

Error that can happen when initializing a VirtualMachinePrototype.

RunErr

Error that can happen when resuming the execution of a function.

StartErr

Error that can happen when calling Prepare::start.

ValueType

Type of a value passed as parameter or returned by a function.

WasmValue

Value that a Wasm function can accept or produce.