smoldot::finality

Module verify

Structs

CommitVerifyConfig

Configuration for a commit verification process.

CommitVerifyIsAuthority

Must return whether a certain public key is in the list of authorities that are allowed to generate pre-commits.

CommitVerifyIsParent

Must return whether a certain block is a descendant of the target block.

JustificationVerifyConfig

Configuration for a justification verification process.

Enums

CommitVerify

Commit verification in progress.

CommitVerifyError

Error that can happen while verifying a commit.

JustificationVerifyError

Error that can happen while verifying a justification.

Functions

verify_commit

Verifies that a commit is valid.

verify_justification

Verifies that a justification is valid.