smoldot::identity

Module keystore

Data structure containing cryptographic key pairs.

The keystore is a shared data structure (i.e. all of its functions accept &self rather than &mut self, making it possible to share it through an Arc for example) containing a list of cryptographic key pairs (i.e. both the public and secret keys).

Each key pair contained within the keystore is identified as a (KeyNamespace, [u8; 32]) tuple, where the [u8; 32] is the public key. See KeyNamespace.

A keystore is optionally associated with a directory of the file system into which it will store secret keys permanently. Keys present in this directory are considered to be the content of the keystore data structure.

For caching reasons, adding and removing keys to the directory manually (for example through the std::fs API) doesn’t automatically propagate to the public API of the keystore. Similarly, it is not intended to be possible to create two Keystore instances associated to the same directory at the same time.

Note: The Substrate framework also has a keystore, however this keystore implementation isn’t compatible with the Substrate keystore implementation. In other words, this keystore cannot load keys found in a directory that was previously associated with a Substrate keystore. This was decided because the Substrate keystore made some questionable decisions that it has to keep for backwards compatibility reasons. This keystore, being newly-written, doesn’t have to follow them.

Structs

Keystore

Collection of key pairs.

VrfSignature

Enums

KeyLoadError

KeyNamespace

Namespace of the key.

SignError

SignVrfError