smoldot::json_rpc

Module methods

List of requests and how to answer them.

Structs

AccountId

Contains the public key of an account.

Block

ChainHeadStorageRequestItem

ChainHeadStorageResponseItem

HashHexString

Header

HeaderDigest

HexString

InvalidParameterError

The parameter of a function call is invalid.

ReadProof

RpcMethods

RuntimeDispatchInfo

RuntimeSpec

RuntimeVersion

StorageChangeSet

SystemHealth

SystemPeer

TransactionWatchEventBlock

Enums

ChainHeadBodyCallReturn

ChainHeadStorageReturn

ChainHeadStorageType

DispatchClass

FollowEvent

HashHexStringSingleOrArray

MaybeRuntimeSpec

MethodCall

MethodError

See ParseClientToServerError::Method or ParseNotificationError::Method.

NetworkEvent

Unstable event. See https://github.com/paritytech/smoldot/issues/2245.

NetworkEventDirection

NetworkEventStatus

NodeRole

ParseClientToServerError

Error produced by parse_jsonrpc_client_to_server.

ParseNotificationError

Error produced by parse_notification.

RemoveMetadataLengthPrefixError

Error potentially returned by remove_metadata_length_prefix.

Response

ServerToClient

ServerToClientResponse

SystemPeerRole

TransactionStatus

TransactionWatchEvent

Functions

build_json_call_object_parameters

Builds a JSON call, to send it to a JSON-RPC server.

parse_jsonrpc_client_to_server

Parses a JSON call (usually sent from a JSON-RPC client and received by a JSON-RPC server).

parse_notification

Parses a JSON notification.

remove_metadata_length_prefix

Removes the length prefix at the beginning of metadata. Used for the Metadata_metadata JSON-RPC request. Returns an error if there is no valid length prefix.