Module smoldot::json_rpc::methods

List of requests and how to answer them.

Structs

AccountId

Contains the public key of an account.

Block

HashHexString

Header

HeaderDigest

HexString

InvalidParameterError

The parameter of a function call is invalid.

NetworkConfig

NumberAsString

RpcMethods

RuntimeDispatchInfo

RuntimeSpec

RuntimeVersion

StorageChangeSet

SystemHealth

SystemPeer

TransactionWatchEventBlock

Enums

ChainHeadBodyEvent

ChainHeadCallEvent

ChainHeadStorageEvent

ChainHeadStorageType

DispatchClass

FollowEvent

MaybeRuntimeSpec

MethodCall

MethodError

See ParseError::Method.

NetworkEvent

Unstable event. See https://github.com/paritytech/smoldot/issues/2245.

NodeRole

ParseError

Error produced by parse_json_call.

RemoveMetadataLengthPrefixError

Error potentially returned by remove_metadata_length_prefix.

Response

ServerToClient

ServerToClientResponse

SystemPeerRole

TransactionStatus

TransactionWatchEvent

Functions

build_json_call_object_parameters

Builds a JSON call, to send it to a JSON-RPC server.

parse_json_call

Parses a JSON call (usually received from a JSON-RPC server).

remove_metadata_length_prefix

Removes the length prefix at the beginning of metadata. Used for the Metadata_metadata JSON-RPC request. Returns an error if there is no valid length prefix.