Enum smoldot::json_rpc::methods::MethodCall

pub enum MethodCall<'a> {

    account_nextIndex {},
    author_hasKey {},
    author_hasSessionKeys {},
    author_insertKey {},
    author_pendingExtrinsics {},
    author_removeExtrinsic {},
    author_rotateKeys {},
    author_submitAndWatchExtrinsic {
        transaction: HexString,
    },
    author_submitExtrinsic {
        transaction: HexString,
    },
    author_unwatchExtrinsic {
        subscription: Cow<'a, str>,
    },
    babe_epochAuthorship {},
    chain_getBlock {
        hash: Option<HashHexString>,
    },
    chain_getBlockHash {
        height: Option<u64>,
    },
    chain_getFinalizedHead {},
    chain_getHeader {
        hash: Option<HashHexString>,
    },
    chain_subscribeAllHeads {},
    chain_subscribeFinalizedHeads {},
    chain_subscribeNewHeads {},
    chain_unsubscribeAllHeads {
        subscription: String,
    },
    chain_unsubscribeFinalizedHeads {
        subscription: String,
    },
    chain_unsubscribeNewHeads {
        subscription: String,
    },
    childstate_getKeys {},
    childstate_getStorage {},
    childstate_getStorageHash {},
    childstate_getStorageSize {},
    grandpa_roundState {},
    offchain_localStorageGet {},
    offchain_localStorageSet {},
    payment_queryInfo {
        extrinsic: HexString,
        hash: Option<HashHexString>,
    },
    rpc_methods {},
    state_call {
        name: Cow<'a, str>,
        parameters: HexString,
        hash: Option<HashHexString>,
    },
    state_getKeys {
        prefix: HexString,
        hash: Option<HashHexString>,
    },
    state_getKeysPaged {
        prefix: Option<HexString>,
        count: u32,
        start_key: Option<HexString>,
        hash: Option<HashHexString>,
    },
    state_getMetadata {
        hash: Option<HashHexString>,
    },
    state_getPairs {},
    state_getReadProof {},
    state_getRuntimeVersion {
        at: Option<HashHexString>,
    },
    state_getStorage {
        key: HexString,
        hash: Option<HashHexString>,
    },
    state_getStorageHash {},
    state_getStorageSize {},
    state_queryStorage {},
    state_queryStorageAt {
        keys: Vec<HexString>,
        at: Option<HashHexString>,
    },
    state_subscribeRuntimeVersion {},
    state_subscribeStorage {
        list: Vec<HexString>,
    },
    state_unsubscribeRuntimeVersion {
        subscription: Cow<'a, str>,
    },
    state_unsubscribeStorage {
        subscription: Cow<'a, str>,
    },
    system_accountNextIndex {
        account: AccountId,
    },
    system_addReservedPeer {},
    system_chain {},
    system_chainType {},
    system_dryRun {},
    system_health {},
    system_localListenAddresses {},
    system_localPeerId {},
    system_name {},
    system_networkState {},
    system_nodeRoles {},
    system_peers {},
    system_properties {},
    system_removeReservedPeer {},
    system_version {},
    chainHead_unstable_body {
        follow_subscription: Cow<'a, str>,
        hash: HashHexString,
        network_config: Option<NetworkConfig>,
    },
    chainHead_unstable_call {
        follow_subscription: Cow<'a, str>,
        hash: HashHexString,
        function: Cow<'a, str>,
        call_parameters: HexString,
        network_config: Option<NetworkConfig>,
    },
    chainHead_unstable_follow {
        runtime_updates: bool,
    },
    chainHead_unstable_genesisHash {},
    chainHead_unstable_header {
        follow_subscription: Cow<'a, str>,
        hash: HashHexString,
    },
    chainHead_unstable_stopBody {
        subscription: Cow<'a, str>,
    },
    chainHead_unstable_stopCall {
        subscription: Cow<'a, str>,
    },
    chainHead_unstable_stopStorage {
        subscription: Cow<'a, str>,
    },
    chainHead_unstable_storage {
        follow_subscription: Cow<'a, str>,
        hash: HashHexString,
        key: HexString,
        child_key: Option<HexString>,
        network_config: Option<NetworkConfig>,
    },
    chainHead_unstable_unfollow {
        follow_subscription: Cow<'a, str>,
    },
    chainHead_unstable_unpin {
        follow_subscription: Cow<'a, str>,
        hash: HashHexString,
    },
    chainSpec_unstable_chainName {},
    chainSpec_unstable_genesisHash {},
    chainSpec_unstable_properties {},
    sudo_unstable_p2pDiscover {
        multiaddr: Cow<'a, str>,
    },
    sudo_unstable_version {},
    transaction_unstable_submitAndWatch {
        transaction: HexString,
    },
    transaction_unstable_unwatch {
        subscription: Cow<'a, str>,
    },
    network_unstable_subscribeEvents {},
    network_unstable_unsubscribeEvents {
        subscription: Cow<'a, str>,
    },
    chainHead_unstable_finalizedDatabase {
        max_size_bytes: Option<u64>,
    },

}

Variants

account_nextIndex

Fields

author_hasKey

Fields

author_hasSessionKeys

Fields

author_insertKey

Fields

author_pendingExtrinsics

Fields

author_removeExtrinsic

Fields

author_rotateKeys

Fields

author_submitAndWatchExtrinsic

Fields

transaction: HexString

author_submitExtrinsic

Fields

transaction: HexString

author_unwatchExtrinsic

Fields

subscription: Cow<'a, str>

babe_epochAuthorship

Fields

chain_getBlock

Fields

hash: Option<HashHexString>

chain_getBlockHash

Fields

height: Option<u64>

chain_getFinalizedHead

Fields

chain_getHeader

Fields

hash: Option<HashHexString>

chain_subscribeAllHeads

Fields

chain_subscribeFinalizedHeads

Fields

chain_subscribeNewHeads

Fields

chain_unsubscribeAllHeads

Fields

subscription: String

chain_unsubscribeFinalizedHeads

Fields

subscription: String

chain_unsubscribeNewHeads

Fields

subscription: String

childstate_getKeys

Fields

childstate_getStorage

Fields

childstate_getStorageHash

Fields

childstate_getStorageSize

Fields

grandpa_roundState

Fields

offchain_localStorageGet

Fields

offchain_localStorageSet

Fields

payment_queryInfo

Fields

extrinsic: HexString

hash: Option<HashHexString>

rpc_methods

Fields

Returns a list of all JSON-RPC methods that are available.

state_call

Fields

name: Cow<'a, str>

parameters: HexString

hash: Option<HashHexString>

state_getKeys

Fields

prefix: HexString

hash: Option<HashHexString>

state_getKeysPaged

Fields

prefix: Option<HexString>

count: u32

start_key: Option<HexString>

hash: Option<HashHexString>

state_getMetadata

Fields

hash: Option<HashHexString>

state_getPairs

Fields

state_getReadProof

Fields

state_getRuntimeVersion

Fields

at: Option<HashHexString>

state_getStorage

Fields

key: HexString

hash: Option<HashHexString>

state_getStorageHash

Fields

state_getStorageSize

Fields

state_queryStorage

Fields

state_queryStorageAt

Fields

keys: Vec<HexString>

at: Option<HashHexString>

state_subscribeRuntimeVersion

Fields

state_subscribeStorage

Fields

list: Vec<HexString>

state_unsubscribeRuntimeVersion

Fields

subscription: Cow<'a, str>

state_unsubscribeStorage

Fields

subscription: Cow<'a, str>

system_accountNextIndex

Fields

account: AccountId

system_addReservedPeer

Fields

system_chain

Fields

system_chainType

Fields

system_dryRun

Fields

system_health

Fields

system_localListenAddresses

Fields

system_localPeerId

Fields

Returns the Base58 encoding of the network identity of the node on the peer-to-peer network.

system_name

Fields

Returns, as an opaque string, the name of the client serving these JSON-RPC requests.

system_networkState

Fields

system_nodeRoles

Fields

system_peers

Fields

system_properties

Fields

system_removeReservedPeer

Fields

system_version

Fields

Returns, as an opaque string, the version of the client serving these JSON-RPC requests.

chainHead_unstable_body

Fields

follow_subscription: Cow<'a, str>

hash: HashHexString

network_config: Option<NetworkConfig>

chainHead_unstable_call

Fields

follow_subscription: Cow<'a, str>

hash: HashHexString

function: Cow<'a, str>

call_parameters: HexString

network_config: Option<NetworkConfig>

chainHead_unstable_follow

Fields

runtime_updates: bool

chainHead_unstable_genesisHash

Fields

chainHead_unstable_header

Fields

follow_subscription: Cow<'a, str>

hash: HashHexString

chainHead_unstable_stopBody

Fields

subscription: Cow<'a, str>

chainHead_unstable_stopCall

Fields

subscription: Cow<'a, str>

chainHead_unstable_stopStorage

Fields

subscription: Cow<'a, str>

chainHead_unstable_storage

Fields

follow_subscription: Cow<'a, str>

hash: HashHexString

key: HexString

child_key: Option<HexString>

network_config: Option<NetworkConfig>

chainHead_unstable_unfollow

Fields

follow_subscription: Cow<'a, str>

chainHead_unstable_unpin

Fields

follow_subscription: Cow<'a, str>

hash: HashHexString

chainSpec_unstable_chainName

Fields

chainSpec_unstable_genesisHash

Fields

chainSpec_unstable_properties

Fields

sudo_unstable_p2pDiscover

Fields

multiaddr: Cow<'a, str>

sudo_unstable_version

Fields

transaction_unstable_submitAndWatch

Fields

transaction: HexString

transaction_unstable_unwatch

Fields

subscription: Cow<'a, str>

network_unstable_subscribeEvents

Fields

network_unstable_unsubscribeEvents

Fields

subscription: Cow<'a, str>

chainHead_unstable_finalizedDatabase

Fields

max_size_bytes: Option<u64>

Implementations

impl<'a> MethodCall<'a>

pub fn method_names() -> impl ExactSizeIterator<Item = &'static str>

Returns a list of RPC method names of all the methods in the enum.

pub fn name(&self) -> &'static str

Returns the name of the method.

pub fn params_to_json_object(&self) -> String

Returns an JSON object containing the list of the parameters of the method.

pub fn to_json_call_object_parameters(&self, id_json: Option<&str>) -> String

Builds a JSON call, to send it to a JSON-RPC server.

Panic

Panics if the id_json isn’t valid JSON.