smoldot::json_rpc::methods

Enum MethodCall

pub enum MethodCall<'a> {

    account_nextIndex {},
    author_hasKey {},
    author_hasSessionKeys {},
    author_insertKey {},
    author_pendingExtrinsics {},
    author_removeExtrinsic {},
    author_rotateKeys {},
    author_submitAndWatchExtrinsic {
        transaction: HexString,
    },
    author_submitExtrinsic {
        transaction: HexString,
    },
    author_unwatchExtrinsic {
        subscription: Cow<'a, str>,
    },
    babe_epochAuthorship {},
    chain_getBlock {
        hash: Option<HashHexString>,
    },
    chain_getBlockHash {
        height: Option<u64>,
    },
    chain_getFinalizedHead {},
    chain_getHeader {
        hash: Option<HashHexString>,
    },
    chain_subscribeAllHeads {},
    chain_subscribeFinalizedHeads {},
    chain_subscribeNewHeads {},
    chain_unsubscribeAllHeads {
        subscription: String,
    },
    chain_unsubscribeFinalizedHeads {
        subscription: String,
    },
    chain_unsubscribeNewHeads {
        subscription: String,
    },
    childstate_getKeys {},
    childstate_getStorage {},
    childstate_getStorageHash {},
    childstate_getStorageSize {},
    grandpa_roundState {},
    offchain_localStorageGet {},
    offchain_localStorageSet {},
    payment_queryInfo {
        extrinsic: HexString,
        hash: Option<HashHexString>,
    },
    rpc_methods {},
    state_call {
        name: Cow<'a, str>,
        parameters: HexString,
        hash: Option<HashHexString>,
    },
    state_getKeys {
        prefix: HexString,
        hash: Option<HashHexString>,
    },
    state_getKeysPaged {
        prefix: Option<HexString>,
        count: u32,
        start_key: Option<HexString>,
        hash: Option<HashHexString>,
    },
    state_getMetadata {
        hash: Option<HashHexString>,
    },
    state_getPairs {},
    state_getReadProof {
        keys: Vec<HexString>,
        at: Option<HashHexString>,
    },
    state_getRuntimeVersion {
        at: Option<HashHexString>,
    },
    state_getStorage {
        key: HexString,
        hash: Option<HashHexString>,
    },
    state_getStorageHash {},
    state_getStorageSize {},
    state_queryStorage {},
    state_queryStorageAt {
        keys: Vec<HexString>,
        at: Option<HashHexString>,
    },
    state_subscribeRuntimeVersion {},
    state_subscribeStorage {
        list: Vec<HexString>,
    },
    state_unsubscribeRuntimeVersion {
        subscription: Cow<'a, str>,
    },
    state_unsubscribeStorage {
        subscription: Cow<'a, str>,
    },
    system_accountNextIndex {
        account: AccountId,
    },
    system_addReservedPeer {},
    system_chain {},
    system_chainType {},
    system_dryRun {},
    system_health {},
    system_localListenAddresses {},
    system_localPeerId {},
    system_name {},
    system_networkState {},
    system_nodeRoles {},
    system_peers {},
    system_properties {},
    system_removeReservedPeer {},
    system_version {},
    chainHead_v1_body {
        follow_subscription: Cow<'a, str>,
        hash: HashHexString,
    },
    chainHead_v1_call {
        follow_subscription: Cow<'a, str>,
        hash: HashHexString,
        function: Cow<'a, str>,
        call_parameters: HexString,
    },
    chainHead_v1_follow {
        with_runtime: bool,
    },
    chainHead_v1_header {
        follow_subscription: Cow<'a, str>,
        hash: HashHexString,
    },
    chainHead_v1_stopOperation {
        follow_subscription: Cow<'a, str>,
        operation_id: Cow<'a, str>,
    },
    chainHead_v1_storage {
        follow_subscription: Cow<'a, str>,
        hash: HashHexString,
        items: Vec<ChainHeadStorageRequestItem>,
        child_trie: Option<HexString>,
    },
    chainHead_v1_continue {
        follow_subscription: Cow<'a, str>,
        operation_id: Cow<'a, str>,
    },
    chainHead_v1_unfollow {
        follow_subscription: Cow<'a, str>,
    },
    chainHead_v1_unpin {
        follow_subscription: Cow<'a, str>,
        hash_or_hashes: HashHexStringSingleOrArray,
    },
    chainSpec_v1_chainName {},
    chainSpec_v1_genesisHash {},
    chainSpec_v1_properties {},
    sudo_unstable_p2pDiscover {
        multiaddr: Cow<'a, str>,
    },
    sudo_unstable_version {},
    transaction_v1_broadcast {
        transaction: HexString,
    },
    transaction_v1_stop {
        operation_id: Cow<'a, str>,
    },
    transactionWatch_v1_submitAndWatch {
        transaction: HexString,
    },
    transactionWatch_v1_unwatch {
        subscription: Cow<'a, str>,
    },
    sudo_network_unstable_watch {},
    sudo_network_unstable_unwatch {
        subscription: Cow<'a, str>,
    },
    chainHead_unstable_finalizedDatabase {
        max_size_bytes: Option<u64>,
    },

}

Variants

account_nextIndex

author_hasKey

author_hasSessionKeys

author_insertKey

author_pendingExtrinsics

author_removeExtrinsic

author_rotateKeys

author_submitAndWatchExtrinsic

Fields

transaction: HexString

author_submitExtrinsic

Fields

transaction: HexString

author_unwatchExtrinsic

Fields

subscription: Cow<'a, str>

babe_epochAuthorship

chain_getBlock

Fields

hash: Option<HashHexString>

chain_getBlockHash

Fields

height: Option<u64>

chain_getFinalizedHead

chain_getHeader

Fields

hash: Option<HashHexString>

chain_subscribeAllHeads

chain_subscribeFinalizedHeads

chain_subscribeNewHeads

chain_unsubscribeAllHeads

Fields

subscription: String

chain_unsubscribeFinalizedHeads

Fields

subscription: String

chain_unsubscribeNewHeads

Fields

subscription: String

childstate_getKeys

childstate_getStorage

childstate_getStorageHash

childstate_getStorageSize

grandpa_roundState

offchain_localStorageGet

offchain_localStorageSet

payment_queryInfo

Fields

extrinsic: HexString

hash: Option<HashHexString>

rpc_methods

Returns a list of all JSON-RPC methods that are available.

state_call

Fields

name: Cow<'a, str>

parameters: HexString

hash: Option<HashHexString>

state_getKeys

Fields

prefix: HexString

hash: Option<HashHexString>

state_getKeysPaged

Fields

prefix: Option<HexString>

count: u32

start_key: Option<HexString>

hash: Option<HashHexString>

state_getMetadata

Fields

hash: Option<HashHexString>

state_getPairs

state_getReadProof

Fields

keys: Vec<HexString>

at: Option<HashHexString>

state_getRuntimeVersion

Fields

at: Option<HashHexString>

state_getStorage

Fields

key: HexString

hash: Option<HashHexString>

state_getStorageHash

state_getStorageSize

state_queryStorage

state_queryStorageAt

Fields

keys: Vec<HexString>

at: Option<HashHexString>

state_subscribeRuntimeVersion

state_subscribeStorage

Fields

list: Vec<HexString>

state_unsubscribeRuntimeVersion

Fields

subscription: Cow<'a, str>

state_unsubscribeStorage

Fields

subscription: Cow<'a, str>

system_accountNextIndex

Fields

account: AccountId

system_addReservedPeer

system_chain

system_chainType

system_dryRun

system_health

system_localListenAddresses

system_localPeerId

Returns the Base58 encoding of the network identity of the node on the peer-to-peer network.

system_name

Returns, as an opaque string, the name of the client serving these JSON-RPC requests.

system_networkState

system_nodeRoles

system_peers

system_properties

system_removeReservedPeer

system_version

Returns, as an opaque string, the version of the client serving these JSON-RPC requests.

chainHead_v1_body

Fields

follow_subscription: Cow<'a, str>

hash: HashHexString

chainHead_v1_call

Fields

follow_subscription: Cow<'a, str>

hash: HashHexString

function: Cow<'a, str>

call_parameters: HexString

chainHead_v1_follow

Fields

with_runtime: bool

chainHead_v1_header

Fields

follow_subscription: Cow<'a, str>

hash: HashHexString

chainHead_v1_stopOperation

Fields

follow_subscription: Cow<'a, str>

operation_id: Cow<'a, str>

chainHead_v1_storage

Fields

follow_subscription: Cow<'a, str>

hash: HashHexString

items: Vec<ChainHeadStorageRequestItem>

child_trie: Option<HexString>

chainHead_v1_continue

Fields

follow_subscription: Cow<'a, str>

operation_id: Cow<'a, str>

chainHead_v1_unfollow

Fields

follow_subscription: Cow<'a, str>

chainHead_v1_unpin

Fields

follow_subscription: Cow<'a, str>

hash_or_hashes: HashHexStringSingleOrArray

chainSpec_v1_chainName

chainSpec_v1_genesisHash

chainSpec_v1_properties

sudo_unstable_p2pDiscover

Fields

multiaddr: Cow<'a, str>

sudo_unstable_version

transaction_v1_broadcast

Fields

transaction: HexString

transaction_v1_stop

Fields

operation_id: Cow<'a, str>

transactionWatch_v1_submitAndWatch

Fields

transaction: HexString

transactionWatch_v1_unwatch

Fields

subscription: Cow<'a, str>

sudo_network_unstable_watch

sudo_network_unstable_unwatch

Fields

subscription: Cow<'a, str>

chainHead_unstable_finalizedDatabase

Fields

max_size_bytes: Option<u64>

Implementations

impl<'a> MethodCall<'a>

pub fn method_names() -> impl ExactSizeIterator<Item = &'static str>

Returns a list of RPC method names of all the methods in the enum.

pub fn name(&self) -> &'static str

Returns the name of the method.

pub fn params_to_json_object(&self) -> String

Returns an JSON object containing the list of the parameters of the method.

pub fn to_json_request_object_parameters(&self, id_json: Option<&str>) -> String

Builds a JSON request, to send it to a JSON-RPC server.

Panic

Panics if the id_json isn’t valid JSON.