Module smoldot::json_rpc::payment_info

Enums

DecodeError

Potential error when decoding payment information runtime output.

Constants

PAYMENT_FEES_FUNCTION_NAME

Name of the runtime function to call in order to obtain the payment fees.

Functions

decode_payment_info

Attempt to decode the output of the runtime call.

payment_info_parameters

Produces the input to pass to the TransactionPaymentApi_query_info runtime call.