Module smoldot::json_rpc::requests_subscriptions

The RequestsSubscriptions state machine holds a list of clients, pending outgoing messages, pending requests, and active subscriptions.

The code in this module is the front line of the JSON-RPC server. It can be subject to DoS attacks, and is therefore designed to properly distribute resources between JSON-RPC clients. If you use this data structure as intended, your design is safe from DoS attacks.

Usage

The RequestsSubscriptions is meant to be shared (through an Arc or similar) between many different asynchronous tasks that call its methods.

Note: While off-topic for this module, you are strongly encouraged to put all these asynchronous tasks within a single FuturesUnordered. This ensure that no two tasks can be processed at the same time, and thus limits the total CPU usage of all these tasks combined to 1.0 CPU cores. This leaves other CPU cores free for the more urgent processing.

There should be:

One lightweight task for each client currently connected to the server.

A fixed number of lightweight tasks (e.g. 16) dedicated to answering requests.

Clients

Whenever a new client connects to the server, spawn a new task dedicated to this client, that:

Calls RequestsSubscriptions::add_client, denying the client if the function returns an error.

Repeatedly polls the socket for a new request then calls RequestsSubscriptions::queue_client_request.

At the same time (for example in a select! block) calls RequestsSubscriptions::next_response then sends the response to the socket.

When the client disconnects, calls RequestsSubscriptions::remove_client.

It is important that no new request is polled from the socket as long as RequestsSubscriptions::queue_client_request hasn’t been able to queue the previous request. This makes it possible to back-pressure the JSON-RPC client in case when the queue is slow to be processed.

Similarly, do not call RequestsSubscriptions::next_response before the socket has been able to send the previous response. Not calling RequestsSubscriptions::next_response often enough will lead to back-pressure being applied onto RequestsSubscriptions::queue_client_request, which will in turn back-pressure the sending side of the JSON-RPC client.

Note that if a client is removed at the same time as a call to RequestsSubscriptions::next_response is in progress, the call will never return It is your responsibility to interrupt this function call when the client is disconnected. If, as advised above, everything is contained within a single task, this is normally not a problem as you simply stop the task altogether after removing the client.

Requests

There should be a certain, fixed, number of lightweight tasks dedicated to pulling requests from the state machine and answering them.

Each of these lightweight tasks should:

Call RequestsSubscriptions::next_request. This function call sleeps until there is a request available.

Parse the request that was returned and generate its response. This step should be relatively fast (e.g. not more than one second), but can liberally perform asynchronous requests, lock mutexes, etc.

Call RequestsSubscriptions::respond.

Jump back to step 1.

If these tasks are too busy and don’t call RequestsSubscriptions::next_request often enough, back-pressure will be applied onto RequestsSubscriptions::queue_client_request, which in turn applies back-pressure onto the JSON-RPC clients.

Subscriptions

If a client-sent request requires starting a subscription, one of the requests-pulling-dedicated tasks should call RequestsSubscriptions::start_subscription.

It is the responsibility of the higher-level code to generate the JSON-RPC-client-facing identifier of the subscription.

When a subscription is started, the higher-level code should spawn a new task dedicated to sending back notifications to the client using RequestsSubscriptions::push_notification, RequestsSubscriptions::try_push_notification, or RequestsSubscriptions::set_queued_notification.

The code on top should maintain a map of JSON-RPC-client-facing identifier to SubscriptionId. When the JSON-RPC client wants to unsubscribe, call RequestsSubscriptions::stop_subscription. This map should also contain a way to abort the task dedicated to that subscription.

Structs

ClientId

Config

Configuration to pass to RequestsSubscriptions::new.

RequestId

RequestsSubscriptions

SubscriptionId

TryQueueClientRequestError

Error returned by RequestsSubscriptions::try_queue_client_request.

Enums

AddClientError

Error returned by RequestsSubscriptions::add_client and RequestsSubscriptions::add_client_mut.

StartSubscriptionError

Error returned by RequestsSubscriptions::start_subscription.