Module smoldot::json_rpc::service::client_main_task

Re-exports

pub use crate::json_rpc::parse::ErrorResponse;

pub use crate::json_rpc::parse::ParseError;

Structs

ClientMainTask

See module-level-documentation.

Config

Configuration for client_main_task.

RequestProcess

Object connected to the ClientMainTask and containing a request expecting an answer.

SendRequestError

Error returned by SerializedRequestsIo::send_request.

SerializedRequestsIo

Object connected to the ClientMainTask that allows sending requests to the task and receiving responses.

Subscription

Object connected to the ClientMainTask representing an active subscription.

SubscriptionStartProcess

Object connected to the ClientMainTask and containing a request that leads to the creation of a subscription.

TrySendRequestError

Error returned by SerializedRequestsIo::try_send_request.

Enums

Event

Outcome of the processing of ClientMainTask::run_until_event.

SendRequestErrorCause

See SendRequestError::cause.

TrySendRequestErrorCause

See TrySendRequestError::cause.

WaitNextResponseError

See SerializedRequestsIo::wait_next_response.

Functions

client_main_task

Creates a new ClientMainTask and a SerializedRequestsIo connected to it.