Struct smoldot::json_rpc::websocket_server::WsServer

pub struct WsServer<T> { /* private fields */ }

Listening socket and list of open connections.

Implementations

impl<T> WsServer<T>

pub async fn new(config: Config) -> Result<Self, Error>

Try opening a TCP listening socket.

Returns an error if the listening socket fails to open.

pub fn local_addr(&self) -> Result<SocketAddr, Error>

Address of the local TCP listening socket, as provided by the operating system.

pub fn accept(&mut self, user_data: T) -> ConnectionId

Accepts the pending connection.

Either WsServer::accept or WsServer::reject must be called after a Event::ConnectionOpen event is returned.

Panic

Panics if no connection is pending.

pub fn reject(&mut self)

Reject the pending connection.

Either WsServer::accept or WsServer::reject must be called after a Event::ConnectionOpen event is returned.

Panic

Panics if no connection is pending.

pub fn is_empty(&self) -> bool

Returns true if there isn’t any active healthy connection.

pub fn len(&self) -> usize

Returns the number of active healthy connections.

pub fn close(&mut self, connection_id: ConnectionId) -> T

Destroys a connection.

The connection will be cleanly shut down in the background, but for API purposes this ConnectionId is now no longer valid.

Panic

Panics if the ConnectionId is invalid.

pub fn queue_send(&mut self, connection: ConnectionId, message: String)

Queues a text frame to be sent on the given connection.

If more than Config::send_buffer_len messages are already buffered, the message is silently discarded and a Event::ConnectionError will soon be generated for this connection.

Panic

Panics if the ConnectionId is invalid.

pub async fn next_event(&mut self) -> Event<'_, T>

Returns the next event happening on the server.