smoldot::libp2p

Module collection

Collection of libp2p connections.

The Network struct in this module is a collection of libp2p connections. In the documentation below, it is also called the coordinator.

When a connection is inserted in the collection with Network::insert_single_stream or Network::insert_multi_stream, two objects are returned: an identifier for this new connection assigned by the collection, and a SingleStreamConnectionTask or MultiStreamConnectionTask.

All the SingleStreamConnectionTasks/MultiStreamConnectionTask created by the Network communicate with that Network by passing messages. Passing the messages has to be done explicitly by the API user. It is the responsibility of the API user to pull messages from the coordinator (i.e. the Network) and push them onto the SingleStreamConnectionTask or MultiStreamConnectionTask and vice-versa.

Usage

Whenever a new connection is established, use Network::insert_single_stream or Network::insert_multi_stream to allocate a connection in the collection.

When a connection has received data or is ready to send more data, use SingleStreamConnectionTask::read_write, SingleStreamConnectionTask::reset, MultiStreamConnectionTask::substream_read_write, MultiStreamConnectionTask::reset, MultiStreamConnectionTask::add_substream, and/or MultiStreamConnectionTask::desired_outbound_substreams to synchronize the state of the task with the actual state of the connection.

Pull messages from the SingleStreamConnectionTasks and MultiStreamConnectionTasks and inject them into the Network, and vice versa.

In parallel, continuously call Network::next_event to process the events generated by the calls to Network::inject_connection_message.

Re-exports

pub use super::peer_id::PeerId;

pub use super::read_write::ReadWrite;

pub use established::InboundError;

pub use established::InboundTy;

pub use established::SubstreamFate;

pub use single_stream_handshake::HandshakeError;

Structs

Config

Configuration for a Network.

ConnectionId

Identifier of a connection spawned by the Network.

ConnectionState

See Network::connection_state.

ConnectionToCoordinator

Message from a connection task destined to the coordinator.

CoordinatorToConnection

Message from the coordinator destined to a connection task.

MultiStreamConnectionTask

State machine dedicated to a single multi-stream connection.

Network

Data structure containing the list of all connections, pending or not, and their latest known state. See also the module-level documentation.

SingleStreamConnectionTask

State machine dedicated to a single single-stream connection.

SubstreamId

Identifier of a request, or an inbound substream, or an outbound substream.

Enums

Event

Event generated by Network::next_event.

MultiStreamHandshakeKind

What kind of handshake to perform on the newly-added connection.

NotificationsInClosedErr

NotificationsOutErr

QueueNotificationError

Error potentially returned by Network::queue_notification.

RequestError

ShutdownCause

Reason why a connection is shutting down. See Event::StartShutdown.

SingleStreamHandshakeKind

What kind of handshake to perform on the newly-added connection.