smoldot::libp2p::collection

Enum Event

pub enum Event<TConn> {

    HandshakeFinished {
        id: ConnectionId,
        peer_id: PeerId,
    },
    StartShutdown {
        id: ConnectionId,
        reason: ShutdownCause,
    },
    Shutdown {
        id: ConnectionId,
        was_established: bool,
        user_data: TConn,
    },
    InboundError {
        id: ConnectionId,
        error: InboundError,
    },
    InboundNegotiated {
        id: ConnectionId,
        substream_id: SubstreamId,
        protocol_name: String,
    },
    InboundNegotiatedCancel {
        substream_id: SubstreamId,
    },
    InboundAcceptedCancel {
        substream_id: SubstreamId,
    },
    Response {
        substream_id: SubstreamId,
        response: Result<Vec<u8>, RequestError>,
    },
    RequestIn {
        substream_id: SubstreamId,
        request_payload: Vec<u8>,
    },
    RequestInCancel {
        substream_id: SubstreamId,
    },
    NotificationsOutResult {
        substream_id: SubstreamId,
        result: Result<Vec<u8>, NotificationsOutErr>,
    },
    NotificationsOutCloseDemanded {
        substream_id: SubstreamId,
    },
    NotificationsOutReset {
        substream_id: SubstreamId,
    },
    NotificationsInOpen {
        substream_id: SubstreamId,
        remote_handshake: Vec<u8>,
    },
    NotificationsInOpenCancel {
        substream_id: SubstreamId,
    },
    NotificationsIn {
        substream_id: SubstreamId,
        notification: Vec<u8>,
    },
    NotificationsInClose {
        substream_id: SubstreamId,
        outcome: Result<(), NotificationsInClosedErr>,
    },
    PingOutSuccess {
        id: ConnectionId,
        ping_time: Duration,
    },
    PingOutFailed {
        id: ConnectionId,
    },

}

Event generated by Network::next_event.

Variants

HandshakeFinished

Handshake of the given connection has completed.

This event can only happen once per connection and only for single-stream connections.

Fields

id: ConnectionId

Identifier of the connection whose handshake is finished.

peer_id: PeerId

Identity of the peer on the other side of the connection.

StartShutdown

A transport-level connection (e.g. a TCP socket) is starting its shutdown.

It is no longer possible to start requests, open notification substreams, or open notifications on this connection, and no new incoming requests or notification substreams will be reported as events.

Further events will close all existing substreams (requests and notifications) one by one. Once all substreams have been closed, a Event::Shutdown is reported.

Keep in mind that this event can also happen for connections that haven’t finished their handshake.

This event is not generated when Network::start_shutdown is called.

Fields

id: ConnectionId

Identifier of the connection that is starting its shutdown.

reason: ShutdownCause

Reason why the connection is starting its shutdown. Because this event is not generated when the shutdown is initiated locally, the reason is always cause by the remote.

Shutdown

A transport-level connection (e.g. a TCP socket) has been shut down.

This ConnectionId is no longer valid, and using it will result in panics.

Fields

id: ConnectionId

Identifier of the connection that has finished its shutdown.

was_established: bool

true if the connection was in its established phase before the shutdown.

user_data: TConn

User data that was stored in the state machine for this connection.

InboundError

Received an incoming substream, but this substream has produced an error.

Note: This event exists only for diagnostic purposes. No action is expected in return.

Fields

id: ConnectionId

Identifier of the connection that has received the substream.

error: InboundError

Error that happened.

InboundNegotiated

An inbound substream has requested to use a protocol. Call Network::accept_inbound or Network::reject_inbound in order to accept or reject this substream.

Fields

id: ConnectionId

Identifier of the connection that has received the substream.

substream_id: SubstreamId

Identifier of the substream. Needs to be provided back when accepting or rejecting the substream.

protocol_name: String

Name of the protocol requested by the remote.

InboundNegotiatedCancel

An inbound substream that is waiting for a call to Network::accept_inbound or Network::reject_inbound has been abruptly closed.

Fields

substream_id: SubstreamId

Identifier of the substream.

InboundAcceptedCancel

An inbound substream that was previously accepted using Network::accept_inbound was closed by the remote or has generated an error.

Fields

substream_id: SubstreamId

Identifier of the substream.

Response

Outcome of a request started using Network::start_request.

All requests always lead to an outcome, even if the connection has been closed while the request was in progress.

Fields

substream_id: SubstreamId

Substream that was returned by Network::start_request.

response: Result<Vec<u8>, RequestError>

If the request is successful, contains the response sent back by the remote. Otherwise, contains the reason why the request isn’t successful.

RequestIn

Received a request from a request-response protocol.

Fields

substream_id: SubstreamId

Substream on which the request has been received. Must be passed back when providing the response.

request_payload: Vec<u8>

Payload that has been sent by the remote. Its interpretation is beyond the scope of this module.

RequestInCancel

Request received earlier has been canceled by the remote.

The SubstreamId is now invalid.

Fields

substream_id: SubstreamId

NotificationsOutResult

Outcome of trying to open a substream with Network::open_out_notifications.

If Ok, it is now possible to send notifications on this substream. If Err, the substream no longer exists and the SubstreamId becomes invalid.

Fields

substream_id: SubstreamId

result: Result<Vec<u8>, NotificationsOutErr>

If Ok, contains the handshake sent back by the remote. Its interpretation is out of scope of this module.

NotificationsOutCloseDemanded

Remote has closed an outgoing notifications substream, meaning that it demands the closing of the substream. Use Network::close_out_notifications as soon as possible, which is typically after all outbound notifications that need to be queued have been queued.

This event is only generated for notification substreams that are fully open.

Fields

substream_id: SubstreamId

NotificationsOutReset

A previously open outbound substream has been closed, by the remote or as a consequence of the connection shutting down.

This event is only generated for notification substreams that are fully open.

The substream no longer exists and the SubstreamId becomes invalid.

Fields

substream_id: SubstreamId

NotificationsInOpen

The remote would like to open a notifications substream.

The substream needs to be accepted or refused using Network::accept_in_notifications or Network::reject_in_notifications.

Fields

substream_id: SubstreamId

Newly-generated identifier for the substream on which the request has been received. Must be passed back when accepting or refusing the substream.

remote_handshake: Vec<u8>

Handshake that has been sent by the remote. Its interpretation is beyond the scope of this module.

NotificationsInOpenCancel

The remote has canceled the opening an incoming notifications substream.

This can only happen before the notification substream has been accepted or refused.

Fields

substream_id: SubstreamId

Substream that has been closed. Guaranteed to match a substream that was earlier reported with a Event::NotificationsInOpen.

NotificationsIn

Received a notification on a notifications substream of a connection.

Fields

substream_id: SubstreamId

Substream on which the notification has been received. Guaranteed to be a substream that has been accepted with Network::accept_in_notifications.

notification: Vec<u8>

Notification that the remote has sent. The meaning of this data is out of scope of this module.

NotificationsInClose

The remote has closed an incoming notifications substream.

This can only happen after the notification substream has been accepted.

Fields

substream_id: SubstreamId

Substream that has been closed. Guaranteed to match a substream that was earlier reported with a Event::NotificationsInOpen.

outcome: Result<(), NotificationsInClosedErr>

Reason why the substream has been closed.

PingOutSuccess

An outgoing ping has succeeded. This event is generated automatically over time for each connection in the collection.

Fields

id: ConnectionId

ping_time: Duration

Time between sending the ping and receiving the pong.

PingOutFailed

An outgoing ping has failed. This event is generated automatically over time for each connection in the collection.

Fields

id: ConnectionId