Struct smoldot::libp2p::collection::Network

pub struct Network<TConn, TNow> { /* private fields */ }

Data structure containing the list of all connections, pending or not, and their latest known state. See also the module-level documentation.

Implementations

impl<TConn, TNow> Network<TConn, TNow>where TNow: Clone + Add<Duration, Output = TNow> + Sub<TNow, Output = Duration> + Ord,

pub fn new(config: Config) -> Self

Initializes a new network data structure.

pub fn insert_single_stream( &mut self, when_connected: TNow, handshake_kind: SingleStreamHandshakeKind<'_>, is_initiator: bool, user_data: TConn ) -> (ConnectionId, SingleStreamConnectionTask<TNow>)

Adds a new single-stream connection to the collection.

Must be passed the moment (as a TNow) when the connection as been established, in order to determine when the handshake timeout expires.

is_initiator must be true if the connection has been initiated locally, or false if it has been initiated by the remote.

pub fn insert_multi_stream<TSubId>( &mut self, now: TNow, handshake_kind: MultiStreamHandshakeKind<'_>, is_initiator: bool, user_data: TConn ) -> (ConnectionId, MultiStreamConnectionTask<TNow, TSubId>)where TSubId: Clone + PartialEq + Eq + Hash,

Adds a new multi-stream connection to the collection.

Must be passed the moment (as a TNow) when the connection as been established, in order to determine when the handshake timeout expires.

pub fn start_shutdown(&mut self, connection_id: ConnectionId)

Switches the connection to a state where it will shut down soon.

Calling this function does not generate a Event::StartShutdown event for this connection. The event is implied.

It is no longer possible to start requests, open notifications substreams, or send notifications. No new incoming requests or notification substreams will be reported. The incoming notifications that were sent by the remote before the shutdown started will still be reported.

It is possible to call this method on connection that are still in their handshaking phase.

This function generates a message destined to the connection. Use Network::pull_message_to_connection to process these messages after it has returned.

Panic

Panics if the connection is already shutting down, either because Network::start_shutdown was called or a Event::StartShutdown event was yielded.

pub fn is_empty(&self) -> bool

Returns true if the collection doesn’t contain any connection.

pub fn len(&self) -> usize

Returns the number of connections in the collection.

pub fn connection_state(&self, connection_id: ConnectionId) -> ConnectionState

Returns the state of the given connection.

Panic

Panics if the identifier is invalid or corresponds to a connection that has already entirely shut down.

pub fn notification_protocols( &self ) -> impl ExactSizeIterator<Item = &NotificationProtocolConfig>

Returns the list the overlay networks originally passed as Config::notification_protocols.

pub fn request_response_protocols( &self ) -> impl ExactSizeIterator<Item = &ConfigRequestResponse>

Returns the list the request-response protocols originally passed as Config::request_response_protocols.

pub fn start_request( &mut self, target: ConnectionId, protocol_index: usize, request_data: impl Into<Vec<u8>>, timeout: TNow ) -> Result<SubstreamId, StartRequestError>

Sends a request to the given peer.

A Event::Response event will later be generated containing the result of the request. This event is generated even if the connection the request was sent on has been closed.

It is invalid to start a request on a connection before a Event::HandshakeFinished or after a Event::StartShutdown has been generated, or after Network::start_shutdown has been called.

Returns a newly-allocated identifier for this substream.

This function generates a message destined to the connection. Use Network::pull_message_to_connection to process these messages after it has returned.

Requests

A request consists in:

Opening a substream on an established connection with the target.

Negotiating the requested protocol (protocol_index) on this substream using the multistream-select protocol.

Sending the request (request_data parameter), prefixed with its length.

Waiting for the response (prefixed with its length), which is then returned.

An error happens if the connection closes while the request is in progress, if the remote doesn’t support the given protocol, if the request or response doesn’t respect the protocol limits (see ConfigRequestResponse), or if the remote takes too much time to answer.

The timeout is the time between the moment the substream is opened and the moment the response is sent back. If the emitter doesn’t send the request or if the receiver doesn’t answer during this time window, the request is considered failed.

Panic

Panics if protocol_index isn’t a valid index in Config::request_response_protocols. Panics if the ConnectionId is invalid or is a connection that hasn’t finished its handshake or is shutting down.

pub fn open_out_notifications( &mut self, connection_id: ConnectionId, overlay_network_index: usize, now: TNow, handshake: impl Into<Vec<u8>> ) -> SubstreamId

Start opening a notifications substream.

It is invalid to open a notifications substream on a connection before a Event::HandshakeFinished or after a Event::StartShutdown has been generated, or after Network::start_shutdown has been called.

Returns a newly-allocated identifier for this substream.

This function generates a message destined to the connection. Use Network::pull_message_to_connection to process these messages after it has returned.

Panic

Panics if overlay_network_index isn’t a valid index in Config::notification_protocols. Panics if the ConnectionId is invalid or is a connection that hasn’t finished its handshake or is shutting down.

pub fn close_out_notifications(&mut self, substream_id: SubstreamId)

Start closing a previously-open notifications substream, or cancels opening a notifications substream.

All the notifications that have been queued remain queued. The substream actually closes only once the queue is empty.

Calling this method does not emit any event. The SubstreamId is considered invalid after this function returns.

This function generates a message destined to the connection. Use Network::pull_message_to_connection to process these messages after it has returned.

Panic

Panics if SubstreamId doesn’t correspond to an outbound notifications substream.

pub fn queue_notification( &mut self, substream_id: SubstreamId, notification: impl Into<Vec<u8>> ) -> Result<(), QueueNotificationError>

Adds a notification to the queue of notifications to send to the given peer.

It is invalid to call this on a SubstreamId before a successful Event::NotificationsOutResult has been yielded.

Each substream maintains a queue of notifications to be sent to the remote. This method attempts to push a notification to this queue.

An error is also returned if the queue exceeds a certain size in bytes, for two reasons:

Since the content of the queue is transferred at a limited rate, each notification pushed at the end of the queue will take more time than the previous one to reach the destination. Once the queue reaches a certain size, the time it would take for newly-pushed notifications to reach the destination would start being unreasonably large.

If the remote deliberately applies back-pressure on the substream, it is undesirable to increase the memory usage of the local node.

Similarly, the queue being full is a normal situation and notification protocols should be designed in such a way that discarding notifications shouldn’t have a too negative impact.

Regardless of the success of this function, no guarantee exists about the successful delivery of notifications.

This function generates a message destined to the connection. Use Network::pull_message_to_connection to process these messages after it has returned.

Panics

Panics if SubstreamId is not a fully open outbound notifications substream.

pub fn accept_in_notifications( &mut self, substream_id: SubstreamId, handshake: Vec<u8> )

Accepts a request for an inbound notifications substream reported by an Event::NotificationsInOpen.

If a Event::NotificationsInClose event is yielded, then this function must not be called and will panic.

This function generates a message destined to the connection. Use Network::pull_message_to_connection to process these messages after it has returned.

Panic

Panics if the SubstreamId doesn’t correspond to an inbound notifications substream.

pub fn reject_in_notifications(&mut self, substream_id: SubstreamId)

Rejects a request for an inbound notifications substream reported by an Event::NotificationsInOpen.

If a Event::NotificationsInClose event is yielded, then this function must not be called and will panic.

The SubstreamId is considered no longer valid after this function returns.

This function generates a message destined to the connection. Use Network::pull_message_to_connection to process these messages after it has returned.

Panic

Panics if the SubstreamId doesn’t correspond to an inbound notifications substream.

pub fn respond_in_request( &mut self, substream_id: SubstreamId, response: Result<Vec<u8>, ()> )

Responds to an incoming request. Must be called in response to a Event::RequestIn.

If the substream was in the meanwhile yielded in an Event::RequestInCancel, then this function must not be called and will panic.

The SubstreamId is considered no longer valid after this function returns.

This function generates a message destined to the connection. Use Network::pull_message_to_connection to process these messages after it has returned.

Panic

Panics if the SubstreamId doesn’t correspond to an active incoming request.

pub fn pull_message_to_connection( &mut self ) -> Option<(ConnectionId, CoordinatorToConnection<TNow>)>

Pulls a message that must be sent to a connection.

The message must be passed to SingleStreamConnectionTask::inject_coordinator_message or MultiStreamConnectionTask::inject_coordinator_message in the appropriate connection.

This function guarantees that the ConnectionId always refers to a connection that is still alive, in the sense that SingleStreamConnectionTask::inject_coordinator_message or MultiStreamConnectionTask::inject_coordinator_message has never returned None.

pub fn inject_connection_message( &mut self, connection_id: ConnectionId, message: ConnectionToCoordinator )

Injects into the state machine a message generated by SingleStreamConnectionTask::pull_message_to_coordinator or MultiStreamConnectionTask::pull_message_to_coordinator.

This message is queued and is later processed in Network::next_event. This means that it is Network::next_event and not Network::inject_connection_message that updates the internals of the state machine according to the content of the message. For example, if a SingleStreamConnectionTask sends a message to the coordinator indicating that a notifications substream has been closed, the coordinator will still believe that it is open until Network::next_event processes this message and at the same time returns a corresponding Event. Processing messages directly in Network::inject_connection_message would introduce “race conditions” where the API user can’t be sure in which state a connection or a substream is.

pub fn next_event(&mut self) -> Option<Event<TConn>>

Returns the next event produced by the service. Returns None if no event is available.

Call this function in a loop after having injected messages using Network::inject_connection_message.