smoldot::libp2p::connection::established

Enum Event

pub enum Event<TSubUd> {

    NewOutboundSubstreamsForbidden,
    InboundError(InboundError),
    InboundNegotiated {
        id: SubstreamId,
        protocol_name: String,
    },
    InboundNegotiatedCancel {
        id: SubstreamId,
    },
    InboundAcceptedCancel {
        id: SubstreamId,
        user_data: TSubUd,
    },
    RequestIn {
        id: SubstreamId,
        request: Vec<u8>,
    },
    Response {
        response: Result<Vec<u8>, RequestError>,
        id: SubstreamId,
        user_data: TSubUd,
    },
    NotificationsInOpen {
        id: SubstreamId,
        handshake: Vec<u8>,
    },
    NotificationsInOpenCancel {
        id: SubstreamId,
    },
    NotificationIn {
        id: SubstreamId,
        notification: Vec<u8>,
    },
    NotificationsInClose {
        id: SubstreamId,
        outcome: Result<(), NotificationsInClosedErr>,
        user_data: TSubUd,
    },
    NotificationsOutResult {
        id: SubstreamId,
        result: Result<Vec<u8>, (NotificationsOutErr, TSubUd)>,
    },
    NotificationsOutCloseDemanded {
        id: SubstreamId,
    },
    NotificationsOutReset {
        id: SubstreamId,
        user_data: TSubUd,
    },
    PingOutSuccess {
        ping_time: Duration,
    },
    PingOutFailed,

}

Event that happened on the connection. See SingleStream::read_write and MultiStream::pull_event.

Variants

NewOutboundSubstreamsForbidden

The connection is now in a mode where opening new substreams (i.e. starting requests and opening notifications substreams) is forbidden, but the remote is still able to open substreams and messages on existing substreams are still allowed to be sent and received.

InboundError(InboundError)

Received an incoming substream, but this substream has produced an error.

Note: This event exists only for diagnostic purposes. No action is expected in return.

InboundNegotiated

An inbound substream has requested to use a protocol. Call SingleStream::accept_inbound, SingleStream::reject_inbound, MultiStream::accept_inbound, or MultiStream::reject_inbound in order to accept or reject this substream.

Fields

id: SubstreamId

Identifier of the request. Needs to be provided back when accepting or rejecting the protocol.

protocol_name: String

Name of the protocol requested by the remote.

InboundNegotiatedCancel

An inbound substream that is waiting for a call to SingleStream::accept_inbound, SingleStream::reject_inbound, MultiStream::accept_inbound, or MultiStream::reject_inbound has been abruptly closed.

Fields

id: SubstreamId

Identifier of the substream.

InboundAcceptedCancel

An inbound substream that was previously accepted using SingleStream::accept_inbound or MultiStream::accept_inbound was closed by the remote or has generated an error.

Fields

id: SubstreamId

Identifier of the substream.

user_data: TSubUd

Value that was passed to SingleStream::accept_inbound or MultiStream::accept_inbound.

RequestIn

Received a request in the context of a request-response protocol.

Fields

id: SubstreamId

Identifier of the request. Needs to be provided back when answering the request.

request: Vec<u8>

Bytes of the request. Its interpretation is out of scope of this module.

Response

Received a response to a previously emitted request on a request-response protocol.

Fields

response: Result<Vec<u8>, RequestError>

Bytes of the response. Its interpretation is out of scope of this module.

id: SubstreamId

Identifier of the request. Value that was returned by SingleStream::add_request or MultiStream::add_request.

user_data: TSubUd

Value that was passed to SingleStream::add_request or MultiStream::add_request.

NotificationsInOpen

Remote has opened an inbound notifications substream.

Either SingleStream::accept_in_notifications_substream or SingleStream::reject_in_notifications_substream, or MultiStream::accept_in_notifications_substream or MultiStream::reject_in_notifications_substream must be called in the near future in order to accept or reject this substream.

Fields

id: SubstreamId

Identifier of the substream. Needs to be provided back when accept or rejecting the substream.

handshake: Vec<u8>

Handshake sent by the remote. Its interpretation is out of scope of this module.

NotificationsInOpenCancel

Remote has canceled an inbound notifications substream opening.

This can only happen after Event::NotificationsInOpen. SingleStream::accept_in_notifications_substream or SingleStream::reject_in_notifications_substream, or MultiStream::accept_in_notifications_substream or MultiStream::reject_in_notifications_substream should not be called on this substream.

Fields

id: SubstreamId

Identifier of the substream.

NotificationIn

Remote has sent a notification on an inbound notifications substream. Can only happen after the substream has been accepted.

Fields

id: SubstreamId

Identifier of the substream.

notification: Vec<u8>

Notification sent by the remote.

NotificationsInClose

Remote has closed an inbound notifications substream.Can only happen after the substream has been accepted.

Fields

id: SubstreamId

Identifier of the substream.

outcome: Result<(), NotificationsInClosedErr>

If Ok, the substream has been closed gracefully. If Err, a problem happened.

user_data: TSubUd

Value that was passed to SingleStream::accept_inbound or MultiStream::accept_inbound.

NotificationsOutResult

Outcome of trying to open a substream with SingleStream::open_notifications_substream or MultiStream::open_notifications_substream.

If Ok, it is now possible to send notifications on this substream. If Err, the substream no longer exists.

Fields

id: SubstreamId

Identifier of the substream. Value that was returned by SingleStream::open_notifications_substream or MultiStream::open_notifications_substream.

result: Result<Vec<u8>, (NotificationsOutErr, TSubUd)>

If Ok, contains the handshake sent back by the remote. Its interpretation is out of scope of this module.

NotificationsOutCloseDemanded

Remote has closed an outgoing notifications substream, meaning that it demands the closing of the substream.

Fields

id: SubstreamId

Identifier of the substream. Value that was returned by SingleStream::open_notifications_substream or MultiStream::open_notifications_substream.

NotificationsOutReset

Remote has reset an outgoing notifications substream. The substream is instantly closed.

Fields

id: SubstreamId

Identifier of the substream. Value that was returned by SingleStream::open_notifications_substream.

user_data: TSubUd

Value that was passed to SingleStream::open_notifications_substream or MultiStream::open_notifications_substream.

PingOutSuccess

An outgoing ping has succeeded. This event is generated automatically over time.

Fields

ping_time: Duration

Duration between sending the ping and receiving the pong.

PingOutFailed

An outgoing ping has failed. This event is generated automatically over time.