smoldot::libp2p::connection::established::substream

Enum Event

pub enum Event {

    InboundError {
        error: InboundError,
        was_accepted: bool,
    },
    InboundNegotiated(String),
    InboundNegotiatedCancel,
    RequestIn {
        request: Vec<u8>,
    },
    Response {
        response: Result<Vec<u8>, RequestError>,
    },
    NotificationsInOpen {
        handshake: Vec<u8>,
    },
    NotificationsInOpenCancel,
    NotificationIn {
        notification: Vec<u8>,
    },
    NotificationsInClose {
        outcome: Result<(), NotificationsInClosedErr>,
    },
    NotificationsOutResult {
        result: Result<Vec<u8>, NotificationsOutErr>,
    },
    NotificationsOutCloseDemanded,
    NotificationsOutReset,
    PingOutSuccess {
        ping_time: Duration,
    },
    PingOutError {
        num_pings: NonZero<usize>,
    },

}

Event that happened on the connection. See Substream::read_write.

Variants

InboundError

Error while receiving an inbound substream.

Fields

error: InboundError

was_accepted: bool

InboundNegotiated(String)

An inbound substream has successfully negotiated a protocol. Call Substream::accept_inbound or Substream::reject_inbound in order to resume.

InboundNegotiatedCancel

An inbound substream that had successfully negotiated a protocol got abruptly closed while waiting for the call to Substream::accept_inbound or Substream::reject_inbound.

RequestIn

Received a request in the context of a request-response protocol.

Fields

request: Vec<u8>

Bytes of the request. Its interpretation is out of scope of this module.

Response

Received a response to a previously emitted request on a request-response protocol.

Fields

response: Result<Vec<u8>, RequestError>

Bytes of the response. Its interpretation is out of scope of this module.

NotificationsInOpen

Remote has opened an inbound notifications substream.

Either Substream::accept_in_notifications_substream or Substream::reject_in_notifications_substream must be called in the near future in order to accept or reject this substream.

Fields

handshake: Vec<u8>

Handshake sent by the remote. Its interpretation is out of scope of this module.

NotificationsInOpenCancel

Remote has canceled an inbound notifications substream opening.

This can only happen after Event::NotificationsInOpen. Substream::accept_in_notifications_substream or Substream::reject_in_notifications_substream should not be called on this substream.

NotificationIn

Remote has sent a notification on an inbound notifications substream. Can only happen after the substream has been accepted.

Fields

notification: Vec<u8>

Notification sent by the remote.

NotificationsInClose

Remote has closed an inbound notifications substream opening. No more notifications will be received.

This can only happen after the substream has been accepted.

Fields

outcome: Result<(), NotificationsInClosedErr>

If Ok, the substream has been closed gracefully. If Err, a problem happened.

NotificationsOutResult

Remote has accepted or refused a substream opened with Substream::notifications_out.

If Ok, it is now possible to send notifications on this substream.

Fields

result: Result<Vec<u8>, NotificationsOutErr>

If Ok, contains the handshake sent back by the remote. Its interpretation is out of scope of this module.

NotificationsOutCloseDemanded

Remote has closed an outgoing notifications substream, meaning that it demands the closing of the substream.

NotificationsOutReset

Remote has reset an outgoing notifications substream. The substream is instantly closed.

PingOutSuccess

A ping has been successfully answered by the remote.

Fields

ping_time: Duration

Time between sending the ping and receiving the pong.

PingOutError

Remote has failed to answer one or more pings.

Fields

num_pings: NonZero<usize>

Number of pings that the remote has failed to answer.