smoldot::libp2p::connection::established::substream

Struct Substream

pub struct Substream<TNow> { /* private fields */ }

State machine containing the state of a single substream of an established connection.

Implementations

impl<TNow> Substream<TNow>
where TNow: Clone + Add<Duration, Output = TNow> + Sub<TNow, Output = Duration> + Ord,

pub fn ingoing(max_protocol_name_len: usize) -> Self

Initializes an new ingoing substream.

After the remote has requested a protocol, an Event::InboundNegotiated event will be generated, after which Substream::accept_inbound or Substream::reject_inbound must be called in order to indicate whether the protocol is accepted, and if so the nature of the negotiated protocol. A Event::InboundError can also be generated, either before or after the Event::InboundNegotiated, but always before any Event::NotificationsInOpen.

If InboundTy::Notifications is passed, then a Event::NotificationsInOpen will be generated (unless an error happens, in which case Event::InboundError). In response, the API user must call either Substream::accept_in_notifications_substream or Substream::reject_in_notifications_substream. Before one of these two methods is called, it is possible for an Event::NotificationsInOpenCancel to be generated, in which case the inbound request is canceled and the substream closed. After Substream::accept_in_notifications_substream is called, zero or more Event::NotificationIn will be generated, until a Event::NotificationsInClose which indicates the end of the substream.

If InboundTy::Request is passed, then a Event::RequestIn will be generated, after which the API user must call Substream::respond_in_request. An Event::InboundError can happen at any point.

This flow is also true if you call Substream::reset at any point.

pub fn notifications_out( timeout: TNow, requested_protocol: String, handshake: Vec<u8>, max_handshake_size: usize, ) -> Self

Initializes an outgoing notifications substream.

After the remote has sent back a handshake or after an error occurred, an Event::NotificationsOutResult event will be generated locally.

If this event contains an Ok, then Substream::write_notification_unbounded, Substream::notification_substream_queued_bytes and Substream::close_out_notifications_substream can be used, and Event::NotificationsOutCloseDemanded and Event::NotificationsOutReset can be generated.

pub fn request_out( requested_protocol: String, timeout: TNow, request: Option<Vec<u8>>, max_response_size: usize, ) -> Self

Initializes an outgoing request substream.

After the remote has sent back a response or after an error occurred, an Event::Response event will be generated locally. The user_data parameter will be passed back.

If the request is None, then nothing at all will be written out, not even a length prefix. If the request is Some, then a length prefix will be written out. Consequently, Some(&[]) writes a single 0 for the request.

pub fn ping_out(ping_protocol_name: String) -> Self

Initializes an outgoing ping substream.

Call Substream::queue_ping in order to queue an outgoing ping on this substream. This can be done at any time, even immediately after this function has returned.

The substream will attempt to negotiate the ping protocol. No event is reported if the protocol fails to negotiate. Instead, outgoing pings will be transparently failing.

Note: At the time of the writing of this comment, no API method exists to close an outgoing ping substream.

pub fn read_write( self, read_write: &mut ReadWrite<TNow>, ) -> (Option<Self>, Option<Event>)

Reads data coming from the socket, updates the internal state machine, and writes data destined to the socket through the read_write::ReadWrite.

If both the reading side and the writing side are closed and no other event can happen, or if at any point a protocol error happens, then None is returned. In that case, the substream must be reset if it is not closed.

pub fn reset(self) -> Option<Event>

pub fn accept_in_notifications_substream( &mut self, handshake: Vec<u8>, max_notification_size: usize, )

Accepts an inbound notifications protocol. Must be called in response to a Event::NotificationsInOpen.

# Panic

Panics if this substream is not of the correct type.

pub fn reject_in_notifications_substream(&mut self)

Rejects an inbound notifications protocol. Must be called in response to a Event::NotificationsInOpen.

# Panic

Panics if this substream is not of the correct type.

pub fn write_notification_unbounded(&mut self, notification: Vec<u8>)

Queues a notification to be written out on the given substream.

Panic

Panics if the substream isn’t a notifications substream, or if the notifications substream isn’t in the appropriate state.

pub fn notification_substream_queued_bytes(&self) -> usize

Returns the number of bytes waiting to be sent out on that substream.

See the documentation of Substream::write_notification_unbounded for context.

Panic

Panics if the substream isn’t a notifications substream, or if the notifications substream isn’t in the appropriate state.

pub fn close_out_notifications_substream(&mut self)

Closes a outgoing notifications substream opened after a successful Event::NotificationsOutResult.

This can be done even when in the negotiation phase, in other words before the remote has accepted/refused the substream.

Panic

Panics if the substream isn’t a notifications substream, or if the notifications substream isn’t in the appropriate state.

pub fn close_in_notifications_substream(&mut self, timeout: TNow)

Closes an ingoing notifications substream that was accepted using Substream::accept_in_notifications_substream.

Notifications can continue to be received. Calling this function only asynchronously signals to the remote that the substream should be closed. The closing is enforced only after the given timeout elapses.

Panic

Panics if the substream isn’t a notifications substream, or if the notifications substream isn’t in the appropriate state.

pub fn queue_ping(&mut self, payload: &[u8; 32], now: TNow, timeout: Duration)

Queues a ping on the given substream. Must be passed a randomly-generated payload of 32 bytes, the time after which this ping is considered as failed.

Panic

Panics if the substream isn’t an outgoing ping substream.

pub fn respond_in_request( &mut self, response: Result<Vec<u8>, ()>, ) -> Result<(), RespondInRequestError>

Responds to an incoming request. Must be called in response to a Event::RequestIn.

Passing an Err corresponds, on the other side, to a RequestError::SubstreamClosed.

pub fn accept_inbound(&mut self, ty: InboundTy)

Call after an Event::InboundNegotiated has been emitted in order to accept the protocol name and indicate the type of the protocol.

Panic

Panics if the substream is not in the correct state.

pub fn reject_inbound(&mut self)

Call after an Event::InboundNegotiated has been emitted in order to reject the protocol name as not supported.

Panic

Panics if the substream is not in the correct state.