Module smoldot::libp2p::connection::multistream_select

Multistream-select is a protocol whose purpose is to negotiate protocols.

Context

The multistream-select protocol makes it possible for two parties to negotiate a protocol.

When using TCP connections, it is used immediately after a connection opens in order to negotiate which encryption protocol to use, then after the encryption protocol handshake to negotiate which multiplexing protocol to use.

It is also used every time a substream opens in order to negotiate which protocol to use for this substream in particular.

Once a protocol has been negotiated, the connection or substream immediately starts speaking this protocol.

The multistream-select protocol is asymmetric: one side needs to be the dialer and the other side the listener. In the context of a TCP connection, the dialer and listener correspond to the dialer and listener of the connection. In the context of a substream, the dialer is the side that initiated the opening of the substream.

About protocol names

Due to flaws in the wire protocol design, a protocol named ls or na causes an ambiguity in the exchange. Because protocol names are normally decided ahead of time, this situation is expected to never arise, except in the presence of a malicious remote. The decision has been taken that such protocol will always fail to negotiate, but will also not produce any error or panic.

Please don’t intentionally name a protocol ls or na.

Usage

To be written.

See also

Official repository

Structs

InProgress

Negotiation in progress.

Enums

Config

Configuration of a multistream-select protocol.

Error

Error that can happen during the negotiation.

MessageOut

Message on the multistream-select protocol.

Negotiation

Current state of a multistream-select negotiation.