smoldot::libp2p::connection

Module noise

Noise protocol libp2p layer.

The noise protocol is a standard framework for building cryptographic protocols. Libp2p uses the noise protocol to provide an encryption layer on top of which data is exchanged.

Protocol details

Libp2p uses the XX pattern. The handshake consists of three packets:

The initiator generates an ephemeral key pair and sends the public key to the responder.

The responder generates its own ephemeral key pair and sends the public key to the initiator. Afterwards, the responder derives a shared secret and uses it to encrypt all further communications. Now encrypted, the responder also sends back its static noise public key (represented with the NoiseKey type of this module), its libp2p public key, and a signature of the static noise public key made using its libp2p private key.

The initiator, after having received the ephemeral key from the remote, derives the same shared secret. It sends its own static noise public key, libp2p public key, and signature.

After these three packets, the initiator and responder derive another shared secret using both the static and ephemeral keys, which is then used to encrypt communications. Note that the libp2p key isn’t used in the key derivation.

Usage

While this is out of scope of this module, the noise protocol must typically first be negotiated using the multistream-select protocol. The name of the protocol is given by the PROTOCOL_NAME constant.

In order to use noise on top of a connection which has agreed to use noise, create a HandshakeInProgress, passing a NoiseKey. This NoiseKey is typically generated at startup and doesn’t need to be persisted after a restart.

Use HandshakeInProgress::read_write when data is received from the wire or when the remote is ready to receive more data. At every call, a NoiseHandshake is returned, potentially indicating the end of the handshake.

If the handshake is finished, a NoiseHandshake::Success is returned, containing the PeerId of the remote, which is known to be legitimate, and a Noise object through which all further communications should go through.

Structs

Config

Configuration for a Noise handshake.

HandshakeInProgress

Handshake still in progress. More data needs to be sent or received.

InnerReadWrite

Stream of decrypted data. See Noise::read_write.

Noise

State of the noise encryption/decryption cipher.

NoiseKey

The noise key is the key exchanged during the noise handshake. It is not the same as the libp2p key. The libp2p key is used only to sign the noise public key, while the ECDH is performed with the noise key.

PayloadDecodeError

Error while decoding the handshake.

UnsignedNoiseKey

Prototype for a NoiseKey.

Enums

CipherError

Error while decoding data.

EncryptError

Error while encrypting data.

HandshakeError

Potential error during the noise handshake.

NoiseHandshake

State of a Noise handshake.

Constants

PROTOCOL_NAME

Name of the protocol, typically used when negotiated it using multistream-select.