Module smoldot::libp2p::connection::yamux

Yamux multiplexing protocol.

The Yamux protocol is a multiplexing protocol. As such, it allows dividing a single stream of data, typically a TCP socket, into multiple individual parallel substreams. The data sent and received over that single stream is divided into frames which, with the exception of ping and goaway frames, belong to a specific substream. In other words, the data transmitted over the substreams is interleaved.

Specification available at https://github.com/hashicorp/yamux/blob/master/spec.md

Usage

The Yamux object holds the state of all yamux-specific information, and the list of all currently-open substreams.

Call Yamux::incoming_data when data is available on the socket. This function parses the received data, updates the internal state machine, and possibly returns an IncomingDataDetail. Call Yamux::extract_out when the remote is ready to accept more data.

The generic parameter of Yamux is an opaque “user data” associated to each substream.

When SubstreamMut::write is called, the buffer of data to send out is stored within the Yamux object. This data will then be progressively returned by Yamux::extract_out.

It is the responsibility of the user to enforce a bound to the amount of enqueued data, as the Yamux itself doesn’t enforce any limit. Enforcing such a bound must be done based on the logic of the higher-level protocols. Failing to do so might lead to potential DoS attack vectors.

Structs

Config

Configuration for a new Yamux.

ExtractOut

IncomingDataOutcome

SubstreamId

Identifier of a substream in the context of a connection.

SubstreamMut

Reference to a substream within the Yamux.

SubstreamRef

Reference to a substream within the Yamux.

Yamux

Enums

DeadSubstreamTy

Error

Error while decoding the Yamux stream.

GoAwayErrorCode

IncomingDataDetail

Details about the incoming data.

Constants

PROTOCOL_NAME

Name of the protocol, typically used when negotiated it using multistream-select.