smoldot::libp2p::connection
Module yamux

Yamux multiplexing protocol.

The Yamux protocol is a multiplexing protocol. As such, it allows dividing a single stream of data, typically a TCP socket, into multiple individual parallel substreams. The data sent and received over that single stream is divided into frames which, with the exception of ping and goaway frames, belong to a specific substream. In other words, the data transmitted over the substreams is interleaved.

Specification available at https://github.com/hashicorp/yamux/blob/master/spec.md

Usage

The Yamux object holds the state of all yamux-specific information, and the list of all currently-open substreams.

The generic parameter of Yamux is an opaque “user data” associated to each substream.

Structs

Config

Configuration for a new Yamux.

SubstreamId

Identifier of a substream in the context of a connection.

SubstreamReadWrite

Yamux

Yamux state machine. See the module-level documentation for more information.

Enums

DeadSubstreamTy

Error

Error while decoding the Yamux stream.

GoAwayErrorCode

OpenSubstreamError

Error potentially returned by Yamux::open_substream.

PendingSubstreamError

Error potentially returned by Yamux::accept_pending_substream or Yamux::reject_pending_substream.

ReadWriteOutcome

Details about the incoming data.

ResetError

Error potentially returned by Yamux::reset.

SendGoAwayError

Error potentially returned by Yamux::send_goaway.

Constants

MAX_PINGS

Maximum number of simultaneous outgoing pings allowed.

NEW_SUBSTREAMS_FRAME_SIZE

By default, all new substreams have this implicit window size.

PROTOCOL_NAME

Name of the protocol, typically used when negotiated it using multistream-select.