smoldot::libp2p::connection::yamux

Struct Yamux

pub struct Yamux<TNow, TSub> { /* private fields */ }

Yamux state machine. See the module-level documentation for more information.

Implementations

impl<TNow, TSub> Yamux<TNow, TSub>

pub fn new(config: Config) -> Yamux<TNow, TSub>

Initializes a new Yamux state machine.

pub fn is_empty(&self) -> bool

Returns true if there is no substream in the state machine.

Note: After a substream has been closed or reset, it must be removed using Yamux::remove_dead_substream before this function can return true.

pub fn len(&self) -> usize

Returns the number of substreams in the Yamux state machine. Includes substreams that are dead but haven’t been removed yet.

pub fn num_inbound(&self) -> usize

Returns the number of inbound substreams in the Yamux state machine. Includes substreams that are dead but haven’t been removed yet.

pub fn received_goaway(&self) -> Option<GoAwayErrorCode>

Returns Some if a ReadWriteOutcome::GoAway event has been generated in the past, in which case the code is returned.

If Some is returned, it is forbidden to open new outbound substreams.

pub fn user_datas(&self) -> impl ExactSizeIterator<Item = (SubstreamId, &TSub)>

Returns an iterator to the list of all substream user datas.

pub fn user_datas_mut( &mut self, ) -> impl ExactSizeIterator<Item = (SubstreamId, &mut TSub)>

Returns an iterator to the list of all substream user datas.

pub fn has_substream(&self, substream_id: SubstreamId) -> bool

Returns true if the given SubstreamId exists.

Also returns true if the substream is in a dead state.

pub fn goaway_queued_or_sent(&self) -> bool

Returns true if Yamux::send_goaway has been called in the past.

In other words, returns true if a GoAway frame has been either queued for sending (and is available through Yamux::read_write) or has already been sent out.

pub fn goaway_sent(&self) -> bool

Returns true if Yamux::send_goaway has been called in the past and that this GoAway frame has been extracted through Yamux::read_write.

impl<TNow, TSub> Yamux<TNow, TSub>
where TNow: Clone + Ord,

pub fn open_substream( &mut self, user_data: TSub, ) -> Result<SubstreamId, OpenSubstreamError>

Opens a new substream.

This method only modifies the state of self and reserves an identifier. No message needs to be sent to the remote before data is actually being sent on the substream.

The substream will be automatically processed by Yamux::read_write in the future.

Note: Importantly, the remote will not be notified of the substream being open before the local side sends data on this substream. As such, protocols where the remote is expected to send data in response to a substream being open, without the local side first sending some data on that substream, will not work. In practice, while this is technically out of concern of the Yamux protocol, all substreams in the context of libp2p start with a multistream-select negotiation, and this scenario can therefore never happen.

Returns an error if a ReadWriteOutcome::GoAway event has been generated. This can also be checked by calling Yamux::received_goaway.

Returns an error if all possible substream IDs are already taken. This happen if there exists more than approximately 2^31 substreams, which is very unlikely to happen unless there exists a bug in the code.

pub fn mark_substream_write_ready(&mut self, substream_id: SubstreamId)

Marks the given substream as being ready to write out data.

Calling this function is necessary in situations where a substream didn’t write any data in the past, but some sort of manual state update will make it write data in the future.

Has no effect if the substream has been reset or closed.

Panic

Panics if the SubstreamId is invalid.

pub fn read_write( self, outer_read_write: &mut ReadWrite<TNow>, ) -> Result<ReadWriteOutcome<'_, TNow, TSub>, Error>

Feeds data coming from a socket and outputs data to write to the socket.

Returns an object that implements Deref<Target = ReadWrite>. This object represents the decrypted stream of data.

An error is returned if the protocol is being violated by the remote or if the nonce overflows. When that happens, the connection should be closed altogether.

pub fn add_remote_window_saturating( &mut self, substream_id: SubstreamId, bytes: u64, )

Adds bytes to the number of bytes the remote is allowed to send at once in the next packet.

Note: The SubstreamReadWrite object ensures that the remote window is big enough to receive the amount of bytes requested by the substream. Calling Yamux::add_remote_window_saturating is therefore not mandatory, but only an optimization to reduce the number of networking round trips.

The counter saturates if its maximum is reached. This could cause stalls if the remote sends a ton of data. However, given that the number of bytes is stored in a u64, the remote would have to send at least 2^64 bytes in order to reach this situation, making it basically impossible.

It is, furthermore, a bad idea to increase this counter by an immense number ahead of time, as the remote can shut down the connection if its own counter overflows. The way this counter is supposed to be used in a “streaming” way.

Increasing the window of a substream by bytes requires sending out 12 * bytes / 2^32 bytes. For example, increasing the window by 2^64 adds an overhead of 48 GiB. Please don’t do that.

Note: When a substream has just been opened or accepted, it starts with an initial window of NEW_SUBSTREAMS_FRAME_SIZE.

Note: It is only possible to add more bytes to the window and not set or reduce this number of bytes, and it is also not possible to obtain the number of bytes the remote is allowed. That’s because it would be ambiguous whether bytes possibly in the send or receive queue should be counted or not.

Has no effect if the remote has already closed their writing side.

Panic

Panics if the SubstreamId is invalid.

pub fn reset(&mut self, substream_id: SubstreamId) -> Result<(), ResetError>

Abruptly shuts down the substream. Sends a frame with the RST flag to the remote.

Use this method when a protocol error happens on a substream.

Returns an error if Yamux::reset has already been called on this substream, or if the remote has reset the substream in the past, or if the substream was closed.

Panic

Panics if the SubstreamId is invalid.

pub fn queue_ping(&mut self)

Queues sending out a ping to the remote.

Panic

Panics if there are already MAX_PINGS pings that have been queued and that the remote hasn’t answered yet. MAX_PINGS is pretty large, and unless there is a bug in the API user’s code causing pings to be allocated in a loop, this limit is not likely to ever be reached.

pub fn send_goaway( &mut self, code: GoAwayErrorCode, ) -> Result<(), SendGoAwayError>

Queues a GoAway frame, requesting the remote to no longer open any substream.

If the state of Yamux is currently waiting for a confirmation to accept/reject a substream, then this function automatically implies calling Yamux::reject_pending_substream.

All follow-up requests for new substreams from the remote are automatically rejected. ReadWriteOutcome::IncomingSubstream events can no longer happen.

pub fn dead_substreams( &self, ) -> impl Iterator<Item = (SubstreamId, DeadSubstreamTy, &TSub)>

Returns the list of all substreams that have been closed or reset.

This function does not remove dead substreams from the state machine. In other words, if this function is called multiple times in a row, it will always return the same substreams. Use Yamux::remove_dead_substream to remove substreams.

pub fn remove_dead_substream(&mut self, id: SubstreamId) -> TSub

Removes a dead substream from the state machine.

Panic

Panics if the substream with that id doesn’t exist or isn’t dead.

pub fn accept_pending_substream( &mut self, user_data: TSub, ) -> Result<SubstreamId, PendingSubstreamError>

Accepts an incoming substream.

Either Yamux::accept_pending_substream or Yamux::reject_pending_substream must be called after ReadWriteOutcome::IncomingSubstream is returned.

Note that there is no expiration window after ReadWriteOutcome::IncomingSubstream is returned until the substream is no longer valid. However, reading will be blocked until the substream is either accepted or rejected. This function should thus be called as soon as possible.

Returns an error if no incoming substream is currently pending.

pub fn reject_pending_substream(&mut self) -> Result<(), PendingSubstreamError>

Rejects an incoming substream.

Either Yamux::accept_pending_substream or Yamux::reject_pending_substream must be called after ReadWriteOutcome::IncomingSubstream is returned.

Note that there is no expiration window after ReadWriteOutcome::IncomingSubstream is returned until the substream is no longer valid. However, reading will be blocked until the substream is either accepted or rejected. This function should thus be called as soon as possible.

Returns an error if no incoming substream is currently pending.