Struct smoldot::libp2p::connection::yamux::Yamux

pub struct Yamux<T> { /* private fields */ }

Yamux state machine. See the module-level documentation for more information.

Implementations

impl<T> Yamux<T>

pub fn new(config: Config) -> Yamux<T>

Initializes a new Yamux state machine.

pub fn is_empty(&self) -> bool

Returns true if there is no substream in the state machine.

Note: After a substream has been closed or reset, it must be removed using Yamux::remove_dead_substream before this function can return true.

pub fn len(&self) -> usize

Returns the number of substreams in the Yamux state machine. Includes substreams that are dead but haven’t been removed yet.

pub fn num_inbound(&self) -> usize

Returns the number of inbound substreams in the Yamux state machine. Includes substreams that are dead but haven’t been removed yet.

pub fn open_substream( &mut self, user_data: T ) -> Result<SubstreamId, OpenSubstreamError>

Opens a new substream.

This method only modifies the state of self and reserves an identifier. No message needs to be sent to the remote before data is actually being sent on the substream.

Note: Importantly, the remote will not be notified of the substream being open before the local side sends data on this substream. As such, protocols where the remote is expected to send data in response to a substream being open, without the local side first sending some data on that substream, will not work. In practice, while this is technically out of concern of the Yamux protocol, all substreams in the context of libp2p start with a multistream-select negotiation, and this scenario can therefore never happen.

Returns an error if a IncomingDataDetail::GoAway event has been generated. This can also be checked by calling Yamux::received_goaway.

Returns an error if all possible substream IDs are already taken. This happen if there exists more than approximately 2^31 substreams, which is very unlikely to happen unless there exists a bug in the code.

pub fn received_goaway(&self) -> Option<GoAwayErrorCode>

Returns Some if a IncomingDataDetail::GoAway event has been generated in the past, in which case the code is returned.

If Some is returned, it is forbidden to open new outbound substreams.

pub fn user_datas(&self) -> impl ExactSizeIterator<Item = (SubstreamId, &T)>

Returns an iterator to the list of all substream user datas.

pub fn user_datas_mut( &mut self ) -> impl ExactSizeIterator<Item = (SubstreamId, &mut T)>

Returns an iterator to the list of all substream user datas.

pub fn has_substream(&self, substream_id: SubstreamId) -> bool

Returns true if the given SubstreamId exists.

Also returns true if the substream is in a dead state.

pub fn user_data(&self, substream_id: SubstreamId) -> &T

Returns the user data associated to a substream.

Panic

Panics if the SubstreamId is invalid.

pub fn user_data_mut(&mut self, substream_id: SubstreamId) -> &mut T

Returns the user data associated to a substream.

Panic

Panics if the SubstreamId is invalid.

pub fn write( &mut self, substream_id: SubstreamId, data: Vec<u8> ) -> Result<(), WriteError>

Appends data to the buffer of data to send out on this substream.

Returns an error if Yamux::close or Yamux::reset has been called on this substream, or if the substream has been reset by the remote.

Panic

Panics if the SubstreamId is invalid.

pub fn add_remote_window_saturating( &mut self, substream_id: SubstreamId, bytes: u64 )

Adds bytes to the number of bytes the remote is allowed to send at once in the next packet.

The counter saturates if its maximum is reached. This could cause stalls if the remote sends a ton of data. However, given that the number of bytes is stored in a u64, the remote would have to send at least 2^64 bytes in order to reach this situation, making it basically impossible.

It is, furthermore, a bad idea to increase this counter by an immense number ahead of time, as the remote can shut down the connection if its own counter overflows. The way this counter is supposed to be used is in a “streaming” way.

Note: When a substream has just been opened or accepted, it starts with an initial window of NEW_SUBSTREAMS_FRAME_SIZE.

Note: It is only possible to add more bytes to the window and not set or reduce this number of bytes, and it is also not possible to obtain the number of bytes the remote is allowed. That’s because it would be ambiguous whether bytes possibly in the send or receive queue should be counted or not.

Has no effect if the remote has already closed their writing side.

Panic

Panics if the SubstreamId is invalid.

pub fn queued_bytes(&self, substream_id: SubstreamId) -> usize

Returns the number of bytes queued for writing on this substream.

Returns 0 if the substream is in a reset state.

Note: Might return non-zero even if Yamux::close has been called, as this counts the number of bytes still waiting to be written out.

Panic

Panics if the SubstreamId is invalid.

pub fn can_receive(&self, substream_id: SubstreamId) -> bool

Returns false if the remote has closed their writing side of this substream, or if Yamux::reset has been called on this substream, or if the substream has been reset by the remote.

Panic

Panics if the SubstreamId is invalid.

pub fn can_send(&self, substream_id: SubstreamId) -> bool

Returns false if Yamux::close or Yamux::reset has been called on this substream, or if the remote has .

Panic

Panics if the SubstreamId is invalid.

pub fn close(&mut self, substream_id: SubstreamId) -> Result<(), CloseError>

Marks the substream as closed. It is no longer possible to write data on it.

Returns an error if the local writing side is already closed, which can happen if Yamux::close has already been called on this substream. Returns an error if Yamux::reset has been called on this substream, or if the remote has reset the substream in the past.

Panic

Panics if the SubstreamId is invalid.

pub fn reset(&mut self, substream_id: SubstreamId) -> Result<(), ResetError>

Abruptly shuts down the substream. Sends a frame with the RST flag to the remote.

Use this method when a protocol error happens on a substream.

Returns an error if Yamux::reset has already been called on this substream or if the remote has reset the substream in the past.

Panic

Panics if the SubstreamId is invalid.

pub fn queue_ping(&mut self)

Queues sending out a ping to the remote.

Panic

Panics if there are already MAX_PINGS pings that have been queued and that the remote hasn’t answered yet. MAX_PINGS is pretty large, and unless there is a bug in the API user’s code causing pings to be allocated in a loop, this limit is not likely to ever be reached.

pub fn goaway_queued_or_sent(&self) -> bool

Returns true if Yamux::send_goaway has been called in the past.

In other words, returns true if a GoAway frame has been either queued for sending (and is available through Yamux::extract_next) or has already been sent out.

pub fn goaway_sent(&self) -> bool

Returns true if Yamux::send_goaway has been called in the past and that this GoAway frame has been extracted through Yamux::extract_next.

pub fn send_goaway( &mut self, code: GoAwayErrorCode ) -> Result<(), SendGoAwayError>

Queues a GoAway frame, requesting the remote to no longer open any substream.

If the state of Yamux is currently waiting for a confirmation to accept/reject a substream, then this function automatically implies calling Yamux::reject_pending_substream.

All follow-up requests for new substreams from the remote are automatically rejected. IncomingDataDetail::IncomingSubstream events can no longer happen.

pub fn dead_substreams( &self ) -> impl Iterator<Item = (SubstreamId, DeadSubstreamTy, &T)> + '_

Returns the list of all substreams that have been closed or reset.

This function does not remove dead substreams from the state machine. In other words, if this function is called multiple times in a row, it will always return the same substreams. Use Yamux::remove_dead_substream to remove substreams.

pub fn remove_dead_substream(&mut self, id: SubstreamId) -> T

Removes a dead substream from the state machine.

Panic

Panics if the substream with that id doesn’t exist or isn’t dead.

pub fn incoming_data(self, data: &[u8]) -> Result<IncomingDataOutcome<T>, Error>

Process some incoming data.

This function takes ownership of self and yields it back if everything goes well. If, on the other hand, a malformed packet is received, an error is yielded and self is destroyed.

This function might not process all the data available for one of the following reasons:

Not all outgoing data has been extracted. In order to process incoming messages, the Yamux might have to queue data to be written out. For example, incoming pings must be replied to. In order to avoid queue an infinite amount of data, processing incoming messages might be blocked if there is data to be sent out.

It is currently waiting for either Yamux::accept_pending_substream or Yamux::reject_pending_substream to be called.

If the remote opens a substream whose ID is equal to a previous substream that is now dead. Use Yamux::dead_substreams and Yamux::remove_dead_substream to remove dead substreams before continuing.

If the return value contains IncomingDataDetail::IncomingSubstream, then either Yamux::accept_pending_substream or Yamux::reject_pending_substream must be called in order to accept or reject the pending substream. API users are encouraged to enforce a limit to the total number of substreams in order to clamp the memory usage of this state machine.

pub fn extract_next( &mut self, size_bytes: usize ) -> Option<impl AsRef<[u8]> + '_>

Builds the next buffer to send out on the socket and returns it.

The buffer will never be larger than size_bytes bytes. The user is expected to pass an exact amount of bytes that the next layer is ready to accept.

Note: Most other objects in the networking code have a “read_write” method that writes the outgoing data to a buffer. This is an idiomatic way to do things in situations where the data is generated on the fly. In the context of Yamux, however, this would be rather sub-optimal considering that buffers to send out are already stored in their final form in the state machine.

pub fn accept_pending_substream( &mut self, user_data: T ) -> Result<SubstreamId, PendingSubstreamError>

Accepts an incoming substream.

Either Yamux::accept_pending_substream or Yamux::reject_pending_substream must be called after IncomingDataDetail::IncomingSubstream is returned.

Note that there is no expiration window after IncomingDataDetail::IncomingSubstream is returned until the substream is no longer valid. However, reading will be blocked until the substream is either accepted or rejected. This function should thus be called as soon as possible.

Returns an error if no incoming substream is currently pending.

pub fn reject_pending_substream(&mut self) -> Result<(), PendingSubstreamError>

Rejects an incoming substream.

Either Yamux::accept_pending_substream or Yamux::reject_pending_substream must be called after IncomingDataDetail::IncomingSubstream is returned.

Note that there is no expiration window after IncomingDataDetail::IncomingSubstream is returned until the substream is no longer valid. However, reading will be blocked until the substream is either accepted or rejected. This function should thus be called as soon as possible.

Returns an error if no incoming substream is currently pending.