Module smoldot::libp2p::peers

Network of peers.

The Peers state machine builds on top of the libp2p module and provides an abstraction over the network based on network identities (i.e. PeerIds). One can set the list of peers to be connected to and through which notification protocols, and the Peers struct will try to open or re-open connections with these peers. Once connected, one can use the Peers to send request or notifications with these peers.

Detailed usage

The Peers struct contains six different collections:

A list, decided by the API user, of peers that are marked as “desired”.

A list, decided by the API user, of (peer_id, notification_protocol) tuples that are marked as “desired”.

A list of connections identified by ConnectionIds.

A list of requests for inbound substreams, identified by a SubstreamId. When a peer desires to open a notification substream with the local node, a SubstreamId is generated. The API user must answer by either accepting or refusing the request.

A list of requests that have been received, identified by a InRequestId. The API user must answer by calling Peers::respond_in_request. Requests can automatically become obsolete if the remote decides to withdraw their request or the connection closes. A request becoming obsolete does not invalidate its InRequestId.

Re-exports

pub use collection::ConfigRequestResponse;

pub use collection::ConfigRequestResponseIn;

pub use collection::ConnectionId;

pub use collection::ConnectionToCoordinator;

pub use collection::CoordinatorToConnection;

pub use collection::MultiStreamHandshakeKind;

pub use collection::NotificationProtocolConfig;

pub use collection::NotificationsInClosedErr;

pub use collection::NotificationsOutErr;

pub use collection::ReadWrite;

pub use collection::RequestError;

pub use collection::SingleStreamHandshakeKind;

pub use collection::SubstreamId;

Structs

Config

Configuration for a Peers.

ConnectionState

See Peers::connection_state.

InRequestId

See Event::RequestIn.

MultiStreamConnectionTask

State machine dedicated to a single multi-stream connection.

OutRequestId

See Peers::start_request.

Peers

SingleStreamConnectionTask

State machine dedicated to a single single-stream connection.

Enums

DesiredState

See Peers::set_peer_notifications_out_desired.

Event

Event happening over the network. See Peers::next_event.

InNotificationAcceptError

Error potentially returned by Peers::in_notification_accept.

InboundError

Error that can happen while processing an inbound substream.

OpenOrPending

QueueNotificationError

Error potentially returned by Peers::queue_notification.

ShutdownCause

Reason why a connection is shutting down. See Event::StartShutdown.

ShutdownPeer

See Event::StartShutdown and Event::Shutdown.