Enum smoldot::libp2p::peers::Event

pub enum Event<TConn> {

    HandshakeFinished {
        connection_id: ConnectionId,
        peer_id: PeerId,
        expected_peer_id: Option<PeerId>,
        num_healthy_peer_connections: NonZeroU32,
    },
    StartShutdown {
        connection_id: ConnectionId,
        peer: ShutdownPeer,
        reason: ShutdownCause,
    },
    Shutdown {
        connection_id: ConnectionId,
        peer: ShutdownPeer,
        user_data: TConn,
    },
    InboundError {
        peer_id: PeerId,
        connection_id: ConnectionId,
        error: InboundError,
    },
    Response {
        request_id: OutRequestId,
        response: Result<Vec<u8>, RequestError>,
    },
    RequestIn {
        request_id: InRequestId,
        peer_id: PeerId,
        connection_id: ConnectionId,
        protocol_index: usize,
        request_payload: Vec<u8>,
    },
    RequestInCancel {
        id: InRequestId,
    },
    NotificationsInOpen {
        id: SubstreamId,
        peer_id: PeerId,
        notifications_protocol_index: usize,
        handshake: Vec<u8>,
    },
    NotificationsInOpenCancel {
        id: SubstreamId,
    },
    NotificationsOutResult {
        peer_id: PeerId,
        notifications_protocol_index: usize,
        result: Result<Vec<u8>, NotificationsOutErr>,
    },
    NotificationsOutClose {
        peer_id: PeerId,
        notifications_protocol_index: usize,
    },
    NotificationsIn {
        peer_id: PeerId,
        notifications_protocol_index: usize,
        notification: Vec<u8>,
    },
    NotificationsInClose {
        peer_id: PeerId,
        notifications_protocol_index: usize,
        outcome: Result<(), NotificationsInClosedErr>,
    },

}

Event happening over the network. See Peers::next_event.

Variants

HandshakeFinished

Fields

connection_id: ConnectionId

Identifier of the connection that has finished its handshake.

peer_id: PeerId

Identity of the peer on the other side of the connection.

expected_peer_id: Option<PeerId>

Identity of the peer that was expected to be reached.

Always Some for outgoing connections and always None for incoming connections.

num_healthy_peer_connections: NonZeroU32

Number of established not-shutting-down connections with the same peer, including the one that has just been established.

Connection has finished its handshake.

Only generated for single-stream connections. The handshake of multi-stream connections is considered to be already finished.

StartShutdown

Fields

connection_id: ConnectionId

Identifier of the connection that has started shutting down.

peer: ShutdownPeer

State of the connection and identity of the remote.

reason: ShutdownCause

Reason for the shutdown.

Shutdown

Fields

connection_id: ConnectionId

Identifier of the connection that has started shutting down.

peer: ShutdownPeer

State of the connection and identity of the remote.

user_data: TConn

User data that was associated to this connection.

A connection has stopped.

InboundError

Fields

peer_id: PeerId

Peer which opened the substream.

connection_id: ConnectionId

Identifier of the connection on which the problem happened.

error: InboundError

Error that happened.

Received an incoming substream, but this substream has produced an error.

Note: This event exists only for diagnostic purposes. No action is expected in return.

Response

Fields

request_id: OutRequestId

Identifier for this request. Was returned by Peers::start_request.

response: Result<Vec<u8>, RequestError>

Outcome of a request started using Peers::start_request.

All requests always lead to an outcome, even if the connection has been closed while the request was in progress.

RequestIn

Fields

request_id: InRequestId

Identifier for this request. Must be passed back when calling Peers::respond_in_request.

peer_id: PeerId

Peer which sent the request.

connection_id: ConnectionId

Identifier of the connection that has sent the request.

protocol_index: usize

Request-response protocol the request is about.

request_payload: Vec<u8>

Payload of the request, opaque to this state machine.

Note: Keep in mind that this data is untrusted.

Received a request from a request-response protocol.

RequestInCancel

Fields

id: InRequestId

Identifier for this request.

A previously-emitted Event::RequestIn is now obsolete.

The InRequestId is now considered dead, and calling Peers::respond_in_request is now invalid.

NotificationsInOpen

Fields

id: SubstreamId

Identifier for this request. Must be passed back when calling Peers::in_notification_accept or Peers::in_notification_refuse.

peer_id: PeerId

Peer which tries to open an inbound substream.

notifications_protocol_index: usize

Notifications protocol the substream is about.

handshake: Vec<u8>

Handshake of the request sent by the peer. Opaque to this state machine.

Note: Keep in mind that this data is untrusted.

A peer would like to open a notifications substream with the local node, in order to send notifications.

Only one inbound notifications substream can exist per peer and per protocol. Any additional one will be automatically refused.

NotificationsInOpenCancel

Fields

id: SubstreamId

Identifier for this request.

A previously-emitted Event::NotificationsInOpen is now obsolete. The SubstreamId is no longer valid.

NotificationsOutResult

Fields

peer_id: PeerId

Peer the substream is open with.

notifications_protocol_index: usize

Notifications protocol the substream is about.

result: Result<Vec<u8>, NotificationsOutErr>

If Ok, contains the handshake sent back by the remote. Its interpretation is out of scope of this module. If Err, the state machine will not automatically try to re-open a substream again. Use Peers::set_peer_notifications_out_desired with DesiredState::DesiredReset in order to try again.

A handshaking outbound substream has been accepted by the remote.

Will happen for combinations of PeerId and notification protocols that have been marked as desired. Can also happen for other combinations, if there were marked as desired in the past but no longer are.

If Ok, it is now possible to send notifications on this substream.

Note: No event if generated when Peers::close_out_notification is called.

NotificationsOutClose

Fields

peer_id: PeerId

Peer the substream is no longer open with.

notifications_protocol_index: usize

Notifications protocol the substream is about.

A previously open outbound substream has been closed by the remote. Can only happen after a corresponding successful Event::NotificationsOutResult event has been emitted in the past.

Note: No event if generated when Peers::close_out_notification is called.

NotificationsIn

Fields

peer_id: PeerId

Peer that sent the notification.

notifications_protocol_index: usize

Notifications protocol the substream is about.

notification: Vec<u8>

Payload of the notification. Opaque to this state machine.

Note: Keep in mind that this data is untrusted.

Received a notification on a notifications substream of a connection.

NotificationsInClose

Fields

peer_id: PeerId

Peer the substream is no longer with.

notifications_protocol_index: usize

Notifications protocol the substream is about.

outcome: Result<(), NotificationsInClosedErr>

If Ok, the substream has been closed gracefully. If Err, a problem happened.

Remote has closed a previously-open inbound notifications substream.