Struct smoldot::libp2p::peers::Peers

pub struct Peers<TConn, TNow> { /* private fields */ }

Implementations

impl<TConn, TNow> Peers<TConn, TNow>where TConn: Clone, TNow: Clone + Add<Duration, Output = TNow> + Sub<TNow, Output = Duration> + Ord,

pub fn new(config: Config) -> Self

Creates a new Peers.

pub fn notification_protocols( &self ) -> impl ExactSizeIterator<Item = &NotificationProtocolConfig>

Returns the list the overlay networks originally passed as Config::notification_protocols.

pub fn request_response_protocols( &self ) -> impl ExactSizeIterator<Item = &ConfigRequestResponse>

Returns the list the request-response protocols originally passed as Config::request_response_protocols.

pub fn noise_key(&self) -> &NoiseKey

Returns the Noise key originally passed as Config::noise_key.

pub fn pull_message_to_connection( &mut self ) -> Option<(ConnectionId, CoordinatorToConnection<TNow>)>

Pulls a message that must be sent to a connection.

The message must be passed to SingleStreamConnectionTask::inject_coordinator_message or MultiStreamConnectionTask::inject_coordinator_message in the appropriate connection.

This function guarantees that the ConnectionId always refers to a connection that is still alive, in the sense that SingleStreamConnectionTask::inject_coordinator_message or MultiStreamConnectionTask::inject_coordinator_message has never returned None.

pub fn inject_connection_message( &mut self, connection_id: ConnectionId, message: ConnectionToCoordinator )

Injects into the state machine a message generated by SingleStreamConnectionTask::pull_message_to_coordinator or MultiStreamConnectionTask::pull_message_to_coordinator.

This message is queued and is later processed in Peers::next_event. This means that it is Peers::next_event and not Peers::inject_connection_message that updates the internals of the state machine according to the content of the message. For example, if a SingleStreamConnectionTask sends a message to the coordinator indicating that a notifications substream has been closed, the coordinator will still believe that it is open until Peers::next_event processes this message and at the same time returns a corresponding Event. Processing messages directly in Peers::inject_connection_message would introduce “race conditions” where the API user can’t be sure in which state a connection or a substream is.

pub fn next_event(&mut self) -> Option<Event<TConn>>

Returns the next event produced by the service.

pub fn add_single_stream_incoming_connection( &mut self, when_connected: TNow, handshake_kind: SingleStreamHandshakeKind, user_data: TConn ) -> (ConnectionId, SingleStreamConnectionTask<TNow>)

Inserts a single-stream incoming connection in the state machine.

This connection hasn’t finished handshaking and the PeerId of the remote isn’t known yet.

Must be passed the moment (as a TNow) when the connection as been established, in order to determine when the handshake timeout expires.

pub fn add_single_stream_outgoing_connection( &mut self, when_connected: TNow, handshake_kind: SingleStreamHandshakeKind, expected_peer_id: &PeerId, user_data: TConn ) -> (ConnectionId, SingleStreamConnectionTask<TNow>)

Inserts a single-stream outgoing connection in the state machine.

This connection hasn’t finished handshaking, and the PeerId of the remote isn’t known yet, but it is expected to be in unfulfilled_desired_peers. After this function has been called, the provided expected_peer_id will no longer be part of the return value of Peers::unfulfilled_desired_peers.

Must be passed the moment (as a TNow) when the connection as been established, in order to determine when the handshake timeout expires.

pub fn add_multi_stream_incoming_connection<TSubId>( &mut self, when_connected: TNow, handshake_kind: MultiStreamHandshakeKind, user_data: TConn ) -> (ConnectionId, MultiStreamConnectionTask<TNow, TSubId>)where TSubId: Clone + PartialEq + Eq + Hash,

Inserts a multi-stream incoming connection in the state machine.

This connection hasn’t finished handshaking and the PeerId of the remote isn’t known yet.

Must be passed the moment (as a TNow) when the connection as been established, in order to determine when the handshake timeout expires.

pub fn add_multi_stream_outgoing_connection<TSubId>( &mut self, when_connected: TNow, handshake_kind: MultiStreamHandshakeKind, expected_peer_id: &PeerId, user_data: TConn ) -> (ConnectionId, MultiStreamConnectionTask<TNow, TSubId>)where TSubId: Clone + PartialEq + Eq + Hash,

Inserts a multi-stream outgoing connection in the state machine.

This connection hasn’t finished handshaking, and the PeerId of the remote isn’t known yet, but it is expected to be in unfulfilled_desired_peers. After this function has been called, the provided expected_peer_id will no longer be part of the return value of Peers::unfulfilled_desired_peers.

Must be passed the moment (as a TNow) when the connection as been established, in order to determine when the handshake timeout expires.

pub fn established_peer_connections( &self, peer_id: &PeerId ) -> impl Iterator<Item = ConnectionId> + '_

Returns all the non-handshaking connections that are connected to the given peer. The list also includes connections that are shutting down.

pub fn handshaking_peer_connections( &self, peer_id: &PeerId ) -> impl Iterator<Item = ConnectionId> + '_

Returns all the handshaking connections that are expected to reach the given peer. The list also includes connections that are shutting down.

pub fn opened_out_notifications( &self, notifications_protocol_index: usize ) -> impl Iterator<Item = &PeerId> + '_

Returns the list of peers for which we have a fully established notifications protocol of the given protocol.

pub fn connection_state(&self, connection_id: ConnectionId) -> ConnectionState

Returns the state of the given connection.

Panic

Panics if the identifier is invalid or corresponds to a connection that has already entirely shut down.

pub fn unfulfilled_desired_peers(&self) -> impl Iterator<Item = &PeerId> + '_

Returns the list of PeerIds that have been marked as desired, but that don’t have any associated connection. An associated connection is either a fully established connection with that peer, or an outgoing connection that is still handshaking but expects to reach that peer.

pub fn set_peer_desired(&mut self, peer_id: &PeerId, desired: bool)

Sets the “desired” flag of the given PeerId.

When a peer is marked as “desired” and there isn’t any pending or established connection towards it, it is returned when calling Peers::unfulfilled_desired_peers.

pub fn set_peer_notifications_out_desired( &mut self, peer_id: &PeerId, notification_protocol: usize, new_desired_state: DesiredState )

Sets the given combinations of notification protocol and PeerId as “desired”.

When a peer is marked as “desired” and there isn’t any pending or established connection towards it, it is returned when calling Peers::unfulfilled_desired_peers.

When a combination of network protocol and PeerId is marked as “desired”, it will be returned by Peers::unfulfilled_desired_outbound_substream.

When a combination of network protocol and PeerId is no longer marked as “desired”, it will be returned by Peers::fulfilled_undesired_outbound_substreams.

This function might generate a message destined to a connection. Use Peers::pull_message_to_connection to process these messages after it has returned.

pub fn in_notification_accept( &mut self, id: SubstreamId, handshake_back: Vec<u8> )

Responds to a Event::NotificationsInOpen by accepting the request for an inbound substream.

This function might generate a message destined to a connection. Use Peers::pull_message_to_connection to process these messages after it has returned.

Panic

Panics if the SubstreamId is invalid, for example if it was yielded in a Event::NotificationsInOpenCancel.

pub fn in_notification_refuse(&mut self, id: SubstreamId)

Responds to a Event::NotificationsInOpen by refusing the request for an inbound substream.

This function might generate a message destined to a connection. Use Peers::pull_message_to_connection to process these messages after it has returned.

Panic

Panics if the SubstreamId is invalid, for example if it was yielded in a Event::NotificationsInOpenCancel.

pub fn unfulfilled_desired_outbound_substream( &self, include_already_tried: bool ) -> impl Iterator<Item = (&PeerId, usize)> + '_

Returns the list of peer-substream combinations marked as desired where there exists a non-shutting-down connection to this peer but the desired substream hasn’t been opened yet.

If include_already_tried, this function returns substreams that were attempted before and have been refused by the remote.

Use Peers::open_out_notification to actually open a substream.

pub fn open_out_notification( &mut self, peer_id: &PeerId, notifications_protocol_index: usize, now: TNow, handshake: Vec<u8> )

Open a new outgoing substream to the given peer. The peer-protocol combination must have been marked as desired, and a non-shutting-down connection must exist with the current peer.

Must be passed the current moment in time in order to determine when the operation of opening the substream times out.

Use Peers::unfulfilled_desired_outbound_substream in order to determine which substreams should be opened.

This function might generate a message destined to a connection. Use Peers::pull_message_to_connection to process these messages after it has returned.

Panic

Panics if this combination of peer-protocol isn’t desired. Panics if there is no established non-shutting-down connection with the given peer. Panics if there is already an outgoing substream with this peer-protocol combination.

pub fn fulfilled_undesired_outbound_substreams( &self ) -> impl Iterator<Item = (&PeerId, usize, OpenOrPending)> + '_

Returns the list of peer-substream combinations not marked as desired but where there exists an open substream or a substream currently being opened.

Use Peers::close_out_notification to actually close the substream.

pub fn close_out_notification( &mut self, peer_id: &PeerId, notifications_protocol_index: usize ) -> OpenOrPending

Close an existing outgoing substream to the given peer, or cancel opening an outgoing substream.

Use Peers::fulfilled_undesired_outbound_substreams in order to determine which substreams should be closed. However, this function can be used to close any substream even if it marked as desired. If it is marked as desired, the substream will subsequently be returned by Peers::unfulfilled_desired_outbound_substream.

This function might generate a message destined to a connection. Use Peers::pull_message_to_connection to process these messages after it has returned.

Returns whether the substream was open or still being opened.

Note: This function does not generate a Event::NotificationsOutResult or Event::NotificationsOutClose event. Calling this function is equivalent to such an event being instantaneously generated.

Panic

Panics if this combination of peer-protocol isn’t open or opening.

pub fn queue_notification( &mut self, target: &PeerId, notifications_protocol_index: usize, notification: impl Into<Vec<u8>> ) -> Result<(), QueueNotificationError>

Adds a notification to the queue of notifications to send to the given peer.

It is invalid to call this on a PeerId before a successful Event::NotificationsOutResult has been yielded.

Each substream maintains a queue of notifications to be sent to the remote. This method attempts to push a notification to this queue.

An error is also returned if the queue exceeds a certain size in bytes, for two reasons:

Since the content of the queue is transferred at a limited rate, each notification pushed at the end of the queue will take more time than the previous one to reach the destination. Once the queue reaches a certain size, the time it would take for newly-pushed notifications to reach the destination would start being unreasonably large.

If the remote deliberately applies back-pressure on the substream, it is undesirable to increase the memory usage of the local node.

Similarly, the queue being full is a normal situation and notification protocols should be designed in such a way that discarding notifications shouldn’t have a too negative impact.

Regardless of the success of this function, no guarantee exists about the successful delivery of notifications.

This function generates a message destined to the connection. Use Peers::pull_message_to_connection to process these messages after it has returned.

Panics

Panics if there is no fully-open outbound substream with that peer-protocol combination. This can be checked using Peers::can_queue_notification.

pub fn can_queue_notification( &self, target: &PeerId, notifications_protocol_index: usize ) -> bool

Returns true if it is allowed to call Peers::queue_notification, in other words if there is an outbound notifications substream currently open with the target.

If this function returns false, calling Peers::queue_notification will panic.

pub fn broadcast_notification( &mut self, notifications_protocol_index: usize, notification: impl Into<Vec<u8>> )

Equivalent to calling Peers::queue_notification for all peers an outbound notifications substream is open with.

Individual errors that would have occurred when calling Peers::queue_notification are silently discarded.

This function might generate messages destined to connections. Use Peers::pull_message_to_connection to process these messages after it has returned.

pub fn start_request( &mut self, target: &PeerId, protocol_index: usize, request_data: Vec<u8>, timeout: TNow ) -> Result<OutRequestId, StartRequestError>

Sends a request to the given peer.

A Event::Response event will later be generated containing the result of the request.

It is invalid to start a request on a peer before an Event::HandshakeFinished event has been generated, or after a Event::StartShutdown event has been generated where ShutdownPeer::Established::num_healthy_peer_connections is 0.

Returns a newly-allocated identifier for this request.

This function generates a message destined to a connection. Use Peers::pull_message_to_connection to process these messages after it has returned.

Requests

A request consists in:

Opening a substream on an established connection with the target.

Negotiating the requested protocol (protocol_index) on this substream using the multistream-select protocol.

Sending the request (request_data parameter), prefixed with its length.

Waiting for the response (prefixed with its length), which is then returned.

An error happens if the connection closes while the request is in progress, if the remote doesn’t support the given protocol, if the request or response doesn’t respect the protocol limits (see ConfigRequestResponse), or if the remote takes too much time to answer.

The timeout is the time between the moment the substream is opened and the moment the response is sent back. If the emitter doesn’t send the request or if the receiver doesn’t answer during this time window, the request is considered failed.

Panic

Panics if protocol_index isn’t a valid index in Config::request_response_protocols. Panics if there is no open connection with the target or if all connections are shutting down. Use Peers::can_start_requests to check if this is the case.

pub fn can_start_requests(&self, peer_id: &PeerId) -> bool

Returns true if if it possible to send requests (i.e. through Peers::start_request) to the given peer.

If false is returned, then starting a request will panic.

In other words, returns true if there exists an established connection non-shutting-down connection with the given peer.

pub fn respond_in_request( &mut self, id: InRequestId, response: Result<Vec<u8>, ()> )

Responds to a previously-emitted Event::RequestIn.

Panic

Panics if the InRequestId is invalid. Note that these ids remain valid forever until Peers::respond_in_request is called or a Event::RequestInCancel is generated.

pub fn peers_list(&self) -> impl Iterator<Item = &PeerId>

Returns an iterator to the list of PeerIds that we have an established connection with.

pub fn num_outgoing_substreams( &self, notifications_protocol_index: usize ) -> usize

Returns the number of connections we have a substream with.