Struct smoldot::libp2p::read_write::ReadWrite

pub struct ReadWrite<'a, TNow> {
    pub now: TNow,
    pub incoming_buffer: Option<&'a [u8]>,
    pub outgoing_buffer: Option<(&'a mut [u8], &'a mut [u8])>,
    pub read_bytes: usize,
    pub written_bytes: usize,
    pub wake_up_after: Option<TNow>,
}

Fields
now: TNowincoming_buffer: Option<&'a [u8]>

Pointer to a buffer of socket data ready to be processed.

Contains None if the remote has closed their writing side of the socket.

outgoing_buffer: Option<(&'a mut [u8], &'a mut [u8])>

Pointer to two consecutive buffers of uninitialized data. Can be written to in order to write data out towards the socket.

Contains None if the writing side of the socket has been closed or must be closed.

read_bytes: usize

Total number of bytes that have been read from ReadWrite::incoming_buffer.

ReadWrite::incoming_buffer must have been advanced after these bytes.

written_bytes: usize

Total number of bytes that have been written to ReadWrite::outgoing_buffer.

ReadWrite::outgoing_buffer must have been advanced after these bytes.

wake_up_after: Option<TNow>

If Some, the socket must be waken up after the given TNow is reached.

Implementations

impl<'a, TNow> ReadWrite<'a, TNow>

pub fn is_dead(&self) -> bool

Returns true if the connection should be considered dead. That is, both ReadWrite::incoming_buffer and ReadWrite::outgoing_buffer are None.

pub fn advance_read(&mut self, num: usize)

Discards the first num bytes of ReadWrite::incoming_buffer and adds them to ReadWrite::read_bytes.

Panic

Panics if num is superior to the size of the available buffer.

pub fn advance_write(&mut self, num: usize)

Discards the first num bytes of ReadWrite::outgoing_buffer and adds them to ReadWrite::written_bytes.

Panic

Panics if num is superior to the size of the available buffer.

pub fn close_write_if_empty(&mut self)

Sets the writing side of the connection to closed.

This is simply a shortcut for setting ReadWrite::outgoing_buffer to None if ReadWrite::written_bytes is equal to 0.

pub fn incoming_buffer_available(&self) -> usize

Returns the size of the data available in the incoming buffer.

pub fn discard_all_incoming(&mut self)

Shortcut to ReadWrite::advance_read, passing as parameter the value of ReadWrite::incoming_buffer_available. This discards all the incoming data.

pub fn incoming_bytes_iter<'b>(&'b mut self) -> IncomingBytes<'a, 'b, TNow>

Returns an iterator that pops bytes from ReadWrite::incoming_buffer. Whenever the iterator advances, ReadWrite::read_bytes is increased by 1.

pub fn read_bytes<const N: usize>(&mut self) -> [u8; N]

Extracts a certain number of bytes from ReadWrite::incoming_buffer and updates ReadWrite::read_bytes.

Panic

Panics if N is super to the number of bytes available.

pub fn outgoing_buffer_available(&self) -> usize

Returns the size of the available outgoing buffer.

pub fn write_out(&mut self, data: &[u8])

Copies the content of data to ReadWrite::outgoing_buffer and increases ReadWrite::written_bytes.

Panic

Panics if data.len() > self.outgoing_buffer_available().

pub fn write_from_vec_deque(&mut self, data: &mut VecDeque<u8>)

Copies as much as possible from the content of data to ReadWrite::outgoing_buffer and increases ReadWrite::written_bytes. The bytes that have been written are removed from data.

pub fn wake_up_after(&mut self, after: &TNow)where
TNow: Clone + Ord,

Sets ReadWrite::wake_up_after to min(wake_up_after, after).