smoldot::network

Module basic_peering_strategy

Basic address book and slots assignments algorithm.

The BasicPeeringStrategy contains a collection of network identities, identified by a PeerId. Each network identity is associated to one or more chains, identified by a TChainId.

Each network-identity-chain association can be in one of these three states:

Normal.

Banned until a certain instant represented by TInstant.

Has a slot.

“Normal” and “banned” network-identity-chain associations represent the potential peers to connect to, while “slot” represent pending or established gossip slots.

Use BasicPeeringStrategy::pick_assignable_peer in order to assign a slot to a randomly-chosen network-identity that doesn’t currently have one.

If a gossip slot fails to be established with a certain peer, or if the peer misbehaves, use BasicPeeringStrategy::unassign_slot_and_ban to ban the peer, preventing it from obtaining a slot for a certain amount of time.

Each network identity that is associated with at least one chain is associated with zero or more addresses. It is not possible to insert addresses to peers that aren’t associated to at least one chain. The number of active connections of each address is also tracked.

There exists a limit to the number of peers per chain and the number of addresses per peer, guaranteeing that the data structure only uses a bounded amount of memory. If these limits are reached, peers and addresses are removed randomly. Peers that have a slot and at least one connected address are never removed.

Re-exports

pub use crate::libp2p::PeerId;

Structs

BasicPeeringStrategy

Config

Configuration passed to BasicPeeringStrategy::new.

UnassignSlotsAndBanIter

See BasicPeeringStrategy::unassign_slots_and_ban.

Enums

AssignablePeer

See BasicPeeringStrategy::pick_assignable_peer.

DecreaseAddressConnectionsError

See BasicPeeringStrategy::decrease_address_connections.

InsertAddressConnectionsResult

See BasicPeeringStrategy::increase_address_connections.

InsertAddressResult

See BasicPeeringStrategy::insert_address.

InsertChainPeerResult

See BasicPeeringStrategy::insert_chain_peer.

UnassignSlotAndBan

See BasicPeeringStrategy::unassign_slot_and_ban.

UnassignSlotAndRemoveChainPeer

See BasicPeeringStrategy::unassign_slot_and_remove_chain_peer.

UnassignSlotsAndBan

See BasicPeeringStrategy::unassign_slots_and_ban.