Module smoldot::network::service

Re-exports

pub use crate::libp2p::collection::ReadWrite;

pub use crate::libp2p::peers::ConnectionId;

pub use crate::libp2p::peers::InboundError;

Structs

ChainConfig

Configuration for a specific overlay network.

ChainNetwork

Data structure containing the list of all connections, pending or not, and their latest known state. See also the module-level documentation.

Config

Configuration for a ChainNetwork.

DiscoveryInsert

Successfull outcome to ChainNetwork::kademlia_discovery_round.

EncodedBlockAnnounce

Undecoded but valid block announce.

EncodedBlockAnnounceHandshake

Undecoded but valid block announce handshake.

EncodedGrandpaCommitMessage

Undecoded but valid GrandPa commit message.

GrandpaState

IdentifyRequestIn

See Event::IdentifyRequestIn.

PendingId

Identifier of a pending connection requested by the network through a StartConnect.

StartConnect

User must start connecting to the given multiaddress.

Enums

BlocksRequestError

Error returned by ChainNetwork::blocks_request.

CallProofRequestError

Error returned by ChainNetwork::call_proof_request.

DiscoveryError

Error during ChainNetwork::kademlia_discovery_round.

Event

Event generated by ChainNetwork::next_event.

GrandpaWarpSyncRequestError

Error returned by ChainNetwork::grandpa_warp_sync_request.

KademliaFindNodeError

Error during ChainNetwork::kademlia_find_node.

NotificationsOutErr

Error that can happen when trying to open an outbound notifications substream.

ProtocolError

See Event::ProtocolError.

StorageProofRequestError

Error returned by ChainNetwork::storage_proof_request.