smoldot::network

Module service

State machine of all networking connections.

The ChainNetwork struct is a state machine containing multiple collections that are controlled by the API user:

A list of networking connections, identified by a ConnectionId. Each connection is either in the handshake phase, healthy, or shutting down. Each connection in the handshake phase has an optional expected PeerId representing the identity of the node that is expected to be reached once the handshake is finished. Each connection that is healthy or shutting down has an (actual) PeerId associated to it.

A list of chains, identified by a ChainId.

A set of “desired” (ChainId, PeerId, GossipKind) tuples representing, for each chain, the identities of the nodes that the API user wants to establish a gossip link with.

In addition to this, the ChainNetwork also exposes:

A set of (ChainId, PeerId, GossipKind) tuples representing the gossip links that have been established.

A list of outgoing requests, identified by a SubstreamId, that have been sent to a peer and that are awaiting a response.

A list of ingoing requests, identified by a SubstreamId, that have been received from a peer and that must be answered by the API user.

A list of outgoing gossip link connection attempts, identified by a SubstreamId, that must be answered by the peer.

A set of (ChainId, PeerId, GossipKind) tuples representing the peers that would like to establish a gossip link with the local node, and that are awaiting a response by the API user.

Usage

At initialization, create a new ChainNetwork with ChainNetwork::new, and add chains using ChainNetwork::add_chain.

The ChainNetwork doesn’t automatically open connections to peers. This must be done manually using ChainNetwork::add_single_stream_connection or ChainNetwork::add_multi_stream_connection. Choosing which peer to connect to and through which address is outside of the scope of this module.

Adding a connection using ChainNetwork::add_single_stream_connection or ChainNetwork::add_multi_stream_connection returns a “connection task”. This connection task must be processed. TODO: expand explanation here

After a message has been injected using ChainNetwork::inject_connection_message, repeatedly ChainNetwork::next_event until it returns None in order to determine what has happened.

Once a connection has been established (which is indicated by a Event::HandshakeFinished event), one can open a gossip link to this peer using ChainNetwork::gossip_open.

In order to facilitate this process, the ChainNetwork provides a “desired gossip links” system. Use ChainNetwork::gossip_insert_desired and ChainNetwork::gossip_remove_desired to insert or remove (ChainId, PeerId, GossipKind) tuples into the state machine. You can then use ChainNetwork::unconnected_desired to obtain a list of PeerIds that are marked as desired and for which a connection should be opened, and ChainNetwork::connected_unopened_gossip_desired to obtain a list of PeerIds that are marked as desired and that have a healthy connection and for which a gossip link should be opened. Marking peers as desired only influences the return values of ChainNetwork::unconnected_desired and ChainNetwork::connected_unopened_gossip_desired and has no other effect.

Re-exports

pub use crate::libp2p::collection::ConnectionId;

pub use crate::libp2p::collection::ConnectionToCoordinator;

pub use crate::libp2p::collection::CoordinatorToConnection;

pub use crate::libp2p::collection::InboundError;

pub use crate::libp2p::collection::MultiStreamConnectionTask;

pub use crate::libp2p::collection::MultiStreamHandshakeKind;

pub use crate::libp2p::collection::NotificationsOutErr;

pub use crate::libp2p::collection::ReadWrite;

pub use crate::libp2p::collection::RequestError;

pub use crate::libp2p::collection::SingleStreamConnectionTask;

pub use crate::libp2p::collection::SingleStreamHandshakeKind;

pub use crate::libp2p::collection::SubstreamId;

pub use crate::libp2p::connection::noise;

pub use crate::libp2p::connection::noise::NoiseKey;

pub use crate::libp2p::multiaddr;

pub use crate::libp2p::multiaddr::Multiaddr;

pub use crate::libp2p::peer_id;

pub use crate::libp2p::peer_id::PeerId;

pub use crate::network::codec::BlockAnnouncesHandshakeDecodeError;

pub use crate::network::codec::Role;

Structs

ChainConfig

Configuration for a specific overlay network.

ChainId

Identifier of a chain added through ChainNetwork::add_chain.

ChainNetwork

Data structure containing the list of all connections and their latest known state. See also the module-level documentation.

Config

Configuration for a ChainNetwork.

EncodedBlockAnnounce

Undecoded but valid block announce.

EncodedBlockAnnounceHandshake

Undecoded but valid block announce handshake.

EncodedGrandpaCommitMessage

Undecoded but valid GrandPa commit message.

EncodedGrandpaWarpSyncResponse

Undecoded but valid GrandPa warp sync response.

EncodedMerkleProof

Undecoded but valid Merkle proof.

EncodedStateResponse

Undecoded but valid state response.

GrandpaState

Enums

AddChainError

Error returned by ChainNetwork::add_chain.

BlocksRequestError

Error returned by ChainNetwork::start_blocks_request.

CallProofRequestError

Error returned by ChainNetwork::start_call_proof_request.

CloseGossipError

Error potentially returned by ChainNetwork::gossip_close.

Event

Event generated by ChainNetwork::next_event.

GossipConnectError

Error that can happen when trying to open an outbound block announces notifications substream.

GossipKind

GrandpaWarpSyncRequestError

Error returned by ChainNetwork::start_grandpa_warp_sync_request.

KademliaFindNodeError

Error during ChainNetwork::start_kademlia_find_node_request.

OpenGossipError

Error potentially returned by ChainNetwork::gossip_open.

ProtocolError

See Event::ProtocolError.

QueueNotificationError

Error potentially returned when queueing a notification.

RemoveChainError

Error returned by ChainNetwork::remove_chain.

RequestResult

Response to an outgoing request.

StartRequestError

Error potentially returned when starting a request.

StartRequestMaybeTooLargeError

Error potentially returned when starting a request that might be too large.

StateRequestError

Error returned by ChainNetwork::start_state_request.

StorageProofRequestError

Error returned by ChainNetwork::start_storage_proof_request.