smoldot::network::service

Enum Event

pub enum Event<TConn> {

    HandshakeFinished {
        id: ConnectionId,
        expected_peer_id: Option<PeerId>,
        peer_id: PeerId,
    },
    PreHandshakeDisconnected {
        id: ConnectionId,
        address: Vec<u8>,
        expected_peer_id: Option<PeerId>,
        user_data: TConn,
    },
    Disconnected {
        id: ConnectionId,
        address: Vec<u8>,
        peer_id: PeerId,
        user_data: TConn,
    },
    PingOutSuccess {
        id: ConnectionId,
        peer_id: PeerId,
        ping_time: Duration,
    },
    GossipConnected {
        peer_id: PeerId,
        chain_id: ChainId,
        kind: GossipKind,
        role: Role,
        best_number: u64,
        best_hash: [u8; 32],
    },
    GossipOpenFailed {
        peer_id: PeerId,
        chain_id: ChainId,
        kind: GossipKind,
        error: GossipConnectError,
    },
    GossipDisconnected {
        peer_id: PeerId,
        chain_id: ChainId,
        kind: GossipKind,
    },
    GossipInDesired {
        peer_id: PeerId,
        chain_id: ChainId,
        kind: GossipKind,
    },
    GossipInDesiredCancel {
        peer_id: PeerId,
        chain_id: ChainId,
        kind: GossipKind,
    },
    RequestResult {
        peer_id: PeerId,
        chain_id: ChainId,
        substream_id: SubstreamId,
        response: RequestResult,
    },
    BlockAnnounce {
        peer_id: PeerId,
        chain_id: ChainId,
        announce: EncodedBlockAnnounce,
    },
    GrandpaNeighborPacket {
        peer_id: PeerId,
        chain_id: ChainId,
        state: GrandpaState,
    },
    GrandpaCommitMessage {
        peer_id: PeerId,
        chain_id: ChainId,
        message: EncodedGrandpaCommitMessage,
    },
    ProtocolError {
        peer_id: PeerId,
        error: ProtocolError,
    },
    IdentifyRequestIn {
        peer_id: PeerId,
        substream_id: SubstreamId,
    },
    BlocksRequestIn {
        peer_id: PeerId,
        chain_id: ChainId,
        config: BlocksRequestConfig,
        substream_id: SubstreamId,
    },
    RequestInCancel {
        substream_id: SubstreamId,
    },

}

Event generated by ChainNetwork::next_event.

Variants

HandshakeFinished

A connection that was added with ChainNetwork::add_single_stream_connection or ChainNetwork::add_multi_stream_connection has now finished its handshake phase. Its PeerId is now known which certainty.

Fields

id: ConnectionId

Identifier of the connection.

expected_peer_id: Option<PeerId>

Parameter that was passed to ChainNetwork::add_single_stream_connection or ChainNetwork::add_multi_stream_connection.

peer_id: PeerId

Actual PeerId of the connection.

PreHandshakeDisconnected

A connection has shut down before finishing its handshake.

Fields

id: ConnectionId

Identifier of the connection.

address: Vec<u8>

Parameter that was passed to ChainNetwork::add_single_stream_connection or ChainNetwork::add_multi_stream_connection.

expected_peer_id: Option<PeerId>

Parameter that was passed to ChainNetwork::add_single_stream_connection or ChainNetwork::add_multi_stream_connection.

user_data: TConn

User data that was passed to ChainNetwork::add_single_stream_connection or ChainNetwork::add_multi_stream_connection.

Disconnected

A connection has shut down after finishing its handshake.

Fields

id: ConnectionId

Identifier of the connection.

address: Vec<u8>

Parameter that was passed to ChainNetwork::add_single_stream_connection or ChainNetwork::add_multi_stream_connection.

peer_id: PeerId

Peer that was connected.

user_data: TConn

User data that was passed to ChainNetwork::add_single_stream_connection or ChainNetwork::add_multi_stream_connection.

PingOutSuccess

The ping of a connection has been measured.

Fields

id: ConnectionId

Identifier of the connection.

peer_id: PeerId

PeerId of the connection.

ping_time: Duration

Time between sending the ping and receiving the pong.

GossipConnected

Now connected to the given peer for gossiping purposes.

This event can only happen as a result of a call to ChainNetwork::gossip_open.

Fields

peer_id: PeerId

Peer we are now connected to.

chain_id: ChainId

Chain of the gossip connection.

kind: GossipKind

Which kind of gossip link is concerned.

role: Role

Role the node reports playing on the network.

best_number: u64

Height of the best block according to this node.

best_hash: [u8; 32]

Hash of the best block according to this node.

GossipOpenFailed

An attempt has been made to open the given chain, but something wrong happened.

This event can only happen as a result of a call to ChainNetwork::gossip_open.

Fields

peer_id: PeerId

Peer concerned by the event.

chain_id: ChainId

Chain of the gossip connection.

kind: GossipKind

Which kind of gossip link is concerned.

error: GossipConnectError

Problem that happened.

GossipDisconnected

No longer connected to the given peer for gossiping purposes.

Fields

peer_id: PeerId

Peer we are no longer connected to.

chain_id: ChainId

Chain of the gossip connection.

kind: GossipKind

Which kind of gossip link is concerned.

GossipInDesired

A peer would like to open a gossiping link with the local node.

Fields

peer_id: PeerId

Peer concerned by the event.

chain_id: ChainId

Chain of the gossip connection.

kind: GossipKind

Which kind of gossip link is concerned.

GossipInDesiredCancel

A previously-emitted Event::GossipInDesired is no longer relevant as the peer has stopped the opening attempt.

Fields

peer_id: PeerId

Peer concerned by the event.

chain_id: ChainId

Chain of the gossip connection.

kind: GossipKind

Which kind of gossip link is concerned.

RequestResult

An outgoing request has finished, either successfully or not.

Fields

peer_id: PeerId

Peer that has answered the request.

chain_id: ChainId

Index of the chain the request relates to.

substream_id: SubstreamId

Identifier of the request that was returned by the function that started the request.

response: RequestResult

Outcome of the request.

BlockAnnounce

Received a new block announce from a peer.

Can only happen after a Event::GossipConnected with the given PeerId and ChainId combination has happened.

Fields

peer_id: PeerId

Identity of the sender of the block announce.

chain_id: ChainId

Index of the chain the block relates to.

announce: EncodedBlockAnnounce

GrandpaNeighborPacket

Received a GrandPa neighbor packet from the network. This contains an update to the finality state of the given peer.

Can only happen after a Event::GossipConnected with the given PeerId and ChainId combination has happened.

Fields

peer_id: PeerId

Identity of the sender of the message.

chain_id: ChainId

Index of the chain the message relates to.

state: GrandpaState

State of the remote.

GrandpaCommitMessage

Received a GrandPa commit message from the network.

Can only happen after a Event::GossipConnected with the given PeerId and ChainId combination has happened.

Fields

peer_id: PeerId

Identity of the sender of the message.

chain_id: ChainId

Index of the chain the commit message relates to.

message: EncodedGrandpaCommitMessage

ProtocolError

Error in the protocol in a connection, such as failure to decode a message. This event doesn’t have any consequence on the health of the connection, and is purely for diagnostic purposes.

Fields

peer_id: PeerId

Peer that has caused the protocol error.

error: ProtocolError

Error that happened.

IdentifyRequestIn

A remote has sent a request for identification information.

You are strongly encouraged to call ChainNetwork::respond_identify.

Fields

peer_id: PeerId

Remote that has sent the request.

substream_id: SubstreamId

Identifier of the request. Necessary to send back the answer.

BlocksRequestIn

A remote has sent a request for blocks.

Can only happen for chains where ChainConfig::allow_inbound_block_requests is true.

You are strongly encouraged to call ChainNetwork::respond_blocks.

Fields

peer_id: PeerId

Remote that has sent the request.

chain_id: ChainId

Index of the chain concerned by the request.

config: BlocksRequestConfig

Information about the request.

substream_id: SubstreamId

Identifier of the request. Necessary to send back the answer.

RequestInCancel

A remote is no longer interested in the response to a request.

Calling ChainNetwork::respond_identify, ChainNetwork::respond_blocks, or similar will now panic.

Fields

substream_id: SubstreamId

Identifier of the request.

This SubstreamId is considered dead and no longer valid.