Struct smoldot::network::service::ChainNetwork

pub struct ChainNetwork<TNow> { /* private fields */ }

Data structure containing the list of all connections, pending or not, and their latest known state. See also the module-level documentation.

Implementations

impl<TNow> ChainNetwork<TNow>where
TNow: Clone + Add<Duration, Output = TNow> + Sub<TNow, Output = Duration> + Ord,

pub fn set_local_best_block(
&mut self,
chain_index: usize,
best_hash: [u8; 32],
best_number: u64
)

Modifies the best block of the local node. See ChainConfig::best_hash and ChainConfig::best_number.

Panic

Panics if chain_index is out of range.

pub fn set_local_grandpa_state(
&mut self,
chain_index: usize,
grandpa_state: GrandpaState
)

Update the state of the local node with regards to GrandPa rounds.

Calling this method does two things:

Send on all the active GrandPa substreams a “neighbor packet” indicating the state of the local node.

Update the neighbor packet that is automatically sent to peers when a GrandPa substream gets opened.

In other words, calling this function atomically informs all the present and future peers of the state of the local node regarding the GrandPa protocol.

Note: The information passed as parameter isn’t validated in any way by this method.

This function might generate a message destined to connections. Use ChainNetwork::pull_message_to_connection to process these messages after it has returned.

Panic

Panics if chain_index is out of range, or if the chain has GrandPa disabled.

pub fn send_block_announce(
&mut self,
target: &PeerId,
chain_index: usize,
scale_encoded_header: &[u8],
is_best: bool
) -> Result<(), QueueNotificationError>

This function might generate a message destined a connection. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

pub fn can_send_block_announces(
&self,
target: &PeerId,
chain_index: usize
) -> bool

Returns true if it is allowed to call ChainNetwork::send_block_announce, in other words if there is an outbound block announces substream currently open with the target.

If this function returns false, calling ChainNetwork::send_block_announce will panic.

pub fn opened_transactions_substream(
&self,
chain_index: usize
) -> impl Iterator<Item = &PeerId> + '_

Returns the list of peers for which we have a fully established notifications protocol of the given protocol.

pub fn announce_transaction(
&mut self,
target: &PeerId,
chain_index: usize,
extrinsic: &[u8]
) -> Result<(), QueueNotificationError>

Must be passed the SCALE-encoded transaction.

This function might generate a message destined connections. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

impl<TNow> ChainNetwork<TNow>where
TNow: Clone + Add<Duration, Output = TNow> + Sub<TNow, Output = Duration> + Ord,

pub fn start_blocks_request(
&mut self,
now: TNow,
target: &PeerId,
chain_index: usize,
config: BlocksRequestConfig,
timeout: Duration
) -> OutRequestId

Sends a blocks request to the given peer.

This function might generate a message destined a connection. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

pub fn start_blocks_request_unchecked(
&mut self,
now: TNow,
target: &PeerId,
chain_index: usize,
config: BlocksRequestConfig,
timeout: Duration
) -> OutRequestId

Sends a blocks request to the given peer.

This function might generate a message destined a connection. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

pub fn start_grandpa_warp_sync_request(
&mut self,
now: TNow,
target: &PeerId,
chain_index: usize,
begin_hash: [u8; 32],
timeout: Duration
) -> OutRequestId

This function might generate a message destined a connection. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

pub fn start_state_request(
&mut self,
now: TNow,
target: &PeerId,
chain_index: usize,
block_hash: &[u8; 32],
start_key: StateRequestStart<'_>,
timeout: Duration
) -> OutRequestId

Sends a state request to a peer.

A state request makes it possible to download the storage of the chain at a given block. The response is not unverified by this function. In other words, the peer is free to send back erroneous data. It is the responsibility of the API user to verify the storage by calculating the state trie root hash and comparing it with the value stored in the block’s header.

Because response have a size limit, it is unlikely that a single request will return the entire storage of the chain at once. Instead, call this function multiple times, each call passing a start_key that follows the last key of the previous response.

This function might generate a message destined a connection. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

pub fn start_storage_proof_request(
&mut self,
now: TNow,
target: &PeerId,
chain_index: usize,
config: StorageProofRequestConfig<impl Iterator<Item = impl AsRef<[u8]> + Clone>>,
timeout: Duration
) -> OutRequestId

Sends a storage request to the given peer.

This function might generate a message destined a connection. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

pub fn start_call_proof_request(
&mut self,
now: TNow,
target: &PeerId,
chain_index: usize,
config: CallProofRequestConfig<'_, impl Iterator<Item = impl AsRef<[u8]>>>,
timeout: Duration
) -> OutRequestId

Sends a call proof request to the given peer.

This request is similar to ChainNetwork::start_storage_proof_request. Instead of requesting specific keys, we request the list of all the keys that are accessed for a specific runtime call.

There exists no guarantee that the proof is complete (i.e. that it contains all the necessary entries), as it is impossible to know this from just the proof itself. As such, this method is just an optimization. When performing the actual call, regular storage proof requests should be performed if the key is not present in the call proof response.

This function might generate a message destined a connection. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

pub fn discover(
&mut self,
now: &TNow,
chain_index: usize,
peer_id: PeerId,
discovered_addrs: impl IntoIterator<Item = Multiaddr>
)

Inserts the given list of nodes into the list of known nodes held within the state machine.

pub fn start_kademlia_discovery_round(
&mut self,
now: TNow,
chain_index: usize
) -> KademliaOperationId

Performs a round of Kademlia discovery.

This future yields once a list of nodes on the network has been discovered, or a problem happened.

This function might generate a message destined a connection. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

pub fn start_kademlia_find_node(
&mut self,
target: &PeerId,
now: TNow,
chain_index: usize,
close_to_key: &[u8]
) -> OutRequestId

Sends a Kademlia “find node” request to a single peer, and waits for it to answer.

Returns an error if there is no active connection with that peer.

This function might generate a message destined a connection. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

pub fn can_start_requests(&self, peer_id: &PeerId) -> bool

Returns true if if it possible to send requests (i.e. through ChainNetwork::start_grandpa_warp_sync_request, ChainNetwork::start_blocks_request, etc.) to the given peer.

If false is returned, starting a request will panic.

In other words, returns true if there exists an established connection non-shutting-down connection with the given peer.

pub fn respond_identify(&mut self, request_id: InRequestId, agent_version: &str)

This function might generate a message destined a connection. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

pub fn respond_blocks(
&mut self,
request_id: InRequestId,
response: Option<Vec<BlockData>>
)

Queue the response to send back.

Pass None in order to deny the request. Do this if blocks aren’t available locally.

Has no effect if the connection that sends the request no longer exists.

This function might generate a message destined a connection. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

impl<TNow> ChainNetwork<TNow>where
TNow: Clone + Add<Duration, Output = TNow> + Sub<TNow, Output = Duration> + Ord,

pub fn new(config: Config<TNow>) -> Self

Initializes a new ChainNetwork.

pub fn num_established_connections(&self) -> usize

Returns the number of established TCP connections, both incoming and outgoing.

pub fn num_peers(&self, chain_index: usize) -> usize

Returns the number of peers we have a substream with.

pub fn num_chains(&self) -> usize

Returns the number of chains. Always equal to the length of Config::chains.

pub fn block_number_bytes(&self, chain_index: usize) -> usize

Returns the value passed as ChainConfig::block_number_bytes for the given chain.

Panic

Panics if chain_index is out of range.

pub fn noise_key(&self) -> &NoiseKey

Returns the Noise key originally passed as Config::noise_key.

pub fn add_single_stream_incoming_connection(
&mut self,
when_connected: TNow,
handshake_kind: SingleStreamHandshakeKind,
remote_addr: Multiaddr
) -> (ConnectionId, SingleStreamConnectionTask<TNow>)

Adds a single-stream incoming connection to the state machine.

This connection hasn’t finished handshaking and the PeerId of the remote isn’t known yet.

Must be passed the moment (as a TNow) when the connection as been established, in order to determine when the handshake timeout expires.

The remote_addr is the address used to reach back the remote. In the case of TCP, it contains the TCP dialing port of the remote. The remote can ask, through the identify libp2p protocol, its own address, in which case we send it.

pub fn add_multi_stream_incoming_connection<TSubId>(
&mut self,
when_connected: TNow,
handshake_kind: MultiStreamHandshakeKind,
remote_addr: Multiaddr
) -> (ConnectionId, MultiStreamConnectionTask<TNow, TSubId>)where
TSubId: Clone + PartialEq + Eq + Hash,

Adds a multi-stream incoming connection to the state machine.

This connection hasn’t finished handshaking and the PeerId of the remote isn’t known yet.

Must be passed the moment (as a TNow) when the connection as been established, in order to determine when the handshake timeout expires.

The remote_addr is the address used to reach back the remote. In the case of TCP, it contains the TCP dialing port of the remote. The remote can ask, through the identify libp2p protocol, its own address, in which case we send it.

pub fn pull_message_to_connection(
&mut self
) -> Option<(ConnectionId, CoordinatorToConnection<TNow>)>

pub fn inject_connection_message(
&mut self,
connection_id: ConnectionId,
message: ConnectionToCoordinator
)

Injects into the state machine a message generated by SingleStreamConnectionTask::pull_message_to_coordinator or MultiStreamConnectionTask::pull_message_to_coordinator.

pub fn discovered_nodes(
&self,
chain_index: usize
) -> impl Iterator<Item = (&PeerId, impl Iterator<Item = &Multiaddr>)> + '_

Returns a list of nodes (their PeerId and multiaddresses) that we know are part of the network.

Nodes that are discovered might disappear over time. In other words, there is no guarantee that a node that has been added through ChainNetwork::discover will later be returned by ChainNetwork::discovered_nodes.

pub fn pending_outcome_ok_single_stream(
&mut self,
id: PendingId,
handshake_kind: SingleStreamHandshakeKind
) -> (ConnectionId, SingleStreamConnectionTask<TNow>)

After calling ChainNetwork::next_start_connect, notifies the ChainNetwork of the success of the dialing attempt.

See also ChainNetwork::pending_outcome_err.

Panic

Panics if the PendingId is invalid.

pub fn pending_outcome_ok_multi_stream<TSubId>(
&mut self,
id: PendingId,
handshake_kind: MultiStreamHandshakeKind
) -> (ConnectionId, MultiStreamConnectionTask<TNow, TSubId>)where
TSubId: Clone + PartialEq + Eq + Hash,

After calling ChainNetwork::next_start_connect, notifies the ChainNetwork of the success of the dialing attempt.

See also ChainNetwork::pending_outcome_err.

Panic

Panics if the PendingId is invalid.

pub fn pending_outcome_err(&mut self, id: PendingId, is_bad_address: bool)

After calling ChainNetwork::next_start_connect, notifies the ChainNetwork of the failure of the dialing attempt.

See also ChainNetwork::pending_outcome_ok_single_stream and ChainNetwork::pending_outcome_ok_multi_stream.

is_bad_address should be true if the address is invalid or definitely unreachable and should thus not be attempted again. If false is passed, the address might be attempted again.

Panic

Panics if the PendingId is invalid.

pub fn next_event(&mut self, now: TNow) -> Option<Event>

Returns the next event produced by the service.

pub fn next_start_connect(
&mut self,
now: impl FnOnce() -> TNow
) -> Option<StartConnect<TNow>>

Allocates a PendingId and returns a StartConnect indicating a multiaddress that the API user must try to dial.

Later, the API user must use ChainNetwork::pending_outcome_ok_single_stream, ChainNetwork::pending_outcome_ok_multi_stream, or ChainNetwork::pending_outcome_err to report how the connection attempt went.

The returned StartConnect contains the StartConnect::timeout field. It is the responsibility of the API user to ensure that ChainNetwork::pending_outcome_err is called if this timeout is reached.

pub fn peers_list(&self) -> impl Iterator<Item = &PeerId>

Returns an iterator to the list of PeerIds that we have an established connection with.

pub fn slots_to_assign(
&self,
chain_index: usize
) -> impl Iterator<Item = &PeerId> + '_

pub fn assign_out_slot(&mut self, chain_index: usize, peer_id: PeerId)

pub fn unassign_slot(
&mut self,
chain_index: usize,
peer_id: &PeerId
) -> Option<SlotTy>

Removes the slot assignment of the given peer, if any.