smoldot::network::service

Struct ChainNetwork

pub struct ChainNetwork<TChain, TConn, TNow> { /* private fields */ }

Data structure containing the list of all connections and their latest known state. See also the module-level documentation.

Implementations

impl<TChain, TConn, TNow> ChainNetwork<TChain, TConn, TNow>
where TNow: Clone + Add<Duration, Output = TNow> + Sub<TNow, Output = Duration> + Ord,

pub fn new(config: Config) -> Self

Initializes a new ChainNetwork.

pub fn add_chain( &mut self, config: ChainConfig<TChain>, ) -> Result<ChainId, AddChainError>

Adds a chain to the list of chains that is handled by the ChainNetwork.

It is not possible to add a chain if its protocol names would conflict with an existing chain.

pub fn remove_chain( &mut self, chain_id: ChainId, ) -> Result<TChain, RemoveChainError>

Removes a chain previously added with ChainNetwork::add_chain.

This function will return an error if any gossip link is currently open through the given chain. Gossip links should be closed prior to calling this function.

Any request using the given chain will continue as normal.

Panic

Panics if the ChainId is out of range.

pub fn set_chain_local_best_block( &mut self, chain_id: ChainId, best_hash: [u8; 32], best_number: u64, )

Modifies the best block of the local node for the given chain. See ChainConfig::best_hash and ChainConfig::best_number.

This information is sent to remotes whenever a block announces substream is opened.

Panic

Panics if the ChainId is out of range.

pub fn chains(&self) -> impl ExactSizeIterator<Item = ChainId>

Returns the list of all the chains that have been added.

pub fn block_number_bytes(&self, chain_id: ChainId) -> usize

Returns the value passed as ChainConfig::block_number_bytes for the given chain.

Panic

Panics if the given ChainId is invalid.

pub fn gossip_insert_desired( &mut self, chain_id: ChainId, peer_id: PeerId, kind: GossipKind, ) -> bool

Marks the given chain-peer combination as “desired”.

Has no effect if it was already marked as desired.

Returns true if the peer has been marked as desired, and false if it was already marked as desired.

Panic

Panics if the given ChainId is invalid.

pub fn gossip_remove_desired( &mut self, chain_id: ChainId, peer_id: &PeerId, kind: GossipKind, ) -> bool

Removes the given chain-peer combination from the list of desired chain-peers.

Has no effect if it was not marked as desired.

Returns true if the peer was desired on this chain.

Panic

Panics if the given ChainId is invalid.

pub fn gossip_remove_desired_all(&mut self, peer_id: &PeerId, kind: GossipKind)

Removes the given peer from the list of desired chain-peers of all the chains that exist.

Has no effect if it was not marked as desired.

pub fn gossip_desired_iter( &self, chain_id: ChainId, kind: GossipKind, ) -> impl Iterator<Item = &PeerId>

Returns the list of gossip-desired peers for the given chain, in no specific order.

Panic

Panics if the given ChainId is invalid.

pub fn gossip_desired_num(&self, chain_id: ChainId, kind: GossipKind) -> usize

Returns the number of gossip-desired peers for the given chain.

Panic

Panics if the given ChainId is invalid.

pub fn unconnected_desired( &self, ) -> impl ExactSizeIterator<Item = &PeerId> + Clone

Returns the list of PeerIds that are desired (for any chain) but for which no connection exists.

Note: Connections that are currently in the process of shutting down are also ignored for the purpose of this function.

pub fn connected_unopened_gossip_desired( &self, ) -> impl ExactSizeIterator<Item = (&PeerId, ChainId, GossipKind)> + Clone

Returns the list of PeerIds that are marked as desired, and for which a healthy connection exists, but for which no substream connection attempt exists.

pub fn opened_gossip_undesired( &self, ) -> impl ExactSizeIterator<Item = (&PeerId, ChainId, GossipKind)> + Clone

Returns the list of PeerIds for which a substream connection or connection attempt exists but that are not marked as desired.

pub fn opened_gossip_undesired_by_chain( &self, chain_id: ChainId, ) -> impl Iterator<Item = (&PeerId, GossipKind)> + Clone

Returns the list of PeerIds for which a substream connection or connection attempt exists against the given chain but that are not marked as desired.

Panic

Panics if the ChainId is invalid.

pub fn add_single_stream_connection( &mut self, when_connection_start: TNow, handshake_kind: SingleStreamHandshakeKind<'_>, remote_addr: Vec<u8>, expected_peer_id: Option<PeerId>, user_data: TConn, ) -> (ConnectionId, SingleStreamConnectionTask<TNow>)

Adds a single-stream connection to the state machine.

This connection hasn’t finished handshaking and the PeerId of the remote isn’t known yet.

If expected_peer_id is Some, this connection is expected to reach the given PeerId. The expected_peer_id is only used to influence the result of ChainNetwork::unconnected_desired.

Must be passed the moment (as a TNow) when the connection has first been opened, in order to determine when the handshake timeout expires.

The remote_addr is the multiaddress used to reach back the remote. In the case of TCP, it contains the TCP dialing port of the remote. The remote can ask, through the identify libp2p protocol, its own address, in which case we send it. Because the multiaddress specification is flexible, this module doesn’t attempt to parse the address.

pub fn add_multi_stream_connection<TSubId>( &mut self, when_connection_start: TNow, handshake_kind: MultiStreamHandshakeKind<'_>, remote_addr: Vec<u8>, expected_peer_id: Option<PeerId>, user_data: TConn, ) -> (ConnectionId, MultiStreamConnectionTask<TNow, TSubId>)
where TSubId: Clone + PartialEq + Eq + Hash,

Adds a multi-stream connection to the state machine.

This connection hasn’t finished handshaking and the PeerId of the remote isn’t known yet.

If expected_peer_id is Some, this connection is expected to reach the given PeerId. The expected_peer_id is only used to influence the result of ChainNetwork::unconnected_desired.

Must be passed the moment (as a TNow) when the connection has first been opened, in order to determine when the handshake timeout expires.

The remote_addr is the multiaddress used to reach back the remote. In the case of TCP, it contains the TCP dialing port of the remote. The remote can ask, through the identify libp2p protocol, its own address, in which case we send it. Because the multiaddress specification is flexible, this module doesn’t attempt to parse the address.

pub fn num_connections(&self) -> usize

Returns the number of connections, both handshaking or established.

pub fn connection_remote_addr(&self, id: ConnectionId) -> &[u8]

Returns the remote address that was passed to ChainNetwork::add_single_stream_connection or ChainNetwork::add_multi_stream_connection for the given connection.

Note: This module does in no way attempt to parse the address. This function simply returns the value that was provided by the API user, whatever it is.

Panic

Panics if the ConnectionId is invalid.

pub fn num_potential_and_established_connections( &self, peer_id: &PeerId, ) -> usize

Returns the number of connections with the given peer.

Both connections that have and have not finished their handshaking phase are considered.

pub fn pull_message_to_connection( &mut self, ) -> Option<(ConnectionId, CoordinatorToConnection)>

Pulls a message that must be sent to a connection.

The message must be passed to SingleStreamConnectionTask::inject_coordinator_message or MultiStreamConnectionTask::inject_coordinator_message in the appropriate connection.

This function guarantees that the ConnectionId always refers to a connection that is still alive, in the sense that SingleStreamConnectionTask::inject_coordinator_message or MultiStreamConnectionTask::inject_coordinator_message has never returned None and that no Event::Disconnected or Event::PreHandshakeDisconnected has been generated for this connection.

pub fn inject_connection_message( &mut self, connection_id: ConnectionId, message: ConnectionToCoordinator, )

Injects into the state machine a message generated by SingleStreamConnectionTask::pull_message_to_coordinator or MultiStreamConnectionTask::pull_message_to_coordinator.

pub fn next_event(&mut self) -> Option<Event<TConn>>

Returns the next event produced by the service.

pub fn start_blocks_request( &mut self, target: &PeerId, chain_id: ChainId, config: BlocksRequestConfig, timeout: Duration, ) -> Result<SubstreamId, StartRequestError>

Sends a blocks request to the given peer.

The code in this module does not verify the response in any way. The blocks might be completely different from the ones requested, or might be missing some information. In other words, the response is completely untrusted.

This function might generate a message destined a connection. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

Panic

Panics if the ChainId is invalid.

pub fn start_grandpa_warp_sync_request( &mut self, target: &PeerId, chain_id: ChainId, begin_hash: [u8; 32], timeout: Duration, ) -> Result<SubstreamId, StartRequestError>

Panic

Panics if the ChainId is invalid.

pub fn start_state_request( &mut self, target: &PeerId, chain_id: ChainId, block_hash: &[u8; 32], start_key: StateRequestStart<'_>, timeout: Duration, ) -> Result<SubstreamId, StartRequestError>

Sends a state request to a peer.

A state request makes it possible to download the storage of the chain at a given block. The response is not unverified by this function. In other words, the peer is free to send back erroneous data. It is the responsibility of the API user to verify the storage by calculating the state trie root hash and comparing it with the value stored in the block’s header.

Because response have a size limit, it is unlikely that a single request will return the entire storage of the chain at once. Instead, call this function multiple times, each call passing a start_key that follows the last key of the previous response.

This function might generate a message destined a connection. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

Panic

Panics if the ChainId is invalid.

pub fn start_storage_proof_request( &mut self, target: &PeerId, chain_id: ChainId, config: StorageProofRequestConfig<impl Iterator<Item = impl AsRef<[u8]> + Clone>>, timeout: Duration, ) -> Result<SubstreamId, StartRequestMaybeTooLargeError>

Sends a storage request to the given peer.

This function might generate a message destined a connection. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

Panic

Panics if the ChainId is invalid.

pub fn start_call_proof_request( &mut self, target: &PeerId, chain_id: ChainId, config: CallProofRequestConfig<'_, impl Iterator<Item = impl AsRef<[u8]>>>, timeout: Duration, ) -> Result<SubstreamId, StartRequestMaybeTooLargeError>

Sends a call proof request to the given peer.

This request is similar to ChainNetwork::start_storage_proof_request. Instead of requesting specific keys, we request the list of all the keys that are accessed for a specific runtime call.

There exists no guarantee that the proof is complete (i.e. that it contains all the necessary entries), as it is impossible to know this from just the proof itself. As such, this method is just an optimization. When performing the actual call, regular storage proof requests should be performed if the key is not present in the call proof response.

This function might generate a message destined a connection. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

Panic

Panics if the ChainId is invalid.

pub fn start_child_storage_proof_request( &mut self, target: &PeerId, chain_id: ChainId, config: ChildStorageProofRequestConfig<impl AsRef<[u8]> + Clone, impl Iterator<Item = impl AsRef<[u8]> + Clone>>, timeout: Duration, ) -> Result<SubstreamId, StartRequestMaybeTooLargeError>

Sends a child storage proof request to the given peer.

This is similar to ChainNetwork::start_storage_proof_request but for child tries.

This function might generate a message destined a connection. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

Panic

Panics if the ChainId is invalid.

pub fn start_kademlia_find_node_request( &mut self, target: &PeerId, chain_id: ChainId, peer_id_to_find: &PeerId, timeout: Duration, ) -> Result<SubstreamId, StartRequestError>

Sends a Kademlia find node request to the given peer.

This function might generate a message destined a connection. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

Panic

Panics if the ChainId is invalid.

pub fn respond_identify( &mut self, substream_id: SubstreamId, agent_version: &str, )

Responds to an identify request. Call this function in response to a Event::IdentifyRequestIn.

Only the agent_version needs to be specified. The other fields are automatically filled by the ChainNetwork.

This function might generate a message destined a connection. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

Panic

Panics if the SubstreamId is invalid or doesn’t correspond to a blocks request or if the request has been cancelled with a Event::RequestInCancel.

pub fn respond_blocks( &mut self, substream_id: SubstreamId, response: Option<Vec<BlockData>>, )

Responds to a blocks request. Call this function in response to a Event::BlocksRequestIn.

Pass None in order to deny the request. Do this if blocks aren’t available locally.

This function might generate a message destined a connection. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

Panic

Panics if the SubstreamId is invalid or doesn’t correspond to a blocks request or if the request has been cancelled with a Event::RequestInCancel.

pub fn gossip_connected_peers( &self, chain_id: ChainId, kind: GossipKind, ) -> impl Iterator<Item = &PeerId>

Returns the list of all peers for a Event::GossipConnected event of the given kind has been emitted. It is possible to send gossip notifications to these peers.

Panic

Panics if the ChainId is invalid.

pub fn gossip_is_connected( &self, chain_id: ChainId, target: &PeerId, kind: GossipKind, ) -> bool

Returns the list of all peers for a Event::GossipConnected event of the given kind has been emitted. It is possible to send gossip notifications to these peers.

Panic

Panics if the ChainId is invalid.

pub fn gossip_open( &mut self, chain_id: ChainId, target: &PeerId, kind: GossipKind, ) -> Result<(), OpenGossipError>

Open a gossiping substream with the given peer on the given chain.

Either a Event::GossipConnected or Event::GossipOpenFailed is guaranteed to later be generated, unless ChainNetwork::gossip_close is called in the meanwhile.

Panic

Panics if the ChainId is invalid.

pub fn gossip_close( &mut self, chain_id: ChainId, peer_id: &PeerId, kind: GossipKind, ) -> Result<(), CloseGossipError>

Switches the gossip link to the given peer to the “closed” state.

This can be used:

To close a opening in progress after having called ChainNetwork::gossip_open, in which case no Event::GossipConnected or Event::GossipOpenFailed is generated.

To close a fully open gossip link. All the notifications that have been queued are still delivered. No event is generated.

To respond to a Event::GossipInDesired by rejecting the request.

Panic

Panics if ChainId is invalid.

pub fn gossip_broadcast_grandpa_state_and_update( &mut self, chain_id: ChainId, grandpa_state: GrandpaState, )

Update the state of the local node with regards to GrandPa rounds.

Calling this method does two things:

Send on all the active GrandPa substreams a “neighbor packet” indicating the state of the local node.

Update the neighbor packet that is automatically sent to peers when a GrandPa substream gets opened.

In other words, calling this function atomically informs all the present and future peers of the state of the local node regarding the GrandPa protocol.

Note: The information passed as parameter isn’t validated in any way by this method.

This function might generate a message destined to connections. Use ChainNetwork::pull_message_to_connection to process these messages after it has returned.

Panic

Panics if ChainId is invalid, or if the chain has GrandPa disabled.

pub fn gossip_send_block_announce( &mut self, target: &PeerId, chain_id: ChainId, scale_encoded_header: &[u8], is_best: bool, ) -> Result<(), QueueNotificationError>

Sends a block announce gossip message to the given peer.

If no Event::GossipConnected event of kind GossipKind::ConsensusTransactions has been emitted for the given peer, then a QueueNotificationError::NoConnection will be returned.

This function might generate a message destined a connection. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

Panic

Panics if the ChainId is invalid.

pub fn gossip_send_transaction( &mut self, target: &PeerId, chain_id: ChainId, extrinsic: &[u8], ) -> Result<(), QueueNotificationError>

Sends a transaction gossip message to the given peer.

Must be passed the SCALE-encoded transaction.

If no Event::GossipConnected event of kind GossipKind::ConsensusTransactions has been emitted for the given peer, then a QueueNotificationError::NoConnection will be returned.

This function might generate a message destined connections. Use ChainNetwork::pull_message_to_connection to process messages after it has returned.

Panic

Panics if the ChainId is invalid.