Module smoldot::sync::all

All syncing strategies grouped together.

This state machine combines GrandPa warp syncing and all forks syncing into one state machine.

Overview

This state machine acts as a container of sources, blocks (verified or not), and requests. In order to initialize it, you need to pass, amongst other things, a chain_information::ChainInformation struct indicating the known state of the finality of the chain.

A request represents a query for information from a source. Once the request has finished, call one of the methods of the AllSync in order to notify the state machine of the outcome.

Re-exports

pub use crate::executor::vm::ExecHint;

pub use blocks_tree::CommitVerifyError;

pub use blocks_tree::JustificationVerifyError;

pub use warp_sync::BuildChainInformationError as WarpSyncBuildChainInformationError;

pub use warp_sync::BuildRuntimeError as WarpSyncBuildRuntimeError;

pub use warp_sync::ConfigCodeTrieNodeHint;

pub use warp_sync::VerifyFragmentError;

pub use warp_sync::WarpSyncFragment;

Structs

AddSourceKnown

See AddSource and AllSync::prepare_add_source.

AddSourceOldBlock

See AddSource and AllSync::prepare_add_source.

AddSourceUnknown

See AddSource and AllSync::prepare_add_source.

AllSync

AnnouncedBlockKnown

See BlockAnnounceOutcome and AllSync::block_announce.

AnnouncedBlockUnknown

See BlockAnnounceOutcome and AllForksSync::block_announce.

Block

BlockRequestSuccessBlock

BlockVerify

Config

Configuration for the AllSync.

FinalityProofVerify

HeaderVerifySuccess

Justification

See BlockRequestSuccessBlock::scale_encoded_justifications.

RequestId

Identifier for a request in the AllSync.

SourceId

Identifier for a source in the AllSync.

WarpSyncBuildChainInformation

Building the chain information is necessary for the warp sync process.

WarpSyncBuildRuntime

Compiling a new runtime is necessary for the warp sync process.

WarpSyncFragmentVerify

WarpSyncResponseFragment

Response to a GrandPa warp sync request.

Enums

AddSource

Outcome of calling AllSync::prepare_add_source.

BlockAnnounceOutcome

Outcome of calling AllSync::block_announce.

DesiredRequest

See AllSync::desired_requests.

FinalityProofVerifyOutcome

Information about the outcome of verifying a finality proof.

GrandpaCommitMessageOutcome

See AllSync::grandpa_commit_message.

HeaderVerifyError

Error that can happen when verifying a block header.

HeaderVerifyOutcome

Outcome of calling BlockVerify::verify_header.

ProcessOne

Outcome of calling AllSync::process_one.

RequestDetail

See AllSync::desired_requests.

ResponseOutcome

Outcome of injecting a response in the AllSync.

Status

Status of the synchronization.