Module smoldot::sync::all

All syncing strategies (optimistic, warp sync, all forks) grouped together.

This state machine combines GrandPa warp syncing, optimistic syncing, and all forks syncing into one state machine.

Overview

This state machine acts as a container of sources, blocks (verified or not), and requests. In order to initialize it, you need to pass, amongst other things, a chain_information::ChainInformation struct indicating the known state of the finality of the chain.

A request represents a query for information from a source. Once the request has finished, call one of the methods of the AllSync in order to notify the state machine of the outcome.

Re-exports

pub use crate::executor::vm::ExecHint;

pub use optimistic::Nibble;

pub use optimistic::TrieEntryVersion;

pub use warp_sync::FragmentError as WarpSyncFragmentError;

pub use warp_sync::WarpSyncFragment;

Structs

AllSync

Block

BlockFull

BlockRequestSuccessBlock

BlockVerificationRuntimeCompilation

Compiling a new runtime is necessary as part of the verification.

Config

Configuration for the AllSync.

FinalityProofVerify

GrandpaWarpSyncResponseFragment

Response to a GrandPa warp sync request.

HeaderBodyVerify

HeaderVerify

Justification

See BlockRequestSuccessBlock::scale_encoded_justifications.

RequestId

Identifier for a request in the AllSync.

SourceId

Identifier for a source in the AllSync.

StorageClosestDescendantMerkleValue

Obtaining the Merkle value of the closest descendant of a trie node is required in order to continue.

StorageGet

Loading a storage value is required in order to continue.

StorageNextKey

Fetching the key that follows a given one is required in order to continue.

WarpSyncBuildChainInformation

Building the chain information is necessary for the warp sync process.

WarpSyncBuildRuntime

Compiling a new runtime is necessary for the warp sync process.

WarpSyncFragmentVerify

Enums

BlockAnnounceOutcome

Outcome of calling AllSync::block_announce.

BlockVerification

State of the processing of blocks.

BlockVerificationError

Error that can happen when verifying a block body.

DesiredRequest

See AllSync::desired_requests.

FinalityProofVerifyOutcome

Information about the outcome of verifying a finality proof.

GrandpaCommitMessageOutcome

See AllSync::grandpa_commit_message.

HeaderVerifyError

Error that can happen when verifying a block header.

HeaderVerifyOutcome

Outcome of calling HeaderVerify::perform.

ProcessOne

Outcome of calling AllSync::process_one.

RequestDetail

See AllSync::desired_requests.

ResponseOutcome

Outcome of injecting a response in the AllSync.

Status

Status of the synchronization.