Struct smoldot::sync::all::AllSync

pub struct AllSync<TRq, TSrc, TBl> { /* private fields */ }

Implementations

impl<TRq, TSrc, TBl> AllSync<TRq, TSrc, TBl>

pub fn new(config: Config) -> Self

Initializes a new state machine.

pub fn block_number_bytes(&self) -> usize

Returns the value that was initially passed in Config::block_number_bytes.

pub fn as_chain_information(&self) -> ValidChainInformationRef<'_>

Builds a chain_information::ChainInformationRef struct corresponding to the current latest finalized block. Can later be used to reconstruct a chain.

pub fn status(&self) -> Status<'_, TSrc>

Returns the current status of the syncing.

pub fn finalized_block_header(&self) -> HeaderRef<'_>

Returns the header of the finalized block.

pub fn best_block_header(&self) -> HeaderRef<'_>

Returns the header of the best block.

Note: This value is provided only for informative purposes. Keep in mind that this best block might be reverted in the future.

pub fn best_block_number(&self) -> u64

Returns the number of the best block.

Note: This value is provided only for informative purposes. Keep in mind that this best block might be reverted in the future.

pub fn best_block_hash(&self) -> [u8; 32]

Returns the hash of the best block.

Note: This value is provided only for informative purposes. Keep in mind that this best block might be reverted in the future.

pub fn best_block_consensus(&self) -> ChainInformationConsensusRef<'_>

Returns consensus information about the current best block of the chain.

pub fn non_finalized_blocks_unordered( &self ) -> impl Iterator<Item = HeaderRef<'_>>

Returns the header of all known non-finalized blocks in the chain without any specific order.

pub fn non_finalized_blocks_ancestry_order( &self ) -> impl Iterator<Item = HeaderRef<'_>>

Returns the header of all known non-finalized blocks in the chain.

The returned items are guaranteed to be in an order in which the parents are found before their children.

pub fn is_near_head_of_chain_heuristic(&self) -> bool

Returns true if it is believed that we are near the head of the chain.

The way this method is implemented is opaque and cannot be relied on. The return value should only ever be shown to the user and not used for any meaningful logic.

pub fn add_source( &mut self, user_data: TSrc, best_block_number: u64, best_block_hash: [u8; 32] ) -> SourceId

Adds a new source to the sync state machine.

Must be passed the best block number and hash of the source, as usually reported by the source itself.

Returns an identifier for this new source, plus a list of requests to start or cancel.

pub fn remove_source( &mut self, source_id: SourceId ) -> (TSrc, impl Iterator<Item = (RequestId, TRq)>)

Removes a source from the state machine. Returns the user data of this source, and all the requests that this source were expected to perform.

Panic

Panics if the SourceId doesn’t correspond to a valid source.

pub fn sources(&self) -> impl Iterator<Item = SourceId> + '_

Returns the list of sources in this state machine.

pub fn source_num_ongoing_requests(&self, source_id: SourceId) -> usize

Returns the number of ongoing requests that concern this source.

Panic

Panics if the SourceId is invalid.

pub fn source_best_block(&self, source_id: SourceId) -> (u64, &[u8; 32])

Returns the current best block of the given source.

This corresponds either the latest call to AllSync::block_announce where is_best was true, or to the parameter passed to AllSync::add_source.

Panic

Panics if the SourceId is invalid.

pub fn source_knows_non_finalized_block( &self, source_id: SourceId, height: u64, hash: &[u8; 32] ) -> bool

Returns true if the source has earlier announced the block passed as parameter or one of its descendants.

Panic

Panics if the SourceId is out of range.

Panics if height is inferior or equal to the finalized block height. Finalized blocks are intentionally not tracked by this data structure, and panicking when asking for a potentially-finalized block prevents potentially confusing or erroneous situations.

pub fn knows_non_finalized_block( &self, height: u64, hash: &[u8; 32] ) -> impl Iterator<Item = SourceId> + '_

Returns the list of sources for which AllSync::source_knows_non_finalized_block would return true.

Panic

Panics if height is inferior or equal to the finalized block height. Finalized blocks are intentionally not tracked by this data structure, and panicking when asking for a potentially-finalized block prevents potentially confusing or erroneous situations.

pub fn try_add_known_block_to_source( &mut self, source_id: SourceId, height: u64, hash: [u8; 32] )

Try register a new block that the source is aware of.

Some syncing strategies do not track blocks known to sources, in which case this function has no effect

Has no effect if height is inferior or equal to the finalized block height, or if the source was already known to know this block.

The block does not need to be known by the data structure.

This is automatically done for the blocks added through block announces or block requests..

Panic

Panics if the SourceId is out of range.

pub fn desired_requests( &self ) -> impl Iterator<Item = (SourceId, &TSrc, DesiredRequest)> + '_

Returns the details of a request to start towards a source.

This method doesn’t modify the state machine in any way. AllSync::add_request must be called in order for the request to actually be marked as started.

pub fn add_request( &mut self, source_id: SourceId, detail: RequestDetail, user_data: TRq ) -> RequestId

Inserts a new request in the data structure.

Note: The request doesn’t necessarily have to match a request returned by AllSync::desired_requests.

Panic

Panics if the SourceId is out of range.

pub fn obsolete_requests(&self) -> impl Iterator<Item = RequestId> + '_

Returns a list of requests that are considered obsolete and can be removed using AllSync::blocks_request_response or similar.

A request becomes obsolete if the state of the request blocks changes in such a way that they don’t need to be requested anymore. The response to the request will be useless.

Note: It is in no way mandatory to actually call this function and cancel the requests that are returned.

pub fn process_one(self) -> ProcessOne<TRq, TSrc, TBl>

Process the next block in the queue of verification.

This method takes ownership of the AllSync and starts a verification process. The AllSync is yielded back at the end of this process.

pub fn block_announce( &mut self, source_id: SourceId, announced_scale_encoded_header: Vec<u8>, is_best: bool ) -> BlockAnnounceOutcome

Injects a block announcement made by a source into the state machine.

pub fn update_source_finality_state( &mut self, source_id: SourceId, finalized_block_height: u64 )

Update the finalized block height of the given source.

Panic

Panics if source_id is invalid.

pub fn grandpa_commit_message( &mut self, source_id: SourceId, scale_encoded_message: Vec<u8> ) -> GrandpaCommitMessageOutcome

Update the state machine with a Grandpa commit message received from the network.

This function only inserts the commit message into the state machine, and does not immediately verify it.

pub fn blocks_request_response( &mut self, request_id: RequestId, blocks: Result<impl Iterator<Item = BlockRequestSuccessBlock<TBl>>, ()> ) -> (TRq, ResponseOutcome)

Inject a response to a previously-emitted blocks request.

Panic

Panics if the RequestId doesn’t correspond to any request, or corresponds to a request of a different type.

pub fn grandpa_warp_sync_response_ok( &mut self, request_id: RequestId, fragments: Vec<WarpSyncFragment>, is_finished: bool ) -> (TRq, ResponseOutcome)

Inject a successful response to a previously-emitted GrandPa warp sync request.

Panic

Panics if the RequestId doesn’t correspond to any request, or corresponds to a request of a different type.

pub fn grandpa_warp_sync_response_err( &mut self, request_id: RequestId ) -> (TRq, ResponseOutcome)

Inject a failure to a previously-emitted GrandPa warp sync request.

Panic

Panics if the RequestId doesn’t correspond to any request, or corresponds to a request of a different type.

pub fn storage_get_response( &mut self, request_id: RequestId, response: Result<Vec<u8>, ()> ) -> (TRq, ResponseOutcome)

Inject a response to a previously-emitted storage proof request.

Panic

Panics if the RequestId doesn’t correspond to any request, or corresponds to a request of a different type.

Panics if the number of items in the response doesn’t match the number of keys that have been requested.

pub fn call_proof_response( &mut self, request_id: RequestId, response: Result<Vec<u8>, ()> ) -> (TRq, ResponseOutcome)

Inject a response to a previously-emitted call proof request.

On success, must contain the encoded Merkle proof. See the trie module for a description of the format of Merkle proofs.

Panic

Panics if the RequestId doesn’t correspond to any request, or corresponds to a request of a different type.