Struct smoldot::sync::all_forks::AllForksSync

pub struct AllForksSync<TBl, TRq, TSrc> { /* fields omitted */ }

Implementations

impl<TBl, TRq, TSrc> AllForksSync<TBl, TRq, TSrc>

pub fn new(config: Config) -> Self

Initializes a new AllForksSync.

pub fn as_chain_information(&self) -> ValidChainInformationRef<'_>

Builds a chain_information::ChainInformationRef struct corresponding to the current latest finalized block. Can later be used to reconstruct a chain.

pub fn finalized_block_header(&self) -> HeaderRef<'_>

Returns the header of the finalized block.

pub fn best_block_header(&self) -> HeaderRef<'_>

Returns the header of the best block.

Note: This value is provided only for informative purposes. Keep in mind that this best block might be reverted in the future.

pub fn best_block_number(&self) -> u64

Returns the number of the best block.

Note: This value is provided only for informative purposes. Keep in mind that this best block might be reverted in the future.

pub fn best_block_hash(&self) -> [u8; 32]

Returns the hash of the best block.

Note: This value is provided only for informative purposes. Keep in mind that this best block might be reverted in the future.

pub fn non_finalized_blocks_ancestry_order(
&self
) -> impl Iterator<Item = HeaderRef<'_>> + '_

Returns the header of all known non-finalized blocks in the chain.

The returned items are guaranteed to be in an order in which the parents are found before their children.

pub fn add_source(
&mut self,
user_data: TSrc,
best_block_number: u64,
best_block_hash: [u8; 32]
) -> SourceId

Inform the AllForksSync of a new potential source of blocks.

The user_data parameter is opaque and decided entirely by the user. It can later be retrieved using AllForksSync::source_user_data.

Returns the newly-created source entry, plus optionally a request that should be started towards this source.

pub fn remove_source(
&mut self,
source_id: SourceId
) -> (TSrc, impl Iterator<Item = (RequestId, RequestParams, TRq)>)

Removes the source from the AllForksSync.

Removing the source implicitly cancels the request that is associated to it (if any).

Returns the user data that was originally passed to AllForksSync::add_source, plus an Option. If this Option is Some, it contains a request that must be started towards the source indicated by the SourceId.

Note: For example, if the source that has just been removed was performing an ancestry search, the Option might contain that same ancestry search.

Panic

Panics if the SourceId is out of range.

pub fn sources(&self) -> impl ExactSizeIterator<Item = SourceId> + '_

Returns the list of sources in this state machine.

pub fn source_knows_non_finalized_block(
&self,
source_id: SourceId,
height: u64,
hash: &[u8; 32]
) -> bool

Returns true if the source has earlier announced the block passed as parameter or one of its descendants.

Also returns true if the requested block is inferior or equal to the known finalized block and the source has announced a block higher or equal to the known finalized block.

Panic

Panics if the SourceId is out of range.

Panics if height is inferior or equal to the finalized block height. Finalized blocks are intentionally not tracked by this data structure, and panicking when asking for a potentially-finalized block prevents potentially confusing or erroneous situations.

pub fn knows_non_finalized_block<'a>(
&'a self,
height: u64,
hash: &[u8; 32]
) -> impl Iterator<Item = SourceId> + 'a

Returns the list of sources for which AllForksSync::source_knows_non_finalized_block would return true.

Panic

Panics if height is inferior or equal to the finalized block height. Finalized blocks are intentionally not tracked by this data structure, and panicking when asking for a potentially-finalized block prevents potentially confusing or erroneous situations.

pub fn source_best_block(&self, source_id: SourceId) -> (u64, &[u8; 32])

Returns the current best block of the given source.

This corresponds either the latest call to AllForksSync::block_announce where is_best was true, or to the parameter passed to AllForksSync::add_source.

Panic

Panics if the SourceId is invalid.

pub fn source_user_data(&self, source_id: SourceId) -> &TSrc

Returns the user data associated to the source. This is the value originally passed through AllForksSync::add_source.

Panic

Panics if the SourceId is out of range.

pub fn source_user_data_mut(&mut self, source_id: SourceId) -> &mut TSrc

Returns the user data associated to the source. This is the value originally passed through AllForksSync::add_source.

Panic

Panics if the SourceId is out of range.

pub fn source_num_ongoing_requests(&self, source_id: SourceId) -> usize

Returns the number of ongoing requests that concern this source.

Panic

Panics if the SourceId is invalid.

pub fn desired_requests(
&self
) -> impl Iterator<Item = (SourceId, &TSrc, RequestParams)> + '_

Returns the details of a request to start towards a source.

This method doesn’t modify the state machine in any way. AllForksSync::add_request must be called in order for the request to actually be marked as started.

pub fn add_request(
&mut self,
source_id: SourceId,
detail: RequestParams,
user_data: TRq
) -> RequestId

Inserts a new request in the data structure.

Note: The request doesn’t necessarily have to match a request returned by AllForksSync::desired_requests.

Panic

Panics if the SourceId is out of range.

pub fn obsolete_requests(&self) -> impl Iterator<Item = (RequestId, &TRq)> + '_

Returns a list of requests that are considered obsolete and can be removed using AllForksSync::finish_ancestry_search or similar.

A request becomes obsolete if the state of the request blocks changes in such a way that they don’t need to be requested anymore. The response to the request will be useless.

Note: It is in no way mandatory to actually call this function and cancel the requests that are returned.

pub fn finish_ancestry_search(
&mut self,
request_id: RequestId,
received_blocks: Result<impl Iterator<Item = RequestSuccessBlock<impl AsRef<[u8]>, impl AsRef<[u8]>>>, ()>
) -> (TRq, AncestrySearchResponseOutcome)

Call in response to a response being finished.

The headers are expected to be sorted in decreasing order. The first element of the iterator should be the block with the hash that was referred by RequestParams::first_block_hash. Each subsequent element is then expected to be the parent of the previous one.

It is legal for the iterator to be shorter than the number of blocks that were requested through RequestParams::num_blocks.

Panic

Panics if the RequestId is invalid.

pub fn block_announce(
&mut self,
source_id: SourceId,
announced_scale_encoded_header: Vec<u8>,
is_best: bool
) -> BlockAnnounceOutcome

Update the source with a newly-announced block.

Note: This information is normally reported by the source itself. In the case of a a networking peer, call this when the source sent a block announce.

Panic

Panics if source_id is invalid.

pub fn grandpa_commit_message(
&mut self,
scale_encoded_message: &[u8]
) -> Result<(), CommitVerifyError>

Update the state machine with a Grandpa commit message received from the network.

On success, the finalized block has been updated.

pub fn process_one(self) -> ProcessOne<TBl, TRq, TSrc>

Process the next block in the queue of verification.

This method takes ownership of the AllForksSync and starts a verification process. The AllForksSync is yielded back at the end of this process.