Struct smoldot::sync::all_forks::AllForksSync

pub struct AllForksSync<TBl, TRq, TSrc> { /* private fields */ }

Implementations

impl<TBl, TRq, TSrc> AllForksSync<TBl, TRq, TSrc>

pub fn new(config: Config) -> Self

Initializes a new AllForksSync.

pub fn block_number_bytes(&self) -> usize

Returns the value that was initially passed in Config::block_number_bytes.

pub fn as_chain_information(&self) -> ValidChainInformationRef<'_>

Builds a chain_information::ChainInformationRef struct corresponding to the current latest finalized block. Can later be used to reconstruct a chain.

pub fn finalized_block_header(&self) -> HeaderRef<'_>

Returns the header of the finalized block.

pub fn best_block_header(&self) -> HeaderRef<'_>

Returns the header of the best block.

Note: This value is provided only for informative purposes. Keep in mind that this best block might be reverted in the future.

pub fn best_block_number(&self) -> u64

Returns the number of the best block.

Note: This value is provided only for informative purposes. Keep in mind that this best block might be reverted in the future.

pub fn best_block_hash(&self) -> [u8; 32]

Returns the hash of the best block.

Note: This value is provided only for informative purposes. Keep in mind that this best block might be reverted in the future.

pub fn non_finalized_blocks_unordered(
&self
) -> impl Iterator<Item = HeaderRef<'_>> + '_

Returns the header of all known non-finalized blocks in the chain without any specific order.

pub fn non_finalized_blocks_ancestry_order(
&self
) -> impl Iterator<Item = HeaderRef<'_>> + '_

Returns the header of all known non-finalized blocks in the chain.

The returned items are guaranteed to be in an order in which the parents are found before their children.

pub fn block_user_data(&self, height: u64, hash: &[u8; 32]) -> &TBl

Gives access to the user data stored for a block of the data structure.

Panic

Panics if the block wasn’t present in the data structure.

pub fn block_user_data_mut(&mut self, height: u64, hash: &[u8; 32]) -> &mut TBl

Gives access to the user data stored for a block of the data structure.

Panic

Panics if the block wasn’t present in the data structure.

pub fn prepare_add_source(
&mut self,
best_block_number: u64,
best_block_hash: [u8; 32]
) -> AddSource<'_, TBl, TRq, TSrc>

Starts the process of inserting a new source in the AllForksSync.

This function doesn’t modify the state machine, but only looks at the current state of the block referenced by best_block_number and best_block_hash. It returns an enum that allows performing the actual insertion.

pub fn remove_source(
&mut self,
source_id: SourceId
) -> (TSrc, impl Iterator<Item = (RequestId, RequestParams, TRq)>)

Removes the source from the AllForksSync.

Removing the source implicitly cancels the request that is associated to it (if any).

Returns the user data that was originally passed when inserting the source, plus an Option. If this Option is Some, it contains a request that must be started towards the source indicated by the SourceId.

Note: For example, if the source that has just been removed was performing an ancestry search, the Option might contain that same ancestry search.

Panic

Panics if the SourceId is out of range.

pub fn sources(&self) -> impl ExactSizeIterator<Item = SourceId> + '_

Returns the list of sources in this state machine.

pub fn source_knows_non_finalized_block(
&self,
source_id: SourceId,
height: u64,
hash: &[u8; 32]
) -> bool

Returns true if the source has earlier announced the block passed as parameter or one of its descendants.

Also returns true if the requested block is inferior or equal to the known finalized block and the source has announced a block higher or equal to the known finalized block.

Panic

Panics if the SourceId is out of range.

Panics if height is inferior or equal to the finalized block height. Finalized blocks are intentionally not tracked by this data structure, and panicking when asking for a potentially-finalized block prevents potentially confusing or erroneous situations.

pub fn knows_non_finalized_block<'a>(
&'a self,
height: u64,
hash: &[u8; 32]
) -> impl Iterator<Item = SourceId> + 'a

Returns the list of sources for which AllForksSync::source_knows_non_finalized_block would return true.

Panic

Panics if height is inferior or equal to the finalized block height. Finalized blocks are intentionally not tracked by this data structure, and panicking when asking for a potentially-finalized block prevents potentially confusing or erroneous situations.

pub fn add_known_block_to_source(
&mut self,
source_id: SourceId,
height: u64,
hash: [u8; 32]
)

Registers a new block that the source is aware of.

Has no effect if height is inferior or equal to the finalized block height, or if the source was already known to know this block.

The block does not need to be known by the data structure.

This is automatically done for the blocks added through AllForksSync::block_announce, AllForksSync::prepare_add_source or FinishAncestrySearch::add_block.

Panic

Panics if the SourceId is out of range.

pub fn source_best_block(&self, source_id: SourceId) -> (u64, &[u8; 32])

Returns the current best block of the given source.

This corresponds either the latest call to AllForksSync::block_announce where is_best was true, or to the parameter passed to AllForksSync::prepare_add_source.

Panic

Panics if the SourceId is invalid.

pub fn source_num_ongoing_requests(&self, source_id: SourceId) -> usize

Returns the number of ongoing requests that concern this source.

Panic

Panics if the SourceId is invalid.

pub fn desired_requests(
&self
) -> impl Iterator<Item = (SourceId, &TSrc, RequestParams)> + '_

Returns the details of a request to start towards a source.

This method doesn’t modify the state machine in any way. AllForksSync::add_request must be called in order for the request to actually be marked as started.

pub fn add_request(
&mut self,
source_id: SourceId,
detail: RequestParams,
user_data: TRq
) -> RequestId

Inserts a new request in the data structure.

Note: The request doesn’t necessarily have to match a request returned by AllForksSync::desired_requests.

Panic

Panics if the SourceId is out of range.

pub fn obsolete_requests(&self) -> impl Iterator<Item = (RequestId, &TRq)> + '_

Returns a list of requests that are considered obsolete and can be removed using AllForksSync::finish_ancestry_search or similar.

A request becomes obsolete if the state of the request blocks changes in such a way that they don’t need to be requested anymore. The response to the request will be useless.

Note: It is in no way mandatory to actually call this function and cancel the requests that are returned.

pub fn finish_ancestry_search(
self,
request_id: RequestId
) -> (TRq, FinishAncestrySearch<TBl, TRq, TSrc>)

Call in response to a blocks request being successful.

This method takes ownership of the AllForksSync and puts it in a mode where the blocks of the response can be added one by one.

The added blocks are expected to be sorted in decreasing order. The first block should be the block with the hash that was referred by RequestParams::first_block_hash. Each subsequent element is then expected to be the parent of the previous one.

Panic

Panics if the RequestId is invalid.

pub fn ancestry_search_failed(
self,
request_id: RequestId
) -> (TRq, AllForksSync<TBl, TRq, TSrc>)

Call in response to a blocks request having failed.

This removes the request from the state machine and returns its user data.

Panic

Panics if the RequestId is invalid.

pub fn block_announce(
&mut self,
source_id: SourceId,
announced_scale_encoded_header: Vec<u8>,
is_best: bool
) -> BlockAnnounceOutcome<'_, TBl, TRq, TSrc>

Update the source with a newly-announced block.

Note: This information is normally reported by the source itself. In the case of a a networking peer, call this when the source sent a block announce.

Panic

Panics if source_id is invalid.

pub fn grandpa_commit_message(
&mut self,
source_id: SourceId,
scale_encoded_commit: Vec<u8>
) -> GrandpaCommitMessageOutcome

Update the state machine with a Grandpa commit message received from the network.

This function only inserts the commit message into the state machine, and does not immediately verify it.

Panic

Panics if source_id is invalid.

pub fn process_one(self) -> ProcessOne<TBl, TRq, TSrc>

Process the next block in the queue of verification.

This method takes ownership of the AllForksSync and starts a verification process. The AllForksSync is yielded back at the end of this process.