Module smoldot::sync::optimistic

Optimistic header and body syncing.

This state machine builds, from a set of sources, a fully verified chain of blocks headers and bodies.

Overview

The algorithm used by this state machine is called “optimistic syncing”. It consists in sending requests for blocks to a certain list of sources, aggregating the answers, and verifying them.

The OptimisticSync struct holds a list of sources, a list of pending block requests, a chain, and a list of blocks received as answers and waiting to be verified.

The requests are emitted ahead of time, so that they can be answered asynchronously while blocks in the verification queue are being processed.

The syncing is said to be optimistic because it is assumed that all sources will provide correct blocks. In the case where the verification of a block fails, the state machine jumps back to the latest known finalized block and resumes syncing from there, possibly using different sources this time.

The optimism aspect comes from the fact that, while a bad source can’t corrupt the state of the local chain, and can’t stall the syncing process (unless there isn’t any other source available), it can still slow it down.

Re-exports

pub use blocks_tree::Nibble;

pub use blocks_tree::TrieEntryVersion;

Structs

Block

BlockFull

BlockVerificationSuccessFull

BlockVerify

Start the processing of a block verification.

Config

Configuration for the OptimisticSync.

Disassemble

Output of OptimisticSync::disassemble.

DisassembleSource

See Disassemble::sources.

JustificationVerify

Start the processing of a justification verification.

OptimisticSync

Optimistic headers-only syncing.

RequestDetail

Request that should be emitted towards a certain source.

RequestId

Identifier for an ongoing request in the OptimisticSync.

RequestSuccessBlock

RequestsDrain

Iterator that drains requests after a source has been removed.

RuntimeCompilation

Compiling a new runtime is necessary as part of the verification.

SourceId

Identifier for a source in the OptimisticSync.

StorageClosestDescendantMerkleValue

Obtaining the Merkle value of the closest descendant of a trie node is required in order to continue.

StorageGet

Loading a storage value is required in order to continue.

StorageNextKey

Fetching the key that follows a given one is required in order to continue.

Enums

BlockVerification

State of the processing of blocks.

FinishRequestOutcome

JustificationVerification

Outcome of the verification of a justification.

ProcessOne

State of the processing of blocks.

ResetCause

Problem that happened and caused the reset.