Struct smoldot::sync::optimistic::OptimisticSync

pub struct OptimisticSync<TRq, TSrc, TBl> { /* private fields */ }

Optimistic headers-only syncing.

Implementations

impl<TRq, TSrc, TBl> OptimisticSync<TRq, TSrc, TBl>

pub fn new(config: Config) -> Self

Builds a new OptimisticSync.

pub fn block_number_bytes(&self) -> usize

Returns the value that was initially passed in Config::block_number_bytes.

pub fn as_chain_information(&self) -> ValidChainInformationRef<'_>

Builds a chain_information::ChainInformationRef struct corresponding to the current latest finalized block. Can later be used to reconstruct a chain.

pub fn finalized_block_header(&self) -> HeaderRef<'_>

Returns the header of the finalized block.

pub fn best_block_header(&self) -> HeaderRef<'_>

Returns the header of the best block.

Note: This value is provided only for informative purposes. Keep in mind that this best block might be reverted in the future.

pub fn best_block_number(&self) -> u64

Returns the number of the best block.

Note: This value is provided only for informative purposes. Keep in mind that this best block might be reverted in the future.

pub fn best_block_hash(&self) -> [u8; 32]

Returns the hash of the best block.

Note: This value is provided only for informative purposes. Keep in mind that this best block might be reverted in the future.

pub fn best_block_consensus(&self) -> ChainInformationConsensusRef<'_>

Returns consensus information about the current best block of the chain.

pub fn best_block_storage(&self) -> Option<BlockStorage<'_, TRq, TSrc, TBl>>

Returns access to the storage of the best block.

Returns None if Config::full was None.

pub fn non_finalized_blocks_unordered(
&self
) -> impl Iterator<Item = HeaderRef<'_>> + '_

Returns the header of all known non-finalized blocks in the chain without any specific order.

pub fn non_finalized_blocks_ancestry_order(
&self
) -> impl Iterator<Item = HeaderRef<'_>> + '_

Returns the header of all known non-finalized blocks in the chain.

The returned items are guaranteed to be in an order in which the parents are found before their children.

pub fn disassemble(self) -> Disassemble<TRq, TSrc>

Disassembles the state machine into its raw components.

pub fn add_source(&mut self, source: TSrc, best_block_number: u64) -> SourceId

Inform the OptimisticSync of a new potential source of blocks.

pub fn source_best_block(&self, source_id: SourceId) -> u64

Returns the current best block of the given source.

This corresponds either the latest call to OptimisticSync::raise_source_best_block, or to the parameter passed to OptimisticSync::add_source.

Panic

Panics if the SourceId is invalid.

pub fn raise_source_best_block(&mut self, id: SourceId, best_block_number: u64)

Updates the best known block of the source.

Has no effect if the previously-known best block is lower than the new one.

Panic

Panics if the SourceId is invalid.

pub fn remove_source(
&mut self,
source_id: SourceId
) -> (TSrc, impl Iterator<Item = (RequestId, TRq)> + '_)

Inform the OptimisticSync that a source of blocks is no longer available.

This automatically cancels all the requests that have been emitted for this source. This list of requests is returned as part of this function.

Panic

Panics if the SourceId is invalid.

pub fn sources(&self) -> impl ExactSizeIterator<Item = SourceId> + '_

Returns the list of sources in this state machine.

pub fn source_num_ongoing_requests(&self, source_id: SourceId) -> usize

Returns the number of ongoing requests that concern this source.

Panic

Panics if the SourceId is invalid.

pub fn obsolete_requests(&self) -> impl Iterator<Item = (RequestId, &TRq)> + '_

Returns an iterator that yields all the requests whose outcome is no longer desired.

pub fn desired_requests(&self) -> impl Iterator<Item = RequestDetail> + '_

Returns an iterator that yields all requests that could be started.

pub fn insert_request(
&mut self,
detail: RequestDetail,
user_data: TRq
) -> RequestId

Updates the OptimisticSync with the fact that a request has been started.

Returns the identifier for the request that must later be passed back to OptimisticSync::finish_request_success or OptimisticSync::finish_request_failed.

Panic

Panics if the SourceId is invalid.

pub fn finish_request_success(
&mut self,
request_id: RequestId,
blocks: impl Iterator<Item = RequestSuccessBlock<TBl>>
) -> (TRq, FinishRequestOutcome)

Update the OptimisticSync with the successful outcome of a request.

Returns the user data that was associated to that request.

If the state machine only handles light clients, that is if Config::full was false, then the values of RequestSuccessBlock::scale_encoded_extrinsics are silently ignored.

Note: If Config::full is false, you are encouraged to not request the block’s body from the source altogether, and to fill the RequestSuccessBlock::scale_encoded_extrinsics fields with Vec::new().

Panic

Panics if the RequestId is invalid.

pub fn finish_request_failed(&mut self, request_id: RequestId) -> TRq

Update the OptimisticSync with the information that the given request has failed.

Returns the user data that was associated to that request.

Panic

Panics if the RequestId is invalid.

pub fn process_one(self) -> ProcessOne<TRq, TSrc, TBl>

Process the next block in the queue of verification.

This method takes ownership of the OptimisticSync. The OptimisticSync is yielded back in the returned value.