Module smoldot::transactions::light_pool

Transactions pool specialized for light clients usage.

See the pool module documentation for details about the transactions pool.

Contrary to super::pool::Pool, this data structure is opinionated towards a certain light client usage. This means:

Block bodies are initially unknown and can be added later.

Transactions included in block bodies that weren’t already in the pool aren’t added, and thus also don’t need to be validated.

The LightPool tracks all forks, not just the best chain, so as to not require fetching again later the block bodies that are already known in case of a double re-org.

Usage

A LightPool is a collection of transactions and a tree of blocks.

Blocks can be added to the tree using LightPool::add_block. When a block is added, its body is unknown. You can add a body to a block using LightPool::set_block_body. The pool also tracks a best block and a finalized block. Use LightPool::set_best_block and LightPool::set_finalized_block to match the light pool with the state of the chain.

Blocks that have been finalized can be removed with LightPool::prune_finalized_with_body. This method only removes blocks whose body is known. You are encouraged to track the value of LightPool::oldest_block_finality_lag and make sure that it doesn’t go above a certain threshold, in order to avoid adding too many blocks to this pool.

Each transaction in the pool exposes three properties:

Whether or not it has been validated, and if yes, the block against which it has been validated and the characteristics of the transaction (as provided by the runtime): the tags it provides and requires, its longevity, and its priority. See the validate module for more information.

The block of the best chain, if any, in which the transaction has been included.

A so-called user data, an opaque field controller by the API user, of type TTx.

Use LightPool::add_unvalidated to add to the pool a transaction that should be included in a block at a later point in time.

Use LightPool::unvalidated_transactions to obtain the list of transactions that should be validated. Validation should be performed using the validate module, and the result reported with LightPool::set_validation_result.

Structs

Config

Configuration for LightPool::new.

LightPool

Data structure containing transactions. See the module-level documentation for more info.

PruneBodyFinalized

See LightPool::prune_finalized_with_body.

RemovedTransaction

See PruneBodyFinalized.

SetBestBlock

See LightPool::set_best_block.

TransactionId

Identifier of a transaction stored within the LightPool.