Struct smoldot::transactions::light_pool::LightPool

pub struct LightPool<TTx, TBl, TErr> { /* private fields */ }

Data structure containing transactions. See the module-level documentation for more info.

Implementations

impl<TTx, TBl, TErr> LightPool<TTx, TBl, TErr>where
TErr: Clone,

pub fn new(config: Config) -> Self

Initializes a new transactions pool.

pub fn num_transactions(&self) -> usize

Returns the number of transactions in the pool.

pub fn add_unvalidated(
&mut self,
scale_encoded: Vec<u8>,
user_data: TTx
) -> TransactionId

Inserts a new unverified transaction in the pool.

Must be passed as parameter the SCALE-encoded transaction.

pub fn remove_transaction(&mut self, id: TransactionId) -> (Vec<u8>, TTx)

Removes from the pool the transaction with the given identifier.

Panic

Panics if the identifier is invalid.

pub fn unvalidated_transactions(
&self
) -> impl Iterator<Item = (TransactionId, &TTx)> + '_

Returns a list of transactions whose state is “not validated”, and their user data.

These transactions should always be validated against the current best block.

pub fn transactions_iter(
&self
) -> impl Iterator<Item = (TransactionId, &TTx)> + '_

Returns the list of all transactions within the pool.

pub fn transactions_iter_mut(
&mut self
) -> impl Iterator<Item = (TransactionId, &mut TTx)> + '_

Returns the list of all transactions within the pool.

pub fn transaction_user_data(&self, id: TransactionId) -> Option<&TTx>

Returns the user data associated with a given transaction.

Returns None if the identifier is invalid.

pub fn transaction_user_data_mut(
&mut self,
id: TransactionId
) -> Option<&mut TTx>

Returns the user data associated with a given transaction.

Returns None if the identifier is invalid.

pub fn scale_encoding(&self, id: TransactionId) -> Option<&[u8]>

Returns the bytes associated with a given transaction.

Returns None if the identifier is invalid.

pub fn find_transaction(
&self,
scale_encoded: &[u8]
) -> impl Iterator<Item = TransactionId> + '_

Tries to find the transactions in the pool whose bytes are scale_encoded.

pub fn is_included_best_chain(&self, id: TransactionId) -> bool

Returns true if the given transaction has been included in the past in an ancestor of the current best block.

Panic

Panics if the transaction with the given id is invalid.

pub fn is_valid_against_best_block(&self, id: TransactionId) -> bool

Returns true if the given transaction has been successfully validated in the past against an ancestor of the best block and is still within its longevity period.

Returns false either if the given transaction hasn’t been validated, or if its validation resulted in an error.

Note: This function might return true or false independently of whether or not the transaction has already been included in the best chain. You might want to call LightPool::is_included_best_chain as well.

Panic

Panics if the transaction with the given id is invalid.

pub fn invalid_transactions_best_block(
&self
) -> impl Iterator<Item = (TransactionId, &TTx, &TErr)> + '_

Returns a list of transactions which have been validated against an ancestor of the current best block but have encountered an error during their validation (i.e. LightPool::set_validation_result was called with Err).

Note: Contrary to LightPool::invalid_transactions_finalized_block, it is possible for the transaction to become valid again in the future if a reorg happens and removes the block the transaction was validated against.

pub fn invalid_transactions_finalized_block(
&self
) -> impl Iterator<Item = (TransactionId, &TTx, &TErr)> + '_

Returns a list of transactions which have been validated against an ancestor of the current finalized block but have encountered an error during their validation (i.e. LightPool::set_validation_result was called with Err).

Note: Once a transaction is considered as invalid, it can be assumed that this transaction will be invalid if verified against any of the descendants of the block it was verified against. In other words, it can assumed that transactions returned here will never be valid.

pub fn set_validation_result(
&mut self,
id: TransactionId,
block_hash_validated_against: &[u8; 32],
result: Result<ValidTransaction, TErr>
)

Sets the outcome of validating the transaction with the given identifier.

The block hash must be the block hash against which the transaction has been validated.

Panic

Panics if the transaction with the given id is invalid. Panics if no block with that hash has been inserted before, or the block has been pruned.

pub fn add_block(&mut self, hash: [u8; 32], parent_hash: &[u8; 32], user_data: TBl)

Adds a block to the collection of blocks.

Has no effect if that block was already present in the collection.

If there is no transaction in the pool, then the block is marked as “doesn’t need a body”, meaning that it will not be returned by LightPool::missing_block_bodies.

Panic

Panics if the parent block cannot be found in the collection.

pub fn set_best_block(&mut self, new_best_block_hash: &[u8; 32]) -> SetBestBlock

Sets the passed block as the new best block of the chain.

Panic

Panics if no block with the given hash has been inserted before.

pub fn has_block(&self, hash: &[u8; 32]) -> bool

Returns true if the block with the given hash is present in the pool.

pub fn best_block_hash(&self) -> &[u8; 32]

Returns the hash of the best block.

Please note that the block with the given hash might not have an associated user data in case the best block is equal to the finalized block and all finalized blocks have been pruned.

pub fn block_user_data(&self, hash: &[u8; 32]) -> Option<&TBl>

Returns the user data associated with a given block.

Returns None if the block hash doesn’t correspond to a known block.

pub fn block_user_data_mut(&mut self, hash: &[u8; 32]) -> Option<&mut TBl>

Returns the user data associated with a given block.

Returns None if the block hash doesn’t correspond to a known block.

pub fn set_block_body(
&mut self,
block_hash: &[u8; 32],
body: impl Iterator<Item = impl AsRef<[u8]>>
) -> impl Iterator<Item = (TransactionId, usize)> + '_

Sets the list of single-SCALE-encoded transactions that are present in the body of a block.

If the block is part of the best chain, returns the list of transactions that are in the pool, that were found in the body, and that weren’t part of the best chain before, alongside with their index in the body.

Panic

Panics if no block with the given hash has been inserted before.

pub fn missing_block_bodies(
&self
) -> impl Iterator<Item = (&[u8; 32], &TBl)> + '_

Returns the list of blocks whose bodies aren’t present in this data structure.

Blocks that were inserted when there wasn’t any transaction in the pool are never returned.

pub fn set_finalized_block(
&mut self,
new_finalized_block_hash: &[u8; 32]
) -> impl Iterator<Item = ([u8; 32], TBl)>

Sets the finalized block of the chain.

Removes and returns the blocks that are not part of the finalized chain. Please note that the finalized chain itself, however, isn’t removed.

The current best block (set using LightPool::set_best_block) must be a descendant of or equal to the node passed as parameter. This guarantees that no transaction gets retracted.

Panic

Panics if no block with the given hash has been inserted before. Panics if the current best block isn’t a descendant of or equal to the new finalized block. Panics if the current finalized block isn’t an ancestor of or equal to the new finalized block.

pub fn prune_finalized_with_body(
&mut self
) -> impl Iterator<Item = PruneBodyFinalized<TTx, TBl>> + '_

Removes from the pool as many blocks as possible from the finalized chain. Blocks are removed from parent to child until either the first non-finalized block or a block whose body is missing is encountered.

Also removes the transactions from the pool that were included in these blocks.

pub fn oldest_block_finality_lag(&self) -> usize

Returns the number of blocks between the oldest block stored in this data structure and the finalized block.