Struct smoldot::transactions::pool::Pool

pub struct Pool<TTx> { /* private fields */ }

Data structure containing transactions. See the module-level documentation for more info.

Implementations

impl<TTx> Pool<TTx>

pub fn new(config: Config) -> Self

Initializes a new transactions pool.

pub fn is_empty(&self) -> bool

Returns true if the pool is empty.

pub fn len(&self) -> usize

Returns the number of transactions in the pool.

pub fn add_unvalidated(
&mut self,
scale_encoded: Vec<u8>,
user_data: TTx
) -> TransactionId

Inserts a new unverified transaction in the pool.

pub fn remove(&mut self, id: TransactionId) -> TTx

Removes from the pool the transaction with the given identifier.

Panic

Panics if the identifier is invalid.

pub fn remove_included(
&mut self,
block_inferior_of_equal: u64
) -> impl Iterator<Item = (TransactionId, TTx)> + '_

Removes from the pool all the transactions that are included in a block whose height is inferior or equal to the one passed as parameter.

Use this method when a block has been finalized.

The returned iterator is guaranteed to remove all transactions even if it is dropped eagerly.

pub fn unvalidated_transactions(
&self
) -> impl ExactSizeIterator<Item = (TransactionId, &TTx, u64)> + '_

Returns a list of transactions whose state is “not validated”, their user data, and the height of the block they should be validated against.

The block height a transaction should be validated against is always equal to either the block at which it has been included minus one, or the current best block. It is yielded by the iterator for convenience, to avoid writing error-prone code.

pub fn inclusion_order(&self) -> impl Iterator<Item = TransactionId> + '_

Returns the transactions from the pool that haven’t been included yet in the order in which they should be inserted in authored blocks.

pub fn iter(&self) -> impl Iterator<Item = (TransactionId, &TTx)> + '_

Returns the list of all transactions within the pool.

pub fn iter_mut(
&mut self
) -> impl Iterator<Item = (TransactionId, &mut TTx)> + '_

Returns the list of all transactions within the pool.

pub fn user_data(&self, id: TransactionId) -> Option<&TTx>

Returns the user data associated with a given transaction.

Returns None if the identifier is invalid.

pub fn user_data_mut(&mut self, id: TransactionId) -> Option<&mut TTx>

Returns the user data associated with a given transaction.

Returns None if the identifier is invalid.

pub fn included_block_height(&self, id: TransactionId) -> Option<u64>

Returns the block height at which the given transaction has been included.

A transaction has been included if it has been added to the pool with Pool::append_block.

Returns None if the identifier is invalid or the transaction doesn’t belong to any block.

pub fn scale_encoding(&self, id: TransactionId) -> Option<&[u8]>

Returns the bytes associated with a given transaction.

Returns None if the identifier is invalid.

pub fn find(
&self,
scale_encoded: &[u8]
) -> impl Iterator<Item = TransactionId> + '_

Tries to find a transaction in the pool whose bytes are scale_encoded.

pub fn best_block_height(&self) -> u64

Returns the best block height according to the pool.

This initially corresponds to the value in Config::finalized_block_height, is incremented by one every time Pool::append_block, and is decreased when Pool::retract_blocks is called.

pub fn append_block(self) -> AppendBlock<TTx>

Adds a block to the chain tracked by the transactions pool.

This function returns an AppendBlock struct that wraps around the Pool and lets you insert transactions that belong to the body of the new block.

pub fn retract_blocks(
&mut self,
num_to_retract: u64
) -> impl Iterator<Item = (TransactionId, u64)>

Pop a certain number of blocks from the list of blocks.

Transactions that were included in these blocks remain in the transactions pool.

Returns the list of transactions that were in blocks that have been retracted, with the height of the block at which they were.

Panic

Panics if num_to_retract > self.best_block_height(), in other words if the block number would go in the negative.

pub fn set_validation_result(
&mut self,
id: TransactionId,
block_number_validated_against: u64,
result: Result<ValidTransaction, TransactionValidityError>
)

Sets the outcome of validating the transaction with the given identifier.

The block number must be the block number against which the transaction has been validated.

The validation result might be ignored if it doesn’t match one of the entries returned by Pool::unvalidated_transactions.

Panic

Panics if the transaction with the given id is invalid.