smoldot::transactions::pool

Struct Pool

pub struct Pool<TTx> { /* private fields */ }

Data structure containing transactions. See the module-level documentation for more info.

Implementations

impl<TTx> Pool<TTx>

pub fn new(config: Config) -> Self

Initializes a new transactions pool.

pub fn is_empty(&self) -> bool

Returns true if the pool is empty.

pub fn len(&self) -> usize

Returns the number of transactions in the pool.

pub fn add_unvalidated( &mut self, scale_encoded: Vec<u8>, user_data: TTx, ) -> TransactionId

Inserts a new non-validated transaction in the pool.

pub fn remove(&mut self, id: TransactionId) -> TTx

Removes from the pool the transaction with the given identifier.

Panic

Panics if the identifier is invalid.

pub fn remove_included( &mut self, block_inferior_of_equal: u64, ) -> impl Iterator<Item = (TransactionId, TTx)>

Removes from the pool all the transactions that are included in a block whose height is inferior or equal to the one passed as parameter.

Use this method when a block has been finalized.

The returned iterator is guaranteed to remove all transactions even if it is dropped eagerly.

pub fn unvalidated_transactions( &self, ) -> impl ExactSizeIterator<Item = (TransactionId, &TTx, u64)>

Returns a list of transactions whose state is “not validated”, their user data, and the height of the block they should be validated against.

The block height a transaction should be validated against is always equal to either the block at which it has been included minus one, or the current best block. It is yielded by the iterator for convenience, to avoid writing error-prone code.

pub fn iter(&self) -> impl Iterator<Item = (TransactionId, &TTx)>

Returns the list of all transactions within the pool.

pub fn iter_mut(&mut self) -> impl Iterator<Item = (TransactionId, &mut TTx)>

Returns the list of all transactions within the pool.

pub fn included_block_height(&self, id: TransactionId) -> Option<u64>

Returns the block height at which the given transaction has been included.

A transaction has been included if it has been added to the pool with Pool::best_block_add_transaction_by_scale_encoding or Pool::best_block_add_transaction_by_id.

Returns None if the identifier is invalid or the transaction doesn’t belong to any block.

pub fn scale_encoding(&self, id: TransactionId) -> Option<&[u8]>

Returns the bytes associated with a given transaction.

Returns None if the identifier is invalid.

pub fn transactions_by_scale_encoding( &self, scale_encoded: &[u8], ) -> impl Iterator<Item = TransactionId>

Finds the transactions in the pool whose bytes are scale_encoded.

This operation has a complexity of O(log n) where n is the number of entries in the pool.

pub fn best_block_height(&self) -> u64

Returns the best block height according to the pool.

This initially corresponds to the value in Config::finalized_block_height, is incremented by one every time Pool::append_empty_block, and is decreased when Pool::retract_blocks is called.

pub fn append_empty_block(&mut self)

Adds a block to the chain tracked by the transactions pool.

pub fn retract_blocks( &mut self, num_to_retract: u64, ) -> impl Iterator<Item = (TransactionId, u64)>

Pop a certain number of blocks from the list of blocks.

Transactions that were included in these blocks remain in the transactions pool.

Returns the list of transactions that were in blocks that have been retracted, with the height of the block at which they were.

Panic

Panics if num_to_retract > self.best_block_height(), in other words if the block number would go in the negative.

pub fn best_block_includable_transactions( &self, ) -> impl Iterator<Item = (TransactionId, &TTx)>

Returns all the transactions that can be included in the highest block.

Use this function if you are currently authoring a block.

The transactions are returned by decreasing priority. Re-ordering the transactions might lead to the runtime returning errors. It is safe, however, to skip some transactions altogether if desired.

pub fn best_block_add_transaction_by_scale_encoding<'a, 'b>( &'a mut self, bytes: &'b [u8], ) -> AppendBlockTransaction<'a, 'b, TTx>

Adds a single-SCALE-encoded transaction to the highest block.

The transaction is compared against the list of non-included transactions that are already in the pool. If a non-included transaction with the same bytes is found, it is switched to the “included” state and AppendBlockTransaction::NonIncludedUpdated is returned. Otherwise, AppendBlockTransaction::Unknown is returned and the transaction can be inserted in the pool.

Note: This function is equivalent to calling Pool::transactions_by_scale_encoding, removing the transactions that are already included (see Pool::included_block_height), then calling Pool::best_block_add_transaction_by_id with one of the transactions that remain.

pub fn best_block_add_transaction_by_id(&mut self, id: TransactionId)

Adds a transaction to the block being appended.

Panic

Panics if the transaction with the given id is invalid. Panics if the transaction with the given id was already included in the chain.

pub fn set_validation_result( &mut self, id: TransactionId, block_number_validated_against: u64, result: ValidTransaction, )

Sets the outcome of validating the transaction with the given identifier.

The block number must be the block number against which the transaction has been validated.

Has no effect if the transaction has been included in the chain and the validation has been performed against a block other than the parent of the block in which it was included.

Has no effect if the transaction has already been validated against the same or a higher block.

Note: If the transaction validation fails, use Pool::remove to remove the transaction instead. Invalid transactions stay invalid forever and thus aren’t meant to be left in the pool.

Panic

Panics if the transaction with the given id is invalid.