smoldot::transactions
Module validate

Runtime call to obtain the transactions validity status.

Structs

DecodeError

Error that can happen during the decoding.

ValidTransaction

Information concerning a valid transaction.

Enums

InvalidTransaction

An invalid transaction validity.

TransactionSource

Source of the transaction.

TransactionValidityError

Errors that can occur while checking the validity of a transaction.

UnknownTransaction

An unknown transaction validity.

Constants

VALIDATION_FUNCTION_NAME

Name of the runtime function to call in order to validate a transaction.

Functions

decode_validate_transaction_return_value

Attempt to decode the return value of the TaggedTransactionQueue_validate_transaction runtime call.

validate_transaction_runtime_parameters_v2

Produces the input to pass to the TaggedTransactionQueue_validate_transaction runtime call.

validate_transaction_runtime_parameters_v3

Produces the input to pass to the TaggedTransactionQueue_validate_transaction runtime call.