smoldot::trie
Module calculate_root

Freestanding function that calculates the root of a radix-16 Merkle-Patricia trie.

See the parent module documentation for an explanation of what the trie is.

This module is meant to be used in situations where all the nodes of the trie that have a storage value associated to them are known and easily accessible, and that no cache is available.

Usage

Calling the root_merkle_value function creates a RootMerkleValueCalculation object which you have to drive to completion.

Example:

use std::{collections::BTreeMap, ops::Bound};
use smoldot::trie::{HashFunction, TrieEntryVersion, calculate_root};

// In this example, the storage consists in a binary tree map.
let mut storage = BTreeMap::<Vec<u8>, (Vec<u8>, TrieEntryVersion)>::new();
storage.insert(b"foo".to_vec(), (b"bar".to_vec(), TrieEntryVersion::V1));

let trie_root = {
    let mut calculation = calculate_root::root_merkle_value(HashFunction::Blake2);
    loop {
        match calculation {
            calculate_root::RootMerkleValueCalculation::Finished { hash, .. } => break hash,
            calculate_root::RootMerkleValueCalculation::NextKey(next_key) => {
                let key_before = next_key.key_before().collect::<Vec<_>>();
                let lower_bound = if next_key.or_equal() {
                    Bound::Included(key_before)
                } else {
                    Bound::Excluded(key_before)
                };
                let outcome = storage
                    .range((lower_bound, Bound::Unbounded))
                    .next()
                    .filter(|(k, _)| {
                        k.iter()
                            .copied()
                            .zip(next_key.prefix())
                            .all(|(a, b)| a == b)
                    })
                    .map(|(k, _)| k);
                calculation = next_key.inject_key(outcome.map(|k| k.iter().copied()));
            }
            calculate_root::RootMerkleValueCalculation::StorageValue(value_request) => {
                let key = value_request.key().collect::<Vec<u8>>();
                calculation = value_request.inject(storage.get(&key).map(|(val, v)| (val, *v)));
            }
        }
    }
};

assert_eq!(
    trie_root,
    [204, 86, 28, 213, 155, 206, 247, 145, 28, 169, 212, 146, 182, 159, 224, 82,
     116, 162, 143, 156, 19, 43, 183, 8, 41, 178, 204, 69, 41, 37, 224, 91]
);

Structs

NextKey

Request to return the key that follows (in lexicographic order) a given one in the storage. Call NextKey::inject_key to indicate this list.

StorageValue

Request the value of the node with a specific key. Call StorageValue::inject to indicate the value.

Enums

RootMerkleValueCalculation

Current state of the RootMerkleValueCalculation and how to continue.

Functions

root_merkle_value

Start calculating the Merkle value of the root node.