smoldot::trie

Function nibbles_to_bytes_suffix_extend

pub fn nibbles_to_bytes_suffix_extend<I: Iterator<Item = Nibble>>(
    nibbles: I,
) -> impl Iterator<Item = u8>

Turns an iterator of nibbles into an iterator of bytes.

If the number of nibbles is uneven, adds a 0 nibble at the end.

Examples

use smoldot::trie::{Nibble, nibbles_to_bytes_suffix_extend};

let input = [Nibble::try_from(0x5).unwrap(), Nibble::try_from(0xa).unwrap()];
assert_eq!(nibbles_to_bytes_suffix_extend(input.into_iter()).collect::<Vec<_>>(), &[0x5a]);

use smoldot::trie::{Nibble, nibbles_to_bytes_suffix_extend};

let input = [Nibble::try_from(0x5).unwrap(), Nibble::try_from(0xa).unwrap(), Nibble::try_from(0x9).unwrap()];
assert_eq!(nibbles_to_bytes_suffix_extend(input.into_iter()).collect::<Vec<_>>(), &[0x5a, 0x90]);