Module smoldot::trie::proof_decode

Decodes and verifies a trie proof.

A trie proof is a proof that a certain key in the trie has a certain storage value (or lacks a storage value). The proof can be verified by knowing only the Merkle value of the root node.

Details

Note: For reminder, the Merkle value of a node is the hash of its node value, or the node value directly if its length is smaller than 32 bytes.

A trie proof consists in a list of node values of nodes in the trie. For the proof to be valid, the hash of one of these node values must match the expected trie root node value. Since a node value contains the Merkle values of the children of the node, it is possible to iterate down the hierarchy of nodes until the one closest to the desired key is found.

Usage

This modules provides the decode_and_verify_proof function that decodes a proof and verifies whether it is correct.

Once decoded, one can examine the content of the proof, in other words the list of storage items and values.

Structs

Children

See TrieNodeInfo::children.

Config

Configuration to pass to decode_and_verify_proof.

DecodedTrieProof

Decoded Merkle proof. The proof is guaranteed valid.

EntryKey

Identifier for an entry within a decoded proof.

EntryKeyIter

IncompleteProofError

Proof doesn’t contain enough information to answer the request.

ProofEntry

Information about an entry in the proof.

TrieNodeInfo

Information about a node of the trie.

Enums

Child

Information about a specific child in the list of children.

Error

Possible error returned by decode_and_verify_proof.

StorageValue

Storage value of the node.

Functions

decode_and_verify_proof

Verifies whether a proof is correct and returns an object that allows examining its content.