smoldot::trie::proof_decode

Struct DecodedTrieProof

pub struct DecodedTrieProof<T> { /* private fields */ }

Decoded Merkle proof. The proof is guaranteed valid.

Implementations

impl<T: AsRef<[u8]>> DecodedTrieProof<T>

pub fn iter_runtime_context_ordered( &self, ) -> impl Iterator<Item = (EntryKey<'_, Vec<u8>>, StorageValue<'_>)>

Returns a list of all elements of the proof, ordered by key in lexicographic order.

This function is a convenient wrapper around DecodedTrieProof::iter_ordered that converts the keys into arrays of bytes. If a key can’t be represented as an array of bytes, then it is filtered out. Assuming that the trie has only ever been used in the context of the runtime, then this cannot happen. See the section below for an explanation.

The iterator might include branch nodes. It is not possible for this function to differentiate between value-less nodes that are present in the proof only because they are branch nodes, and value-less nodes that are present in the proof because the fact that they have no value is important for the proof.

Detailed explanation

The trie consists of nodes, each with a key and a value. The keys consist of an array of “nibbles”, which are 4 bits each.

When the runtime writes a value in the trie, it passes a key as an array a bytes. In order to know where to write this value, this array of bytes is converted into an array of nibbles by turning each byte into two nibbles.

Due to the fact that the host-runtime interface only ever uses arrays of bytes, it is not possible for the runtime to store a value or read a value in the trie at a key that consists in an uneven number of nibbles, as an uneven number of nibbles cannot be converted to an array of bytes.

In other words, if a trie has only ever been used in the context of a runtime, then it is guaranteed to not contain any storage value at key that consists in an uneven number of nibbles.

The trie format itself, however, technically doesn’t forbid storing reading and writing values at keys that consist in an uneven number of nibbles. For this reason, a proof containing a value at a key that consists in an uneven number of nibbles is considered as valid according to decode_and_verify_proof.

However, given that decode_and_verify_proof verifies the trie proof against the state trie root hash, we are also guaranteed that this proof reflects the actual trie. If the actual trie can’t contain any storage value at a key that consists in an uneven number of nibbles, then the proof is also guaranteed to not contain any storage value at a key that consists in an uneven number of nibbles. Importantly, this is only true if we are sure that the block is valid, in other words that it has indeed been created using a runtime. Blocks that are invalid might have been created through a fake trie.

As a conclusion, if this proof is made against a trie that has only ever been used in the context of a runtime, and that the block using this trie is guaranteed to be valid, then this function will work as intended and return the entire content of the proof.

pub fn iter_ordered( &self, ) -> impl Iterator<Item = (EntryKey<'_, EntryKeyIter<'_, T>>, ProofEntry<'_, T>)>

Returns a list of all elements of the proof, ordered by key in lexicographic order.

The iterator includes branch nodes.

pub fn trie_root_proof_entry<'a>( &'a self, trie_root_merkle_value: &'a [u8; 32], ) -> Option<ProofEntry<'a, T>>

Returns the ProofEntry of the proof entry whose Merkle value is given as parameter.

The returned ProofEntry is guaranteed to have no parent.

Returns None if the proof doesn’t contain any sub-trie with the given Merkle value.

pub fn proof_entry<'a>( &'a self, trie_root_merkle_value: &'a [u8; 32], key: impl Iterator<Item = Nibble>, ) -> Option<ProofEntry<'a, T>>

Returns the ProofEntry of the given key.

Returns None if the key doesn’t have any entry in the proof.

Note: In situations where the key isn’t in the proof but the proof contains enough information about the trie in order to infer some information about that key, the DecodedTrieProof::trie_node_info function will be successful whereas the DecodedTrieProof::proof_entry function will fail.

pub fn closest_ancestor_in_proof<'a>( &'a self, trie_root_merkle_value: &[u8; 32], key: impl Iterator<Item = Nibble>, ) -> Result<Option<EntryKeyIter<'a, T>>, IncompleteProofError>

Returns the key of the closest ancestor to the given key that can be found in the proof. If key is in the proof, returns key.

Returns None if the key is completely outside of the trie (i.e. the trie root is not an ancestor of the key).

pub fn trie_node_info( &self, trie_root_merkle_value: &[u8; 32], key: impl Iterator<Item = Nibble>, ) -> Result<TrieNodeInfo<'_, T>, IncompleteProofError>

Returns information about a trie node.

Returns an error if the proof doesn’t contain enough information about this trie node.

This function will return Ok even if there is no node in the trie for key, in which case the returned TrieNodeInfo will indicate no storage value and no children.

pub fn storage_value( &self, trie_root_merkle_value: &[u8; 32], key: &[u8], ) -> Result<Option<(&[u8], TrieEntryVersion)>, IncompleteProofError>

Queries from the proof the storage value at the given key.

Returns an error if the storage value couldn’t be determined from the proof. Returns Ok(None) if the storage value is known to have no value.

Note: This function is a convenient wrapper around DecodedTrieProof::trie_node_info.

pub fn next_key( &self, trie_root_merkle_value: &[u8; 32], key_before: impl Iterator<Item = Nibble>, or_equal: bool, prefix: impl Iterator<Item = Nibble>, branch_nodes: bool, ) -> Result<Option<EntryKeyIter<'_, T>>, IncompleteProofError>

Find in the proof the trie node that follows key_before in lexicographic order.

If or_equal is true, then key_before is returned if it is equal to a node in the trie. If false, then only keys that are strictly superior are returned.

The returned value must always start with prefix. Note that the value of prefix is important as it can be the difference between Err(IncompleteProofError) and Ok(None).

If branch_nodes is false, then trie nodes that don’t have a storage value are skipped.

Returns an error if the proof doesn’t contain enough information to determine the next key. Returns Ok(None) if the proof indicates that there is no next key (within the given prefix).

pub fn closest_descendant_merkle_value( &self, trie_root_merkle_value: &[u8; 32], key: impl Iterator<Item = Nibble>, ) -> Result<Option<&[u8]>, IncompleteProofError>

Find in the proof the closest trie node that descends from key and returns its Merkle value.

Returns an error if the proof doesn’t contain enough information to determine the Merkle value. Returns Ok(None) if the proof indicates that there is no descendant.