smoldot::trie

Module trie_structure

Manages the structure of a trie. Allows inserting and removing nodes, but does not store any value. Only the structure is stored.

See the TrieStructure struct.

Structs

BranchNodeAccess

Access to a node within the TrieStructure that is known to not have any storage value associated to it.

NodeIndex

Index of a node in the trie. Never invalidated, except when if node in question is destroyed.

PrepareInsertOne

One node will be inserted in the trie.

PrepareInsertTwo

Two nodes will be inserted in the trie.

StorageNodeAccess

Access to a node within the TrieStructure that is known to have a storage value associated to it.

TrieStructure

Stores the structure of a trie, including branch nodes that have no storage value.

Vacant

Access to a non-existing node within the TrieStructure.

Enums

Entry

Access to a entry for a potential node within the TrieStructure.

NodeAccess

Access to a node within the TrieStructure.

PrepareInsert

Preparation for a new node insertion.

Remove

Outcome of the removal of a storage value.