Module smoldot::verify::aura

AURA consensus.

AURA, for Authority Round, is one of the consensus algorithm available to Substrate-based chains in order to determine who is authorized to generate a block.

Every block (with the exception of the genesis block) must contain, in its header, some data that makes it possible to verify that it has been generated by a legitimate author.

Overview of AURA

In the AURA algorithm, time is divided into non-overlapping slots. How long a slot is never changes, and is determined by calling the AuraApi_slot_duration runtime entry point, typically at initialization. The current slot number is equal to unix_time / slot_duration. The slot number of a block can be found in its header.

Note: Slot durations values are usually around 3 to 20 seconds.

A list of authorities (each authority being represented as a public key) is maintained by the chain. The current authorities of a block can be found by calling the AuraApi_authorities runtime entry point. When this list is modified, a consensus log item is added to the header of the block, meaning that this runtime entry point doesn’t have to be called every single time.

During a slot, only the authority whose public key is found at authorities[slot_number % authorities.len()] is allowed to produce a block.

Each block being produced must include an Aura seal in its header containing a signature of the block header (with the exclusion of the seal itself) made using the public key in question.

Structs

VerifyConfig

Configuration for verify_header.

VerifySuccess

Information yielded back after successfully verifying a block.

Enums

VerifyError

Failure to verify a block.

Functions

verify_header

Verifies whether a block header provides a correct proof of the legitimacy of the authorship.