smoldot::verify

Module babe

BABE consensus.

BABE, or Blind Assignment for Blockchain Extension, is the consensus algorithm used by Polkadot in order to determine who is authorized to generate a block.

Every block (with the exception of the genesis block) must contain, in its header, some data that makes it possible to verify that it has been generated by a legitimate author.

References:

https://research.web3.foundation/en/latest/polkadot/BABE/Babe.html

Overview of BABE

In the BABE algorithm, time is divided into non-overlapping epochs, themselves divided into slots. How long an epoch and a slot are is determined by calling the BabeApi_configuration runtime entry point.

Note: As example values, in the Polkadot genesis, a slot lasts for 6 seconds and an epoch consists of 2400 slots (in other words, four hours).

Every block that is produced must belong to a specific slot. This slot number can be found in the block header, with the exception of the genesis block which is considered timeless and doesn’t have any slot number.

At the moment, the current slot number is determined purely based on the slot duration (e.g. 6 seconds for Polkadot) and the local clock based on the UNIX EPOCH. The current slot number is unix_timestamp / duration_per_slot. This might change in the future.

The first epoch (epoch number 0) starts at slot_number(block #1) and ends at slot_number(block #1) + slots_per_epoch. The second epoch (epoch #1) starts at slot end_of_epoch_0 + 1. All epochs end at start_of_new_epoch + slots_per_epoch. Block #0 doesn’t belong to any epoch.

The header of first block produced after a transition to a new epoch (including block #1) must contain a log entry indicating the public keys that are allowed to sign blocks, alongside with a weight for each of them, and a “randomness value”. This information does not concern the newly-started epoch, but the one immediately after. In other words, the first block of epoch N contains the information about epoch N+1.

Note: The way the list of authorities and their weights is determined is at the discretion of the runtime code and is out of scope of this module, but it normally corresponds to the list of validators and how much stake is available to them.

In order to produce a block, one must generate, using a VRF (Verifiable Random Function), and based on the slot number, genesis hash, and aforementioned “randomness value”, a number whose value is lower than a certain threshold.

The number that has been generated must be included in the header of the authored block, alongside with the proof of the correct generation that can be verified using one of the public keys allowed to generate blocks in that epoch. The weight associated to that public key determines the allowed threshold.

The “randomness value” of an epoch N is calculated by combining the generated numbers of all the blocks of the epoch N - 2.

Secondary slots

While all slots can be claimed by generating a number below a certain threshold, each slot is additionally assigned to a specific public key amongst the ones allowed. The owner of a public key is always allowed to generate a block during the slot assigned to it.

The mechanism of attributing each slot to a public key is called “secondary slot claims”, while the mechanism of generating a number below a certain threshold is called a “primary slot claim”. As their name indicates, primary slot claims have a higher priority over secondary slot claims.

Secondary slot claims are a way to guarantee that all slots can potentially lead to a block being produced.

Chain selection

The “best” block of a chain in the BABE algorithm is the one with the highest slot number. If there exists multiple blocks on the same slot, the best block is one with the highest number of primary slot claims. In other words, if two blocks have the same parent, but one is a primary slot claim and the other is a secondary slot claim, we prefer the one with the primary slot claim.

Keep in mind that there can still be draws in terms of primary slot claims count, in which case the winning block is the one upon which the next block author builds upon.

Epochs 0 and 1

The information about an epoch N is provided by the first block of the epoch N-1.

Because of this, we need to special-case epoch 0. The information about epoch 0 is contained in the chain-wide BABE configuration found in the runtime. The first block of epoch 0 is the block number #1. The information about epoch 1 is therefore contained in block #1.

Usage

In order to verify a Babe block, two of the main information to pass are:

The chain_information::BabeEpochInformationRef struct corresponding to the epoch the parent block belongs to.

The chain_information::BabeEpochInformationRef struct of the epoch that follows.

When verifying block number 1, VerifyConfig::parent_block_epoch must be set to None and VerifyConfig::parent_block_next_epoch must be set to the definition of epoch #0 as determined by performing runtime calls.

Any time verifying a block produces a Some in VerifySuccess::epoch_transition_target, which is guaranteed to be the case when verifying block number 1, an epoch transition occurs. When verifying a child of such block, the value formerly passed as VerifyConfig::parent_block_next_epoch must now be passed as VerifyConfig::parent_block_epoch, and the value in VerifySuccess::epoch_transition_target becomes VerifyConfig::parent_block_next_epoch.

When designing around these rules, be aware of forks: there can be multiple blocks at the same height performing epoch transitions.

See also the crate::chain::chain_information module for more help.

Structs

VerifyConfig

Configuration for verify_header.

VerifySuccess

Information yielded back after successfully verifying a block.

Enums

InvalidChainConfiguration

See VerifyError::InvalidChainConfiguration

VerifyError

Failure to verify a block.

Functions

verify_header

Verifies whether a block header provides a correct proof of the legitimacy of the authorship.