smoldot::verify

Module body_only

Verifying a block body. This operation is also called executing a block.

In order to execute a block, one must perform two runtime calls, to CHECK_INHERENTS_FUNCTION_NAME and to EXECUTE_BLOCK_FUNCTION_NAME (in that order).

The parameter to pass for these runtime calls can be determined using check_inherents_parameter and execute_block_parameter. When execution succeeds, the output of the runtime call must be checked using check_check_inherents_output and check_execute_block_output. The storage changes must be preserved between the two calls.

Any error during the execution or the output verification means that the block is invalid.

Enums

ExecuteBlockOutputError

Error potentially returned by check_execute_block_output.

ExecuteBlockParameterError

Error potentially returned by execute_block_parameter.

InherentsOutputError

Error potentially returned by check_check_inherents_output.

Constants

CHECK_INHERENTS_FUNCTION_NAME

EXECUTE_BLOCK_FUNCTION_NAME

Functions

check_check_inherents_output

Checks the output of the BlockBuilder_check_inherents runtime call.

check_execute_block_output

Checks the output of the Core_execute_block runtime call.

check_inherents_parameter

Returns a list of buffers that, when concatenated together, forms the parameter to pass to the BlockBuilder_check_inherents function in order to verify the inherents of a block.

execute_block_parameter

Returns a list of buffers that, when concatenated together, forms the parameter to pass to the Core_execute_block function in order to verify the inherents of a block.