smoldot::verify

Module header_only

Structs

Config

Configuration for a block verification.

Enums

ConfigConsensus

Extra items of Config that are dependant on the consensus engine of the chain.

ConfigFinality

Extra items of Config that are dependant on the finality engine of the chain.

Error

Error that can happen during the verification.

Success

Block successfully verified.

Functions

verify

Verifies whether a block is valid.