Module smoldot::verify::inherents

Inherents, together with transactions, form the body of a block.

The body of a block consists of a list of what is called extrinsics. An extrinsic can be either a transaction, when it was submitted by a user, or an inherent, which is what this module is about.

When a block is authored, one of the first steps is for the block author to generate the list of inherents. This is done by calling a runtime function, passing as parameter an encoded InherentData.

When a block is later verified, the inherents are verified by calling a runtime function and passing as parameter an encoded InherentData as well.

Structs

InherentData

Values of the inherents to pass to the runtime.